Sumulated project database by Boctor (1993)

Sumulated project database of multimode, individual projects by Boctor (1993)

data("Boctor")

Format

A list (Collection_PDM) of PDM_list and PDM_constraint types of lists, where

PDM_list

contains PDM_matrix (PDM) and the number of completion modes (w), and number of resources (Rs)

PDM_const

a list of (renewable resource) constraints

References

Boctor, F. F. (1993). Heuristics for scheduling projects with resource restrictions and several resource-duration modes. The international journal of production research, 31(11), 2547–2558, https://doi.org/10.1080/00207549308956882.

Examples

data(Boctor)

# Summary of data collection
summary(Boctor)
#> 
#> 
#> 
#> Summary of PDM collection:
#> 
#> Number of projects:  240
#> List of projects: 
#>  Project name:  boct1, w:  4, Rs:  1
#>  Project name:  boct10, w:  4, Rs:  2
#>  Project name:  boct100, w:  4, Rs:  2
#>  Project name:  boct101, w:  4, Rs:  4
#>  Project name:  boct102, w:  4, Rs:  4
#>  Project name:  boct103, w:  4, Rs:  1
#>  Project name:  boct104, w:  4, Rs:  1
#>  Project name:  boct105, w:  4, Rs:  2
#>  Project name:  boct106, w:  4, Rs:  2
#>  Project name:  boct107, w:  4, Rs:  4
#>  Project name:  boct108, w:  4, Rs:  4
#>  Project name:  boct109, w:  4, Rs:  1
#>  Project name:  boct11, w:  4, Rs:  4
#>  Project name:  boct110, w:  4, Rs:  1
#>  Project name:  boct111, w:  4, Rs:  2
#>  Project name:  boct112, w:  4, Rs:  2
#>  Project name:  boct113, w:  4, Rs:  4
#>  Project name:  boct114, w:  4, Rs:  4
#>  Project name:  boct115, w:  4, Rs:  1
#>  Project name:  boct116, w:  4, Rs:  1
#>  Project name:  boct117, w:  4, Rs:  2
#>  Project name:  boct118, w:  4, Rs:  2
#>  Project name:  boct119, w:  4, Rs:  4
#>  Project name:  boct12, w:  4, Rs:  4
#>  Project name:  boct120, w:  4, Rs:  4
#>  Project name:  boct121, w:  4, Rs:  1
#>  Project name:  boct122, w:  4, Rs:  1
#>  Project name:  boct123, w:  4, Rs:  2
#>  Project name:  boct124, w:  4, Rs:  2
#>  Project name:  boct125, w:  4, Rs:  4
#>  Project name:  boct126, w:  4, Rs:  4
#>  Project name:  boct127, w:  4, Rs:  1
#>  Project name:  boct128, w:  4, Rs:  1
#>  Project name:  boct129, w:  4, Rs:  2
#>  Project name:  boct13, w:  4, Rs:  1
#>  Project name:  boct130, w:  4, Rs:  2
#>  Project name:  boct131, w:  4, Rs:  4
#>  Project name:  boct132, w:  4, Rs:  4
#>  Project name:  boct133, w:  4, Rs:  1
#>  Project name:  boct134, w:  4, Rs:  1
#>  Project name:  boct135, w:  4, Rs:  2
#>  Project name:  boct136, w:  4, Rs:  2
#>  Project name:  boct137, w:  4, Rs:  4
#>  Project name:  boct138, w:  4, Rs:  4
#>  Project name:  boct139, w:  4, Rs:  1
#>  Project name:  boct14, w:  4, Rs:  1
#>  Project name:  boct140, w:  4, Rs:  1
#>  Project name:  boct141, w:  4, Rs:  2
#>  Project name:  boct142, w:  4, Rs:  2
#>  Project name:  boct143, w:  4, Rs:  4
#>  Project name:  boct144, w:  4, Rs:  4
#>  Project name:  boct145, w:  4, Rs:  1
#>  Project name:  boct146, w:  4, Rs:  1
#>  Project name:  boct147, w:  4, Rs:  2
#>  Project name:  boct148, w:  4, Rs:  2
#>  Project name:  boct149, w:  4, Rs:  4
#>  Project name:  boct15, w:  4, Rs:  2
#>  Project name:  boct150, w:  4, Rs:  4
#>  Project name:  boct151, w:  4, Rs:  1
#>  Project name:  boct152, w:  4, Rs:  1
#>  Project name:  boct153, w:  4, Rs:  2
#>  Project name:  boct154, w:  4, Rs:  2
#>  Project name:  boct155, w:  4, Rs:  4
#>  Project name:  boct156, w:  4, Rs:  4
#>  Project name:  boct157, w:  4, Rs:  1
#>  Project name:  boct158, w:  4, Rs:  1
#>  Project name:  boct159, w:  4, Rs:  2
#>  Project name:  boct16, w:  4, Rs:  2
#>  Project name:  boct160, w:  4, Rs:  2
#>  Project name:  boct161, w:  4, Rs:  4
#>  Project name:  boct162, w:  4, Rs:  4
#>  Project name:  boct163, w:  4, Rs:  1
#>  Project name:  boct164, w:  4, Rs:  1
#>  Project name:  boct165, w:  4, Rs:  2
#>  Project name:  boct166, w:  4, Rs:  2
#>  Project name:  boct167, w:  4, Rs:  4
#>  Project name:  boct168, w:  4, Rs:  4
#>  Project name:  boct169, w:  4, Rs:  1
#>  Project name:  boct17, w:  4, Rs:  4
#>  Project name:  boct170, w:  4, Rs:  1
#>  Project name:  boct171, w:  4, Rs:  2
#>  Project name:  boct172, w:  4, Rs:  2
#>  Project name:  boct173, w:  4, Rs:  4
#>  Project name:  boct174, w:  4, Rs:  4
#>  Project name:  boct175, w:  4, Rs:  1
#>  Project name:  boct176, w:  4, Rs:  1
#>  Project name:  boct177, w:  4, Rs:  2
#>  Project name:  boct178, w:  4, Rs:  2
#>  Project name:  boct179, w:  4, Rs:  4
#>  Project name:  boct18, w:  4, Rs:  4
#>  Project name:  boct180, w:  4, Rs:  4
#>  Project name:  boct181, w:  4, Rs:  1
#>  Project name:  boct182, w:  4, Rs:  1
#>  Project name:  boct183, w:  4, Rs:  2
#>  Project name:  boct184, w:  4, Rs:  2
#>  Project name:  boct185, w:  4, Rs:  4
#>  Project name:  boct186, w:  4, Rs:  4
#>  Project name:  boct187, w:  4, Rs:  1
#>  Project name:  boct188, w:  4, Rs:  1
#>  Project name:  boct189, w:  4, Rs:  2
#>  Project name:  boct19, w:  4, Rs:  1
#>  Project name:  boct190, w:  4, Rs:  2
#>  Project name:  boct191, w:  4, Rs:  4
#>  Project name:  boct192, w:  4, Rs:  4
#>  Project name:  boct193, w:  4, Rs:  1
#>  Project name:  boct194, w:  4, Rs:  1
#>  Project name:  boct195, w:  4, Rs:  2
#>  Project name:  boct196, w:  4, Rs:  2
#>  Project name:  boct197, w:  4, Rs:  4
#>  Project name:  boct198, w:  4, Rs:  4
#>  Project name:  boct199, w:  4, Rs:  1
#>  Project name:  boct2, w:  4, Rs:  1
#>  Project name:  boct20, w:  4, Rs:  1
#>  Project name:  boct200, w:  4, Rs:  1
#>  Project name:  boct201, w:  4, Rs:  2
#>  Project name:  boct202, w:  4, Rs:  2
#>  Project name:  boct203, w:  4, Rs:  4
#>  Project name:  boct204, w:  4, Rs:  4
#>  Project name:  boct205, w:  4, Rs:  1
#>  Project name:  boct206, w:  4, Rs:  1
#>  Project name:  boct207, w:  4, Rs:  2
#>  Project name:  boct208, w:  4, Rs:  2
#>  Project name:  boct209, w:  4, Rs:  4
#>  Project name:  boct21, w:  4, Rs:  2
#>  Project name:  boct210, w:  4, Rs:  4
#>  Project name:  boct211, w:  4, Rs:  1
#>  Project name:  boct212, w:  4, Rs:  1
#>  Project name:  boct213, w:  4, Rs:  2
#>  Project name:  boct214, w:  4, Rs:  2
#>  Project name:  boct215, w:  4, Rs:  4
#>  Project name:  boct216, w:  4, Rs:  4
#>  Project name:  boct217, w:  4, Rs:  1
#>  Project name:  boct218, w:  4, Rs:  1
#>  Project name:  boct219, w:  4, Rs:  2
#>  Project name:  boct22, w:  4, Rs:  2
#>  Project name:  boct220, w:  4, Rs:  2
#>  Project name:  boct221, w:  4, Rs:  4
#>  Project name:  boct222, w:  4, Rs:  4
#>  Project name:  boct223, w:  4, Rs:  1
#>  Project name:  boct224, w:  4, Rs:  1
#>  Project name:  boct225, w:  4, Rs:  2
#>  Project name:  boct226, w:  4, Rs:  2
#>  Project name:  boct227, w:  4, Rs:  4
#>  Project name:  boct228, w:  4, Rs:  4
#>  Project name:  boct229, w:  4, Rs:  1
#>  Project name:  boct23, w:  4, Rs:  4
#>  Project name:  boct230, w:  4, Rs:  1
#>  Project name:  boct231, w:  4, Rs:  2
#>  Project name:  boct232, w:  4, Rs:  2
#>  Project name:  boct233, w:  4, Rs:  4
#>  Project name:  boct234, w:  4, Rs:  4
#>  Project name:  boct235, w:  4, Rs:  1
#>  Project name:  boct236, w:  4, Rs:  1
#>  Project name:  boct237, w:  4, Rs:  2
#>  Project name:  boct238, w:  4, Rs:  2
#>  Project name:  boct239, w:  4, Rs:  4
#>  Project name:  boct24, w:  4, Rs:  4
#>  Project name:  boct240, w:  4, Rs:  4
#>  Project name:  boct25, w:  4, Rs:  1
#>  Project name:  boct26, w:  4, Rs:  1
#>  Project name:  boct27, w:  4, Rs:  2
#>  Project name:  boct28, w:  4, Rs:  2
#>  Project name:  boct29, w:  4, Rs:  4
#>  Project name:  boct3, w:  4, Rs:  2
#>  Project name:  boct30, w:  4, Rs:  4
#>  Project name:  boct31, w:  4, Rs:  1
#>  Project name:  boct32, w:  4, Rs:  1
#>  Project name:  boct33, w:  4, Rs:  2
#>  Project name:  boct34, w:  4, Rs:  2
#>  Project name:  boct35, w:  4, Rs:  4
#>  Project name:  boct36, w:  4, Rs:  4
#>  Project name:  boct37, w:  4, Rs:  1
#>  Project name:  boct38, w:  4, Rs:  1
#>  Project name:  boct39, w:  4, Rs:  2
#>  Project name:  boct4, w:  4, Rs:  2
#>  Project name:  boct40, w:  4, Rs:  2
#>  Project name:  boct41, w:  4, Rs:  4
#>  Project name:  boct42, w:  4, Rs:  4
#>  Project name:  boct43, w:  4, Rs:  1
#>  Project name:  boct44, w:  4, Rs:  1
#>  Project name:  boct45, w:  4, Rs:  2
#>  Project name:  boct46, w:  4, Rs:  2
#>  Project name:  boct47, w:  4, Rs:  4
#>  Project name:  boct48, w:  4, Rs:  4
#>  Project name:  boct49, w:  4, Rs:  1
#>  Project name:  boct5, w:  4, Rs:  4
#>  Project name:  boct50, w:  4, Rs:  1
#>  Project name:  boct51, w:  4, Rs:  2
#>  Project name:  boct52, w:  4, Rs:  2
#>  Project name:  boct53, w:  4, Rs:  4
#>  Project name:  boct54, w:  4, Rs:  4
#>  Project name:  boct55, w:  4, Rs:  1
#>  Project name:  boct56, w:  4, Rs:  1
#>  Project name:  boct57, w:  4, Rs:  2
#>  Project name:  boct58, w:  4, Rs:  2
#>  Project name:  boct59, w:  4, Rs:  4
#>  Project name:  boct6, w:  4, Rs:  4
#>  Project name:  boct60, w:  4, Rs:  4
#>  Project name:  boct61, w:  4, Rs:  1
#>  Project name:  boct62, w:  4, Rs:  1
#>  Project name:  boct63, w:  4, Rs:  2
#>  Project name:  boct64, w:  4, Rs:  2
#>  Project name:  boct65, w:  4, Rs:  4
#>  Project name:  boct66, w:  4, Rs:  4
#>  Project name:  boct67, w:  4, Rs:  1
#>  Project name:  boct68, w:  4, Rs:  1
#>  Project name:  boct69, w:  4, Rs:  2
#>  Project name:  boct7, w:  4, Rs:  1
#>  Project name:  boct70, w:  4, Rs:  2
#>  Project name:  boct71, w:  4, Rs:  4
#>  Project name:  boct72, w:  4, Rs:  4
#>  Project name:  boct73, w:  4, Rs:  1
#>  Project name:  boct74, w:  4, Rs:  1
#>  Project name:  boct75, w:  4, Rs:  2
#>  Project name:  boct76, w:  4, Rs:  2
#>  Project name:  boct77, w:  4, Rs:  4
#>  Project name:  boct78, w:  4, Rs:  4
#>  Project name:  boct79, w:  4, Rs:  1
#>  Project name:  boct8, w:  4, Rs:  1
#>  Project name:  boct80, w:  4, Rs:  1
#>  Project name:  boct81, w:  4, Rs:  2
#>  Project name:  boct82, w:  4, Rs:  2
#>  Project name:  boct83, w:  4, Rs:  4
#>  Project name:  boct84, w:  4, Rs:  4
#>  Project name:  boct85, w:  4, Rs:  1
#>  Project name:  boct86, w:  4, Rs:  1
#>  Project name:  boct87, w:  4, Rs:  2
#>  Project name:  boct88, w:  4, Rs:  2
#>  Project name:  boct89, w:  4, Rs:  4
#>  Project name:  boct9, w:  4, Rs:  2
#>  Project name:  boct90, w:  4, Rs:  4
#>  Project name:  boct91, w:  4, Rs:  1
#>  Project name:  boct92, w:  4, Rs:  1
#>  Project name:  boct93, w:  4, Rs:  2
#>  Project name:  boct94, w:  4, Rs:  2
#>  Project name:  boct95, w:  4, Rs:  4
#>  Project name:  boct96, w:  4, Rs:  4
#>  Project name:  boct97, w:  4, Rs:  1
#>  Project name:  boct98, w:  4, Rs:  1
#>  Project name:  boct99, w:  4, Rs:  2

# Summary of the first project structure
summary(Boctor[[1]]$PDM_list)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  4
#> Number of resources (Rs):  1
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15 a_16
#> a_1    1   0   1   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_2    0   1   0   1   0   1   0   1   0    0    0    0    0    0    0    0
#> a_3    0   0   1   1   1   0   0   0   0    0    0    1    0    0    0    0
#> a_4    0   0   0   1   0   0   1   0   0    0    0    0    0    0    0    0
#> a_5    0   0   0   0   1   1   0   1   0    0    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   1   0   1    0    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0   0    1    0    0    1    0    0    0
#> a_8    0   0   0   0   0   0   0   1   0    0    1    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0   1    1    0    0    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0   0    1    1    0    0    0    0    1
#> a_11   0   0   0   0   0   0   0   0   0    0    1    0    0    1    0    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    1    0    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    1    1    0
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1    0    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0    1    1
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    1
#>      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28 a_29 a_30 a_31
#> a_1     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_11    1    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_12    1    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_14    0    1    1    0    0    0    0    0    0    0    0    0    0    0    0
#> a_15    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_16    1    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#>      a_32 a_33 a_34 a_35 a_36 a_37 a_38 a_39 a_40 a_41 a_42 a_43 a_44 a_45 a_46
#> a_1     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_15    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_16    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#>      a_47 a_48 a_49 a_50 t_1 t_2 t_3 t_4 c_1 c_2 c_3 c_4 r_1.1 r_2.1 r_3.1
#> a_1     0    0    0    0  13  17  26  28   0   0   0   0     4     3     2
#> a_2     0    0    0    0  15  18  24  26   0   0   0   0     5     4     3
#> a_3     0    0    0    0  11  19  20   0   0   0   0   0     3     2     1
#> a_4     0    0    0    0   9  15   0   0   0   0   0   0     2     1     0
#> a_5     0    0    0    0   5   0   0   0   0   0   0   0     5     0     0
#> a_6     0    0    0    0   2   4   5   6   0   0   0   0     4     3     2
#> a_7     0    0    0    0   8  16  24  25   0   0   0   0     5     3     2
#> a_8     0    0    0    0   3   5   7   0   0   0   0   0     3     2     1
#> a_9     0    0    0    0  11  16   0   0   0   0   0   0     5     3     0
#> a_10    0    0    0    0   6  12   0   0   0   0   0   0     5     4     0
#> a_11    0    0    0    0  10   0   0   0   0   0   0   0     5     0     0
#> a_12    0    0    0    0   1   0   0   0   0   0   0   0     3     0     0
#> a_13    0    0    0    0   9   0   0   0   0   0   0   0     2     0     0
#> a_14    0    0    0    0   8  16   0   0   0   0   0   0     3     2     0
#> a_15    0    0    0    0   3   0   0   0   0   0   0   0     1     0     0
#> a_16    0    0    0    0   7  11   0   0   0   0   0   0     3     1     0
#>      r_4.1
#> a_1      1
#> a_2      1
#> a_3      0
#> a_4      0
#> a_5      0
#> a_6      1
#> a_7      1
#> a_8      0
#> a_9      0
#> a_10     0
#> a_11     0
#> a_12     0
#> a_13     0
#> a_14     0
#> a_15     0
#> a_16     0
#>  [ reached getOption("max.print") -- omitted 34 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  91
#> Const constraint (Cc):  0
#> Score/scope constraint (Cs):  1
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  427
#> Const constraint (Cc):  0
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#>     R_1
#> TPR  10

# Summary of the first project constraints
summary(Boctor[[1]]$PDM_const)
#> 
#> Summary of the PDM constraints structure:
#> 
#> Resource constraint(s) (CR):
#>     R_1
#> TPR   7