Function to calculate minimal/maximal/most likely project structures.

Specify minimal, maximal, maximin, minimax, and most likely/most desired project structures.

get.structures(x,type=c("min","max","minimax","maximin","most"))

Arguments

x

N by M PDM matrix or a PDM list.

type

type of specified project structure. "min" is a minimal strucuture, where the time, cost demands, and as well as the quality parameters are minimal; "max" where the time, cost demands, and as well as the quality parameters are maximal; "minimax", where the cost and resource demands are minimal; "maximix", where the cost and resource demands are maxnimal; "most", where the total project score is maximal.

Value

PDMs

set of PDM_matrix or a set of PDM_list (depending on the input type)

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Generation of PDM matrix for flexible project planning MFPP package.

# Define number of modes, flexibility factor and connectivity factor of a project scenerio.
N=4;ff=0.05;cf=0

# Define maximum value of time domain, Cost domain and Resourcces domain of a project scenerio.
mTD=3;mCD=4;mRD=3

# Define number of modes, number of resources,
# number of possible extra tasks, scale and quality domain of a project scenerio.

w=2;nR=2;nW=2
scale=1.4  #(default value)

# Generation of PDM matrix for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   0   0   0   0 1.873757 1.928120 3.116710 1.928120 2.058529
#> a_2   0   1   1   0   0   0 2.428791 2.492385 3.159771 2.492385 2.067952
#> a_3   0   0   1   0   0   0 2.252833 2.687654 2.352789 2.687654 2.004576
#> a_4   0   0   0   1   0   0 1.065629 1.116171 1.882180 1.116171 1.203337
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.509467 2.334200 2.476161
#> a_2 2.185418 2.400045 2.866924
#> a_3 2.158552 2.398464 2.685312
#> a_4 1.278828 2.010649 2.404273
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   0   0   0   0 1.873757 1.928120 3.116710 1.928120 2.058529
#> a_2   0   1   1   0   0   0 2.428791 2.492385 3.159771 2.492385 2.067952
#> a_3   0   0   1   0   0   0 2.252833 2.687654 2.352789 2.687654 2.004576
#> a_4   0   0   0   1   0   0 1.065629 1.116171 1.882180 1.116171 1.203337
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.509467 2.334200 2.476161
#> a_2 2.185418 2.400045 2.866924
#> a_3 2.158552 2.398464 2.685312
#> a_4 1.278828 2.010649 2.404273
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $moststruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   0   0   0   0 1.873757 1.928120 3.116710 1.928120 2.058529
#> a_2   0   1   1   0   0   0 2.428791 2.492385 3.159771 2.492385 2.067952
#> a_3   0   0   1   0   0   0 2.252833 2.687654 2.352789 2.687654 2.004576
#> a_4   0   0   0   1   0   0 1.065629 1.116171 1.882180 1.116171 1.203337
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.509467 2.334200 2.476161
#> a_2 2.185418 2.400045 2.866924
#> a_3 2.158552 2.398464 2.685312
#> a_4 1.278828 2.010649 2.404273
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $minimaxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   0   0   0   0 1.873757 1.928120 3.116710 1.928120 2.058529
#> a_2   0   1   1   0   0   0 2.428791 2.492385 3.159771 2.492385 2.067952
#> a_3   0   0   1   0   0   0 2.252833 2.687654 2.352789 2.687654 2.004576
#> a_4   0   0   0   1   0   0 1.065629 1.116171 1.882180 1.116171 1.203337
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.509467 2.334200 2.476161
#> a_2 2.185418 2.400045 2.866924
#> a_3 2.158552 2.398464 2.685312
#> a_4 1.278828 2.010649 2.404273
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maximinstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   0   0   0   0 1.873757 1.928120 3.116710 1.928120 2.058529
#> a_2   0   1   1   0   0   0 2.428791 2.492385 3.159771 2.492385 2.067952
#> a_3   0   0   1   0   0   0 2.252833 2.687654 2.352789 2.687654 2.004576
#> a_4   0   0   0   1   0   0 1.065629 1.116171 1.882180 1.116171 1.203337
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.509467 2.334200 2.476161
#> a_2 2.185418 2.400045 2.866924
#> a_3 2.158552 2.398464 2.685312
#> a_4 1.278828 2.010649 2.404273
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> attr(,"class")
#> [1] "Set_PDM_matrix"

# Generation of PDM list for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=FALSE,lst=TRUE)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.540627 2.855874 2.655598 2.855874 2.2796946
#> a_2   0   1   1   1   0   0 2.531194 2.542678 3.267832 2.542678 2.0871409
#> a_3   0   0   0   0   0   0 1.747697 1.842291 3.509658 1.842291 2.5904967
#> a_4   0   0   0   1   0   0 1.814196 2.085296 3.412143 2.085296 0.9539114
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1    r_1.2    r_2.2
#> a_1 2.4652461 2.295387 2.751307
#> a_2 2.4985480 2.376014 2.920906
#> a_3 2.7953832 1.729249 1.944163
#> a_4 0.9736737 2.846764 2.956645
#> a_5 0.0000000 0.000000 0.000000
#> a_6 0.0000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.540627 2.855874 2.655598 2.855874 2.2796946
#> a_2   0   1   1   1   0   0 2.531194 2.542678 3.267832 2.542678 2.0871409
#> a_3   0   0   1   0   0   0 1.747697 1.842291 3.509658 1.842291 2.5904967
#> a_4   0   0   0   1   0   0 1.814196 2.085296 3.412143 2.085296 0.9539114
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1    r_1.2    r_2.2
#> a_1 2.4652461 2.295387 2.751307
#> a_2 2.4985480 2.376014 2.920906
#> a_3 2.7953832 1.729249 1.944163
#> a_4 0.9736737 2.846764 2.956645
#> a_5 0.0000000 0.000000 0.000000
#> a_6 0.0000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $moststruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.540627 2.855874 2.655598 2.855874 2.2796946
#> a_2   0   1   1   1   0   0 2.531194 2.542678 3.267832 2.542678 2.0871409
#> a_3   0   0   1   0   0   0 1.747697 1.842291 3.509658 1.842291 2.5904967
#> a_4   0   0   0   1   0   0 1.814196 2.085296 3.412143 2.085296 0.9539114
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1    r_1.2    r_2.2
#> a_1 2.4652461 2.295387 2.751307
#> a_2 2.4985480 2.376014 2.920906
#> a_3 2.7953832 1.729249 1.944163
#> a_4 0.9736737 2.846764 2.956645
#> a_5 0.0000000 0.000000 0.000000
#> a_6 0.0000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $minimaxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.540627 2.855874 2.655598 2.855874 2.2796946
#> a_2   0   1   1   1   0   0 2.531194 2.542678 3.267832 2.542678 2.0871409
#> a_3   0   0   0   0   0   0 1.747697 1.842291 3.509658 1.842291 2.5904967
#> a_4   0   0   0   1   0   0 1.814196 2.085296 3.412143 2.085296 0.9539114
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1    r_1.2    r_2.2
#> a_1 2.4652461 2.295387 2.751307
#> a_2 2.4985480 2.376014 2.920906
#> a_3 2.7953832 1.729249 1.944163
#> a_4 0.9736737 2.846764 2.956645
#> a_5 0.0000000 0.000000 0.000000
#> a_6 0.0000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maximinstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.540627 2.855874 2.655598 2.855874 2.2796946
#> a_2   0   1   1   1   0   0 2.531194 2.542678 3.267832 2.542678 2.0871409
#> a_3   0   0   1   0   0   0 1.747697 1.842291 3.509658 1.842291 2.5904967
#> a_4   0   0   0   1   0   0 1.814196 2.085296 3.412143 2.085296 0.9539114
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1    r_1.2    r_2.2
#> a_1 2.4652461 2.295387 2.751307
#> a_2 2.4985480 2.376014 2.920906
#> a_3 2.7953832 1.729249 1.944163
#> a_4 0.9736737 2.846764 2.956645
#> a_5 0.0000000 0.000000 0.000000
#> a_6 0.0000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> attr(,"class")
#> [1] "Set_PDM_list"