Calculate Pareto-optimal resource allocation

Calculate Pareto-optimal resource allocation

paretores(DSM,TD,RD)

Arguments

DSM

An N by N upper triangular adjacency matrix of logic network with fixed dependencies (a binary matrix).

TD

N by 1 matrix of task durations (a numeric matrix)

RD

N by r matrix of task resources, where the number of resources is r.

Value

RD

Pareto-optimal TPR (1 by r numeric vector)).

SST

Scheduled Start Time (N by N numeric matrix)).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Calculation of Pareto-optimal resource allocation using MFPP package.

# Define a 3 by 3 upper traingular logic domain of a project structure.

# Specification of Logic Domain
DSM<-matrix(c(1,1,0,1,0,0,
              0,1,0,0,0,0,
              0,0,1,1,0,0,
              0,0,0,1,0,0,
              0,0,0,0,0,0,
              0,0,0,0,0,0),nrow=6, byrow=TRUE)

# Specification of Time Domain
TD<-matrix(c(2,
             3,
             1,
             4,
             0,
             0),nrow=6, byrow=TRUE)

# Specification of Resource Domain
RD<-matrix(c(4,2,
             3,2,
             5,1,
             6,4,
             0,0,
             0,0),nrow=6, byrow=TRUE)

RES<-paretores(DSM,TD,RD)
#> ********** R based Nondominated Sorting Genetic Algorithm II *********
#> initializing the population
#> ranking the initial population
#> crowding distance calculation
#> ---------------generation--------------- 1 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 1 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 2 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 2 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 3 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 3 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 4 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 4 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 5 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 5 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 6 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 6 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 7 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 7 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 8 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 8 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 9 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 9 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 10 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 10 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 11 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 11 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 12 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 12 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 13 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 13 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 14 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 14 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 15 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 15 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 16 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 16 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 17 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 17 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 18 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 18 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 19 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 19 ends
#> 
#> ********** new iteration *********
#> ---------------generation--------------- 20 starts
#> tournament selection
#> crossover operator
#> mutation operator
#> evaluate the objective fns of childAfterM
#> Rt = Pt + Qt
#> ranking again
#> crowded comparison again
#> environmental selection
#> ---------------generation--------------- 20 ends
#> ********** stop the evolution *********
RES
#> $RD
#>     R_1 R_2
#> TPR   9   6
#> 
#> $SST
#>         SST
#>  0.96913869
#>  0.01657898
#>  0.00000000
#>  1.00000000
#>  1.29597053
#>  0.22573108
#> 
# Verify results
tpr(RES$SST,DSM,TD,RD)
#>     R_1 R_2
#> TPR   9   6