Function to simulate estimation uncertainty.

Simulating the effects of the estimatio uncertainty.

phase1(x,a=-0.1,b=0.30,pdftype="uniform")

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

a

optimistic duration, the negative percent deviation from the actual duration (a negative value). The default value is a=-0.1, which means if the actual duration is 1, the minimal possible duration is 0.9.

b

pessimistic duration, the positive percent deviation from the actual duration (a positive value). The default value is b=0.3, which means if the actual duration is 1, the maximal possible duration is 1.3.

pdftype

the pdftype is either 'uniform' or 'beta' (the default value is 'uniform'), which means the generated distribution either follows uniform or a beta distribution betweem interval [a,b]

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2      a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13      a_14
#> a_1    1   1 0.000000   1   1   1   0   0   0    0    0    0    0 0.0000000
#> a_2    0   1 1.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_3    0   0 0.806829   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_4    0   0 0.000000   1   1   0   0   0   0    0    0    0    0 0.0000000
#> a_5    0   0 0.000000   0   1   0   0   0   1    0    0    0    0 0.0000000
#> a_6    0   0 0.000000   0   0   1   0   0   0    0    0    0    0 0.0000000
#> a_7    0   0 0.000000   0   0   0   1   0   0    1    0    0    0 0.0000000
#> a_8    0   0 0.000000   0   0   0   0   1   0    0    0    1    0 0.0000000
#> a_9    0   0 0.000000   0   0   0   0   0   1    0    0    0    0 0.0000000
#> a_10   0   0 0.000000   0   0   0   0   0   0    1    0    0    0 0.0000000
#> a_11   0   0 0.000000   0   0   0   0   0   0    0    1    1    0 0.0000000
#> a_12   0   0 0.000000   0   0   0   0   0   0    0    0    1    1 0.0000000
#> a_13   0   0 0.000000   0   0   0   0   0   0    0    0    0    1 1.0000000
#> a_14   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.8989999
#> a_15   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_16   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_17   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_18   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_19   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_20   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_21   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_22   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_23   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_24   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_25   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_26   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_27   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_28   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_29   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_30   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_31   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_32   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#>      a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_2     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_7     1    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0 0.0000000    1    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_14    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_15    1    1 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_16    0    1 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_17    0    0 0.7010317    0    0    0    0    0    0    0    0    0    0    0
#> a_18    0    0 0.0000000    1    0    1    0    1    0    0    0    0    0    0
#> a_19    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_20    0    0 0.0000000    0    0    1    1    0    0    0    0    0    0    0
#> a_21    0    0 0.0000000    0    0    0    1    0    1    0    0    0    0    0
#> a_22    0    0 0.0000000    0    0    0    0    1    0    1    0    0    1    0
#> a_23    0    0 0.0000000    0    0    0    0    0    1    0    0    0    0    0
#> a_24    0    0 0.0000000    0    0    0    0    0    0    1    0    0    0    0
#> a_25    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0    1
#> a_26    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_27    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1    0
#> a_28    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    1
#> a_29    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#>      a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2        q_1
#> a_1     0    0    0    0 11.822129 12.675231 21.611949 23.846201 0.74944953
#> a_2     0    0    0    0 16.507186 19.840959 12.063318 13.420962 0.78956001
#> a_3     0    0    0    0  7.606723  8.226604 20.410482 21.219761 0.58526344
#> a_4     0    0    0    0 18.203500 19.986367 24.360343 25.848348 0.62126608
#> a_5     0    0    0    0  5.272433  6.071614 25.274916 29.090610 0.66648739
#> a_6     0    0    0    0 10.890426 13.504482 25.800383 25.836661 0.60839832
#> a_7     0    0    0    0 16.710533 16.970201 24.631510 28.812009 0.52291700
#> a_8     0    0    0    0  7.272821  8.040311 12.925094 15.374841 0.65735984
#> a_9     0    0    0    0 16.944250 19.404009 24.703954 28.882378 0.52971488
#> a_10    0    0    0    0 11.308447 12.173605 24.239267 26.106953 0.03294064
#> a_11    0    0    0    0 10.945165 11.840139 26.028483 27.524269 0.66888513
#> a_12    0    0    0    0 14.800725 16.134976 20.007011 21.870251 0.85545168
#> a_13    0    0    0    0 15.785557 16.465403  9.589183 10.953800 0.84708393
#> a_14    0    0    0    0 12.775573 15.451755 11.474250 12.176814 0.85552474
#> a_15    0    0    0    0 17.468029 19.915686 10.566670 11.005065 0.67347687
#> a_16    0    0    0    0 17.842190 18.158003 22.981740 28.441177 0.42354981
#> a_17    0    0    0    0 10.894846 11.523669 15.542406 15.768455 0.51248200
#> a_18    0    0    0    0 12.322000 14.565129 15.137579 16.624389 0.73520071
#> a_19    0    0    0    0 12.350835 14.486641 15.263562 18.820086 0.66651279
#> a_20    0    0    0    0 10.377507 12.941202 22.643177 24.174604 0.79472313
#> a_21    0    0    0    0 16.918287 19.978976 12.966632 13.605463 0.75558655
#> a_22    0    0    0    0 12.671095 12.976515 17.578898 20.287704 0.35811437
#> a_23    0    0    0    0 15.448850 15.695624 22.775063 28.331342 0.88966917
#> a_24    0    0    0    0 19.469139 19.726742  8.963825 10.110919 0.49087500
#> a_25    0    0    0    0  8.786977  9.363707 26.773754 29.863225 0.89195994
#> a_26    0    0    0    0 14.319861 17.469934 25.847372 28.826439 0.12121700
#> a_27    0    0    0    0  9.943809 10.025231  2.527764  2.541299 0.48860594
#> a_28    0    0    0    0 16.078979 16.477761  8.196306 10.102226 0.53893350
#> a_29    1    0    0    0 12.940406 13.267783 26.787954 27.503375 0.51411883
#> a_30    0    1    0    0 14.748218 16.008071  7.281673  7.474241 0.58008437
#> a_31    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.00000000
#> a_32    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.00000000
#>             q_2     r_1.1      r_2.1     r_1.2     r_2.2
#> a_1  0.91087021 14.349104 16.6841153 14.372374 14.747200
#> a_2  0.93413990 16.603935 18.8668197 14.422264 16.882671
#> a_3  0.72740001 10.042552 12.5274538  4.039188  4.465396
#> a_4  0.76796660 15.826708 19.0307697 10.188093 12.015816
#> a_5  0.67858243  9.382517 10.6564582  5.252162  6.276426
#> a_6  0.69551745 11.350936 11.9858840 15.997469 16.121093
#> a_7  0.57272605 15.850606 16.1313738  7.309479  7.659326
#> a_8  0.78142803 13.549337 15.9963525 12.302351 13.065944
#> a_9  0.60146274  2.180037  2.5577646  9.290463 10.692974
#> a_10 0.03760558 13.046135 16.0060442 12.348976 14.742475
#> a_11 0.75871573  0.877524  0.9529533  7.634493  7.686852
#> a_12 0.86599690 14.688357 16.0607269  1.012024  1.022350
#> a_13 0.90333138 17.945649 19.2458038  6.882311  7.670866
#> a_14 0.92512953  3.314131  4.0505063 14.646390 16.750416
#> a_15 0.79474014 17.399411 18.3790703 12.232155 12.526418
#> a_16 0.46677000 11.698884 11.9119476 15.812611 19.535873
#> a_17 0.52750011 18.815569 19.1550970 12.538088 14.955887
#> a_18 0.89472464 12.964572 15.7822186 18.295584 18.794869
#> a_19 0.77643073 14.209485 15.6579155  5.692368  6.489086
#> a_20 0.82327708 14.667995 18.0091610  6.061982  6.809302
#> a_21 0.92269408 11.845849 12.6974641 10.852543 12.409485
#> a_22 0.38405795 15.163505 15.6170506 12.574477 14.610420
#> a_23 0.91169663 14.437970 16.1799370  7.019724  8.690424
#> a_24 0.61148380 10.897684 11.5103135  5.611289  5.871938
#> a_25 0.98839425 16.785619 18.8748886 11.976599 13.817696
#> a_26 0.12405961 16.832254 17.8891826 10.396901 12.937932
#> a_27 0.48866767 14.429593 16.1630250  9.703273 11.471818
#> a_28 0.60938275  5.078915  5.6456101 10.851460 11.739412
#> a_29 0.53531891 14.588214 14.8510177 18.054068 18.881421
#> a_30 0.59955602 15.027092 17.3353392 17.162987 19.320218
#> a_31 0.00000000  0.000000  0.0000000  0.000000  0.000000
#> a_32 0.00000000  0.000000  0.0000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define negative percentage deviation from actual duration.
a=-0.20

# Define negative percentage deviation from actual duration.
b=0.40

# Simulate the effects of estimation uncertainty of a project scenario using MFPP package
# with default values.
 Res<-phase1(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2      a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13      a_14
#> a_1    1   1 0.000000   1   1   1   0   0   0    0    0    0    0 0.0000000
#> a_2    0   1 1.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_3    0   0 0.806829   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_4    0   0 0.000000   1   1   0   0   0   0    0    0    0    0 0.0000000
#> a_5    0   0 0.000000   0   1   0   0   0   1    0    0    0    0 0.0000000
#> a_6    0   0 0.000000   0   0   1   0   0   0    0    0    0    0 0.0000000
#> a_7    0   0 0.000000   0   0   0   1   0   0    1    0    0    0 0.0000000
#> a_8    0   0 0.000000   0   0   0   0   1   0    0    0    1    0 0.0000000
#> a_9    0   0 0.000000   0   0   0   0   0   1    0    0    0    0 0.0000000
#> a_10   0   0 0.000000   0   0   0   0   0   0    1    0    0    0 0.0000000
#> a_11   0   0 0.000000   0   0   0   0   0   0    0    1    1    0 0.0000000
#> a_12   0   0 0.000000   0   0   0   0   0   0    0    0    1    1 0.0000000
#> a_13   0   0 0.000000   0   0   0   0   0   0    0    0    0    1 1.0000000
#> a_14   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.8989999
#> a_15   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_16   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_17   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_18   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_19   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_20   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_21   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_22   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_23   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_24   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_25   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_26   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_27   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_28   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_29   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_30   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_31   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_32   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#>      a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_2     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_7     1    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0 0.0000000    1    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_14    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_15    1    1 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_16    0    1 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_17    0    0 0.7010317    0    0    0    0    0    0    0    0    0    0    0
#> a_18    0    0 0.0000000    1    0    1    0    1    0    0    0    0    0    0
#> a_19    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_20    0    0 0.0000000    0    0    1    1    0    0    0    0    0    0    0
#> a_21    0    0 0.0000000    0    0    0    1    0    1    0    0    0    0    0
#> a_22    0    0 0.0000000    0    0    0    0    1    0    1    0    0    1    0
#> a_23    0    0 0.0000000    0    0    0    0    0    1    0    0    0    0    0
#> a_24    0    0 0.0000000    0    0    0    0    0    0    1    0    0    0    0
#> a_25    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0    1
#> a_26    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_27    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1    0
#> a_28    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    1
#> a_29    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#>      a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2        q_1
#> a_1     0    0    0    0 11.607700 12.984105 19.598654 25.756181 0.76415367
#> a_2     0    0    0    0 19.200473 25.278101 12.867832 12.803684 0.84341927
#> a_3     0    0    0    0  9.403588  8.588379 20.010943 22.581933 0.55036370
#> a_4     0    0    0    0 19.846297 22.734401 28.100891 29.585794 0.74275892
#> a_5     0    0    0    0  6.724531  7.276457 32.495667 33.781632 0.68626977
#> a_6     0    0    0    0 10.335244 14.644660 28.677106 24.839948 0.54877890
#> a_7     0    0    0    0 19.919381 17.321095 25.327698 36.909295 0.65867318
#> a_8     0    0    0    0  8.677365  8.544845 14.430461 17.760076 0.78332123
#> a_9     0    0    0    0 20.724501 23.493682 22.895731 31.324912 0.60403823
#> a_10    0    0    0    0 10.199234 13.851807 24.932950 29.704991 0.03109238
#> a_11    0    0    0    0  9.912137 12.644505 25.565257 34.674317 0.80268056
#> a_12    0    0    0    0 16.206979 19.970338 20.603432 20.820334 0.91673170
#> a_13    0    0    0    0 15.927329 18.239855  8.809797 11.261744 1.00000000
#> a_14    0    0    0    0 14.960459 14.071655 14.221339 15.125840 0.87154154
#> a_15    0    0    0    0 22.562090 18.181347 13.521435 12.725151 0.70431972
#> a_16    0    0    0    0 18.169359 17.239023 24.001120 33.825447 0.45744689
#> a_17    0    0    0    0 13.446938 12.966854 15.826301 17.849863 0.62766102
#> a_18    0    0    0    0 13.220446 17.305537 15.033972 17.526221 0.80223432
#> a_19    0    0    0    0 13.367579 16.663446 14.495207 21.155631 0.61813214
#> a_20    0    0    0    0  9.523304 16.261662 25.604677 24.829148 0.76795424
#> a_21    0    0    0    0 15.293127 20.105760 12.196390 17.064699 0.80190039
#> a_22    0    0    0    0 12.328410 15.547564 16.494012 23.628896 0.32962009
#> a_23    0    0    0    0 13.906538 15.729522 24.544561 32.288461 1.00000000
#> a_24    0    0    0    0 22.974101 19.512058  8.645297 10.277344 0.51176733
#> a_25    0    0    0    0 10.094719 12.124203 26.567813 35.974138 0.96890962
#> a_26    0    0    0    0 13.833517 16.185755 30.074723 27.306521 0.14030246
#> a_27    0    0    0    0  9.318106 12.653316  2.471491  2.668891 0.62886343
#> a_28    0    0    0    0 14.638214 15.762158 10.109017 10.090158 0.52411366
#> a_29    1    0    0    0 15.003092 12.097277 27.034756 35.719344 0.46907455
#> a_30    0    1    0    0 14.759866 19.721596  7.140240  7.971924 0.53930876
#> a_31    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.00000000
#> a_32    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.00000000
#>             q_2      r_1.1      r_2.1     r_1.2     r_2.2
#> a_1  0.93477006 15.3865317 16.2981342 18.175750 13.802018
#> a_2  1.00000000 19.9919144 22.3315747 17.520930 20.504410
#> a_3  0.80743909 11.6680145 16.1568848  3.857641  5.012785
#> a_4  0.80592061 19.2439197 20.1650502 11.297775 15.185597
#> a_5  0.72866832 11.8482523 12.7300199  4.879705  7.785053
#> a_6  0.80319099 10.5912348 13.0812812 15.735846 19.825832
#> a_7  0.65802531 16.5288856 17.7661924  9.030786  9.402633
#> a_8  0.82848304 16.9321898 16.0125333 11.181621 15.426573
#> a_9  0.70164051  2.7925052  2.9536492  9.890334 10.052495
#> a_10 0.03620718 14.2048227 18.0864331 14.460631 14.117006
#> a_11 0.69820119  0.7947615  0.8918168  9.792115  9.423369
#> a_12 0.84141703 17.3186188 14.9074578  1.081089  1.081266
#> a_13 0.95942208 19.3705800 19.4813729  8.488216  8.705287
#> a_14 1.00000000  3.7325828  4.7328285 15.241207 21.451139
#> a_15 0.91564919 20.6967578 20.8056603 12.808069 12.536925
#> a_16 0.50007593 12.9644203 12.0903478 15.232571 18.860271
#> a_17 0.57886384 21.9430811 20.5134552 11.752800 13.484797
#> a_18 1.00000000 16.1874392 14.5897881 17.625890 17.713676
#> a_19 0.73008300 16.7636176 15.5787863  6.688022  6.559079
#> a_20 1.00000000 14.7529877 17.7101033  5.810960  6.287305
#> a_21 0.87265394 11.3103611 14.7487441 12.173775 12.853856
#> a_22 0.36382843 14.1824959 18.0651962 12.095863 15.756913
#> a_23 1.00000000 14.8677400 17.0258782  7.506080 11.170009
#> a_24 0.60040478 13.1543464 13.9609439  6.526381  6.445179
#> a_25 1.00000000 19.7212537 21.6192803 10.990118 16.604491
#> a_26 0.14107666 16.3664382 17.1729071  9.364018 13.648663
#> a_27 0.54120319 13.1904261 16.0669515 11.872082 12.599732
#> a_28 0.78335330  5.8838709  6.4944849 12.816621 13.717769
#> a_29 0.58742102 15.4030983 14.9694749 22.525802 18.408779
#> a_30 0.58626914 14.6507221 17.2078236 21.889482 23.542880
#> a_31 0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> a_32 0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  99.23874
#> Const constraint (Cc):  513.98
#> Score/scope constraint (Cs):  0.8514988
#> Quailty constraint (Cq):  0.4597352
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  122.907
#> Const constraint (Cc):  686.0927
#> Score/scope constraint (Cs):  0.9790868
#> Quailty constraint (Cq):  0.6424754
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 200.2709 200.1461

# Simulate the estimation uncertainty of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase1(PDM$PDM,a,b,"beta")
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2      a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13      a_14
#> a_1    1   1 0.000000   1   1   1   0   0   0    0    0    0    0 0.0000000
#> a_2    0   1 1.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_3    0   0 0.806829   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_4    0   0 0.000000   1   1   0   0   0   0    0    0    0    0 0.0000000
#> a_5    0   0 0.000000   0   1   0   0   0   1    0    0    0    0 0.0000000
#> a_6    0   0 0.000000   0   0   1   0   0   0    0    0    0    0 0.0000000
#> a_7    0   0 0.000000   0   0   0   1   0   0    1    0    0    0 0.0000000
#> a_8    0   0 0.000000   0   0   0   0   1   0    0    0    1    0 0.0000000
#> a_9    0   0 0.000000   0   0   0   0   0   1    0    0    0    0 0.0000000
#> a_10   0   0 0.000000   0   0   0   0   0   0    1    0    0    0 0.0000000
#> a_11   0   0 0.000000   0   0   0   0   0   0    0    1    1    0 0.0000000
#> a_12   0   0 0.000000   0   0   0   0   0   0    0    0    1    1 0.0000000
#> a_13   0   0 0.000000   0   0   0   0   0   0    0    0    0    1 1.0000000
#> a_14   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.8989999
#> a_15   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_16   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_17   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_18   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_19   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_20   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_21   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_22   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_23   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_24   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_25   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_26   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_27   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_28   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_29   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_30   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_31   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#> a_32   0   0 0.000000   0   0   0   0   0   0    0    0    0    0 0.0000000
#>      a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_2     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_7     1    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0 0.0000000    1    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_14    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_15    1    1 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_16    0    1 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_17    0    0 0.7010317    0    0    0    0    0    0    0    0    0    0    0
#> a_18    0    0 0.0000000    1    0    1    0    1    0    0    0    0    0    0
#> a_19    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_20    0    0 0.0000000    0    0    1    1    0    0    0    0    0    0    0
#> a_21    0    0 0.0000000    0    0    0    1    0    1    0    0    0    0    0
#> a_22    0    0 0.0000000    0    0    0    0    1    0    1    0    0    1    0
#> a_23    0    0 0.0000000    0    0    0    0    0    1    0    0    0    0    0
#> a_24    0    0 0.0000000    0    0    0    0    0    0    1    0    0    0    0
#> a_25    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0    1
#> a_26    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_27    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1    0
#> a_28    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    1
#> a_29    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#>      a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2        q_1
#> a_1     0    0    0    0 14.449737 14.771025 27.194930 22.658505 0.64207315
#> a_2     0    0    0    0 17.149451 20.319466 12.895188 14.475848 1.00000000
#> a_3     0    0    0    0  8.232908 10.732755 24.464231 22.264086 0.53585667
#> a_4     0    0    0    0 17.871018 22.686168 25.301566 25.004594 0.60429006
#> a_5     0    0    0    0  5.914088  7.086066 26.911578 31.731371 0.67424905
#> a_6     0    0    0    0 12.397021 15.996774 25.989353 31.786991 0.72664485
#> a_7     0    0    0    0 16.932304 16.406666 31.954344 30.923166 0.53844966
#> a_8     0    0    0    0  7.586335 10.070194 13.538573 15.022852 0.61375709
#> a_9     0    0    0    0 16.451453 19.863736 23.801670 27.738103 0.49414150
#> a_10    0    0    0    0 12.199211 14.265565 27.745111 28.246318 0.04335042
#> a_11    0    0    0    0 10.430061 13.011078 32.192967 27.392536 0.85854347
#> a_12    0    0    0    0 14.878424 18.908827 25.262209 23.522577 0.99943516
#> a_13    0    0    0    0 16.568204 17.739905 11.769937 10.952264 1.00000000
#> a_14    0    0    0    0 13.970350 20.262253 13.154739 14.636178 0.79514758
#> a_15    0    0    0    0 16.604320 23.779493 12.058058 13.199050 0.56896879
#> a_16    0    0    0    0 20.652846 19.405399 20.351986 27.126163 0.52024819
#> a_17    0    0    0    0 11.580795 13.275779 16.082952 16.392864 0.53557047
#> a_18    0    0    0    0 15.710840 13.319166 17.604184 18.577791 0.69092100
#> a_19    0    0    0    0 12.210497 13.687114 15.823185 21.208176 0.76370001
#> a_20    0    0    0    0 10.350742 11.950531 27.317927 27.667423 0.92167102
#> a_21    0    0    0    0 20.017027 22.282987 12.857879 13.907277 0.62914386
#> a_22    0    0    0    0 15.058217 16.035970 21.529264 23.376124 0.36782853
#> a_23    0    0    0    0 18.125440 17.332748 25.809574 32.403569 1.00000000
#> a_24    0    0    0    0 22.156044 23.054351 10.409932 10.670283 0.42601445
#> a_25    0    0    0    0 10.320950  8.514740 31.952204 35.115386 1.00000000
#> a_26    0    0    0    0 14.838265 19.169105 28.367855 26.492754 0.12859314
#> a_27    0    0    0    0 10.274949 10.825659  2.501105  2.293294 0.46090588
#> a_28    0    0    0    0 18.622050 18.896758  8.523329  9.521339 0.59441257
#> a_29    1    0    0    0 14.046727 14.535600 29.472519 27.062989 0.48818694
#> a_30    0    1    0    0 15.287966 17.536236  7.645109  7.962277 0.58330450
#> a_31    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.00000000
#> a_32    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.00000000
#>             q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  1.00000000 13.539974 16.300525 17.101837 19.343931
#> a_2  1.00000000 17.623122 22.352246 17.413964 18.940372
#> a_3  0.90286387 11.979045 13.911226  4.971545  4.436220
#> a_4  0.92861750 15.330584 25.861483 12.949083 11.942556
#> a_5  0.82634279 11.185254 10.561438  4.695361  6.247339
#> a_6  0.75795250 11.993831 13.818983 17.437299 16.383271
#> a_7  0.62318954 16.375166 16.135987  8.733954  7.178124
#> a_8  0.96395024 12.593709 18.912936 14.477287 15.072839
#> a_9  0.72208121  2.256184  2.938832  8.725126 10.470803
#> a_10 0.03869591 13.344431 14.674462 13.346301 16.982523
#> a_11 0.85218229  1.000000  1.000000  7.503263  9.812825
#> a_12 0.94583630 16.796479 20.281598  1.154663  1.168480
#> a_13 0.88674146 16.641956 21.292958  7.684635  8.011241
#> a_14 0.97525230  3.686462  4.228614 18.054904 18.780956
#> a_15 0.94134798 20.097424 18.250611 11.144072 15.694302
#> a_16 0.51981655 14.134122 10.198399 17.825304 18.302608
#> a_17 0.52362437 21.972192 21.938416 13.750753 14.760047
#> a_18 1.00000000 15.898939 19.244737 18.540966 17.981620
#> a_19 0.76643538 14.663675 15.013574  5.561417  7.058700
#> a_20 0.81752964 18.107479 17.438135  6.354187  8.139336
#> a_21 1.00000000 14.960962 16.027809  9.649348 14.858080
#> a_22 0.46925635 14.183489 20.918104 16.124647 18.085174
#> a_23 1.00000000 17.953136 19.365615  6.637798 10.993794
#> a_24 0.77561741 11.077595 11.926476  6.616480  7.580283
#> a_25 1.00000000 16.120024 21.312057 15.359944 13.818414
#> a_26 0.10485533 19.838622 22.733067 10.520620 15.218089
#> a_27 0.54130629 16.458714 18.917836 12.402811 11.841104
#> a_28 0.63980082  5.179841  5.421389 13.711863 12.737524
#> a_29 0.59843570 14.554992 16.299937 22.094811 20.059484
#> a_30 0.74482644 19.392733 14.879219 16.350034 23.501145
#> a_31 0.00000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.00000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  101.2562
#> Const constraint (Cc):  539.0668
#> Score/scope constraint (Cs):  0.8514988
#> Quailty constraint (Cq):  0.4469139
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  117.7058
#> Const constraint (Cc):  659.247
#> Score/scope constraint (Cs):  0.9790868
#> Quailty constraint (Cq):  0.6689116
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 206.1171 199.4627