Function to simulate shock effects.

Phase 2 simulates the shock effects. Where p percent of task demands can be increased up to s times.

phase2 (x,p=0.10, s=5.0)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability factor for task selection (a numeric value between 0 and 1). The default value is p=1.00

s

Scale factor: the ratio of the modification (a numeric integer). The default value is s=5.0

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8       a_9 a_10 a_11 a_12 a_13 a_14 a_15
#> a_1    1   1   1   0   1   0   0   0 0.0000000    0    0    0    0    0    0
#> a_2    0   1   1   0   1   0   0   0 0.0000000    0    0    0    0    0    0
#> a_3    0   0   1   1   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_5    0   0   0   0   1   1   0   0 0.0000000    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   1   0 0.0000000    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0 0.0000000    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1 0.0000000    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0 0.7618916    0    0    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0 0.0000000    1    0    0    0    0    0
#> a_11   0   0   0   0   0   0   0   0 0.0000000    0    1    0    0    0    1
#> a_12   0   0   0   0   0   0   0   0 0.0000000    0    0    1    0    1    0
#> a_13   0   0   0   0   0   0   0   0 0.0000000    0    0    0    1    0    0
#> a_14   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    1    1
#> a_15   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    1
#> a_16   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_17   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_24   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_25   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_26   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_27   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_28   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_29   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_30   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_31   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_32   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#>      a_16 a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25      a_26 a_27 a_28 a_29
#> a_1     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_2     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_3     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_4     0    0    0    0    1    0    0    0    0    0 0.0000000    0    0    0
#> a_5     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_6     0    0    0    0    0    0    1    0    0    0 0.0000000    0    0    0
#> a_7     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_8     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_9     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_10    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_11    0    0    0    0    0    0    0    0    0    0 1.0000000    0    0    0
#> a_12    0    0    0    0    0    0    0    0    0    0 1.0000000    0    0    0
#> a_13    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_14    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_15    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_16    1    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_17    0    1    0    1    0    0    0    0    0    0 0.0000000    0    0    0
#> a_18    0    0    1    1    0    1    0    0    0    0 0.0000000    0    0    0
#> a_19    0    0    0    1    1    1    0    0    0    1 0.0000000    0    0    0
#> a_20    0    0    0    0    1    1    0    0    0    0 0.0000000    0    0    0
#> a_21    0    0    0    0    0    1    0    0    0    0 0.0000000    0    0    0
#> a_22    0    0    0    0    0    0    1    0    1    0 0.0000000    0    0    0
#> a_23    0    0    0    0    0    0    0    1    0    0 0.0000000    0    0    0
#> a_24    0    0    0    0    0    0    0    0    1    0 0.0000000    1    0    0
#> a_25    0    0    0    0    0    0    0    0    0    1 0.0000000    0    0    0
#> a_26    0    0    0    0    0    0    0    0    0    0 0.5841634    0    0    0
#> a_27    0    0    0    0    0    0    0    0    0    0 0.0000000    1    0    0
#> a_28    0    0    0    0    0    0    0    0    0    0 0.0000000    0    1    1
#> a_29    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    1
#> a_30    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_31    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_32    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#>      a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1       q_2
#> a_1     0    0    0 16.347869 18.298229 17.155904 19.828587 0.4089204 0.4213035
#> a_2     0    0    0 13.863452 16.243329  8.402097  9.547030 0.9231952 0.9350867
#> a_3     0    0    0  3.480427  3.686522 12.727002 12.769227 0.7129179 0.8573705
#> a_4     0    0    0 13.721209 15.711412 23.062730 25.915935 0.6594185 0.6744472
#> a_5     0    0    0 14.131963 16.727585 21.834141 24.245701 0.1032106 0.1128993
#> a_6     0    0    0 16.413953 16.680300 15.481238 15.990274 0.7411996 0.9165027
#> a_7     0    0    0 13.186623 14.899179 19.225869 21.036511 0.5875751 0.6453637
#> a_8     0    0    0 12.416754 15.159107 24.177449 25.373347 0.8224378 0.9792452
#> a_9     0    0    0 14.317363 14.806764 15.693322 18.164857 0.4776105 0.5125546
#> a_10    0    0    0  6.933816  8.663723 12.290713 13.181177 0.4588683 0.5049861
#> a_11    0    0    0  3.667132  3.817730 25.922503 26.372165 0.3264099 0.3880509
#> a_12    0    0    0  3.850505  4.710925 23.853609 24.264802 0.7299107 0.8193567
#> a_13    0    0    0  8.815423 10.145003 19.267661 21.629470 0.6430413 0.7878548
#> a_14    0    0    0 14.692313 16.952350 12.308258 15.143500 0.8160978 0.8891377
#> a_15    0    0    0 13.137955 13.617844 28.242833 29.641111 0.3739066 0.4386208
#> a_16    1    0    0 13.394942 14.963177 28.870363 29.686101 0.7543239 0.8409803
#> a_17    0    0    0  7.266500  7.472334 15.469767 15.998222 0.8503954 0.9827242
#> a_18    0    0    0 14.908959 18.097767 17.937612 22.379042 0.7717529 0.8882783
#> a_19    0    0    0  8.328546  9.078106 19.047426 20.870878 0.5717522 0.6563525
#> a_20    0    0    0 17.578725 18.094342 13.186107 13.310176 0.7729071 0.7735467
#> a_21    0    0    0 11.302310 12.027360  5.966204  7.430421 0.3951842 0.4231735
#> a_22    0    0    0 11.892810 13.644096  2.794244  2.953205 0.6332488 0.7540137
#> a_23    0    0    0 17.791176 17.839820 17.161617 18.429000 0.7066186 0.8679323
#> a_24    0    0    0  6.683629  7.292560 26.305123 29.225673 0.8819783 0.9140934
#> a_25    0    0    0 13.650855 15.299786 25.723088 26.079152 0.2351954 0.2730400
#> a_26    0    0    0  7.811844  9.669422 13.072637 14.430451 0.8213459 0.8725134
#> a_27    0    0    0 11.409021 12.898317  5.830064  6.594458 0.7239677 0.7678395
#> a_28    0    0    0 10.556697 10.873345 19.143520 22.237036 0.8552030 0.8651854
#> a_29    0    0    0 19.127439 19.966827 23.049431 27.873607 0.8641409 0.9513516
#> a_30    1    0    0  8.755102  9.237438 10.773667 11.019231 0.8676661 0.9060229
#> a_31    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000 0.0000000
#> a_32    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000 0.0000000
#>          r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  17.703216 18.114123 14.842545 17.298473
#> a_2  14.100242 16.224924 14.871066 16.179055
#> a_3  14.246534 17.296766 14.370011 15.333486
#> a_4  10.754806 10.839342  9.272403  9.504462
#> a_5  17.549028 18.604800  8.187283  8.895507
#> a_6  16.240685 18.487362 13.470484 15.617842
#> a_7  16.174895 19.131759 14.938567 17.172125
#> a_8  18.652535 19.784851 18.632289 19.137673
#> a_9   8.905653  9.502539  4.580488  5.287344
#> a_10  1.981629  2.412266  7.432791  8.771176
#> a_11  6.413529  7.587459  5.386400  5.891791
#> a_12 18.277771 18.305161 10.428924 12.680847
#> a_13 10.769736 12.020686 13.251156 14.652635
#> a_14  4.772696  5.649190 13.026433 14.435262
#> a_15 16.830455 18.906727 15.194064 15.694112
#> a_16 13.212976 14.707310  8.914558 10.446126
#> a_17 13.269626 13.862361 16.527386 19.750050
#> a_18  6.049818  6.255375 13.928998 15.904600
#> a_19  8.255751  9.131961 13.425966 15.751406
#> a_20 13.220242 16.288403 11.789797 11.813725
#> a_21  9.476928 11.719104  3.172687  3.689997
#> a_22 13.474330 14.604138 11.629125 12.342363
#> a_23  5.480703  6.214546 17.278826 17.441952
#> a_24  7.785466  7.859236 17.688555 18.654985
#> a_25 12.300738 13.435881  8.293347 10.058709
#> a_26 11.627320 14.448386 16.757830 19.011123
#> a_27  9.147552 11.298398 13.876418 15.567132
#> a_28 15.588071 17.015939 13.294800 14.358246
#> a_29 12.289971 15.175152  8.962725 10.958463
#> a_30  7.449603  9.132087 15.197245 17.280546
#> a_31  0.000000  0.000000  0.000000  0.000000
#> a_32  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=5.0

# Simulate the shock effect of a project scenario using MFPP package
# with default values.
 Res<-phase2(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8       a_9 a_10 a_11 a_12 a_13 a_14 a_15
#> a_1    1   1   1   0   1   0   0   0 0.0000000    0    0    0    0    0    0
#> a_2    0   1   1   0   1   0   0   0 0.0000000    0    0    0    0    0    0
#> a_3    0   0   1   1   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_5    0   0   0   0   1   1   0   0 0.0000000    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   1   0 0.0000000    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0 0.0000000    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1 0.0000000    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0 0.7618916    0    0    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0 0.0000000    1    0    0    0    0    0
#> a_11   0   0   0   0   0   0   0   0 0.0000000    0    1    0    0    0    1
#> a_12   0   0   0   0   0   0   0   0 0.0000000    0    0    1    0    1    0
#> a_13   0   0   0   0   0   0   0   0 0.0000000    0    0    0    1    0    0
#> a_14   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    1    1
#> a_15   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    1
#> a_16   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_17   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_24   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_25   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_26   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_27   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_28   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_29   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_30   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_31   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_32   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#>      a_16 a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25      a_26 a_27 a_28 a_29
#> a_1     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_2     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_3     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_4     0    0    0    0    1    0    0    0    0    0 0.0000000    0    0    0
#> a_5     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_6     0    0    0    0    0    0    1    0    0    0 0.0000000    0    0    0
#> a_7     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_8     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_9     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_10    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_11    0    0    0    0    0    0    0    0    0    0 1.0000000    0    0    0
#> a_12    0    0    0    0    0    0    0    0    0    0 1.0000000    0    0    0
#> a_13    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_14    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_15    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_16    1    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_17    0    1    0    1    0    0    0    0    0    0 0.0000000    0    0    0
#> a_18    0    0    1    1    0    1    0    0    0    0 0.0000000    0    0    0
#> a_19    0    0    0    1    1    1    0    0    0    1 0.0000000    0    0    0
#> a_20    0    0    0    0    1    1    0    0    0    0 0.0000000    0    0    0
#> a_21    0    0    0    0    0    1    0    0    0    0 0.0000000    0    0    0
#> a_22    0    0    0    0    0    0    1    0    1    0 0.0000000    0    0    0
#> a_23    0    0    0    0    0    0    0    1    0    0 0.0000000    0    0    0
#> a_24    0    0    0    0    0    0    0    0    1    0 0.0000000    1    0    0
#> a_25    0    0    0    0    0    0    0    0    0    1 0.0000000    0    0    0
#> a_26    0    0    0    0    0    0    0    0    0    0 0.5841634    0    0    0
#> a_27    0    0    0    0    0    0    0    0    0    0 0.0000000    1    0    0
#> a_28    0    0    0    0    0    0    0    0    0    0 0.0000000    0    1    1
#> a_29    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    1
#> a_30    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_31    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_32    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#>      a_30 a_31 a_32        t_1        t_2       c_1        c_2       q_1
#> a_1     0    0    0  16.347869 109.789371 17.155904  19.828587 0.4089204
#> a_2     0    0    0  13.863452  16.243329  8.402097   9.547030 0.9231952
#> a_3     0    0    0   3.480427   3.686522 12.727002  12.769227 0.7129179
#> a_4     0    0    0  13.721209  15.711412 23.062730  25.915935 0.6594185
#> a_5     0    0    0  14.131963  16.727585 21.834141 145.474209 0.1032106
#> a_6     0    0    0  16.413953 100.081802 15.481238  15.990274 0.7411996
#> a_7     0    0    0  13.186623  14.899179 19.225869  21.036511 0.5875751
#> a_8     0    0    0  12.416754  15.159107 24.177449  25.373347 0.8224378
#> a_9     0    0    0  14.317363  14.806764 15.693322  18.164857 0.4776105
#> a_10    0    0    0   6.933816   8.663723 12.290713  13.181177 0.4588683
#> a_11    0    0    0   3.667132   3.817730 25.922503  26.372165 0.3264099
#> a_12    0    0    0   3.850505   4.710925 23.853609 145.588810 0.7299107
#> a_13    0    0    0   8.815423  10.145003 19.267661 129.776822 0.6430413
#> a_14    0    0    0  14.692313  16.952350 12.308258  15.143500 0.8160978
#> a_15    0    0    0  13.137955  13.617844 28.242833  29.641111 0.3739066
#> a_16    1    0    0  13.394942  14.963177 28.870363  29.686101 0.7543239
#> a_17    0    0    0  43.598998   7.472334 15.469767  15.998222 0.8503954
#> a_18    0    0    0  14.908959  18.097767 17.937612  22.379042 0.7717529
#> a_19    0    0    0   8.328546   9.078106 19.047426  20.870878 0.5717522
#> a_20    0    0    0 105.472349  18.094342 13.186107  13.310176 0.7729071
#> a_21    0    0    0  11.302310  12.027360  5.966204   7.430421 0.3951842
#> a_22    0    0    0  11.892810  81.864573  2.794244   2.953205 0.6332488
#> a_23    0    0    0  17.791176  17.839820 17.161617  18.429000 0.7066186
#> a_24    0    0    0   6.683629   7.292560 26.305123  29.225673 0.8819783
#> a_25    0    0    0  13.650855  15.299786 25.723088  26.079152 0.2351954
#> a_26    0    0    0   7.811844   9.669422 13.072637  14.430451 0.8213459
#> a_27    0    0    0  11.409021  12.898317  5.830064   6.594458 0.7239677
#> a_28    0    0    0  63.340184  10.873345 19.143520  22.237036 0.8552030
#> a_29    0    0    0  19.127439  19.966827 23.049431  27.873607 0.8641409
#> a_30    1    0    0   8.755102   9.237438 10.773667  66.115384 0.8676661
#> a_31    0    0    0   0.000000   0.000000  0.000000   0.000000 0.0000000
#> a_32    0    0    0   0.000000   0.000000  0.000000   0.000000 0.0000000
#>            q_2     r_1.1      r_2.1     r_1.2      r_2.2
#> a_1  0.4213035 17.703216  18.114123 14.842545  17.298473
#> a_2  0.9350867 14.100242  97.349542 14.871066  16.179055
#> a_3  0.8573705 85.479202 103.780594 14.370011  15.333486
#> a_4  0.6744472 10.754806  10.839342  9.272403   9.504462
#> a_5  0.1128993 17.549028  18.604800  8.187283   8.895507
#> a_6  1.0000000 16.240685  18.487362 13.470484  15.617842
#> a_7  0.6453637 16.174895  19.131759 14.938567  17.172125
#> a_8  1.0000000 18.652535  19.784851 18.632289  19.137673
#> a_9  0.5125546  8.905653   9.502539  4.580488   5.287344
#> a_10 0.5049861  1.981629   2.412266  7.432791   8.771176
#> a_11 0.3880509  6.413529   7.587459  5.386400   5.891791
#> a_12 0.8193567 18.277771  18.305161 10.428924  12.680847
#> a_13 0.7878548 10.769736  12.020686 13.251156  14.652635
#> a_14 0.8891377  4.772696   5.649190 78.158596  86.611569
#> a_15 0.4386208 16.830455  18.906727 91.164383  15.694112
#> a_16 0.8409803 13.212976  14.707310  8.914558  10.446126
#> a_17 0.9827242 79.617759  13.862361 16.527386  19.750050
#> a_18 0.8882783  6.049818   6.255375 13.928998  15.904600
#> a_19 0.6563525  8.255751   9.131961 13.425966  15.751406
#> a_20 0.7735467 13.220242  16.288403 11.789797  70.882353
#> a_21 0.4231735  9.476928  11.719104  3.172687   3.689997
#> a_22 0.7540137 13.474330  14.604138 11.629125  12.342363
#> a_23 0.8679323  5.480703   6.214546 17.278826  17.441952
#> a_24 0.9140934  7.785466   7.859236 17.688555 111.929911
#> a_25 0.2730400 12.300738  13.435881  8.293347  10.058709
#> a_26 0.8725134 69.763923  14.448386 16.757830  19.011123
#> a_27 0.7678395  9.147552  11.298398 13.876418  15.567132
#> a_28 0.8651854 15.588071  17.015939 13.294800  86.149476
#> a_29 0.9513516 12.289971  15.175152  8.962725  10.958463
#> a_30 0.9060229  7.449603  54.792521 15.197245  17.280546
#> a_31 0.0000000  0.000000   0.000000  0.000000   0.000000
#> a_32 0.0000000  0.000000   0.000000  0.000000   0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  74.39483
#> Const constraint (Cc):  495.2102
#> Score/scope constraint (Cs):  0.930284
#> Quailty constraint (Cq):  0.5417505
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  235.1082
#> Const constraint (Cc):  977.4164
#> Score/scope constraint (Cs):  0.9750196
#> Quailty constraint (Cq):  0.6497017
#> Resource constraint(s) (CR):
#>          R_1     R_2
#> TPR 340.2939 364.479

# Simulate the effect of a project scenario using MFPP package
# with specified values.
 Res<-phase2(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8       a_9 a_10 a_11 a_12 a_13 a_14 a_15
#> a_1    1   1   1   0   1   0   0   0 0.0000000    0    0    0    0    0    0
#> a_2    0   1   1   0   1   0   0   0 0.0000000    0    0    0    0    0    0
#> a_3    0   0   1   1   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_5    0   0   0   0   1   1   0   0 0.0000000    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   1   0 0.0000000    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0 0.0000000    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1 0.0000000    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0 0.7618916    0    0    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0 0.0000000    1    0    0    0    0    0
#> a_11   0   0   0   0   0   0   0   0 0.0000000    0    1    0    0    0    1
#> a_12   0   0   0   0   0   0   0   0 0.0000000    0    0    1    0    1    0
#> a_13   0   0   0   0   0   0   0   0 0.0000000    0    0    0    1    0    0
#> a_14   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    1    1
#> a_15   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    1
#> a_16   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_17   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_24   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_25   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_26   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_27   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_28   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_29   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_30   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_31   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#> a_32   0   0   0   0   0   0   0   0 0.0000000    0    0    0    0    0    0
#>      a_16 a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25      a_26 a_27 a_28 a_29
#> a_1     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_2     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_3     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_4     0    0    0    0    1    0    0    0    0    0 0.0000000    0    0    0
#> a_5     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_6     0    0    0    0    0    0    1    0    0    0 0.0000000    0    0    0
#> a_7     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_8     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_9     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_10    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_11    0    0    0    0    0    0    0    0    0    0 1.0000000    0    0    0
#> a_12    0    0    0    0    0    0    0    0    0    0 1.0000000    0    0    0
#> a_13    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_14    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_15    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_16    1    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_17    0    1    0    1    0    0    0    0    0    0 0.0000000    0    0    0
#> a_18    0    0    1    1    0    1    0    0    0    0 0.0000000    0    0    0
#> a_19    0    0    0    1    1    1    0    0    0    1 0.0000000    0    0    0
#> a_20    0    0    0    0    1    1    0    0    0    0 0.0000000    0    0    0
#> a_21    0    0    0    0    0    1    0    0    0    0 0.0000000    0    0    0
#> a_22    0    0    0    0    0    0    1    0    1    0 0.0000000    0    0    0
#> a_23    0    0    0    0    0    0    0    1    0    0 0.0000000    0    0    0
#> a_24    0    0    0    0    0    0    0    0    1    0 0.0000000    1    0    0
#> a_25    0    0    0    0    0    0    0    0    0    1 0.0000000    0    0    0
#> a_26    0    0    0    0    0    0    0    0    0    0 0.5841634    0    0    0
#> a_27    0    0    0    0    0    0    0    0    0    0 0.0000000    1    0    0
#> a_28    0    0    0    0    0    0    0    0    0    0 0.0000000    0    1    1
#> a_29    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    1
#> a_30    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_31    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_32    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#>      a_30 a_31 a_32       t_1        t_2        c_1        c_2       q_1
#> a_1     0    0    0 16.347869  18.298229  17.155904  19.828587 0.4089204
#> a_2     0    0    0 13.863452  16.243329   8.402097   9.547030 0.9231952
#> a_3     0    0    0  3.480427   3.686522  12.727002  12.769227 0.7129179
#> a_4     0    0    0 13.721209  15.711412  23.062730 155.495611 0.6594185
#> a_5     0    0    0 14.131963  16.727585  21.834141  24.245701 0.1032106
#> a_6     0    0    0 16.413953  16.680300  15.481238  15.990274 0.7411996
#> a_7     0    0    0 13.186623  14.899179  19.225869  21.036511 0.5875751
#> a_8     0    0    0 12.416754  90.954643  24.177449  25.373347 0.8224378
#> a_9     0    0    0 14.317363  14.806764  15.693322  18.164857 0.4776105
#> a_10    0    0    0 41.602894   8.663723  12.290713  13.181177 0.4588683
#> a_11    0    0    0  3.667132   3.817730  25.922503  26.372165 0.3264099
#> a_12    0    0    0  3.850505   4.710925  23.853609  24.264802 0.7299107
#> a_13    0    0    0  8.815423  10.145003  19.267661  21.629470 0.6430413
#> a_14    0    0    0 14.692313  16.952350  12.308258  90.861001 0.8160978
#> a_15    0    0    0 13.137955  13.617844  28.242833  29.641111 0.3739066
#> a_16    1    0    0 13.394942  14.963177  28.870363  29.686101 0.7543239
#> a_17    0    0    0  7.266500   7.472334  15.469767  15.998222 0.8503954
#> a_18    0    0    0 89.453752 108.586601  17.937612  22.379042 0.7717529
#> a_19    0    0    0  8.328546   9.078106  19.047426  20.870878 0.5717522
#> a_20    0    0    0 17.578725  18.094342  13.186107  13.310176 0.7729071
#> a_21    0    0    0 11.302310  12.027360   5.966204   7.430421 0.3951842
#> a_22    0    0    0 11.892810  13.644096   2.794244   2.953205 0.6332488
#> a_23    0    0    0 17.791176  17.839820  17.161617  18.429000 0.7066186
#> a_24    0    0    0  6.683629   7.292560  26.305123  29.225673 0.8819783
#> a_25    0    0    0 13.650855  15.299786  25.723088  26.079152 0.2351954
#> a_26    0    0    0  7.811844   9.669422  78.435825  14.430451 0.8213459
#> a_27    0    0    0 11.409021  12.898317   5.830064   6.594458 0.7239677
#> a_28    0    0    0 10.556697  10.873345 114.861118  22.237036 0.8552030
#> a_29    0    0    0 19.127439  19.966827 138.296584  27.873607 0.8641409
#> a_30    1    0    0  8.755102   9.237438  10.773667  11.019231 0.8676661
#> a_31    0    0    0  0.000000   0.000000   0.000000   0.000000 0.0000000
#> a_32    0    0    0  0.000000   0.000000   0.000000   0.000000 0.0000000
#>            q_2      r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.4213035 106.219294 18.114123 14.842545 17.298473
#> a_2  0.9350867  14.100242 16.224924 14.871066 16.179055
#> a_3  0.8573705  14.246534 17.296766 14.370011 15.333486
#> a_4  1.0000000  10.754806 10.839342 55.634418  9.504462
#> a_5  0.1128993  17.549028 18.604800  8.187283 53.373043
#> a_6  0.9165027  16.240685 18.487362 13.470484 93.707049
#> a_7  0.6453637  16.174895 19.131759 14.938567 17.172125
#> a_8  0.9792452  18.652535 19.784851 18.632289 19.137673
#> a_9  1.0000000   8.905653  9.502539  4.580488  5.287344
#> a_10 0.5049861   1.981629  2.412266  7.432791  8.771176
#> a_11 1.0000000   6.413529  7.587459  5.386400  5.891791
#> a_12 0.8193567  18.277771 18.305161 10.428924 76.085080
#> a_13 0.7878548  64.618416 12.020686 13.251156 14.652635
#> a_14 0.8891377   4.772696  5.649190 13.026433 14.435262
#> a_15 0.4386208 100.982733 18.906727 15.194064 15.694112
#> a_16 0.8409803  13.212976 14.707310  8.914558 10.446126
#> a_17 0.9827242  13.269626 13.862361 16.527386 19.750050
#> a_18 0.8882783   6.049818  6.255375 13.928998 15.904600
#> a_19 1.0000000  49.534507  9.131961 13.425966 15.751406
#> a_20 1.0000000  13.220242 16.288403 11.789797 11.813725
#> a_21 0.4231735   9.476928 11.719104  3.172687  3.689997
#> a_22 0.7540137  80.845977 14.604138 11.629125 12.342363
#> a_23 0.8679323   5.480703  6.214546 17.278826 17.441952
#> a_24 0.9140934   7.785466  7.859236 17.688555 18.654985
#> a_25 0.2730400  12.300738 80.615288  8.293347 10.058709
#> a_26 0.8725134  11.627320 14.448386 16.757830 19.011123
#> a_27 0.7678395   9.147552 11.298398 13.876418 15.567132
#> a_28 1.0000000  15.588071 17.015939 13.294800 14.358246
#> a_29 0.9513516  12.289971 91.050912  8.962725 10.958463
#> a_30 0.9060229   7.449603  9.132087 15.197245 17.280546
#> a_31 0.0000000   0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000   0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  126.6633
#> Const constraint (Cc):  503.1279
#> Score/scope constraint (Cs):  0.930284
#> Quailty constraint (Cq):  0.5417505
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  147.7864
#> Const constraint (Cc):  1043.97
#> Score/scope constraint (Cs):  0.9750196
#> Quailty constraint (Cq):  0.7090246
#> Resource constraint(s) (CR):
#>          R_1     R_2
#> TPR 440.9113 241.178