Function to simulate the effects of the change of customer claims.

Simulating the effects of the changes of customer claims.

phase3 (x,p=0.10, s=0.50)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability of structural change (default value is p=0.10).

s

Increase/decrease ratio of priorities (default value is s=0.50).

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6       a_7 a_8       a_9      a_10 a_11 a_12 a_13
#> a_1    1   1   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_2    0   1   1   0   0   0 0.0000000   0 0.7068193 0.0000000    0    0    0
#> a_3    0   0   1   1   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_4    0   0   0   1   1   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_5    0   0   0   0   1   0 0.8376003   0 0.0000000 0.0000000    0    1    0
#> a_6    0   0   0   0   0   1 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_7    0   0   0   0   0   0 1.0000000   0 0.0000000 0.0000000    0    0    0
#> a_8    0   0   0   0   0   0 0.0000000   1 0.0000000 0.0000000    0    0    0
#> a_9    0   0   0   0   0   0 0.0000000   0 1.0000000 1.0000000    1    0    0
#> a_10   0   0   0   0   0   0 0.0000000   0 0.0000000 0.8020264    1    1    0
#> a_11   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    1    0    0
#> a_12   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    1    0
#> a_13   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    1
#> a_14   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_15   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_16   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_17   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_18   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_19   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_20   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_21   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_22   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_23   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_24   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_25   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_26   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_27   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_28   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_29   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_30   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_31   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#> a_32   0   0   0   0   0   0 0.0000000   0 0.0000000 0.0000000    0    0    0
#>      a_14 a_15 a_16 a_17 a_18 a_19 a_20 a_21 a_22 a_23      a_24 a_25 a_26 a_27
#> a_1     0    0    0    0    0    0    1    0    0    0 0.0000000    0    0    0
#> a_2     0    1    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_3     1    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_4     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_5     0    0    0    0    0    0    1    0    0    0 0.0000000    0    0    0
#> a_6     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_7     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_8     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_9     0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_10    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_11    0    0    1    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_12    0    0    0    0    0    0    0    0    0    0 0.0000000    0    1    0
#> a_13    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_14    1    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_15    0    1    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_16    0    0    1    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_17    0    0    0    1    0    0    0    0    1    0 0.0000000    0    0    0
#> a_18    0    0    0    0    1    1    0    0    0    0 0.0000000    0    0    0
#> a_19    0    0    0    0    0    1    0    0    0    0 0.0000000    0    0    0
#> a_20    0    0    0    0    0    0    1    1    0    0 0.5783969    0    0    1
#> a_21    0    0    0    0    0    0    0    1    0    1 0.0000000    0    1    0
#> a_22    0    0    0    0    0    0    0    0    1    0 0.0000000    0    0    0
#> a_23    0    0    0    0    0    0    0    0    0    1 0.0000000    0    0    0
#> a_24    0    0    0    0    0    0    0    0    0    0 1.0000000    0    0    0
#> a_25    0    0    0    0    0    0    0    0    0    0 0.0000000    1    0    0
#> a_26    0    0    0    0    0    0    0    0    0    0 0.0000000    0    1    1
#> a_27    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    1
#> a_28    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_29    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_30    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_31    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#> a_32    0    0    0    0    0    0    0    0    0    0 0.0000000    0    0    0
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0  2.301139  2.493263 11.576630 12.624806 0.4691942
#> a_2     0    0    0    0    0 18.099485 18.611744 10.832293 13.338379 0.7781432
#> a_3     0    0    0    0    0 17.584928 18.778192 11.049531 13.751565 0.3444858
#> a_4     0    0    0    0    0 18.187586 18.826325  3.495776  4.221873 0.5579386
#> a_5     0    0    0    0    0  9.285713 11.313341 16.578017 19.410156 0.6933574
#> a_6     0    0    0    0    0 16.529493 19.412737 19.170531 23.240695 0.5130048
#> a_7     0    0    0    0    0  5.328804  5.959527 14.805266 16.715168 0.6137883
#> a_8     0    0    0    0    0 12.697573 13.520908 16.279895 18.408858 0.6470740
#> a_9     0    0    0    0    0 12.960878 14.573120 22.679789 26.668362 0.8465924
#> a_10    0    0    0    0    0 13.048931 15.106848 14.424886 16.509053 0.5883755
#> a_11    0    0    0    0    0 17.032864 19.977627 10.721291 12.281008 0.5989135
#> a_12    0    0    0    0    0 14.606305 17.544339 14.744169 17.344597 0.8561401
#> a_13    0    0    0    0    0 15.067933 17.216002  2.967044  3.331605 0.7049626
#> a_14    0    0    0    0    0 16.087120 18.279094  9.680978 10.900662 0.3744115
#> a_15    0    0    0    0    0 17.239941 18.730956 17.561490 21.830331 0.3914429
#> a_16    0    0    0    0    0 15.299348 16.365430 15.852940 17.240645 0.7900010
#> a_17    0    0    0    0    0  3.221207  3.474191  7.889933  9.059529 0.4991775
#> a_18    0    0    0    0    0 10.725129 11.712008 24.950582 29.918873 0.4613426
#> a_19    0    0    0    0    0 15.560227 17.780721 22.430974 23.200758 0.6057429
#> a_20    0    0    0    0    0  5.953181  6.302124 23.271244 23.696816 0.8518833
#> a_21    0    0    0    0    0 15.913838 16.659591 26.510453 26.547886 0.4315185
#> a_22    0    0    0    0    0  1.853236  2.122454 12.581153 12.780319 0.3338708
#> a_23    0    0    0    0    0 15.752421 18.397711 22.456749 23.036681 0.3853201
#> a_24    0    0    0    0    0 14.379279 17.823137 15.424374 17.804567 0.6959424
#> a_25    0    0    0    0    0 17.150406 19.308790  7.010209  8.403820 0.4222313
#> a_26    0    0    0    0    0 16.663504 17.838040  7.554132  8.803231 0.6447334
#> a_27    0    0    0    0    0 11.688123 12.025986 16.217439 17.395721 0.6072583
#> a_28    1    0    0    0    0 18.177957 18.561772  9.696388 11.007670 0.6035469
#> a_29    0    1    0    0    0 16.358864 16.668290 11.534555 11.811030 0.6512519
#> a_30    0    0    1    0    0 15.457658 18.251179 19.314912 22.986456 0.6556246
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.5174588 13.234122 14.581530 10.566248 13.183811
#> a_2  0.8638524 13.699590 15.422656  7.200756  8.224126
#> a_3  0.4165109  4.911462  5.054926  4.556959  5.190598
#> a_4  0.6226630  9.256895 10.826090 13.972664 14.139643
#> a_5  0.7517785 13.403837 13.668742  2.296732  2.831684
#> a_6  0.5580523 13.149876 15.053310 14.464764 14.607059
#> a_7  0.6842679  4.699537  5.169577 16.890346 18.240664
#> a_8  0.7949430 14.961348 17.242990 15.840271 19.766390
#> a_9  0.8835081 10.853462 12.454891 13.905069 14.459292
#> a_10 0.6151275 14.417906 17.370369 10.098940 10.166560
#> a_11 0.6103385 10.868228 13.155397  4.255330  5.153211
#> a_12 0.8934922 15.020271 18.164702 11.631560 14.167317
#> a_13 0.8342912  6.126484  6.910487 16.978018 16.992262
#> a_14 0.4101690 12.226582 13.789354 12.374986 15.070204
#> a_15 0.4387946 17.982446 17.983544 13.169910 15.142868
#> a_16 0.8870228 17.100462 17.207243 17.401226 18.389631
#> a_17 0.6032608 15.173240 16.539409 14.285318 15.357935
#> a_18 0.4787154  9.317324  9.989277 14.904365 17.550989
#> a_19 0.7192511 15.503287 18.397984 16.897304 17.866218
#> a_20 0.8677862  9.212595 10.065498 16.061832 17.978586
#> a_21 0.4392470  8.363159  8.729066 11.709009 11.759927
#> a_22 0.3689668 13.241447 16.397675  9.592835 10.418009
#> a_23 0.4053924 16.938635 17.557161  5.524634  6.209450
#> a_24 0.7244180 10.631585 12.406955  9.328349 10.946963
#> a_25 0.4806034 13.807524 16.348513 17.823028 19.739994
#> a_26 0.6941722  7.333930  7.940227 12.846634 15.201561
#> a_27 0.7181080 17.232280 18.931795 12.146725 13.441901
#> a_28 0.6520422 14.060112 17.362973  9.119785  9.800199
#> a_29 0.7570994 13.146676 13.562057 16.051606 16.547068
#> a_30 0.7466798 12.828299 14.381995 10.339689 11.048119
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection and dependency selecton of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=0.10

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with default values.

 Res<-phase3(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6       a_7        a_8 a_9 a_10       a_11 a_12 a_13
#> a_1    1   1   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_2    0   1   1   0   0   0 0.0000000 0.00000000   1    0 0.00000000    0    0
#> a_3    0   0   1   1   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_4    0   0   0   1   1   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_5    0   0   0   0   1   0 0.8376003 0.00000000   0    0 0.08446809    1    0
#> a_6    0   0   0   0   0   1 0.2675313 0.00000000   0    0 0.00000000    0    0
#> a_7    0   0   0   0   0   0 1.0000000 0.03055609   0    0 0.00000000    0    0
#> a_8    0   0   0   0   0   0 0.0000000 1.00000000   0    0 0.00000000    0    0
#> a_9    0   0   0   0   0   0 0.0000000 0.00000000   1    1 1.00000000    0    0
#> a_10   0   0   0   0   0   0 0.0000000 0.00000000   0    1 1.00000000    1    0
#> a_11   0   0   0   0   0   0 0.0000000 0.00000000   0    0 1.00000000    0    0
#> a_12   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    1    0
#> a_13   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    1
#> a_14   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_15   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_16   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_17   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_18   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_19   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_20   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_21   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_22   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_23   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_24   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_25   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_26   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_27   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_28   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_29   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_30   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_31   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#> a_32   0   0   0   0   0   0 0.0000000 0.00000000   0    0 0.00000000    0    0
#>           a_14      a_15       a_16 a_17 a_18       a_19 a_20      a_21
#> a_1  0.0000000 0.0000000 0.00000000    0    0 0.37105229    1 0.0000000
#> a_2  0.0000000 1.0000000 0.04677821    0    0 0.06853062    0 0.0000000
#> a_3  1.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_4  0.0000000 0.0000000 0.49742752    0    0 0.00000000    0 0.0000000
#> a_5  0.0000000 0.0000000 0.00000000    0    0 0.22087986    1 0.0000000
#> a_6  0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_7  0.0000000 0.0000000 0.00000000    0    0 0.34338228    0 0.0000000
#> a_8  0.0000000 0.1050232 0.00000000    0    0 0.00000000    0 0.0000000
#> a_9  0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_10 0.1255233 0.0000000 0.00000000    0    0 0.01595260    0 0.0000000
#> a_11 0.0000000 0.0000000 1.00000000    0    0 0.49237024    0 0.0000000
#> a_12 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_13 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_14 1.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_15 0.0000000 1.0000000 0.00000000    0    0 0.00000000    0 0.2596299
#> a_16 0.0000000 0.0000000 1.00000000    0    0 0.00000000    0 0.0000000
#> a_17 0.0000000 0.0000000 0.00000000    1    0 0.00000000    0 0.0000000
#> a_18 0.0000000 0.0000000 0.00000000    0    1 1.00000000    0 0.0000000
#> a_19 0.0000000 0.0000000 0.00000000    0    0 1.00000000    0 0.0000000
#> a_20 0.0000000 0.0000000 0.00000000    0    0 0.00000000    1 1.0000000
#> a_21 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 1.0000000
#> a_22 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_23 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_24 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_25 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_26 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_27 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_28 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_29 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_30 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_31 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#> a_32 0.0000000 0.0000000 0.00000000    0    0 0.00000000    0 0.0000000
#>          a_22      a_23      a_24       a_25       a_26      a_27 a_28
#> a_1  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_2  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_3  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_4  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_5  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_6  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.2172549    0
#> a_7  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_8  0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_9  0.000000 0.0000000 0.0000000 0.00000000 0.47118530 0.0000000    0
#> a_10 0.000000 0.0000000 0.0000000 0.45200966 0.00000000 0.0000000    0
#> a_11 0.000000 0.1110902 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_12 0.000000 0.2661622 0.1000327 0.37067790 1.00000000 0.0000000    0
#> a_13 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_14 0.000000 0.0000000 0.0000000 0.00000000 0.05572548 0.0000000    0
#> a_15 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_16 0.000000 0.2271914 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_17 1.000000 0.1127512 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_18 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0780001    0
#> a_19 0.479424 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_20 0.000000 0.0000000 0.5783969 0.04310953 0.00000000 1.0000000    0
#> a_21 0.000000 1.0000000 0.0000000 0.00000000 1.00000000 0.0000000    0
#> a_22 1.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.1415789    0
#> a_23 0.000000 1.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_24 0.000000 0.0000000 1.0000000 0.00000000 0.00000000 0.0000000    0
#> a_25 0.000000 0.0000000 0.0000000 1.00000000 0.00000000 0.0000000    0
#> a_26 0.000000 0.0000000 0.0000000 0.00000000 1.00000000 1.0000000    0
#> a_27 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 1.0000000    0
#> a_28 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    1
#> a_29 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_30 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_31 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#> a_32 0.000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000    0
#>           a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1  0.0000000    0    0    0  2.301139  2.493263 11.576630 12.624806 0.4691942
#> a_2  0.0000000    0    0    0 18.099485 18.611744 10.832293 13.338379 0.7781432
#> a_3  0.0000000    0    0    0 17.584928 18.778192 11.049531 13.751565 0.3444858
#> a_4  0.0000000    0    0    0 18.187586 18.826325  3.495776  4.221873 0.5579386
#> a_5  0.0000000    0    0    0  9.285713 11.313341 16.578017 19.410156 0.6933574
#> a_6  0.0000000    0    0    0 16.529493 19.412737 19.170531 23.240695 0.5130048
#> a_7  0.0000000    0    0    0  5.328804  5.959527 14.805266 16.715168 0.6137883
#> a_8  0.0000000    0    0    0 12.697573 13.520908 16.279895 18.408858 0.6470740
#> a_9  0.0000000    0    0    0 12.960878 14.573120 22.679789 26.668362 0.8465924
#> a_10 0.0000000    0    0    0 13.048931 15.106848 14.424886 16.509053 0.5883755
#> a_11 0.0000000    0    0    0 17.032864 19.977627 10.721291 12.281008 0.5989135
#> a_12 0.0000000    0    0    0 14.606305 17.544339 14.744169 17.344597 0.8561401
#> a_13 0.0000000    0    0    0 15.067933 17.216002  2.967044  3.331605 0.7049626
#> a_14 0.0000000    0    0    0 16.087120 18.279094  9.680978 10.900662 0.3744115
#> a_15 0.0308229    0    0    0 17.239941 18.730956 17.561490 21.830331 0.3914429
#> a_16 0.0000000    0    0    0 15.299348 16.365430 15.852940 17.240645 0.7900010
#> a_17 0.0000000    0    0    0  3.221207  3.474191  7.889933  9.059529 0.4991775
#> a_18 0.0000000    0    0    0 10.725129 11.712008 24.950582 29.918873 0.4613426
#> a_19 0.0000000    0    0    0 15.560227 17.780721 22.430974 23.200758 0.6057429
#> a_20 0.0000000    0    0    0  5.953181  6.302124 23.271244 23.696816 0.8518833
#> a_21 0.0000000    0    0    0 15.913838 16.659591 26.510453 26.547886 0.4315185
#> a_22 0.0000000    0    0    0  1.853236  2.122454 12.581153 12.780319 0.3338708
#> a_23 0.0000000    0    0    0 15.752421 18.397711 22.456749 23.036681 0.3853201
#> a_24 0.0000000    0    0    0 14.379279 17.823137 15.424374 17.804567 0.6959424
#> a_25 0.0000000    0    0    0 17.150406 19.308790  7.010209  8.403820 0.4222313
#> a_26 0.0000000    0    0    0 16.663504 17.838040  7.554132  8.803231 0.6447334
#> a_27 0.0000000    0    0    0 11.688123 12.025986 16.217439 17.395721 0.6072583
#> a_28 0.3752859    0    0    0 18.177957 18.561772  9.696388 11.007670 0.6035469
#> a_29 1.0000000    0    0    0 16.358864 16.668290 11.534555 11.811030 0.6512519
#> a_30 0.0000000    1    0    0 15.457658 18.251179 19.314912 22.986456 0.6556246
#> a_31 0.0000000    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32 0.0000000    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.5174588 13.234122 14.581530 10.566248 13.183811
#> a_2  0.8638524 13.699590 15.422656  7.200756  8.224126
#> a_3  0.4165109  4.911462  5.054926  4.556959  5.190598
#> a_4  0.6226630  9.256895 10.826090 13.972664 14.139643
#> a_5  0.7517785 13.403837 13.668742  2.296732  2.831684
#> a_6  0.5580523 13.149876 15.053310 14.464764 14.607059
#> a_7  0.6842679  4.699537  5.169577 16.890346 18.240664
#> a_8  0.7949430 14.961348 17.242990 15.840271 19.766390
#> a_9  0.8835081 10.853462 12.454891 13.905069 14.459292
#> a_10 0.6151275 14.417906 17.370369 10.098940 10.166560
#> a_11 0.6103385 10.868228 13.155397  4.255330  5.153211
#> a_12 0.8934922 15.020271 18.164702 11.631560 14.167317
#> a_13 0.8342912  6.126484  6.910487 16.978018 16.992262
#> a_14 0.4101690 12.226582 13.789354 12.374986 15.070204
#> a_15 0.4387946 17.982446 17.983544 13.169910 15.142868
#> a_16 0.8870228 17.100462 17.207243 17.401226 18.389631
#> a_17 0.6032608 15.173240 16.539409 14.285318 15.357935
#> a_18 0.4787154  9.317324  9.989277 14.904365 17.550989
#> a_19 0.7192511 15.503287 18.397984 16.897304 17.866218
#> a_20 0.8677862  9.212595 10.065498 16.061832 17.978586
#> a_21 0.4392470  8.363159  8.729066 11.709009 11.759927
#> a_22 0.3689668 13.241447 16.397675  9.592835 10.418009
#> a_23 0.4053924 16.938635 17.557161  5.524634  6.209450
#> a_24 0.7244180 10.631585 12.406955  9.328349 10.946963
#> a_25 0.4806034 13.807524 16.348513 17.823028 19.739994
#> a_26 0.6941722  7.333930  7.940227 12.846634 15.201561
#> a_27 0.7181080 17.232280 18.931795 12.146725 13.441901
#> a_28 0.6520422 14.060112 17.362973  9.119785  9.800199
#> a_29 0.7570994 13.146676 13.562057 16.051606 16.547068
#> a_30 0.7466798 12.828299 14.381995 10.339689 11.048119
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  113.3764
#> Const constraint (Cc):  439.2636
#> Score/scope constraint (Cs):  1
#> Quailty constraint (Cq):  0.5669616
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  152.2646
#> Const constraint (Cc):  494.2711
#> Score/scope constraint (Cs):  1
#> Quailty constraint (Cq):  0.6260504
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 238.8453 239.1669

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase3(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4         a_5        a_6       a_7 a_8        a_9       a_10
#> a_1    1   1   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_2    0   1   1   0 0.000000000 0.00000000 0.0000000   0 0.70681929 0.00000000
#> a_3    0   0   1   1 0.004435594 0.06999317 0.0000000   0 0.06617519 0.06405649
#> a_4    0   0   0   1 1.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_5    0   0   0   0 1.000000000 0.00000000 0.8376003   0 0.00000000 0.00000000
#> a_6    0   0   0   0 0.000000000 1.00000000 0.0000000   0 0.00000000 0.00000000
#> a_7    0   0   0   0 0.000000000 0.00000000 1.0000000   0 0.00000000 0.00000000
#> a_8    0   0   0   0 0.000000000 0.00000000 0.0000000   1 0.00000000 0.00000000
#> a_9    0   0   0   0 0.000000000 0.00000000 0.0000000   0 1.00000000 1.00000000
#> a_10   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.80202638
#> a_11   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_12   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_13   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_14   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_15   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_16   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_17   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_18   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_19   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_20   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_21   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_22   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_23   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_24   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_25   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_26   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_27   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_28   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_29   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_30   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_31   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#> a_32   0   0   0   0 0.000000000 0.00000000 0.0000000   0 0.00000000 0.00000000
#>      a_11        a_12 a_13 a_14       a_15       a_16       a_17       a_18
#> a_1     0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_2     0 0.001142882    0    0 1.00000000 0.00000000 0.00000000 0.04525178
#> a_3     0 0.073440849    0    1 0.00000000 0.00000000 0.00000000 0.00000000
#> a_4     0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_5     0 1.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_6     0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.05594434
#> a_7     0 0.000000000    0    0 0.07630829 0.00000000 0.00000000 0.00000000
#> a_8     0 0.000000000    0    0 0.00000000 0.09518266 0.00000000 0.00000000
#> a_9     1 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_10    1 1.000000000    0    0 0.00000000 0.00000000 0.04224927 0.00000000
#> a_11    1 0.000000000    0    0 0.00000000 1.00000000 0.00000000 0.00000000
#> a_12    0 1.000000000    0    0 0.00000000 0.01347718 0.00000000 0.00000000
#> a_13    0 0.000000000    1    0 0.00000000 0.00000000 0.07265218 0.00000000
#> a_14    0 0.000000000    0    1 0.00000000 0.00000000 0.00000000 0.00000000
#> a_15    0 0.000000000    0    0 1.00000000 0.00000000 0.00000000 0.00000000
#> a_16    0 0.000000000    0    0 0.00000000 1.00000000 0.00000000 0.00000000
#> a_17    0 0.000000000    0    0 0.00000000 0.00000000 1.00000000 0.00000000
#> a_18    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 1.00000000
#> a_19    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_20    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_21    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_22    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_23    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_24    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_25    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_26    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_27    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_28    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_29    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_30    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_31    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#> a_32    0 0.000000000    0    0 0.00000000 0.00000000 0.00000000 0.00000000
#>             a_19       a_20       a_21        a_22       a_23       a_24
#> a_1  0.000000000 1.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_2  0.000000000 0.00000000 0.02936357 0.000000000 0.06648682 0.00000000
#> a_3  0.000000000 0.05089904 0.00000000 0.000000000 0.00000000 0.02570835
#> a_4  0.000000000 0.00000000 0.00000000 0.028691192 0.00000000 0.00000000
#> a_5  0.000000000 1.00000000 0.00000000 0.020430100 0.00000000 0.00000000
#> a_6  0.000000000 0.00000000 0.00000000 0.009128827 0.00000000 0.00000000
#> a_7  0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_8  0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_9  0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_10 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_11 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_12 0.000000000 0.05159520 0.00000000 0.000000000 0.00000000 0.00000000
#> a_13 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_14 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.04717063
#> a_15 0.000000000 0.00000000 0.07591324 0.000000000 0.00000000 0.00000000
#> a_16 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_17 0.002100126 0.00000000 0.00000000 1.000000000 0.00000000 0.00000000
#> a_18 1.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_19 1.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_20 0.000000000 1.00000000 1.00000000 0.000000000 0.00000000 0.57839692
#> a_21 0.000000000 0.00000000 1.00000000 0.000000000 1.00000000 0.00000000
#> a_22 0.000000000 0.00000000 0.00000000 1.000000000 0.00000000 0.03692915
#> a_23 0.000000000 0.00000000 0.00000000 0.000000000 1.00000000 0.00000000
#> a_24 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 1.00000000
#> a_25 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_26 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_27 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_28 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_29 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_30 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_31 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_32 0.000000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#>            a_25       a_26       a_27       a_28         a_29       a_30 a_31
#> a_1  0.00000000 0.00000000 0.00000000 0.00000000 0.0009536063 0.00000000    0
#> a_2  0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_3  0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_4  0.01554142 0.00000000 0.00000000 0.00000000 0.0000000000 0.02612037    0
#> a_5  0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_6  0.00000000 0.06610816 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_7  0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_8  0.00000000 0.00000000 0.06826361 0.00000000 0.0000000000 0.00000000    0
#> a_9  0.06685250 0.00000000 0.00000000 0.03326300 0.0000000000 0.00000000    0
#> a_10 0.00000000 0.00000000 0.00000000 0.00000000 0.0798857245 0.00000000    0
#> a_11 0.03723593 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_12 0.00000000 1.00000000 0.00000000 0.00000000 0.0000000000 0.05996815    0
#> a_13 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_14 0.00000000 0.00000000 0.09739540 0.00000000 0.0000000000 0.00000000    0
#> a_15 0.00000000 0.00000000 0.00000000 0.08221715 0.0000000000 0.03289764    0
#> a_16 0.00000000 0.07926021 0.00000000 0.03428516 0.0000000000 0.00000000    0
#> a_17 0.00000000 0.00000000 0.00000000 0.02661224 0.0000000000 0.00000000    0
#> a_18 0.00000000 0.00000000 0.08362664 0.00000000 0.0000000000 0.00000000    0
#> a_19 0.00000000 0.00000000 0.00000000 0.02212098 0.0484891972 0.09306896    0
#> a_20 0.00000000 0.00000000 1.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_21 0.02040768 1.00000000 0.00000000 0.00000000 0.0656332273 0.00000000    0
#> a_22 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.03427177    0
#> a_23 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_24 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_25 1.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_26 0.00000000 1.00000000 1.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_27 0.00000000 0.00000000 1.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_28 0.00000000 0.00000000 0.00000000 1.00000000 0.0730341640 0.00000000    0
#> a_29 0.00000000 0.00000000 0.00000000 0.00000000 1.0000000000 0.01950239    0
#> a_30 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 1.00000000    0
#> a_31 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#> a_32 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000000 0.00000000    0
#>      a_32       t_1       t_2       c_1       c_2       q_1       q_2     r_1.1
#> a_1     0  2.301139  2.493263 11.576630 12.624806 0.4691942 0.5174588 13.234122
#> a_2     0 18.099485 18.611744 10.832293 13.338379 0.7781432 0.8638524 13.699590
#> a_3     0 17.584928 18.778192 11.049531 13.751565 0.3444858 0.4165109  4.911462
#> a_4     0 18.187586 18.826325  3.495776  4.221873 0.5579386 0.6226630  9.256895
#> a_5     0  9.285713 11.313341 16.578017 19.410156 0.6933574 0.7517785 13.403837
#> a_6     0 16.529493 19.412737 19.170531 23.240695 0.5130048 0.5580523 13.149876
#> a_7     0  5.328804  5.959527 14.805266 16.715168 0.6137883 0.6842679  4.699537
#> a_8     0 12.697573 13.520908 16.279895 18.408858 0.6470740 0.7949430 14.961348
#> a_9     0 12.960878 14.573120 22.679789 26.668362 0.8465924 0.8835081 10.853462
#> a_10    0 13.048931 15.106848 14.424886 16.509053 0.5883755 0.6151275 14.417906
#> a_11    0 17.032864 19.977627 10.721291 12.281008 0.5989135 0.6103385 10.868228
#> a_12    0 14.606305 17.544339 14.744169 17.344597 0.8561401 0.8934922 15.020271
#> a_13    0 15.067933 17.216002  2.967044  3.331605 0.7049626 0.8342912  6.126484
#> a_14    0 16.087120 18.279094  9.680978 10.900662 0.3744115 0.4101690 12.226582
#> a_15    0 17.239941 18.730956 17.561490 21.830331 0.3914429 0.4387946 17.982446
#> a_16    0 15.299348 16.365430 15.852940 17.240645 0.7900010 0.8870228 17.100462
#> a_17    0  3.221207  3.474191  7.889933  9.059529 0.4991775 0.6032608 15.173240
#> a_18    0 10.725129 11.712008 24.950582 29.918873 0.4613426 0.4787154  9.317324
#> a_19    0 15.560227 17.780721 22.430974 23.200758 0.6057429 0.7192511 15.503287
#> a_20    0  5.953181  6.302124 23.271244 23.696816 0.8518833 0.8677862  9.212595
#> a_21    0 15.913838 16.659591 26.510453 26.547886 0.4315185 0.4392470  8.363159
#> a_22    0  1.853236  2.122454 12.581153 12.780319 0.3338708 0.3689668 13.241447
#> a_23    0 15.752421 18.397711 22.456749 23.036681 0.3853201 0.4053924 16.938635
#> a_24    0 14.379279 17.823137 15.424374 17.804567 0.6959424 0.7244180 10.631585
#> a_25    0 17.150406 19.308790  7.010209  8.403820 0.4222313 0.4806034 13.807524
#> a_26    0 16.663504 17.838040  7.554132  8.803231 0.6447334 0.6941722  7.333930
#> a_27    0 11.688123 12.025986 16.217439 17.395721 0.6072583 0.7181080 17.232280
#> a_28    0 18.177957 18.561772  9.696388 11.007670 0.6035469 0.6520422 14.060112
#> a_29    0 16.358864 16.668290 11.534555 11.811030 0.6512519 0.7570994 13.146676
#> a_30    0 15.457658 18.251179 19.314912 22.986456 0.6556246 0.7466798 12.828299
#> a_31    0  0.000000  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000
#> a_32    0  0.000000  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000
#>          r_2.1     r_1.2     r_2.2
#> a_1  14.581530 10.566248 13.183811
#> a_2  15.422656  7.200756  8.224126
#> a_3   5.054926  4.556959  5.190598
#> a_4  10.826090 13.972664 14.139643
#> a_5  13.668742  2.296732  2.831684
#> a_6  15.053310 14.464764 14.607059
#> a_7   5.169577 16.890346 18.240664
#> a_8  17.242990 15.840271 19.766390
#> a_9  12.454891 13.905069 14.459292
#> a_10 17.370369 10.098940 10.166560
#> a_11 13.155397  4.255330  5.153211
#> a_12 18.164702 11.631560 14.167317
#> a_13  6.910487 16.978018 16.992262
#> a_14 13.789354 12.374986 15.070204
#> a_15 17.983544 13.169910 15.142868
#> a_16 17.207243 17.401226 18.389631
#> a_17 16.539409 14.285318 15.357935
#> a_18  9.989277 14.904365 17.550989
#> a_19 18.397984 16.897304 17.866218
#> a_20 10.065498 16.061832 17.978586
#> a_21  8.729066 11.709009 11.759927
#> a_22 16.397675  9.592835 10.418009
#> a_23 17.557161  5.524634  6.209450
#> a_24 12.406955  9.328349 10.946963
#> a_25 16.348513 17.823028 19.739994
#> a_26  7.940227 12.846634 15.201561
#> a_27 18.931795 12.146725 13.441901
#> a_28 17.362973  9.119785  9.800199
#> a_29 13.562057 16.051606 16.547068
#> a_30 14.381995 10.339689 11.048119
#> a_31  0.000000  0.000000  0.000000
#> a_32  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  113.3764
#> Const constraint (Cc):  424.8387
#> Score/scope constraint (Cs):  0.9506463
#> Quailty constraint (Cq):  0.5304531
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  156.8942
#> Const constraint (Cc):  494.2711
#> Score/scope constraint (Cs):  0.9931295
#> Quailty constraint (Cq):  0.6218218
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 238.8453 239.1669