Print PDM constraints, matrices, lists, sets, collections.

Print PDM constraints, matrices, lists, sets, collections.

# S3 method for PDM_const
summary(object, digits =  getOption("digits"), ...)
# S3 method for PDM_matrix
summary(object, digits =  getOption("digits"),w=getOption("w"),Rs=getOption("Rs"), ...)
# S3 method for PDM_list
summary(object, digits =  getOption("digits"), ...)
# S3 method for Set_PDM_matrix
summary(object, digits =  getOption("digits"),w=getOption("w"),Rs=getOption("Rs"), ...)
# S3 method for Set_PDM_list
summary(object, digits =  getOption("digits"), ...)
# S3 method for Collection_PDM
summary(object, digits =  getOption("digits"), ...)
# S3 method for TPT
summary(object, digits =  getOption("digits"), ...)

Arguments

object

an object of class 'PDM_const', 'PDM_matrix', 'PDM_list', 'Set_PDM_matrix', or 'Set_PDM_list'.

digits

the number of significant digits to use when add.stats = TRUE.

w

number of completion modes (positive integer)

Rs

number of resources (positive integer)

...

additional arguments affecting the print produced.

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Print PDM constriants, matrices, lists, and sets.

# Define number of modes, flexibility factor and connectivity factor of a project scenerio.
N=4;ff=0.05;cf=0

# Define maximum value of time domain, Cost domain and Resourcces domain of a project scenerio.
mTD=3;mCD=4;mRD=3

# Define number of modes, number of resources,
# number of possible extra tasks, scale and quality domain of a project scenerio.

w=2;nR=2;nW=2
scale=1.4  #(default value)

# Generation of PDM list for TCTP a project scenario using MFPP package.

PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,lst=TRUE)

# Generate PDM maximal constraints (ratio=1.0)

CONST<-percent(PDM)

# Calculate main structures

PDMs<-get.structures(PDM)

# print of PDM constraint

summary(CONST)
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  4.719776
#> Const constraint (Cc):  11.83343
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#>          R_1     R_2
#> TPR 4.984869 4.52511

# print of PDM list

summary(PDM)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>     a_1       a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1 0.9807703   1   0   0   0 1.508270 1.731944 2.243927 1.731944 1.646738
#> a_2   0 1.0000000   1   1   0   0 1.521626 1.866585 2.302027 1.866585 1.077810
#> a_3   0 0.0000000   1   0   0   0 1.543632 1.916518 3.560341 1.916518 2.885445
#> a_4   0 0.0000000   0   1   0   0 2.413692 2.853191 3.727138 2.853191 1.762010
#> a_5   0 0.0000000   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0 0.0000000   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.041813 1.333495 1.598965
#> a_2 1.225391 2.167176 2.618126
#> a_3 2.921460 1.838874 2.290482
#> a_4 2.063409 2.044010 2.234628
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  3.935318
#> Const constraint (Cc):  8.368238
#> Score/scope constraint (Cs):  1
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  4.719776
#> Const constraint (Cc):  11.83343
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#>          R_1     R_2
#> TPR 4.984869 4.52511

# print of main structures of PDM matrices

summary(PDMs)
#> 
#> Summary of main structures:
#> 
#> 
#> 
#> Summary of minimal structure:
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   0   1   0   0   0 1.508270 1.731944 2.243927 1.731944 1.646738
#> a_2   0   1   1   1   0   0 1.521626 1.866585 2.302027 1.866585 1.077810
#> a_3   0   0   1   0   0   0 1.543632 1.916518 3.560341 1.916518 2.885445
#> a_4   0   0   0   1   0   0 2.413692 2.853191 3.727138 2.853191 1.762010
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.041813 1.333495 1.598965
#> a_2 1.225391 2.167176 2.618126
#> a_3 2.921460 1.838874 2.290482
#> a_4 2.063409 2.044010 2.234628
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  3.935318
#> Const constraint (Cc):  8.368238
#> Score/scope constraint (Cs):  1
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  4.719776
#> Const constraint (Cc):  11.83343
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#> 
#> 
#> 
#> Summary of maximal structure:
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.508270 1.731944 2.243927 1.731944 1.646738
#> a_2   0   1   1   1   0   0 1.521626 1.866585 2.302027 1.866585 1.077810
#> a_3   0   0   1   0   0   0 1.543632 1.916518 3.560341 1.916518 2.885445
#> a_4   0   0   0   1   0   0 2.413692 2.853191 3.727138 2.853191 1.762010
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.041813 1.333495 1.598965
#> a_2 1.225391 2.167176 2.618126
#> a_3 2.921460 1.838874 2.290482
#> a_4 2.063409 2.044010 2.234628
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  3.935318
#> Const constraint (Cc):  8.368238
#> Score/scope constraint (Cs):  1
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  4.719776
#> Const constraint (Cc):  11.83343
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#> 
#> 
#> 
#> Summary of minimax structure:
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.508270 1.731944 2.243927 1.731944 1.646738
#> a_2   0   1   1   1   0   0 1.521626 1.866585 2.302027 1.866585 1.077810
#> a_3   0   0   1   0   0   0 1.543632 1.916518 3.560341 1.916518 2.885445
#> a_4   0   0   0   1   0   0 2.413692 2.853191 3.727138 2.853191 1.762010
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.041813 1.333495 1.598965
#> a_2 1.225391 2.167176 2.618126
#> a_3 2.921460 1.838874 2.290482
#> a_4 2.063409 2.044010 2.234628
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  3.935318
#> Const constraint (Cc):  8.368238
#> Score/scope constraint (Cs):  1
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  4.719776
#> Const constraint (Cc):  11.83343
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#> 
#> 
#> 
#> Summary of maximin structure:
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   0   1   0   0   0 1.508270 1.731944 2.243927 1.731944 1.646738
#> a_2   0   1   1   1   0   0 1.521626 1.866585 2.302027 1.866585 1.077810
#> a_3   0   0   1   0   0   0 1.543632 1.916518 3.560341 1.916518 2.885445
#> a_4   0   0   0   1   0   0 2.413692 2.853191 3.727138 2.853191 1.762010
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.041813 1.333495 1.598965
#> a_2 1.225391 2.167176 2.618126
#> a_3 2.921460 1.838874 2.290482
#> a_4 2.063409 2.044010 2.234628
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  3.935318
#> Const constraint (Cc):  8.368238
#> Score/scope constraint (Cs):  1
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  4.719776
#> Const constraint (Cc):  11.83343
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#> 
#> 
#> 
#> Summary of most likely/most desired structure:
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.508270 1.731944 2.243927 1.731944 1.646738
#> a_2   0   1   1   1   0   0 1.521626 1.866585 2.302027 1.866585 1.077810
#> a_3   0   0   1   0   0   0 1.543632 1.916518 3.560341 1.916518 2.885445
#> a_4   0   0   0   1   0   0 2.413692 2.853191 3.727138 2.853191 1.762010
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.041813 1.333495 1.598965
#> a_2 1.225391 2.167176 2.618126
#> a_3 2.921460 1.838874 2.290482
#> a_4 2.063409 2.044010 2.234628
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  3.935318
#> Const constraint (Cc):  8.368238
#> Score/scope constraint (Cs):  1
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  4.719776
#> Const constraint (Cc):  11.83343
#> Score/scope constraint (Cs):  1
#> Resource constraint(s) (CR):
#>          R_1     R_2
#> TPR 4.984869 4.52511

# print of PDM collection
data(Boctor)
collection<-Boctor
summary(collection)
#> 
#> 
#> 
#> Summary of PDM collection:
#> 
#> Number of projects:  240
#> List of projects: 
#>  Project name:  boct1, w:  4, Rs:  1
#>  Project name:  boct10, w:  4, Rs:  2
#>  Project name:  boct100, w:  4, Rs:  2
#>  Project name:  boct101, w:  4, Rs:  4
#>  Project name:  boct102, w:  4, Rs:  4
#>  Project name:  boct103, w:  4, Rs:  1
#>  Project name:  boct104, w:  4, Rs:  1
#>  Project name:  boct105, w:  4, Rs:  2
#>  Project name:  boct106, w:  4, Rs:  2
#>  Project name:  boct107, w:  4, Rs:  4
#>  Project name:  boct108, w:  4, Rs:  4
#>  Project name:  boct109, w:  4, Rs:  1
#>  Project name:  boct11, w:  4, Rs:  4
#>  Project name:  boct110, w:  4, Rs:  1
#>  Project name:  boct111, w:  4, Rs:  2
#>  Project name:  boct112, w:  4, Rs:  2
#>  Project name:  boct113, w:  4, Rs:  4
#>  Project name:  boct114, w:  4, Rs:  4
#>  Project name:  boct115, w:  4, Rs:  1
#>  Project name:  boct116, w:  4, Rs:  1
#>  Project name:  boct117, w:  4, Rs:  2
#>  Project name:  boct118, w:  4, Rs:  2
#>  Project name:  boct119, w:  4, Rs:  4
#>  Project name:  boct12, w:  4, Rs:  4
#>  Project name:  boct120, w:  4, Rs:  4
#>  Project name:  boct121, w:  4, Rs:  1
#>  Project name:  boct122, w:  4, Rs:  1
#>  Project name:  boct123, w:  4, Rs:  2
#>  Project name:  boct124, w:  4, Rs:  2
#>  Project name:  boct125, w:  4, Rs:  4
#>  Project name:  boct126, w:  4, Rs:  4
#>  Project name:  boct127, w:  4, Rs:  1
#>  Project name:  boct128, w:  4, Rs:  1
#>  Project name:  boct129, w:  4, Rs:  2
#>  Project name:  boct13, w:  4, Rs:  1
#>  Project name:  boct130, w:  4, Rs:  2
#>  Project name:  boct131, w:  4, Rs:  4
#>  Project name:  boct132, w:  4, Rs:  4
#>  Project name:  boct133, w:  4, Rs:  1
#>  Project name:  boct134, w:  4, Rs:  1
#>  Project name:  boct135, w:  4, Rs:  2
#>  Project name:  boct136, w:  4, Rs:  2
#>  Project name:  boct137, w:  4, Rs:  4
#>  Project name:  boct138, w:  4, Rs:  4
#>  Project name:  boct139, w:  4, Rs:  1
#>  Project name:  boct14, w:  4, Rs:  1
#>  Project name:  boct140, w:  4, Rs:  1
#>  Project name:  boct141, w:  4, Rs:  2
#>  Project name:  boct142, w:  4, Rs:  2
#>  Project name:  boct143, w:  4, Rs:  4
#>  Project name:  boct144, w:  4, Rs:  4
#>  Project name:  boct145, w:  4, Rs:  1
#>  Project name:  boct146, w:  4, Rs:  1
#>  Project name:  boct147, w:  4, Rs:  2
#>  Project name:  boct148, w:  4, Rs:  2
#>  Project name:  boct149, w:  4, Rs:  4
#>  Project name:  boct15, w:  4, Rs:  2
#>  Project name:  boct150, w:  4, Rs:  4
#>  Project name:  boct151, w:  4, Rs:  1
#>  Project name:  boct152, w:  4, Rs:  1
#>  Project name:  boct153, w:  4, Rs:  2
#>  Project name:  boct154, w:  4, Rs:  2
#>  Project name:  boct155, w:  4, Rs:  4
#>  Project name:  boct156, w:  4, Rs:  4
#>  Project name:  boct157, w:  4, Rs:  1
#>  Project name:  boct158, w:  4, Rs:  1
#>  Project name:  boct159, w:  4, Rs:  2
#>  Project name:  boct16, w:  4, Rs:  2
#>  Project name:  boct160, w:  4, Rs:  2
#>  Project name:  boct161, w:  4, Rs:  4
#>  Project name:  boct162, w:  4, Rs:  4
#>  Project name:  boct163, w:  4, Rs:  1
#>  Project name:  boct164, w:  4, Rs:  1
#>  Project name:  boct165, w:  4, Rs:  2
#>  Project name:  boct166, w:  4, Rs:  2
#>  Project name:  boct167, w:  4, Rs:  4
#>  Project name:  boct168, w:  4, Rs:  4
#>  Project name:  boct169, w:  4, Rs:  1
#>  Project name:  boct17, w:  4, Rs:  4
#>  Project name:  boct170, w:  4, Rs:  1
#>  Project name:  boct171, w:  4, Rs:  2
#>  Project name:  boct172, w:  4, Rs:  2
#>  Project name:  boct173, w:  4, Rs:  4
#>  Project name:  boct174, w:  4, Rs:  4
#>  Project name:  boct175, w:  4, Rs:  1
#>  Project name:  boct176, w:  4, Rs:  1
#>  Project name:  boct177, w:  4, Rs:  2
#>  Project name:  boct178, w:  4, Rs:  2
#>  Project name:  boct179, w:  4, Rs:  4
#>  Project name:  boct18, w:  4, Rs:  4
#>  Project name:  boct180, w:  4, Rs:  4
#>  Project name:  boct181, w:  4, Rs:  1
#>  Project name:  boct182, w:  4, Rs:  1
#>  Project name:  boct183, w:  4, Rs:  2
#>  Project name:  boct184, w:  4, Rs:  2
#>  Project name:  boct185, w:  4, Rs:  4
#>  Project name:  boct186, w:  4, Rs:  4
#>  Project name:  boct187, w:  4, Rs:  1
#>  Project name:  boct188, w:  4, Rs:  1
#>  Project name:  boct189, w:  4, Rs:  2
#>  Project name:  boct19, w:  4, Rs:  1
#>  Project name:  boct190, w:  4, Rs:  2
#>  Project name:  boct191, w:  4, Rs:  4
#>  Project name:  boct192, w:  4, Rs:  4
#>  Project name:  boct193, w:  4, Rs:  1
#>  Project name:  boct194, w:  4, Rs:  1
#>  Project name:  boct195, w:  4, Rs:  2
#>  Project name:  boct196, w:  4, Rs:  2
#>  Project name:  boct197, w:  4, Rs:  4
#>  Project name:  boct198, w:  4, Rs:  4
#>  Project name:  boct199, w:  4, Rs:  1
#>  Project name:  boct2, w:  4, Rs:  1
#>  Project name:  boct20, w:  4, Rs:  1
#>  Project name:  boct200, w:  4, Rs:  1
#>  Project name:  boct201, w:  4, Rs:  2
#>  Project name:  boct202, w:  4, Rs:  2
#>  Project name:  boct203, w:  4, Rs:  4
#>  Project name:  boct204, w:  4, Rs:  4
#>  Project name:  boct205, w:  4, Rs:  1
#>  Project name:  boct206, w:  4, Rs:  1
#>  Project name:  boct207, w:  4, Rs:  2
#>  Project name:  boct208, w:  4, Rs:  2
#>  Project name:  boct209, w:  4, Rs:  4
#>  Project name:  boct21, w:  4, Rs:  2
#>  Project name:  boct210, w:  4, Rs:  4
#>  Project name:  boct211, w:  4, Rs:  1
#>  Project name:  boct212, w:  4, Rs:  1
#>  Project name:  boct213, w:  4, Rs:  2
#>  Project name:  boct214, w:  4, Rs:  2
#>  Project name:  boct215, w:  4, Rs:  4
#>  Project name:  boct216, w:  4, Rs:  4
#>  Project name:  boct217, w:  4, Rs:  1
#>  Project name:  boct218, w:  4, Rs:  1
#>  Project name:  boct219, w:  4, Rs:  2
#>  Project name:  boct22, w:  4, Rs:  2
#>  Project name:  boct220, w:  4, Rs:  2
#>  Project name:  boct221, w:  4, Rs:  4
#>  Project name:  boct222, w:  4, Rs:  4
#>  Project name:  boct223, w:  4, Rs:  1
#>  Project name:  boct224, w:  4, Rs:  1
#>  Project name:  boct225, w:  4, Rs:  2
#>  Project name:  boct226, w:  4, Rs:  2
#>  Project name:  boct227, w:  4, Rs:  4
#>  Project name:  boct228, w:  4, Rs:  4
#>  Project name:  boct229, w:  4, Rs:  1
#>  Project name:  boct23, w:  4, Rs:  4
#>  Project name:  boct230, w:  4, Rs:  1
#>  Project name:  boct231, w:  4, Rs:  2
#>  Project name:  boct232, w:  4, Rs:  2
#>  Project name:  boct233, w:  4, Rs:  4
#>  Project name:  boct234, w:  4, Rs:  4
#>  Project name:  boct235, w:  4, Rs:  1
#>  Project name:  boct236, w:  4, Rs:  1
#>  Project name:  boct237, w:  4, Rs:  2
#>  Project name:  boct238, w:  4, Rs:  2
#>  Project name:  boct239, w:  4, Rs:  4
#>  Project name:  boct24, w:  4, Rs:  4
#>  Project name:  boct240, w:  4, Rs:  4
#>  Project name:  boct25, w:  4, Rs:  1
#>  Project name:  boct26, w:  4, Rs:  1
#>  Project name:  boct27, w:  4, Rs:  2
#>  Project name:  boct28, w:  4, Rs:  2
#>  Project name:  boct29, w:  4, Rs:  4
#>  Project name:  boct3, w:  4, Rs:  2
#>  Project name:  boct30, w:  4, Rs:  4
#>  Project name:  boct31, w:  4, Rs:  1
#>  Project name:  boct32, w:  4, Rs:  1
#>  Project name:  boct33, w:  4, Rs:  2
#>  Project name:  boct34, w:  4, Rs:  2
#>  Project name:  boct35, w:  4, Rs:  4
#>  Project name:  boct36, w:  4, Rs:  4
#>  Project name:  boct37, w:  4, Rs:  1
#>  Project name:  boct38, w:  4, Rs:  1
#>  Project name:  boct39, w:  4, Rs:  2
#>  Project name:  boct4, w:  4, Rs:  2
#>  Project name:  boct40, w:  4, Rs:  2
#>  Project name:  boct41, w:  4, Rs:  4
#>  Project name:  boct42, w:  4, Rs:  4
#>  Project name:  boct43, w:  4, Rs:  1
#>  Project name:  boct44, w:  4, Rs:  1
#>  Project name:  boct45, w:  4, Rs:  2
#>  Project name:  boct46, w:  4, Rs:  2
#>  Project name:  boct47, w:  4, Rs:  4
#>  Project name:  boct48, w:  4, Rs:  4
#>  Project name:  boct49, w:  4, Rs:  1
#>  Project name:  boct5, w:  4, Rs:  4
#>  Project name:  boct50, w:  4, Rs:  1
#>  Project name:  boct51, w:  4, Rs:  2
#>  Project name:  boct52, w:  4, Rs:  2
#>  Project name:  boct53, w:  4, Rs:  4
#>  Project name:  boct54, w:  4, Rs:  4
#>  Project name:  boct55, w:  4, Rs:  1
#>  Project name:  boct56, w:  4, Rs:  1
#>  Project name:  boct57, w:  4, Rs:  2
#>  Project name:  boct58, w:  4, Rs:  2
#>  Project name:  boct59, w:  4, Rs:  4
#>  Project name:  boct6, w:  4, Rs:  4
#>  Project name:  boct60, w:  4, Rs:  4
#>  Project name:  boct61, w:  4, Rs:  1
#>  Project name:  boct62, w:  4, Rs:  1
#>  Project name:  boct63, w:  4, Rs:  2
#>  Project name:  boct64, w:  4, Rs:  2
#>  Project name:  boct65, w:  4, Rs:  4
#>  Project name:  boct66, w:  4, Rs:  4
#>  Project name:  boct67, w:  4, Rs:  1
#>  Project name:  boct68, w:  4, Rs:  1
#>  Project name:  boct69, w:  4, Rs:  2
#>  Project name:  boct7, w:  4, Rs:  1
#>  Project name:  boct70, w:  4, Rs:  2
#>  Project name:  boct71, w:  4, Rs:  4
#>  Project name:  boct72, w:  4, Rs:  4
#>  Project name:  boct73, w:  4, Rs:  1
#>  Project name:  boct74, w:  4, Rs:  1
#>  Project name:  boct75, w:  4, Rs:  2
#>  Project name:  boct76, w:  4, Rs:  2
#>  Project name:  boct77, w:  4, Rs:  4
#>  Project name:  boct78, w:  4, Rs:  4
#>  Project name:  boct79, w:  4, Rs:  1
#>  Project name:  boct8, w:  4, Rs:  1
#>  Project name:  boct80, w:  4, Rs:  1
#>  Project name:  boct81, w:  4, Rs:  2
#>  Project name:  boct82, w:  4, Rs:  2
#>  Project name:  boct83, w:  4, Rs:  4
#>  Project name:  boct84, w:  4, Rs:  4
#>  Project name:  boct85, w:  4, Rs:  1
#>  Project name:  boct86, w:  4, Rs:  1
#>  Project name:  boct87, w:  4, Rs:  2
#>  Project name:  boct88, w:  4, Rs:  2
#>  Project name:  boct89, w:  4, Rs:  4
#>  Project name:  boct9, w:  4, Rs:  2
#>  Project name:  boct90, w:  4, Rs:  4
#>  Project name:  boct91, w:  4, Rs:  1
#>  Project name:  boct92, w:  4, Rs:  1
#>  Project name:  boct93, w:  4, Rs:  2
#>  Project name:  boct94, w:  4, Rs:  2
#>  Project name:  boct95, w:  4, Rs:  4
#>  Project name:  boct96, w:  4, Rs:  4
#>  Project name:  boct97, w:  4, Rs:  1
#>  Project name:  boct98, w:  4, Rs:  1
#>  Project name:  boct99, w:  4, Rs:  2

# schedule table of the first completion mode of the first project structure

# get the first project structure

PDM<-collection[[1]]$PDM_list$PDM

# get the logic domain

LD<-PDM[1:nrow(PDM),1:nrow(PDM)]

# get the time demands of the first completion mode

TD<-PDM[1:nrow(PDM),nrow(PDM)+1]

# calculate TPT

TPT<-tpt(LD,TD)

# print schedule table

summary(TPT)
#> 
#> 
#>  Table of schedule
#>      Dur EST EFT LST LFT TF SST SFT SF Is.Crit
#> a_1   13   0  13   0  13  0   0  13  0    TRUE
#> a_2   15   0  15  10  25 10   0  15 10   FALSE
#> a_3   11  13  24  13  24  0  13  24  0    TRUE
#> a_4    9  24  33  25  34  1  24  33  1   FALSE
#> a_5    5  24  29  24  29  0  24  29  0    TRUE
#> a_6    2  29  31  29  31  0  29  31  0    TRUE
#> a_7    8  33  41  34  42  1  33  41  1   FALSE
#> a_8    3  29  32  45  48 16  29  32 16   FALSE
#> a_9   11  31  42  31  42  0  31  42  0    TRUE
#> a_10   6  42  48  42  48  0  42  48  0    TRUE
#> a_11  10  48  58  48  58  0  48  58  0    TRUE
#> a_12   1  24  25  57  58 33  24  25 33   FALSE
#> a_13   9  41  50  43  52  2  41  50  2   FALSE
#> a_14   8  58  66  58  66  0  58  66  0    TRUE
#> a_15   3  50  53  52  55  2  50  53  2   FALSE
#> a_16   7  53  60  55  62  2  53  60  2   FALSE
#> a_17   4  60  64  62  66  2  60  64  2   FALSE
#> a_18   8  66  74  69  77  3  66  74  3   FALSE
#> a_19   4  66  70  66  70  0  66  70  0    TRUE
#> a_20  11  70  81  70  81  0  70  81  0    TRUE
#> a_21   1  74  75  77  78  3  74  75  3   FALSE
#> a_22   9  64  73  72  81  8  64  73  8   FALSE
#> a_23   7  70  77  74  81  4  70  77  4   FALSE
#> a_24   1  81  82  82  83  1  81  82  1   FALSE
#> a_25   2  81  83  81  83  0  81  83  0    TRUE
#> a_26  10  75  85  78  88  3  75  85  3   FALSE
#> a_27   9  85  94  88  97  3  85  94  3   FALSE
#> a_28  11  85  96  92 103  7  85  96  7   FALSE
#> a_29  14  83  97  83  97  0  83  97  0    TRUE
#> a_30   6  97 103  97 103  0  97 103  0    TRUE
#> a_31  13  97 110 105 118  8  97 110  8   FALSE
#> a_32  15 103 118 103 118  0 103 118  0    TRUE
#> a_33  13  96 109 105 118  9  96 109  9   FALSE
#> a_34   8 118 126 118 126  0 118 126  0    TRUE
#> a_35  15 126 141 126 141  0 126 141  0    TRUE
#> a_36   9 141 150 141 150  0 141 150  0    TRUE
#> a_37   7 150 157 150 157  0 150 157  0    TRUE
#> a_38  14  96 110 163 177 67  96 110 67   FALSE
#> a_39  14 157 171 157 171  0 157 171  0    TRUE
#> a_40   6 171 177 171 177  0 171 177  0    TRUE
#> a_41   5 177 182 177 182  0 177 182  0    TRUE
#> a_42  12 157 169 170 182 13 157 169 13   FALSE
#> a_43  15 182 197 182 197  0 182 197  0    TRUE
#> a_44   9 197 206 197 206  0 197 206  0    TRUE
#> a_45  11 177 188 206 217 29 177 188 29   FALSE
#> a_46  11 206 217 206 217  0 206 217  0    TRUE
#> a_47  11 217 228 217 228  0 217 228  0    TRUE
#> a_48  15 228 243 228 243  0 228 243  0    TRUE
#> a_49   2 228 230 241 243 13 228 230 13   FALSE
#> a_50   2 243 245 243 245  0 243 245  0    TRUE