Function to generate a PDM matrix for flexible project planning problems.

Function to generate a PDM matrix for flexible project planning problems.

generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=FALSE,lst=FALSE)

Arguments

N

Number of tasks (a numeric value)

ff

flexibility factor between [0,1] (a numeric value).

cf

connectivity factor (a numeric integer).

mTD

max value of TD (a numeric positive integer).

mCD

max value of CD (a numeric positive integer).

mRD

max value of RD (a numeric positive integer).

w

number of modes (a numeric integer).

nR

number of resources (a numeric integer).

nW

number of possible extra tasks (a positive numeric integer).

scale

scale parameter, the default value is 1.4 (a numeric value).

QD

logic parameter, the default value is FALSE, which means, the quality domain is neglected.

lst

(logic parameter, the default value is FALSE, which means, the output is a PDM matrix, while TRUE means the output is a PDM list including number of modes (w) and the number of resources (Rs).

Value

PDM

[PEM,TD, CD, QD,RD] is a Project Domain Matrix (a numeric matrix), where;

PEM

An N by N upper triangular matrix of logic domain (a numeric matrix).

TD

N by w matrix of task durations (a numeric matrix).

CD

N by w matrix of cost demands (a numeric matrix).

QD

N by w matrix of quality parameters (a numeric matrix).

RD

N by w*nR matrix of resource demands (a numeric matrix).

w

Number of completion modes (scalar, if the output is PDM list (lst=TRUE)).

Rs

Number of resources (scalar, if the output is PDM list (lst=TRUE)).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Generation of PDM matrix for flexible project planning MFPP package.

# Define number of modes, flexibility factor and connectivity factor of a project scenerio.
N=4;ff=0.05;cf=0

# Define maximum value of time domain, Cost domain and Resourcces domain of a project scenerio.
mTD=3;mCD=4;mRD=3

# Define number of modes, number of resources,
# number of possible extra tasks, scale and quality domain of a project scenerio.

w=2;nR=2;nW=2
scale=1.4  #(default value)

# Generation of PDM matrix for TCTP a project scenario using MFPP package.
generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4)
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   1   0   0   0 2.183304 2.534723 2.093986 2.534723 0.8244378
#> a_2   0   1   1   0   0   0 1.268735 1.488854 1.801072 1.488854 1.4300655
#> a_3   0   0   1   0   0   0 1.140419 1.179932 1.562823 1.179932 0.7679955
#> a_4   0   0   0   1   0   0 1.387378 1.409916 3.172352 1.409916 2.0920206
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1    r_1.2    r_2.2
#> a_1 0.9287633 2.398635 2.781072
#> a_2 1.4657121 1.394407 1.556389
#> a_3 0.9066589 2.435842 2.844313
#> a_4 2.5878676 1.085057 1.266863
#> a_5 0.0000000 0.000000 0.000000
#> a_6 0.0000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"

# Generation of PDM list for TCTP a project scenario using MFPP package.
generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=FALSE,lst=TRUE)
#> $PDM
#>     a_1 a_2       a_3 a_4 a_5 a_6       t_1       t_2      c_1       c_2
#> a_1   1   1 1.0000000   0   0   0 1.8575812 2.0248802 3.811871 2.0248802
#> a_2   0   1 0.9870739   0   0   0 1.5719348 1.8894050 3.103401 1.8894050
#> a_3   0   0 1.0000000   0   0   0 0.2996424 0.3626934 3.353155 0.3626934
#> a_4   0   0 0.0000000   1   0   0 0.9016195 0.9267278 1.383481 0.9267278
#> a_5   0   0 0.0000000   0   0   0 0.0000000 0.0000000 0.000000 0.0000000
#> a_6   0   0 0.0000000   0   0   0 0.0000000 0.0000000 0.000000 0.0000000
#>         r_1.1     r_2.1     r_1.2     r_2.2
#> a_1 2.5504827 2.7845305 2.3084644 2.4370241
#> a_2 1.7009593 2.0616352 2.0499208 2.5515469
#> a_3 0.3741524 0.4229811 1.7334682 1.9956576
#> a_4 0.7881620 0.8674833 0.8841895 0.9307351
#> a_5 0.0000000 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Generation of PDM matrix for TQCTP a project scenario using MFPP package.
generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=TRUE)
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2       q_1
#> a_1   1   1   0   0   0   0 2.5530547 2.815419 3.861158 3.985117 0.7505263
#> a_2   0   1   1   0   0   0 0.9238605 1.127368 3.491784 3.499776 0.8473704
#> a_3   0   0   1   0   0   0 2.8973369 2.982854 3.284645 3.592587 0.2209747
#> a_4   0   0   0   1   0   0 2.2215998 2.613421 2.986825 3.292780 0.3174907
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#>           q_2     r_1.1    r_2.1    r_1.2    r_2.2
#> a_1 0.7596991 0.9559042 1.187009 2.604151 2.684440
#> a_2 0.8477051 1.7067390 1.945415 2.495409 2.642168
#> a_3 0.2623166 2.1480861 2.278826 2.076423 2.292816
#> a_4 0.3463995 1.0266346 1.236614 2.826844 2.883968
#> a_5 0.0000000 0.0000000 0.000000 0.000000 0.000000
#> a_6 0.0000000 0.0000000 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"

 # Generation of PDM list for TQCTP a project scenario using MFPP package.
generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=TRUE,lst=TRUE)
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2       q_1
#> a_1   1   1   0   0   0   0 1.228861 1.249019 3.020917 3.565519 0.3834757
#> a_2   0   1   1   0   0   0 2.643617 2.714723 1.645160 1.706281 0.3191197
#> a_3   0   0   1   0   0   0 1.651103 1.847329 3.168348 3.754123 0.9021854
#> a_4   0   0   0   1   0   0 2.836659 2.914067 1.474429 1.596574 0.5018063
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>           q_2    r_1.1    r_2.1     r_1.2     r_2.2
#> a_1 0.4789169 1.837673 2.091709 0.6179477 0.7341536
#> a_2 0.3565975 1.244216 1.294085 0.7871237 0.9311174
#> a_3 0.9106997 1.046323 1.091387 2.3390868 2.5862588
#> a_4 0.5149713 1.807844 2.092207 2.1564354 2.1604424
#> a_5 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#> a_6 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"