Function to calculate minimal/maximal/most likely project structures.

Specify minimal, maximal, maximin, minimax, and most likely/most desired project structures.

get.structures(x,type=c("min","max","minimax","maximin","most"))

Arguments

x

N by M PDM matrix or a PDM list.

type

type of specified project structure. "min" is a minimal strucuture, where the time, cost demands, and as well as the quality parameters are minimal; "max" where the time, cost demands, and as well as the quality parameters are maximal; "minimax", where the cost and resource demands are minimal; "maximix", where the cost and resource demands are maxnimal; "most", where the total project score is maximal.

Value

PDMs

set of PDM_matrix or a set of PDM_list (depending on the input type)

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Generation of PDM matrix for flexible project planning MFPP package.

# Define number of modes, flexibility factor and connectivity factor of a project scenerio.
N=4;ff=0.05;cf=0

# Define maximum value of time domain, Cost domain and Resourcces domain of a project scenerio.
mTD=3;mCD=4;mRD=3

# Define number of modes, number of resources,
# number of possible extra tasks, scale and quality domain of a project scenerio.

w=2;nR=2;nW=2
scale=1.4  #(default value)

# Generation of PDM matrix for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2     r_1.1
#> a_1   0   0   0   0   0   0 2.6071963 2.988387 1.579880 2.988387 2.3986714
#> a_2   0   0   1   1   0   0 1.1794354 1.303977 3.797941 1.303977 2.1920779
#> a_3   0   0   1   1   0   0 2.3344993 2.672038 2.185342 2.672038 0.5555572
#> a_4   0   0   0   1   0   0 0.9653243 1.099360 1.454572 1.099360 2.2482798
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.4080854 0.5034366 0.5222218
#> a_2 2.6991189 2.8354101 2.9550895
#> a_3 0.6389286 2.0968283 2.2197276
#> a_4 2.3653854 2.6904837 2.8478280
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.6071963 2.988387 1.579880 2.988387 2.3986714
#> a_2   0   1   1   1   0   0 1.1794354 1.303977 3.797941 1.303977 2.1920779
#> a_3   0   0   1   1   0   0 2.3344993 2.672038 2.185342 2.672038 0.5555572
#> a_4   0   0   0   1   0   0 0.9653243 1.099360 1.454572 1.099360 2.2482798
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.4080854 0.5034366 0.5222218
#> a_2 2.6991189 2.8354101 2.9550895
#> a_3 0.6389286 2.0968283 2.2197276
#> a_4 2.3653854 2.6904837 2.8478280
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $moststruct
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.6071963 2.988387 1.579880 2.988387 2.3986714
#> a_2   0   1   1   1   0   0 1.1794354 1.303977 3.797941 1.303977 2.1920779
#> a_3   0   0   1   1   0   0 2.3344993 2.672038 2.185342 2.672038 0.5555572
#> a_4   0   0   0   1   0   0 0.9653243 1.099360 1.454572 1.099360 2.2482798
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.4080854 0.5034366 0.5222218
#> a_2 2.6991189 2.8354101 2.9550895
#> a_3 0.6389286 2.0968283 2.2197276
#> a_4 2.3653854 2.6904837 2.8478280
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $minimaxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2     r_1.1
#> a_1   0   0   0   0   0   0 2.6071963 2.988387 1.579880 2.988387 2.3986714
#> a_2   0   0   1   1   0   0 1.1794354 1.303977 3.797941 1.303977 2.1920779
#> a_3   0   0   1   1   0   0 2.3344993 2.672038 2.185342 2.672038 0.5555572
#> a_4   0   0   0   1   0   0 0.9653243 1.099360 1.454572 1.099360 2.2482798
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.4080854 0.5034366 0.5222218
#> a_2 2.6991189 2.8354101 2.9550895
#> a_3 0.6389286 2.0968283 2.2197276
#> a_4 2.3653854 2.6904837 2.8478280
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maximinstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   0   0   0   0   0 2.6071963 2.988387 1.579880 2.988387 2.3986714
#> a_2   0   1   1   1   0   0 1.1794354 1.303977 3.797941 1.303977 2.1920779
#> a_3   0   0   1   1   0   0 2.3344993 2.672038 2.185342 2.672038 0.5555572
#> a_4   0   0   0   1   0   0 0.9653243 1.099360 1.454572 1.099360 2.2482798
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.4080854 0.5034366 0.5222218
#> a_2 2.6991189 2.8354101 2.9550895
#> a_3 0.6389286 2.0968283 2.2197276
#> a_4 2.3653854 2.6904837 2.8478280
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> attr(,"class")
#> [1] "Set_PDM_matrix"

# Generation of PDM list for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=FALSE,lst=TRUE)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1       t_2       c_1       c_2    r_1.1
#> a_1   1   1   0   0   0   0 2.6559571 2.9503724 3.2279502 2.9503724 2.749853
#> a_2   0   1   1   0   0   0 0.9059689 0.9518897 2.2949837 0.9518897 2.444158
#> a_3   0   0   1   1   0   0 1.5367680 1.5938120 0.6267872 1.5938120 2.216591
#> a_4   0   0   0   1   0   0 2.1746359 2.6196449 2.8148645 2.6196449 2.439033
#> a_5   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.897776 2.241339 2.253021
#> a_2 2.697584 1.777302 1.996798
#> a_3 2.672655 2.479161 2.561736
#> a_4 2.840756 1.560074 1.677321
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1       t_2       c_1       c_2    r_1.1
#> a_1   1   1   0   0   0   0 2.6559571 2.9503724 3.2279502 2.9503724 2.749853
#> a_2   0   1   1   0   0   0 0.9059689 0.9518897 2.2949837 0.9518897 2.444158
#> a_3   0   0   1   1   0   0 1.5367680 1.5938120 0.6267872 1.5938120 2.216591
#> a_4   0   0   0   1   0   0 2.1746359 2.6196449 2.8148645 2.6196449 2.439033
#> a_5   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.897776 2.241339 2.253021
#> a_2 2.697584 1.777302 1.996798
#> a_3 2.672655 2.479161 2.561736
#> a_4 2.840756 1.560074 1.677321
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $moststruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1       t_2       c_1       c_2    r_1.1
#> a_1   1   1   0   0   0   0 2.6559571 2.9503724 3.2279502 2.9503724 2.749853
#> a_2   0   1   1   0   0   0 0.9059689 0.9518897 2.2949837 0.9518897 2.444158
#> a_3   0   0   1   1   0   0 1.5367680 1.5938120 0.6267872 1.5938120 2.216591
#> a_4   0   0   0   1   0   0 2.1746359 2.6196449 2.8148645 2.6196449 2.439033
#> a_5   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.897776 2.241339 2.253021
#> a_2 2.697584 1.777302 1.996798
#> a_3 2.672655 2.479161 2.561736
#> a_4 2.840756 1.560074 1.677321
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $minimaxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1       t_2       c_1       c_2    r_1.1
#> a_1   1   1   0   0   0   0 2.6559571 2.9503724 3.2279502 2.9503724 2.749853
#> a_2   0   1   1   0   0   0 0.9059689 0.9518897 2.2949837 0.9518897 2.444158
#> a_3   0   0   1   1   0   0 1.5367680 1.5938120 0.6267872 1.5938120 2.216591
#> a_4   0   0   0   1   0   0 2.1746359 2.6196449 2.8148645 2.6196449 2.439033
#> a_5   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.897776 2.241339 2.253021
#> a_2 2.697584 1.777302 1.996798
#> a_3 2.672655 2.479161 2.561736
#> a_4 2.840756 1.560074 1.677321
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maximinstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1       t_2       c_1       c_2    r_1.1
#> a_1   1   1   0   0   0   0 2.6559571 2.9503724 3.2279502 2.9503724 2.749853
#> a_2   0   1   1   0   0   0 0.9059689 0.9518897 2.2949837 0.9518897 2.444158
#> a_3   0   0   1   1   0   0 1.5367680 1.5938120 0.6267872 1.5938120 2.216591
#> a_4   0   0   0   1   0   0 2.1746359 2.6196449 2.8148645 2.6196449 2.439033
#> a_5   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>        r_2.1    r_1.2    r_2.2
#> a_1 2.897776 2.241339 2.253021
#> a_2 2.697584 1.777302 1.996798
#> a_3 2.672655 2.479161 2.561736
#> a_4 2.840756 1.560074 1.677321
#> a_5 0.000000 0.000000 0.000000
#> a_6 0.000000 0.000000 0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> attr(,"class")
#> [1] "Set_PDM_list"