Function to calculate minimal/maximal/most likely project structures.

Specify minimal, maximal, maximin, minimax, and most likely/most desired project structures.

get.structures(x,type=c("min","max","minimax","maximin","most"))

Arguments

x

N by M PDM matrix or a PDM list.

type

type of specified project structure. "min" is a minimal strucuture, where the time, cost demands, and as well as the quality parameters are minimal; "max" where the time, cost demands, and as well as the quality parameters are maximal; "minimax", where the cost and resource demands are minimal; "maximix", where the cost and resource demands are maxnimal; "most", where the total project score is maximal.

Value

PDMs

set of PDM_matrix or a set of PDM_list (depending on the input type)

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Generation of PDM matrix for flexible project planning MFPP package.

# Define number of modes, flexibility factor and connectivity factor of a project scenerio.
N=4;ff=0.05;cf=0

# Define maximum value of time domain, Cost domain and Resourcces domain of a project scenerio.
mTD=3;mCD=4;mRD=3

# Define number of modes, number of resources,
# number of possible extra tasks, scale and quality domain of a project scenerio.

w=2;nR=2;nW=2
scale=1.4  #(default value)

# Generation of PDM matrix for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.322709 2.641987 1.962372 2.641987 1.8501256
#> a_2   0   1   1   1   0   0 2.180953 2.282715 2.154933 2.282715 0.9753522
#> a_3   0   0   1   1   0   0 2.458183 2.473048 2.818827 2.473048 1.8627066
#> a_4   0   0   0   1   0   0 2.508330 2.536076 1.456664 2.536076 1.8328630
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 2.300128 2.1175003 2.4974447
#> a_2 1.102856 0.8778707 0.8818225
#> a_3 2.003169 2.3405565 2.7494271
#> a_4 2.057720 2.0590506 2.4581996
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.322709 2.641987 1.962372 2.641987 1.8501256
#> a_2   0   1   1   1   0   0 2.180953 2.282715 2.154933 2.282715 0.9753522
#> a_3   0   0   1   1   0   0 2.458183 2.473048 2.818827 2.473048 1.8627066
#> a_4   0   0   0   1   0   0 2.508330 2.536076 1.456664 2.536076 1.8328630
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 2.300128 2.1175003 2.4974447
#> a_2 1.102856 0.8778707 0.8818225
#> a_3 2.003169 2.3405565 2.7494271
#> a_4 2.057720 2.0590506 2.4581996
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $moststruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.322709 2.641987 1.962372 2.641987 1.8501256
#> a_2   0   1   1   1   0   0 2.180953 2.282715 2.154933 2.282715 0.9753522
#> a_3   0   0   1   1   0   0 2.458183 2.473048 2.818827 2.473048 1.8627066
#> a_4   0   0   0   1   0   0 2.508330 2.536076 1.456664 2.536076 1.8328630
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 2.300128 2.1175003 2.4974447
#> a_2 1.102856 0.8778707 0.8818225
#> a_3 2.003169 2.3405565 2.7494271
#> a_4 2.057720 2.0590506 2.4581996
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $minimaxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.322709 2.641987 1.962372 2.641987 1.8501256
#> a_2   0   1   1   1   0   0 2.180953 2.282715 2.154933 2.282715 0.9753522
#> a_3   0   0   1   1   0   0 2.458183 2.473048 2.818827 2.473048 1.8627066
#> a_4   0   0   0   1   0   0 2.508330 2.536076 1.456664 2.536076 1.8328630
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 2.300128 2.1175003 2.4974447
#> a_2 1.102856 0.8778707 0.8818225
#> a_3 2.003169 2.3405565 2.7494271
#> a_4 2.057720 2.0590506 2.4581996
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maximinstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.322709 2.641987 1.962372 2.641987 1.8501256
#> a_2   0   1   1   1   0   0 2.180953 2.282715 2.154933 2.282715 0.9753522
#> a_3   0   0   1   1   0   0 2.458183 2.473048 2.818827 2.473048 1.8627066
#> a_4   0   0   0   1   0   0 2.508330 2.536076 1.456664 2.536076 1.8328630
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 2.300128 2.1175003 2.4974447
#> a_2 1.102856 0.8778707 0.8818225
#> a_3 2.003169 2.3405565 2.7494271
#> a_4 2.057720 2.0590506 2.4581996
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> attr(,"class")
#> [1] "Set_PDM_matrix"

# Generation of PDM list for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=FALSE,lst=TRUE)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.414831 2.515532 1.095182 2.515532 2.1076850
#> a_2   0   1   0   1   0   0 2.312140 2.360251 2.390757 2.360251 2.7746440
#> a_3   0   0   1   0   0   0 1.819430 2.082948 2.072909 2.082948 1.3512096
#> a_4   0   0   0   1   0   0 1.008120 1.238691 2.412375 1.238691 0.6162157
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.3083064 2.2503996 2.8076667
#> a_2 2.9057874 2.4954426 2.9379592
#> a_3 1.4845910 0.1208832 0.1445074
#> a_4 0.6204215 2.5198654 2.6577500
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.414831 2.515532 1.095182 2.515532 2.1076850
#> a_2   0   1   0   1   0   0 2.312140 2.360251 2.390757 2.360251 2.7746440
#> a_3   0   0   1   1   0   0 1.819430 2.082948 2.072909 2.082948 1.3512096
#> a_4   0   0   0   1   0   0 1.008120 1.238691 2.412375 1.238691 0.6162157
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.3083064 2.2503996 2.8076667
#> a_2 2.9057874 2.4954426 2.9379592
#> a_3 1.4845910 0.1208832 0.1445074
#> a_4 0.6204215 2.5198654 2.6577500
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $moststruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.414831 2.515532 1.095182 2.515532 2.1076850
#> a_2   0   1   0   1   0   0 2.312140 2.360251 2.390757 2.360251 2.7746440
#> a_3   0   0   1   1   0   0 1.819430 2.082948 2.072909 2.082948 1.3512096
#> a_4   0   0   0   1   0   0 1.008120 1.238691 2.412375 1.238691 0.6162157
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.3083064 2.2503996 2.8076667
#> a_2 2.9057874 2.4954426 2.9379592
#> a_3 1.4845910 0.1208832 0.1445074
#> a_4 0.6204215 2.5198654 2.6577500
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $minimaxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.414831 2.515532 1.095182 2.515532 2.1076850
#> a_2   0   1   0   1   0   0 2.312140 2.360251 2.390757 2.360251 2.7746440
#> a_3   0   0   1   1   0   0 1.819430 2.082948 2.072909 2.082948 1.3512096
#> a_4   0   0   0   1   0   0 1.008120 1.238691 2.412375 1.238691 0.6162157
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.3083064 2.2503996 2.8076667
#> a_2 2.9057874 2.4954426 2.9379592
#> a_3 1.4845910 0.1208832 0.1445074
#> a_4 0.6204215 2.5198654 2.6577500
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maximinstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2      c_1      c_2     r_1.1
#> a_1   1   1   0   0   0   0 2.414831 2.515532 1.095182 2.515532 2.1076850
#> a_2   0   1   0   1   0   0 2.312140 2.360251 2.390757 2.360251 2.7746440
#> a_3   0   0   1   0   0   0 1.819430 2.082948 2.072909 2.082948 1.3512096
#> a_4   0   0   0   1   0   0 1.008120 1.238691 2.412375 1.238691 0.6162157
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.000000 0.000000 0.0000000
#>         r_2.1     r_1.2     r_2.2
#> a_1 2.3083064 2.2503996 2.8076667
#> a_2 2.9057874 2.4954426 2.9379592
#> a_3 1.4845910 0.1208832 0.1445074
#> a_4 0.6204215 2.5198654 2.6577500
#> a_5 0.0000000 0.0000000 0.0000000
#> a_6 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> attr(,"class")
#> [1] "Set_PDM_list"