Function to calculate minimal/maximal/most likely project structures.

Specify minimal, maximal, maximin, minimax, and most likely/most desired project structures.

get.structures(x,type=c("min","max","minimax","maximin","most"))

Arguments

x

N by M PDM matrix or a PDM list.

type

type of specified project structure. "min" is a minimal strucuture, where the time, cost demands, and as well as the quality parameters are minimal; "max" where the time, cost demands, and as well as the quality parameters are maximal; "minimax", where the cost and resource demands are minimal; "maximix", where the cost and resource demands are maxnimal; "most", where the total project score is maximal.

Value

PDMs

set of PDM_matrix or a set of PDM_list (depending on the input type)

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Generation of PDM matrix for flexible project planning MFPP package.

# Define number of modes, flexibility factor and connectivity factor of a project scenerio.
N=4;ff=0.05;cf=0

# Define maximum value of time domain, Cost domain and Resourcces domain of a project scenerio.
mTD=3;mCD=4;mRD=3

# Define number of modes, number of resources,
# number of possible extra tasks, scale and quality domain of a project scenerio.

w=2;nR=2;nW=2
scale=1.4  #(default value)

# Generation of PDM matrix for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2       c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.967782 1.978390 0.2994121 1.978390 1.433711
#> a_2   0   1   0   1   0   0 1.744913 1.934271 1.6476249 1.934271 2.429572
#> a_3   0   0   0   0   0   0 1.059772 1.215906 1.0815559 1.215906 1.263336
#> a_4   0   0   0   0   0   0 2.690854 2.874048 3.7180723 2.874048 2.341781
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.603267 2.1974535 2.4619021
#> a_2 2.742596 1.8087567 1.9764286
#> a_3 1.363247 1.6463970 1.8961611
#> a_4 2.911649 0.7705028 0.8413075
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2       c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.967782 1.978390 0.2994121 1.978390 1.433711
#> a_2   0   1   0   1   0   0 1.744913 1.934271 1.6476249 1.934271 2.429572
#> a_3   0   0   1   1   0   0 1.059772 1.215906 1.0815559 1.215906 1.263336
#> a_4   0   0   0   1   0   0 2.690854 2.874048 3.7180723 2.874048 2.341781
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.603267 2.1974535 2.4619021
#> a_2 2.742596 1.8087567 1.9764286
#> a_3 1.363247 1.6463970 1.8961611
#> a_4 2.911649 0.7705028 0.8413075
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $moststruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2       c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.967782 1.978390 0.2994121 1.978390 1.433711
#> a_2   0   1   0   1   0   0 1.744913 1.934271 1.6476249 1.934271 2.429572
#> a_3   0   0   1   1   0   0 1.059772 1.215906 1.0815559 1.215906 1.263336
#> a_4   0   0   0   1   0   0 2.690854 2.874048 3.7180723 2.874048 2.341781
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.603267 2.1974535 2.4619021
#> a_2 2.742596 1.8087567 1.9764286
#> a_3 1.363247 1.6463970 1.8961611
#> a_4 2.911649 0.7705028 0.8413075
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $minimaxstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2       c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.967782 1.978390 0.2994121 1.978390 1.433711
#> a_2   0   1   0   1   0   0 1.744913 1.934271 1.6476249 1.934271 2.429572
#> a_3   0   0   0   0   0   0 1.059772 1.215906 1.0815559 1.215906 1.263336
#> a_4   0   0   0   0   0   0 2.690854 2.874048 3.7180723 2.874048 2.341781
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.603267 2.1974535 2.4619021
#> a_2 2.742596 1.8087567 1.9764286
#> a_3 1.363247 1.6463970 1.8961611
#> a_4 2.911649 0.7705028 0.8413075
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $maximinstruct
#>     a_1 a_2 a_3 a_4 a_5 a_6      t_1      t_2       c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.967782 1.978390 0.2994121 1.978390 1.433711
#> a_2   0   1   0   1   0   0 1.744913 1.934271 1.6476249 1.934271 2.429572
#> a_3   0   0   1   0   0   0 1.059772 1.215906 1.0815559 1.215906 1.263336
#> a_4   0   0   0   1   0   0 2.690854 2.874048 3.7180723 2.874048 2.341781
#> a_5   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.000000 0.000000 0.0000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.603267 2.1974535 2.4619021
#> a_2 2.742596 1.8087567 1.9764286
#> a_3 1.363247 1.6463970 1.8961611
#> a_4 2.911649 0.7705028 0.8413075
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> attr(,"class")
#> [1] "Set_PDM_matrix"

# Generation of PDM list for TCTP a project scenario using MFPP package.
PDM<-generatepdm(N,ff,cf,mTD,mCD,mRD,w,nR,nW,scale=1.4,QD=FALSE,lst=TRUE)

# Get main structures.

PDMs<-get.structures(PDM)
PDMs
#> $minstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.4039211 1.587546 3.061663 1.587546 1.743293
#> a_2   0   0   0   0   0   0 0.1262128 0.152930 2.440633 0.152930 2.667325
#> a_3   0   0   1   0   0   0 0.9118529 1.071362 3.474453 1.071362 1.896800
#> a_4   0   0   0   1   0   0 1.7340839 1.806671 2.715694 1.806671 2.283223
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.994056 1.4980821 1.7633467
#> a_2 2.853196 0.8284898 0.8730229
#> a_3 2.152575 2.2544550 2.7945220
#> a_4 2.575660 2.0476160 2.2078800
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.4039211 1.587546 3.061663 1.587546 1.743293
#> a_2   0   1   0   0   0   0 0.1262128 0.152930 2.440633 0.152930 2.667325
#> a_3   0   0   1   0   0   0 0.9118529 1.071362 3.474453 1.071362 1.896800
#> a_4   0   0   0   1   0   0 1.7340839 1.806671 2.715694 1.806671 2.283223
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.994056 1.4980821 1.7633467
#> a_2 2.853196 0.8284898 0.8730229
#> a_3 2.152575 2.2544550 2.7945220
#> a_4 2.575660 2.0476160 2.2078800
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $moststruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.4039211 1.587546 3.061663 1.587546 1.743293
#> a_2   0   1   0   0   0   0 0.1262128 0.152930 2.440633 0.152930 2.667325
#> a_3   0   0   1   0   0   0 0.9118529 1.071362 3.474453 1.071362 1.896800
#> a_4   0   0   0   1   0   0 1.7340839 1.806671 2.715694 1.806671 2.283223
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.994056 1.4980821 1.7633467
#> a_2 2.853196 0.8284898 0.8730229
#> a_3 2.152575 2.2544550 2.7945220
#> a_4 2.575660 2.0476160 2.2078800
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $minimaxstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.4039211 1.587546 3.061663 1.587546 1.743293
#> a_2   0   0   0   0   0   0 0.1262128 0.152930 2.440633 0.152930 2.667325
#> a_3   0   0   1   0   0   0 0.9118529 1.071362 3.474453 1.071362 1.896800
#> a_4   0   0   0   1   0   0 1.7340839 1.806671 2.715694 1.806671 2.283223
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.994056 1.4980821 1.7633467
#> a_2 2.853196 0.8284898 0.8730229
#> a_3 2.152575 2.2544550 2.7945220
#> a_4 2.575660 2.0476160 2.2078800
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> $maximinstruct
#> $PDM
#>     a_1 a_2 a_3 a_4 a_5 a_6       t_1      t_2      c_1      c_2    r_1.1
#> a_1   1   1   1   0   0   0 1.4039211 1.587546 3.061663 1.587546 1.743293
#> a_2   0   1   0   0   0   0 0.1262128 0.152930 2.440633 0.152930 2.667325
#> a_3   0   0   1   0   0   0 0.9118529 1.071362 3.474453 1.071362 1.896800
#> a_4   0   0   0   1   0   0 1.7340839 1.806671 2.715694 1.806671 2.283223
#> a_5   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#> a_6   0   0   0   0   0   0 0.0000000 0.000000 0.000000 0.000000 0.000000
#>        r_2.1     r_1.2     r_2.2
#> a_1 1.994056 1.4980821 1.7633467
#> a_2 2.853196 0.8284898 0.8730229
#> a_3 2.152575 2.2544550 2.7945220
#> a_4 2.575660 2.0476160 2.2078800
#> a_5 0.000000 0.0000000 0.0000000
#> a_6 0.000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"
#> 
#> attr(,"class")
#> [1] "Set_PDM_list"