Function to simulate estimation uncertainty.

Simulating the effects of the estimatio uncertainty.

phase1(x,a=-0.1,b=0.30,pdftype="uniform")

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

a

optimistic duration, the negative percent deviation from the actual duration (a negative value). The default value is a=-0.1, which means if the actual duration is 1, the minimal possible duration is 0.9.

b

pessimistic duration, the positive percent deviation from the actual duration (a positive value). The default value is b=0.3, which means if the actual duration is 1, the maximal possible duration is 1.3.

pdftype

the pdftype is either 'uniform' or 'beta' (the default value is 'uniform'), which means the generated distribution either follows uniform or a beta distribution betweem interval [a,b]

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13      a_14 a_15
#> a_1    1   1   1   0   1   0   0   0   0    0    0    0    0 0.0000000    0
#> a_2    0   1   1   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_3    0   0   1   1   0   0   0   0   1    0    0    0    0 0.0000000    0
#> a_4    0   0   0   1   1   0   0   0   0    0    0    0    0 0.0000000    0
#> a_5    0   0   0   0   1   1   0   1   0    0    0    0    0 0.0000000    1
#> a_6    0   0   0   0   0   1   1   0   0    0    0    0    0 0.0000000    0
#> a_7    0   0   0   0   0   0   1   1   0    0    0    0    0 0.0000000    0
#> a_8    0   0   0   0   0   0   0   1   0    0    0    0    0 0.0000000    0
#> a_9    0   0   0   0   0   0   0   0   1    0    0    0    0 0.0000000    0
#> a_10   0   0   0   0   0   0   0   0   0    1    0    0    0 0.0000000    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    1    0 0.0000000    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    1 0.0000000    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1 1.0000000    1
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0 0.5608771    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    1
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_24   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_25   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_26   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_27   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_28   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_29   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_30   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_31   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_32   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#>      a_16 a_17 a_18 a_19 a_20 a_21 a_22      a_23 a_24 a_25 a_26 a_27 a_28 a_29
#> a_1     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_5     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_6     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_11    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_12    0    1    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_15    0    1    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_16    1    1    0    0    1    0    0 0.0000000    0    0    0    0    0    0
#> a_17    0    1    0    0    0    0    0 1.0000000    0    0    0    0    0    0
#> a_18    0    0    1    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_19    0    0    0    1    0    0    0 0.0000000    0    0    0    0    0    0
#> a_20    0    0    0    0    1    0    1 1.0000000    0    0    0    0    0    0
#> a_21    0    0    0    0    0    1    0 0.6300737    0    0    0    0    0    0
#> a_22    0    0    0    0    0    0    1 0.0000000    0    0    0    0    0    0
#> a_23    0    0    0    0    0    0    0 1.0000000    0    0    0    0    0    0
#> a_24    0    0    0    0    0    0    0 0.0000000    1    0    0    0    0    0
#> a_25    0    0    0    0    0    0    0 0.0000000    0    1    0    0    0    0
#> a_26    0    0    0    0    0    0    0 0.0000000    0    0    1    0    0    0
#> a_27    0    0    0    0    0    0    0 0.0000000    0    0    0    1    0    0
#> a_28    0    0    0    0    0    0    0 0.0000000    0    0    0    0    1    0
#> a_29    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    1
#> a_30    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_31    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_32    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#>           a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1  0.0000000    0    0 19.451970 19.732839 24.210707 24.565609 0.1079700
#> a_2  0.0000000    0    0 11.984019 13.268788 25.318327 28.727520 0.4390708
#> a_3  0.0000000    0    0  6.503091  6.522573 21.720525 26.759489 0.7536477
#> a_4  0.0000000    0    0  9.538294  9.938639 25.109528 26.842028 0.3601677
#> a_5  0.0000000    0    0 14.035488 14.048876  7.524023  7.761622 0.7329649
#> a_6  0.0000000    0    0 16.186470 19.120748 12.894920 14.068824 0.5511407
#> a_7  0.0000000    0    0  5.075112  5.673243 20.166400 20.947034 0.7908588
#> a_8  0.0000000    0    0 12.380647 13.190632 20.771311 21.018764 0.7486008
#> a_9  0.0000000    0    0 15.834809 17.919217 19.311531 23.149941 0.8619554
#> a_10 0.0000000    0    0 14.727965 17.659767 10.101456 11.712561 0.2078817
#> a_11 0.0000000    0    0 13.077511 14.725425 16.794343 19.500528 0.5817767
#> a_12 0.0000000    0    0 16.159327 18.787990 10.298668 10.644173 0.4052653
#> a_13 0.0000000    0    0 16.510984 18.308132  5.687504  6.943255 0.6872698
#> a_14 0.0000000    0    0 15.060364 16.411476 23.080764 26.169502 0.7494951
#> a_15 0.0000000    0    0  8.486682  9.793860 21.085573 25.405599 0.6260970
#> a_16 0.0000000    0    0 17.378737 19.237990 17.437956 21.013880 0.5268720
#> a_17 0.0000000    0    0 12.929059 14.965752 21.418380 26.501508 0.8098453
#> a_18 0.6463093    0    0 14.377032 17.624179 19.768495 24.487091 0.6075537
#> a_19 0.0000000    0    0  8.365947  8.737247 11.533052 11.938338 0.7615413
#> a_20 0.0000000    0    0 11.423162 11.820048 15.265866 17.323716 0.7132936
#> a_21 0.0000000    0    0  9.780740 10.097973 16.742029 20.080425 0.5637616
#> a_22 0.0000000    0    0  5.096220  5.680859 26.552382 28.548891 0.7162401
#> a_23 0.0000000    0    0 18.250297 19.503409 22.602027 24.916400 0.6047823
#> a_24 0.0000000    0    0 14.981410 18.000946 22.097193 26.160092 0.5683737
#> a_25 0.0000000    0    0 10.406906 11.453202 13.375784 13.613626 0.8199386
#> a_26 0.0000000    0    0 16.322983 19.981212 22.067586 25.830942 0.5623763
#> a_27 0.0000000    0    0 15.972680 19.356662 21.393726 25.700329 0.7501718
#> a_28 0.0000000    0    0 13.369046 15.918640 13.696711 16.931654 0.6495740
#> a_29 1.0000000    0    0 11.514493 12.476673 18.935601 19.642817 0.3644508
#> a_30 1.0000000    0    0 15.802304 19.395201 16.466595 16.874751 0.8129073
#> a_31 0.0000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32 0.0000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1      r_2.1     r_1.2     r_2.2
#> a_1  0.1316775 15.377856 18.6668316 16.733567 16.865290
#> a_2  0.5120590 15.577600 18.5778266 16.804780 19.552446
#> a_3  0.8409793 15.157581 17.7152933 15.089920 17.066702
#> a_4  0.3637204 13.840084 16.1912745 13.746393 17.143720
#> a_5  0.8485438 15.939456 19.0560818  7.460760  7.976366
#> a_6  0.5885333 16.303734 18.8515081  3.502632  3.921777
#> a_7  0.7995329 18.163695 18.8930664 13.863299 16.748249
#> a_8  0.7730184 18.741973 19.3314495 16.331749 18.347278
#> a_9  0.8818430  2.551904  2.5519560  8.449601  9.757946
#> a_10 0.2315428 11.878159 12.6989955  5.914368  5.953518
#> a_11 0.6721677  1.636396  1.7949998 12.477868 15.441920
#> a_12 0.4795022 17.715187 19.1193619 15.935171 16.253525
#> a_13 0.7691840  6.431350  7.5594113 16.410184 19.110124
#> a_14 0.7645804 10.318443 12.2015027 10.509328 11.255758
#> a_15 0.6496711 15.826067 18.3851462 11.927158 12.331605
#> a_16 0.6393813 11.109925 12.5620386 14.798304 17.493529
#> a_17 0.8717740 13.476069 14.5496679  9.430317  9.928526
#> a_18 0.6227924 16.841561 16.9017725 12.726605 14.241107
#> a_19 0.7896962 12.650479 13.9154198  3.267269  3.497518
#> a_20 0.8133253 13.167181 14.0893156 12.603056 14.021544
#> a_21 0.6011813 14.098135 15.2182329 16.674817 17.336548
#> a_22 0.7206234 14.676951 15.8068405  9.295914  9.537268
#> a_23 0.7332705 10.294989 10.9898450 11.020173 12.888695
#> a_24 0.5944095  9.714647 11.2146887 13.623282 16.503304
#> a_25 0.9046314 13.892757 16.2012152 14.639706 16.731647
#> a_26 0.6945366 10.753137 12.9251243 19.277416 19.987215
#> a_27 0.7747052 11.559832 13.8705816 10.886686 11.871822
#> a_28 0.7689992 15.325306 17.0845126 17.406762 19.796541
#> a_29 0.4464060  0.915927  0.9482604 15.396988 15.911811
#> a_30 0.8554823 15.928969 17.3304136 17.160157 19.602449
#> a_31 0.0000000  0.000000  0.0000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.0000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define negative percentage deviation from actual duration.
a=-0.20

# Define negative percentage deviation from actual duration.
b=0.40

# Simulate the effects of estimation uncertainty of a project scenario using MFPP package
# with default values.
 Res<-phase1(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13      a_14 a_15
#> a_1    1   1   1   0   1   0   0   0   0    0    0    0    0 0.0000000    0
#> a_2    0   1   1   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_3    0   0   1   1   0   0   0   0   1    0    0    0    0 0.0000000    0
#> a_4    0   0   0   1   1   0   0   0   0    0    0    0    0 0.0000000    0
#> a_5    0   0   0   0   1   1   0   1   0    0    0    0    0 0.0000000    1
#> a_6    0   0   0   0   0   1   1   0   0    0    0    0    0 0.0000000    0
#> a_7    0   0   0   0   0   0   1   1   0    0    0    0    0 0.0000000    0
#> a_8    0   0   0   0   0   0   0   1   0    0    0    0    0 0.0000000    0
#> a_9    0   0   0   0   0   0   0   0   1    0    0    0    0 0.0000000    0
#> a_10   0   0   0   0   0   0   0   0   0    1    0    0    0 0.0000000    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    1    0 0.0000000    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    1 0.0000000    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1 1.0000000    1
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0 0.5608771    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    1
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_24   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_25   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_26   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_27   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_28   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_29   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_30   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_31   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_32   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#>      a_16 a_17 a_18 a_19 a_20 a_21 a_22      a_23 a_24 a_25 a_26 a_27 a_28 a_29
#> a_1     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_5     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_6     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_11    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_12    0    1    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_15    0    1    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_16    1    1    0    0    1    0    0 0.0000000    0    0    0    0    0    0
#> a_17    0    1    0    0    0    0    0 1.0000000    0    0    0    0    0    0
#> a_18    0    0    1    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_19    0    0    0    1    0    0    0 0.0000000    0    0    0    0    0    0
#> a_20    0    0    0    0    1    0    1 1.0000000    0    0    0    0    0    0
#> a_21    0    0    0    0    0    1    0 0.6300737    0    0    0    0    0    0
#> a_22    0    0    0    0    0    0    1 0.0000000    0    0    0    0    0    0
#> a_23    0    0    0    0    0    0    0 1.0000000    0    0    0    0    0    0
#> a_24    0    0    0    0    0    0    0 0.0000000    1    0    0    0    0    0
#> a_25    0    0    0    0    0    0    0 0.0000000    0    1    0    0    0    0
#> a_26    0    0    0    0    0    0    0 0.0000000    0    0    1    0    0    0
#> a_27    0    0    0    0    0    0    0 0.0000000    0    0    0    1    0    0
#> a_28    0    0    0    0    0    0    0 0.0000000    0    0    0    0    1    0
#> a_29    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    1
#> a_30    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_31    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_32    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#>           a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1  0.0000000    0    0 20.139429 19.766766 23.527511 28.589364 0.1094311
#> a_2  0.0000000    0    0 14.022868 14.701689 23.969701 31.121429 0.5683870
#> a_3  0.0000000    0    0  8.086662  6.726055 21.938092 25.142380 0.8006643
#> a_4  0.0000000    0    0 12.315425 10.783991 28.762815 28.937289 0.3583518
#> a_5  0.0000000    0    0 17.035919 15.232274  9.075755  8.512095 0.6718737
#> a_6  0.0000000    0    0 18.419235 21.954371 15.010996 14.644436 0.5758921
#> a_7  0.0000000    0    0  6.594794  6.327410 23.009535 21.128946 0.9287700
#> a_8  0.0000000    0    0 12.708409 15.651365 21.414418 26.615837 0.9179128
#> a_9  0.0000000    0    0 19.129656 22.063377 17.799577 21.771148 0.8222458
#> a_10 0.0000000    0    0 18.403396 18.839491 11.718355 14.795306 0.2492076
#> a_11 0.0000000    0    0 15.789624 15.394025 19.525742 20.764547 0.7227299
#> a_12 0.0000000    0    0 19.302849 19.897815 10.749160 10.235306 0.5117510
#> a_13 0.0000000    0    0 20.515953 18.823874  6.277708  7.746781 0.8634534
#> a_14 0.0000000    0    0 17.569276 19.502257 27.028617 23.701949 0.7379734
#> a_15 0.0000000    0    0 10.933432 12.121553 26.806487 28.105139 0.7542018
#> a_16 0.0000000    0    0 22.376056 21.475412 19.965843 19.170184 0.6148958
#> a_17 0.0000000    0    0 12.838390 17.307220 23.410609 33.244829 0.9369654
#> a_18 0.6463093    0    0 17.620131 19.906853 19.602383 24.881631 0.5878850
#> a_19 0.0000000    0    0  9.313833 10.901175 13.852344 12.519241 0.8850994
#> a_20 0.0000000    0    0 12.908977 14.555001 19.061082 21.181985 0.8402992
#> a_21 0.0000000    0    0  9.289138  9.537158 19.929574 21.524381 0.5105258
#> a_22 0.0000000    0    0  5.245687  5.718109 32.095407 36.100727 0.6671267
#> a_23 0.0000000    0    0 21.615516 24.491676 20.351289 28.831457 0.7164059
#> a_24 0.0000000    0    0 17.558272 16.215091 25.396796 24.887372 0.5999795
#> a_25 0.0000000    0    0  9.682981 13.107864 17.238021 15.132410 0.8728165
#> a_26 0.0000000    0    0 16.821410 22.508609 21.056686 25.393283 0.5254138
#> a_27 0.0000000    0    0 16.000244 18.211383 26.500391 23.852563 0.7452311
#> a_28 0.0000000    0    0 13.438327 16.922137 14.023138 16.300637 0.6769738
#> a_29 1.0000000    0    0 11.437802 11.465786 20.176593 23.889330 0.4562900
#> a_30 1.0000000    0    0 17.117184 24.283149 19.952178 19.402165 0.9224140
#> a_31 0.0000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32 0.0000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1      r_2.1     r_1.2     r_2.2
#> a_1  0.1560524 13.901358 18.3931085 17.641727 16.696789
#> a_2  0.5613695 14.049166 17.1348502 17.570631 22.231923
#> a_3  0.9612720 14.453822 21.5026651 16.695586 15.654144
#> a_4  0.4375859 13.344484 16.1652408 17.144617 21.228588
#> a_5  1.0000000 15.197767 24.7603203  7.120022  7.828897
#> a_6  0.7128719 20.903586 21.2800235  3.900575  5.069044
#> a_7  1.0000000 18.292307 20.3285015 13.106932 19.999587
#> a_8  0.7597497 22.616499 23.2802382 20.579127 21.195849
#> a_9  0.8027676  2.899143  3.2295231  9.758779 10.286497
#> a_10 0.2840358 13.450406 16.4421323  7.682903  7.080732
#> a_11 0.8495788  1.565154  1.8592943 15.570534 18.482080
#> a_12 0.5063133 17.085942 19.9581012 15.261217 18.840467
#> a_13 0.8653623  5.810989  7.8849410 16.451164 20.684495
#> a_14 0.9452213 10.006069 15.6045792 13.285985 10.624634
#> a_15 0.6830960 20.202284 17.5859862 13.756058 14.986380
#> a_16 0.7250348 13.481458 12.8950375 13.825853 21.610982
#> a_17 1.0000000 14.433718 16.4303278 11.680532 10.548464
#> a_18 0.5798046 16.335318 18.3032842 15.164457 13.616474
#> a_19 0.8261980 14.252134 13.6971438  3.631087  3.317532
#> a_20 0.8133793 14.330547 13.6134157 14.645121 16.092859
#> a_21 0.6743051 14.356461 13.9970676 20.589198 17.610777
#> a_22 0.8276122 17.815641 19.4552401 11.418815 11.201043
#> a_23 0.9315036 10.549577 13.6954951 11.124959 11.738511
#> a_24 0.6813728 12.371045 10.7337803 16.949230 18.800046
#> a_25 1.0000000 12.984750 20.0220710 15.499697 20.506145
#> a_26 0.8412259 12.180804 16.4933303 23.621214 18.144903
#> a_27 0.7493771 12.541869 12.6066129 12.027028 13.937420
#> a_28 0.9133842 19.168562 16.3591451 21.699118 21.934443
#> a_29 0.5710892  1.000000  0.8717702 19.644637 20.433606
#> a_30 0.8868123 17.131710 21.0355849 21.587430 21.343158
#> a_31 0.0000000  0.000000  0.0000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.0000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  98.90808
#> Const constraint (Cc):  560.9319
#> Score/scope constraint (Cs):  0.9746099
#> Quailty constraint (Cq):  0.592491
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  110.1238
#> Const constraint (Cc):  672.7171
#> Score/scope constraint (Cs):  0.9820919
#> Quailty constraint (Cq):  0.7040745
#> Resource constraint(s) (CR):
#>          R_1    R_2
#> TPR 240.3163 235.38

# Simulate the estimation uncertainty of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase1(PDM$PDM,a,b,"beta")
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13      a_14 a_15
#> a_1    1   1   1   0   1   0   0   0   0    0    0    0    0 0.0000000    0
#> a_2    0   1   1   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_3    0   0   1   1   0   0   0   0   1    0    0    0    0 0.0000000    0
#> a_4    0   0   0   1   1   0   0   0   0    0    0    0    0 0.0000000    0
#> a_5    0   0   0   0   1   1   0   1   0    0    0    0    0 0.0000000    1
#> a_6    0   0   0   0   0   1   1   0   0    0    0    0    0 0.0000000    0
#> a_7    0   0   0   0   0   0   1   1   0    0    0    0    0 0.0000000    0
#> a_8    0   0   0   0   0   0   0   1   0    0    0    0    0 0.0000000    0
#> a_9    0   0   0   0   0   0   0   0   1    0    0    0    0 0.0000000    0
#> a_10   0   0   0   0   0   0   0   0   0    1    0    0    0 0.0000000    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    1    0 0.0000000    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    1 0.0000000    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1 1.0000000    1
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0 0.5608771    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    1
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_24   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_25   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_26   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_27   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_28   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_29   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_30   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_31   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#> a_32   0   0   0   0   0   0   0   0   0    0    0    0    0 0.0000000    0
#>      a_16 a_17 a_18 a_19 a_20 a_21 a_22      a_23 a_24 a_25 a_26 a_27 a_28 a_29
#> a_1     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_5     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_6     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_11    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_12    0    1    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_15    0    1    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_16    1    1    0    0    1    0    0 0.0000000    0    0    0    0    0    0
#> a_17    0    1    0    0    0    0    0 1.0000000    0    0    0    0    0    0
#> a_18    0    0    1    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_19    0    0    0    1    0    0    0 0.0000000    0    0    0    0    0    0
#> a_20    0    0    0    0    1    0    1 1.0000000    0    0    0    0    0    0
#> a_21    0    0    0    0    0    1    0 0.6300737    0    0    0    0    0    0
#> a_22    0    0    0    0    0    0    1 0.0000000    0    0    0    0    0    0
#> a_23    0    0    0    0    0    0    0 1.0000000    0    0    0    0    0    0
#> a_24    0    0    0    0    0    0    0 0.0000000    1    0    0    0    0    0
#> a_25    0    0    0    0    0    0    0 0.0000000    0    1    0    0    0    0
#> a_26    0    0    0    0    0    0    0 0.0000000    0    0    1    0    0    0
#> a_27    0    0    0    0    0    0    0 0.0000000    0    0    0    1    0    0
#> a_28    0    0    0    0    0    0    0 0.0000000    0    0    0    0    1    0
#> a_29    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    1
#> a_30    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_31    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#> a_32    0    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0
#>           a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1  0.0000000    0    0 20.607563 19.358267 28.770788 27.270482 0.1020398
#> a_2  0.0000000    0    0 11.763170 16.408186 32.455811 33.475875 0.4586504
#> a_3  0.0000000    0    0  7.451107  6.480360 20.130942 27.210301 0.7986348
#> a_4  0.0000000    0    0  9.461431 11.254584 33.032786 34.490240 0.4477136
#> a_5  0.0000000    0    0 15.111001 16.025857 10.171654  9.760259 0.9667108
#> a_6  0.0000000    0    0 16.784154 23.636437 12.820557 14.094344 0.5582137
#> a_7  0.0000000    0    0  5.115624  6.308935 24.611681 26.467020 1.0000000
#> a_8  0.0000000    0    0 14.804214 15.833916 22.278351 18.454103 0.9024264
#> a_9  0.0000000    0    0 14.444970 19.557617 19.282129 27.848047 0.7568391
#> a_10 0.0000000    0    0 18.237731 20.687079 12.599478 12.316381 0.2053350
#> a_11 0.0000000    0    0 14.566513 17.970547 20.204662 23.296794 0.6474857
#> a_12 0.0000000    0    0 16.305725 20.511957 13.162895 11.250215 0.4555463
#> a_13 0.0000000    0    0 17.405625 16.427934  5.760477  8.112863 0.8703168
#> a_14 0.0000000    0    0 15.518615 18.447063 27.291660 25.115905 0.8468365
#> a_15 0.0000000    0    0  8.970486  9.282228 20.831574 23.995286 0.7046467
#> a_16 0.0000000    0    0 17.860159 20.641920 18.653651 20.885738 0.6431406
#> a_17 0.0000000    0    0 12.392880 15.280533 23.919079 32.292740 0.8833128
#> a_18 0.6463093    0    0 14.529919 18.777637 22.099099 25.258861 0.6564483
#> a_19 0.0000000    0    0  9.187087 10.056851 11.320421 11.583978 0.8349941
#> a_20 0.0000000    0    0 13.389491 13.985840 14.547233 19.951263 0.7670613
#> a_21 0.0000000    0    0 10.873219 10.440405 18.938705 22.449142 0.6766760
#> a_22 0.0000000    0    0  5.754513  6.776174 27.473271 33.372581 0.9114531
#> a_23 0.0000000    0    0 23.011464 22.587616 23.519986 23.891121 0.7162642
#> a_24 0.0000000    0    0 18.813783 18.261579 22.096535 29.626948 0.5758966
#> a_25 0.0000000    0    0 12.893786 13.847934 16.349890 16.326235 0.8178943
#> a_26 0.0000000    0    0 20.474908 24.438433 24.973164 27.745985 0.6449082
#> a_27 0.0000000    0    0 17.941858 23.043968 20.026465 31.057583 0.7714918
#> a_28 0.0000000    0    0 14.190699 20.293903 13.648147 16.370443 0.6942306
#> a_29 1.0000000    0    0 13.141124 13.280253 19.951418 21.857853 0.3783749
#> a_30 1.0000000    0    0 16.746727 22.922744 20.135051 20.240592 0.9760838
#> a_31 0.0000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32 0.0000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2      r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.1665178 18.4726139 22.395957 19.588180 20.764421
#> a_2  0.6613476 16.3929343 21.128187 20.376930 17.293839
#> a_3  1.0000000 19.7123758 17.094280 19.100203 21.324825
#> a_4  0.3450501 17.3848635 18.887999 14.088684 18.756705
#> a_5  1.0000000 20.5184157 21.480238  8.123785  8.792504
#> a_6  0.6146336 18.8817088 20.864981  3.701474  4.991386
#> a_7  0.8603273 20.1075116 19.755164 17.776355 21.253660
#> a_8  0.9279008 19.5949275 21.215793 17.629702 17.202789
#> a_9  0.9371373  2.7438947  2.899534  9.260344  9.614937
#> a_10 0.2603377 13.1647541 15.482748  6.323733  6.676436
#> a_11 0.7832731  2.0230764  1.987272 13.196372 17.821016
#> a_12 0.5467570 18.8789210 23.087194 19.469305 18.878966
#> a_13 0.8340713  5.7123401  8.414797 17.581349 19.369527
#> a_14 0.8706410 12.4058258 11.646952 12.564339 11.039645
#> a_15 0.8326609 16.5285892 16.499905 14.063364 13.529213
#> a_16 0.7195960 11.2310676 11.357618 15.320622 16.036583
#> a_17 1.0000000 17.3358189 16.729638 11.509279 11.785766
#> a_18 0.6597699 20.3226187 19.642148 12.561231 16.407244
#> a_19 0.9542970 13.1470181 16.193728  3.399659  4.086919
#> a_20 0.9943368 14.0542337 15.204568 13.580313 15.659158
#> a_21 0.6717489 15.2271522 14.715161 22.793814 19.830701
#> a_22 0.7423331 16.5453239 20.941128 10.976630  8.954184
#> a_23 0.8910197 10.9590281 12.925522 12.008598 11.754642
#> a_24 0.7513689  9.8487676 12.710551 14.322907 21.058063
#> a_25 1.0000000 14.7880298 17.436904 15.948988 18.146234
#> a_26 0.6213699 12.1402463 15.755218 19.325142 19.133498
#> a_27 0.9786297 11.8046095 15.882320 12.015453 10.131903
#> a_28 0.8994563 15.0272493 21.679655 16.165782 18.068695
#> a_29 0.5361770  0.9643757  1.000000 17.583628 16.308664
#> a_30 0.9785610 17.0076149 14.209282 21.137238 20.503533
#> a_31 0.0000000  0.0000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.0000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  91.25115
#> Const constraint (Cc):  565.8105
#> Score/scope constraint (Cs):  0.9746099
#> Quailty constraint (Cq):  0.5964812
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  103.4624
#> Const constraint (Cc):  686.2003
#> Score/scope constraint (Cs):  0.9820919
#> Quailty constraint (Cq):  0.7213131
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 238.0863 222.5094