Function to simulate estimation uncertainty.

Simulating the effects of the estimation uncertainty.

phase1(x,a=-0.1,b=0.30,pdftype="uniform")

Details

The changed demands are generated by the interval between [o+a,o+b], where o is the original value. The random generator can follow uniform (=default), or beta distribution.

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

a

optimistic duration, the negative percent deviation from the actual duration (a negative value). The default value is a=-0.1, which means if the actual duration is 1, the minimal possible duration is 0.9.

b

pessimistic duration, the positive percent deviation from the actual duration (a positive value). The default value is b=0.3, which means if the actual duration is 1, the maximal possible duration is 1.3.

pdftype

the pdftype is either 'uniform' or 'beta' (the default value is 'uniform'), which means the generated distribution either follows uniform or a beta distribution betweem interval [a,b]

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7      a_8 a_9      a_10 a_11 a_12 a_13 a_14
#> a_1    1   1   0   0   0   0   0 0.000000   0 1.0000000    0    0    0    0
#> a_2    0   1   1   0   1   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_3    0   0   1   0   0   0   0 0.000000   0 0.0000000    1    0    0    0
#> a_4    0   0   0   1   0   1   0 0.828566   0 0.0000000    0    0    0    0
#> a_5    0   0   0   0   1   0   0 0.000000   0 0.0000000    0    0    1    0
#> a_6    0   0   0   0   0   1   0 0.000000   0 0.0000000    0    0    0    0
#> a_7    0   0   0   0   0   0   1 0.000000   0 0.0000000    0    0    0    0
#> a_8    0   0   0   0   0   0   0 1.000000   0 0.0000000    0    1    0    0
#> a_9    0   0   0   0   0   0   0 0.000000   1 0.9172083    0    1    0    0
#> a_10   0   0   0   0   0   0   0 0.000000   0 1.0000000    0    0    0    0
#> a_11   0   0   0   0   0   0   0 0.000000   0 0.0000000    1    0    1    0
#> a_12   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    1    0    0
#> a_13   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    1    1
#> a_14   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    1
#> a_15   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_16   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_17   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_18   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_19   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_20   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_21   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_22   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_23   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_24   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_25   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_26   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_27   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_28   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_29   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_30   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_31   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_32   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#>           a_15 a_16 a_17      a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27
#> a_1  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_2  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_3  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_4  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_5  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_6  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_7  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_8  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_9  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_10 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_11 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_12 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_13 0.0000000    0    0 0.0000000    0    0    0    0    0    0    1    0    0
#> a_14 0.0000000    1    1 0.0000000    0    0    0    0    0    0    0    0    1
#> a_15 0.7502203    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_16 0.0000000    1    1 0.0000000    0    0    0    0    0    0    0    0    0
#> a_17 0.0000000    0    1 0.0000000    0    0    0    0    0    0    0    0    0
#> a_18 0.0000000    0    0 0.7519418    1    0    0    0    0    0    0    0    0
#> a_19 0.0000000    0    0 0.0000000    1    0    0    1    1    0    0    0    0
#> a_20 0.0000000    0    0 0.0000000    0    1    0    0    0    0    0    0    0
#> a_21 0.0000000    0    0 0.0000000    0    0    1    0    0    0    0    0    0
#> a_22 0.0000000    0    0 0.0000000    0    0    0    1    0    0    0    0    0
#> a_23 0.0000000    0    0 0.0000000    0    0    0    0    1    0    0    0    0
#> a_24 0.0000000    0    0 0.0000000    0    0    0    0    0    1    0    0    0
#> a_25 0.0000000    0    0 0.0000000    0    0    0    0    0    0    1    1    1
#> a_26 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    1    1
#> a_27 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    1
#> a_28 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_29 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_30 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_31 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_32 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0  4.953396  5.312656 16.917284 17.178685 0.2917716
#> a_2     0    0    0    0    0 16.133923 16.246272 25.108960 28.813532 0.3147737
#> a_3     0    0    0    0    0 11.841105 14.055775 15.858383 17.898370 0.8411432
#> a_4     0    0    0    0    0 13.938643 15.875849 10.377052 11.368583 0.8011702
#> a_5     0    0    0    0    0 19.720322 19.827863 19.272096 24.070414 0.5918244
#> a_6     0    0    0    0    0 11.080842 11.284910 25.153281 26.725291 0.8221149
#> a_7     0    0    0    0    0 13.568460 16.331650 12.401253 14.094304 0.4921004
#> a_8     0    0    0    0    0 10.335461 12.303948  5.593909  6.021169 0.5257080
#> a_9     0    0    1    0    0 13.336094 13.874610 12.509837 15.461037 0.8366310
#> a_10    0    0    0    0    0  9.716405 10.809432 21.672720 24.765451 0.5040460
#> a_11    0    0    0    0    0 18.156534 19.801812 19.305346 21.778518 0.5101033
#> a_12    0    0    0    0    0 13.729555 16.397378 12.469315 12.692611 0.7815476
#> a_13    0    0    0    0    0 13.527654 15.677964 14.796630 15.605627 0.1136548
#> a_14    0    0    0    0    0 14.377178 15.381250 19.934451 21.265269 0.9646900
#> a_15    0    0    0    0    0 13.035132 13.259190 14.702916 15.833949 0.6311049
#> a_16    0    0    0    0    0  4.026063  4.845432 25.562088 29.073806 0.8021829
#> a_17    0    0    0    0    0 10.447932 10.685968 12.830738 15.677288 0.7312623
#> a_18    0    0    0    0    0 12.823374 13.424289 20.707829 25.016140 0.6618646
#> a_19    0    0    0    0    0 16.771673 18.270303 16.879453 18.455670 0.5625938
#> a_20    0    0    0    0    0  7.953811  8.664292 24.445750 25.705849 0.2966870
#> a_21    0    0    0    0    0 16.855203 18.272147 12.916649 14.032633 0.5510372
#> a_22    0    0    0    0    0  3.286380  3.723979  4.136774  5.168235 0.8641867
#> a_23    0    0    0    0    0  9.249326 10.440103 22.759205 22.764547 0.6683735
#> a_24    0    0    0    0    0 16.905412 17.018275 23.409185 26.743113 0.6593265
#> a_25    0    0    0    0    0  7.826906  7.893834 14.015362 14.718108 0.7720903
#> a_26    0    0    0    0    0  1.836105  2.053866  3.182530  3.195504 0.4483456
#> a_27    0    0    0    0    0 12.855885 14.625315 19.030855 22.076466 0.5918191
#> a_28    1    0    0    0    0  5.993771  6.385492 27.718023 29.168386 0.3986242
#> a_29    0    1    0    0    0 14.464525 16.444338 14.255792 17.242016 0.8585182
#> a_30    0    0    1    0    0 12.042274 13.226908 23.860940 25.112452 0.7272887
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.2939660 11.393654 13.461157  9.565689 11.658482
#> a_2  0.3371243  5.840528  6.956143 13.847636 15.072832
#> a_3  0.9391955 14.702418 18.175030 12.828142 12.983710
#> a_4  0.9869701  4.992859  5.109457  6.962723  8.148678
#> a_5  0.7197083 13.732000 15.372640  6.441395  6.554086
#> a_6  0.9127639 14.384108 14.915964 14.982767 16.515264
#> a_7  0.5111777 14.765548 14.880700 11.075364 13.503518
#> a_8  0.5950674 15.487718 15.922670 15.559663 17.563179
#> a_9  0.9935634 14.004128 15.310406  9.291085 11.302603
#> a_10 0.5272821 18.131680 18.665640 16.289149 19.793050
#> a_11 0.5999780 18.835146 19.837816 10.876692 13.102033
#> a_12 0.8526849  7.896563  8.869063 18.047143 19.596691
#> a_13 0.1168151 13.645378 16.961010 11.585539 13.365893
#> a_14 0.9876359 15.328522 17.923421  6.925036  7.964187
#> a_15 0.7448666 13.989840 14.230651 16.005498 19.631788
#> a_16 0.8841916  7.399613  7.799914 13.951664 15.112830
#> a_17 0.8749183 11.104080 13.677949 10.935797 13.661434
#> a_18 0.7216033 14.839931 14.853135  6.089469  6.213873
#> a_19 0.5932565 15.090518 17.512090 10.704870 12.221355
#> a_20 0.3476999 15.986445 16.492272 12.918404 14.986834
#> a_21 0.5749067 14.661159 15.973420 14.728185 16.041736
#> a_22 0.8665585 17.049889 17.699706 15.861701 16.064104
#> a_23 0.7290210 11.178565 11.940493 16.641469 19.672473
#> a_24 0.8071835 15.509888 16.184701 14.772334 15.167467
#> a_25 0.9623772 14.350402 16.670086 15.475282 16.535284
#> a_26 0.4752062 11.098225 13.499836  7.837407  9.208262
#> a_27 0.7384378 18.307214 19.100740 11.453143 13.581648
#> a_28 0.4424770 15.448851 15.680870 11.193920 13.794337
#> a_29 0.8765475 10.096284 11.745396 14.903474 17.233019
#> a_30 0.9031633 16.737420 19.945762 13.490176 15.858805
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define negative percentage deviation from actual duration.
a=-0.20

# Define negative percentage deviation from actual duration.
b=0.40

# Simulate the effects of estimation uncertainty of a project scenario using MFPP package
# with default values.
 Res<-phase1(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7      a_8 a_9      a_10 a_11 a_12 a_13 a_14
#> a_1    1   1   0   0   0   0   0 0.000000   0 1.0000000    0    0    0    0
#> a_2    0   1   1   0   1   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_3    0   0   1   0   0   0   0 0.000000   0 0.0000000    1    0    0    0
#> a_4    0   0   0   1   0   1   0 0.828566   0 0.0000000    0    0    0    0
#> a_5    0   0   0   0   1   0   0 0.000000   0 0.0000000    0    0    1    0
#> a_6    0   0   0   0   0   1   0 0.000000   0 0.0000000    0    0    0    0
#> a_7    0   0   0   0   0   0   1 0.000000   0 0.0000000    0    0    0    0
#> a_8    0   0   0   0   0   0   0 1.000000   0 0.0000000    0    1    0    0
#> a_9    0   0   0   0   0   0   0 0.000000   1 0.9172083    0    1    0    0
#> a_10   0   0   0   0   0   0   0 0.000000   0 1.0000000    0    0    0    0
#> a_11   0   0   0   0   0   0   0 0.000000   0 0.0000000    1    0    1    0
#> a_12   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    1    0    0
#> a_13   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    1    1
#> a_14   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    1
#> a_15   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_16   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_17   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_18   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_19   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_20   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_21   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_22   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_23   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_24   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_25   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_26   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_27   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_28   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_29   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_30   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_31   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_32   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#>           a_15 a_16 a_17      a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27
#> a_1  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_2  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_3  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_4  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_5  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_6  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_7  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_8  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_9  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_10 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_11 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_12 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_13 0.0000000    0    0 0.0000000    0    0    0    0    0    0    1    0    0
#> a_14 0.0000000    1    1 0.0000000    0    0    0    0    0    0    0    0    1
#> a_15 0.7502203    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_16 0.0000000    1    1 0.0000000    0    0    0    0    0    0    0    0    0
#> a_17 0.0000000    0    1 0.0000000    0    0    0    0    0    0    0    0    0
#> a_18 0.0000000    0    0 0.7519418    1    0    0    0    0    0    0    0    0
#> a_19 0.0000000    0    0 0.0000000    1    0    0    1    1    0    0    0    0
#> a_20 0.0000000    0    0 0.0000000    0    1    0    0    0    0    0    0    0
#> a_21 0.0000000    0    0 0.0000000    0    0    1    0    0    0    0    0    0
#> a_22 0.0000000    0    0 0.0000000    0    0    0    1    0    0    0    0    0
#> a_23 0.0000000    0    0 0.0000000    0    0    0    0    1    0    0    0    0
#> a_24 0.0000000    0    0 0.0000000    0    0    0    0    0    1    0    0    0
#> a_25 0.0000000    0    0 0.0000000    0    0    0    0    0    0    1    1    1
#> a_26 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    1    1
#> a_27 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    1
#> a_28 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_29 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_30 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_31 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_32 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0  5.739858  5.921637 20.377452 20.725568 0.3117883
#> a_2     0    0    0    0    0 19.726208 20.113822 23.876852 27.138518 0.3193362
#> a_3     0    0    0    0    0 13.156621 15.137291 15.164951 19.923076 0.7671788
#> a_4     0    0    0    0    0 13.645868 16.532715 11.704473 11.139176 0.8787400
#> a_5     0    0    0    0    0 20.481286 23.443949 18.496243 28.633512 0.7605995
#> a_6     0    0    0    0    0 13.147398 14.495800 32.059453 31.586465 1.0000000
#> a_7     0    0    0    0    0 14.992348 16.096290 14.724268 15.386099 0.5760378
#> a_8     0    0    0    0    0 12.636614 14.447755  6.278483  6.397691 0.5106020
#> a_9     0    0    1    0    0 14.119888 17.914989 14.741428 16.497073 0.8374664
#> a_10    0    0    0    0    0 11.913933  9.741836 27.805332 23.763616 0.6438583
#> a_11    0    0    0    0    0 21.869312 23.946801 22.525739 22.075721 0.4869750
#> a_12    0    0    0    0    0 12.421968 16.440525 13.100892 15.749651 0.8435129
#> a_13    0    0    0    0    0 15.613262 19.105366 14.832958 17.612354 0.1434050
#> a_14    0    0    0    0    0 14.618621 14.750753 17.979707 27.129136 1.0000000
#> a_15    0    0    0    0    0 15.373800 16.157374 16.503636 15.723981 0.8084332
#> a_16    0    0    0    0    0  4.450763  4.781085 27.503459 26.261743 0.9878676
#> a_17    0    0    0    0    0 12.385719 12.870681 14.454951 17.508242 0.8016293
#> a_18    0    0    0    0    0 13.722112 15.383811 20.749227 25.055201 0.6217618
#> a_19    0    0    0    0    0 21.788029 17.082719 18.671221 21.595421 0.5885144
#> a_20    0    0    0    0    0  9.839270  8.082546 30.925944 32.870043 0.3696150
#> a_21    0    0    0    0    0 19.901951 20.131023 12.441141 16.177861 0.5973404
#> a_22    0    0    0    0    0  3.115704  4.276175  4.606496  5.562668 1.0000000
#> a_23    0    0    0    0    0  8.392652 12.003050 26.742691 28.933961 0.6194271
#> a_24    0    0    0    0    0 19.290258 19.253804 25.357928 24.535446 0.6789419
#> a_25    0    0    0    0    0 10.082277  9.605063 16.143259 16.006332 0.8768355
#> a_26    0    0    0    0    0  2.268197  2.394205  3.528788  3.476344 0.4884135
#> a_27    0    0    0    0    0 13.371049 16.910362 18.263411 26.959011 0.6111104
#> a_28    1    0    0    0    0  7.618552  5.804982 27.869940 35.696622 0.3968428
#> a_29    0    1    0    0    0 14.103450 15.493657 14.826996 18.301576 1.0000000
#> a_30    0    0    1    0    0 12.352204 16.572039 23.554327 30.551823 0.7022382
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.3141400 10.568160 15.088878  9.322845 11.479009
#> a_2  0.3983071  6.729117  7.437890 15.137250 18.695027
#> a_3  1.0000000 15.865437 19.033562 13.898094 16.717009
#> a_4  0.9702289  5.044912  6.005804  8.577798  8.217870
#> a_5  0.8446114 17.201454 14.375275  7.395089  8.236542
#> a_6  0.9156174 17.588507 17.102203 15.279217 18.211681
#> a_7  0.6576615 17.319229 18.512848 11.779128 17.015442
#> a_8  0.6553406 15.161825 17.551917 20.087684 17.094351
#> a_9  1.0000000 14.867487 17.312911 11.063785 11.003632
#> a_10 0.5578217 22.027085 21.743614 14.958596 22.243415
#> a_11 0.5441771 22.687055 18.923152 12.846411 12.125514
#> a_12 0.8284106  8.808680  8.937795 17.618190 24.260464
#> a_13 0.1468030 12.861253 21.161129 13.056545 12.861064
#> a_14 1.0000000 18.718671 21.855109  7.711553  8.015865
#> a_15 0.7534422 17.950486 15.885971 20.688834 21.593254
#> a_16 0.8772753  9.075698  7.350130 17.303603 18.944548
#> a_17 1.0000000 13.080329 17.244434 12.322453 13.789111
#> a_18 0.6523044 18.467082 19.052973  7.027595  6.405776
#> a_19 0.6353903 19.285219 16.152101 10.003746 15.713403
#> a_20 0.3733036 14.919358 20.144588 11.758803 18.772483
#> a_21 0.6963502 15.723666 15.934539 18.264912 15.817918
#> a_22 1.0000000 21.541270 18.432833 17.340583 20.003118
#> a_23 0.8947234 14.132774 11.948123 16.127916 18.124137
#> a_24 0.8925177 14.810233 19.804222 14.366787 16.733632
#> a_25 1.0000000 14.278136 18.475850 17.180933 18.938424
#> a_26 0.5235531 14.339045 15.746215  7.868737  8.456179
#> a_27 0.7443963 22.271202 21.462503 11.021341 12.654329
#> a_28 0.4389396 16.057469 16.829239 14.006944 16.814185
#> a_29 1.0000000 10.050837 15.028582 17.348827 18.751017
#> a_30 1.0000000 19.065971 18.527728 14.792483 14.903284
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  101.8205
#> Const constraint (Cc):  510.7752
#> Score/scope constraint (Cs):  0.9167554
#> Quailty constraint (Cq):  0.5492387
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  110.7058
#> Const constraint (Cc):  637.5372
#> Score/scope constraint (Cs):  0.9822689
#> Quailty constraint (Cq):  0.6871557
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 267.0135 272.1345

# Simulate the estimation uncertainty of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase1(PDM$PDM,a,b,"beta")
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7      a_8 a_9      a_10 a_11 a_12 a_13 a_14
#> a_1    1   1   0   0   0   0   0 0.000000   0 1.0000000    0    0    0    0
#> a_2    0   1   1   0   1   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_3    0   0   1   0   0   0   0 0.000000   0 0.0000000    1    0    0    0
#> a_4    0   0   0   1   0   1   0 0.828566   0 0.0000000    0    0    0    0
#> a_5    0   0   0   0   1   0   0 0.000000   0 0.0000000    0    0    1    0
#> a_6    0   0   0   0   0   1   0 0.000000   0 0.0000000    0    0    0    0
#> a_7    0   0   0   0   0   0   1 0.000000   0 0.0000000    0    0    0    0
#> a_8    0   0   0   0   0   0   0 1.000000   0 0.0000000    0    1    0    0
#> a_9    0   0   0   0   0   0   0 0.000000   1 0.9172083    0    1    0    0
#> a_10   0   0   0   0   0   0   0 0.000000   0 1.0000000    0    0    0    0
#> a_11   0   0   0   0   0   0   0 0.000000   0 0.0000000    1    0    1    0
#> a_12   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    1    0    0
#> a_13   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    1    1
#> a_14   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    1
#> a_15   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_16   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_17   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_18   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_19   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_20   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_21   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_22   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_23   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_24   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_25   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_26   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_27   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_28   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_29   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_30   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_31   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#> a_32   0   0   0   0   0   0   0 0.000000   0 0.0000000    0    0    0    0
#>           a_15 a_16 a_17      a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27
#> a_1  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_2  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_3  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_4  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_5  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_6  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_7  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_8  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_9  0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_10 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_11 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_12 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_13 0.0000000    0    0 0.0000000    0    0    0    0    0    0    1    0    0
#> a_14 0.0000000    1    1 0.0000000    0    0    0    0    0    0    0    0    1
#> a_15 0.7502203    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_16 0.0000000    1    1 0.0000000    0    0    0    0    0    0    0    0    0
#> a_17 0.0000000    0    1 0.0000000    0    0    0    0    0    0    0    0    0
#> a_18 0.0000000    0    0 0.7519418    1    0    0    0    0    0    0    0    0
#> a_19 0.0000000    0    0 0.0000000    1    0    0    1    1    0    0    0    0
#> a_20 0.0000000    0    0 0.0000000    0    1    0    0    0    0    0    0    0
#> a_21 0.0000000    0    0 0.0000000    0    0    1    0    0    0    0    0    0
#> a_22 0.0000000    0    0 0.0000000    0    0    0    1    0    0    0    0    0
#> a_23 0.0000000    0    0 0.0000000    0    0    0    0    1    0    0    0    0
#> a_24 0.0000000    0    0 0.0000000    0    0    0    0    0    1    0    0    0
#> a_25 0.0000000    0    0 0.0000000    0    0    0    0    0    0    1    1    1
#> a_26 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    1    1
#> a_27 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    1
#> a_28 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_29 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_30 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_31 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#> a_32 0.0000000    0    0 0.0000000    0    0    0    0    0    0    0    0    0
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0  5.597118  4.567088 16.384247 18.996597 0.3017222
#> a_2     0    0    0    0    0 20.200658 14.527222 31.285961 37.010377 0.2938401
#> a_3     0    0    0    0    0 11.338029 14.343026 15.048060 19.857337 0.8547331
#> a_4     0    0    0    0    0 16.240785 19.328432 10.436377 10.542440 0.7612067
#> a_5     0    0    0    0    0 19.935013 24.856970 22.151495 27.038685 0.5929812
#> a_6     0    0    0    0    0 13.096472 13.262046 31.562173 28.528415 0.9419611
#> a_7     0    0    0    0    0 11.529772 17.145440 12.937872 16.481771 0.5782710
#> a_8     0    0    0    0    0 11.018889 14.225925  5.639419  6.845755 0.6281893
#> a_9     0    0    1    0    0 14.799228 12.230149 14.843935 18.213396 0.8746073
#> a_10    0    0    0    0    0 10.881286 13.683846 19.842396 29.508117 0.5172238
#> a_11    0    0    0    0    0 19.323046 19.717921 25.104573 24.399204 0.5519640
#> a_12    0    0    0    0    0 15.505485 14.780334 13.376411 11.390236 0.7262281
#> a_13    0    0    0    0    0 14.409201 20.322873 16.247447 18.088611 0.1103836
#> a_14    0    0    0    0    0 17.015222 18.518221 22.454374 20.298997 1.0000000
#> a_15    0    0    0    0    0 13.612090 15.136734 13.788100 15.782628 0.6411747
#> a_16    0    0    0    0    0  4.996537  4.822043 27.336767 26.898370 1.0000000
#> a_17    0    0    0    0    0 10.749865 11.538869 12.340693 18.579770 0.8067767
#> a_18    0    0    0    0    0 12.809044 12.478994 20.515711 28.727264 0.6746654
#> a_19    0    0    0    0    0 19.596908 16.010746 20.241436 18.141965 0.6933710
#> a_20    0    0    0    0    0  9.426557 10.417329 25.531348 28.093485 0.3421270
#> a_21    0    0    0    0    0 17.157934 19.309169 15.683177 16.178481 0.5455447
#> a_22    0    0    0    0    0  4.153976  3.814698  5.343794  5.074268 0.9321959
#> a_23    0    0    0    0    0  9.726717 10.464669 29.135948 22.127239 0.8384866
#> a_24    0    0    0    0    0 15.605077 18.257071 30.406893 28.359133 0.6806934
#> a_25    0    0    0    0    0  8.758697  7.426732 14.839927 15.725316 0.9935621
#> a_26    0    0    0    0    0  1.847314  2.584320  3.840643  3.882053 0.4796414
#> a_27    0    0    0    0    0 12.322347 15.055077 20.274325 24.142164 0.6394773
#> a_28    1    0    0    0    0  6.227172  7.022646 35.025524 32.740320 0.3741084
#> a_29    0    1    0    0    0 19.263542 15.229513 14.350510 20.808258 0.8450397
#> a_30    0    0    1    0    0 11.139375 17.840149 31.426166 24.525730 0.6699301
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.3309121 12.482479 15.575960 10.648892 12.443900
#> a_2  0.4008187  5.745029  8.453073 14.026193 17.030352
#> a_3  1.0000000 19.291399 17.965929 13.060682 16.316094
#> a_4  1.0000000  6.262124  4.655538  7.031240  9.169610
#> a_5  0.7706734 16.428478 19.042903  7.567524  8.140803
#> a_6  0.9360685 13.497440 15.740582 14.456531 19.680606
#> a_7  0.5132783 16.016825 17.415832 12.594327 15.529535
#> a_8  0.6951769 15.121992 18.596512 16.594999 18.624733
#> a_9  1.0000000 17.745423 17.157853 11.780376 13.912314
#> a_10 0.5616077 23.554208 20.248364 17.653246 21.060700
#> a_11 0.6856335 20.626514 18.319805 12.620538 15.677501
#> a_12 0.8028190  9.920352  7.974783 23.062755 18.798828
#> a_13 0.1264113 15.430187 18.358929 14.151113 13.951633
#> a_14 1.0000000 18.659614 22.355192  7.526805 10.686754
#> a_15 0.7102802 16.203726 13.991471 19.024023 21.841108
#> a_16 1.0000000  6.936324  8.471891 15.234189 15.634072
#> a_17 0.9080316 12.444406 14.502425 13.835195 18.057600
#> a_18 0.6326047 14.566232 14.293520  6.832813  7.600231
#> a_19 0.6516664 14.992035 17.439048 11.238508 13.703595
#> a_20 0.3876587 17.626589 18.338748 13.920563 16.579169
#> a_21 0.6476352 17.402445 16.426261 14.295475 14.649792
#> a_22 1.0000000 17.787238 17.196551 13.907641 18.799893
#> a_23 0.8692546 14.891513 12.559398 19.986291 21.026853
#> a_24 0.9490546 16.357077 19.178073 15.043664 14.644285
#> a_25 1.0000000 12.475936 20.871502 16.201086 15.835831
#> a_26 0.6296881 14.128210 14.283897  7.239839  8.497060
#> a_27 0.9514568 20.296940 19.269984 11.475184 15.458219
#> a_28 0.5065319 19.187409 18.267110 14.140273 14.218338
#> a_29 1.0000000 10.609855 14.047648 17.536345 19.552896
#> a_30 1.0000000 14.968941 25.558619 11.653076 17.012050
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  92.18463
#> Const constraint (Cc):  514.1617
#> Score/scope constraint (Cs):  0.9167554
#> Quailty constraint (Cq):  0.5393139
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  109.6381
#> Const constraint (Cc):  645.9166
#> Score/scope constraint (Cs):  0.9822689
#> Quailty constraint (Cq):  0.6933594
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 264.0258 271.5732