Function to simulate estimation uncertainty.

Simulating the effects of the estimation uncertainty.

phase1(x,a=-0.1,b=0.30,pdftype="uniform")

Details

The changed demands are generated by the interval between [o+a,o+b], where o is the original value. The random generator can follow uniform (=default), or beta distribution.

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

a

optimistic duration, the negative percent deviation from the actual duration (a negative value). The default value is a=-0.1, which means if the actual duration is 1, the minimal possible duration is 0.9.

b

pessimistic duration, the positive percent deviation from the actual duration (a positive value). The default value is b=0.3, which means if the actual duration is 1, the maximal possible duration is 1.3.

pdftype

the pdftype is either 'uniform' or 'beta' (the default value is 'uniform'), which means the generated distribution either follows uniform or a beta distribution betweem interval [a,b]

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4       a_5      a_6 a_7 a_8 a_9 a_10 a_11      a_12 a_13
#> a_1    1   1   1   0 0.0000000 0.000000   0   1   0    0    0 0.0000000    0
#> a_2    0   1   1   1 0.0000000 1.000000   0   0   0    1    0 0.0000000    0
#> a_3    0   0   1   1 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_4    0   0   0   1 0.0000000 0.000000   0   0   0    0    0 0.0000000    1
#> a_5    0   0   0   0 0.7081613 0.000000   0   0   0    0    0 0.0000000    0
#> a_6    0   0   0   0 0.0000000 0.544443   1   0   0    0    0 0.0000000    0
#> a_7    0   0   0   0 0.0000000 0.000000   1   1   0    0    0 0.0000000    1
#> a_8    0   0   0   0 0.0000000 0.000000   0   1   0    0    0 0.0000000    0
#> a_9    0   0   0   0 0.0000000 0.000000   0   0   1    1    0 0.0000000    0
#> a_10   0   0   0   0 0.0000000 0.000000   0   0   0    1    0 0.6107276    0
#> a_11   0   0   0   0 0.0000000 0.000000   0   0   0    0    1 1.0000000    0
#> a_12   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 1.0000000    0
#> a_13   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    1
#> a_14   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_15   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_16   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_17   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_18   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_19   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_20   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_21   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_22   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_23   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_24   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_25   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_26   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_27   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_28   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_29   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_30   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_31   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_32   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#>      a_14 a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27
#> a_1     0    0    0 0.0000000    0    0    0    0    1    0    0    0    0    0
#> a_2     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_3     0    1    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0    1 0.0000000    0    0    0    0    0    0    0    0    1    0
#> a_9     0    0    1 1.0000000    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0    0 1.0000000    0    0    0    0    0    0    0    0    0    0
#> a_14    1    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_15    0    1    1 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_16    0    0    1 0.6301399    0    1    0    0    0    0    0    0    0    0
#> a_17    0    0    0 1.0000000    0    1    1    0    0    0    0    0    0    0
#> a_18    0    0    0 0.0000000    1    0    0    0    0    0    0    0    0    0
#> a_19    0    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0
#> a_20    0    0    0 0.0000000    0    0    1    0    0    0    0    0    0    0
#> a_21    0    0    0 0.0000000    0    0    0    1    0    1    0    0    0    0
#> a_22    0    0    0 0.0000000    0    0    0    0    1    0    0    0    0    0
#> a_23    0    0    0 0.0000000    0    0    0    0    0    1    0    0    0    0
#> a_24    0    0    0 0.0000000    0    0    0    0    0    0    1    1    0    0
#> a_25    0    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0
#> a_26    0    0    0 0.0000000    0    0    0    0    0    0    0    0    1    1
#> a_27    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1
#> a_28    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_29    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0 14.329341 15.902055 13.602217 14.759813 0.7307740
#> a_2     0    0    0    0    0 16.680861 19.583498 19.545537 19.927889 0.7757293
#> a_3     0    0    0    0    0  1.584959  1.781548 19.253011 22.567656 0.2718807
#> a_4     0    0    0    0    0  2.689703  2.749608 27.712342 27.991772 0.6678560
#> a_5     0    0    0    0    0 19.383257 19.933039 21.883587 25.240180 0.5234242
#> a_6     0    0    0    0    0 17.131321 18.080286 12.282860 14.787420 0.1833533
#> a_7     0    0    0    0    0  9.723309  9.756303 19.791911 20.770466 0.4701090
#> a_8     0    0    0    0    0 15.236986 17.646069 19.391832 19.954940 0.7352177
#> a_9     0    0    0    0    0 12.217787 12.425113 16.882372 17.723932 0.6565486
#> a_10    0    0    0    0    0 11.855605 14.652357 18.377602 21.106006 0.2557494
#> a_11    0    0    0    0    0  3.920360  3.939042  2.067325  2.365861 0.5212337
#> a_12    0    0    0    0    0 15.778277 17.783100 20.171597 22.131937 0.3735557
#> a_13    0    0    0    0    0 17.872400 19.858198 24.462124 25.362776 0.5633802
#> a_14    0    0    0    0    0  8.102393  8.746197 12.939686 12.995509 0.6823251
#> a_15    0    0    0    0    0 14.077288 15.680047 22.519089 22.946630 0.5653050
#> a_16    0    0    0    0    0 11.001576 12.826176 17.411530 19.832808 0.6236006
#> a_17    0    0    0    0    0  5.238539  5.615857 12.075636 14.944949 0.7535428
#> a_18    0    0    0    0    0  6.554969  6.742764 19.364672 22.342816 0.8578747
#> a_19    0    0    0    0    0  9.289805 11.301952 25.825259 29.078828 0.1845822
#> a_20    0    0    0    0    0 15.789859 16.330827 16.770504 17.439494 0.7667012
#> a_21    0    0    0    0    0 14.131338 16.140098 12.918117 15.569162 0.9445380
#> a_22    0    0    0    0    0 19.785141 19.823685 23.972335 29.391862 0.5827484
#> a_23    0    0    0    0    0 16.538007 18.337822  2.595881  3.244028 0.8722820
#> a_24    0    0    0    0    0 10.411860 11.511336 16.926825 18.421288 0.7281190
#> a_25    0    0    0    0    0 11.878368 12.056251  4.078627  4.127621 0.2731049
#> a_26    0    0    0    0    0 15.382423 15.971679 25.836362 28.004089 0.1223019
#> a_27    0    0    0    0    0  8.402372  9.578437 20.944489 21.838982 0.2943156
#> a_28    1    1    0    0    0 14.701055 17.417597 10.334537 11.622160 0.5226775
#> a_29    0    1    0    0    0 16.027677 18.746397 19.752982 20.636466 0.8861753
#> a_30    0    0    1    0    0  5.407548  6.215389 19.234316 20.515193 0.7177193
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.7782727  4.926133  5.982489  6.578715  7.254562
#> a_2  0.9444973 16.546218 19.214830 16.536267 19.028660
#> a_3  0.3005633  7.822696  7.825903 11.623063 14.479020
#> a_4  0.6919807 15.775072 17.127815 13.133385 16.378168
#> a_5  0.5915635 16.082209 17.087411 17.008463 18.985012
#> a_6  0.1862723 16.987126 17.706585 15.535069 16.276374
#> a_7  0.5652956 14.073236 15.431016 17.119034 18.363043
#> a_8  0.7849336 16.710157 18.648171  3.219537  3.611160
#> a_9  0.6641059 10.261800 11.324946  9.220678 10.603112
#> a_10 0.2723105 15.042347 15.212092  8.025689  9.992732
#> a_11 0.5568126 14.161545 15.103542  7.165890  8.212789
#> a_12 0.3937241  9.496522 11.529051 11.245508 12.701866
#> a_13 0.5867407  6.068244  6.266636 12.439531 14.929364
#> a_14 0.8269248 15.934305 17.634738 17.503461 17.617267
#> a_15 0.6539569  7.268955  7.412404 18.757008 18.939477
#> a_16 0.6459539  7.627054  9.167719  7.687699  7.906435
#> a_17 0.9118395 10.552092 12.148853 12.133267 14.378494
#> a_18 0.8942652 17.527652 19.352188 17.505772 19.084555
#> a_19 0.2242663 15.069570 16.341858 16.809156 18.580217
#> a_20 0.8116729 11.071918 12.589663 16.135108 16.372872
#> a_21 0.9453640  7.444686  8.792864  9.939516 11.621130
#> a_22 0.6501023 10.621147 13.046417  4.453822  5.391131
#> a_23 0.9218489 17.801321 19.968462  8.954814  9.469246
#> a_24 0.7734360 12.970415 13.774967 11.339952 11.885091
#> a_25 0.2806231  8.081576  8.692893  8.060972  8.586326
#> a_26 0.1490529 15.484819 18.664348  8.519372  9.697777
#> a_27 0.3190336  5.223924  5.357834 14.630046 15.162339
#> a_28 0.5492164  9.479008  9.602663 15.540208 15.974530
#> a_29 0.9323108 15.831004 18.595104 17.008789 19.757750
#> a_30 0.7862619 10.197982 12.494091 11.116391 12.570166
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define negative percentage deviation from actual duration.
a=-0.20

# Define negative percentage deviation from actual duration.
b=0.40

# Simulate the effects of estimation uncertainty of a project scenario using MFPP package
# with default values.
 Res<-phase1(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4       a_5      a_6 a_7 a_8 a_9 a_10 a_11      a_12 a_13
#> a_1    1   1   1   0 0.0000000 0.000000   0   1   0    0    0 0.0000000    0
#> a_2    0   1   1   1 0.0000000 1.000000   0   0   0    1    0 0.0000000    0
#> a_3    0   0   1   1 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_4    0   0   0   1 0.0000000 0.000000   0   0   0    0    0 0.0000000    1
#> a_5    0   0   0   0 0.7081613 0.000000   0   0   0    0    0 0.0000000    0
#> a_6    0   0   0   0 0.0000000 0.544443   1   0   0    0    0 0.0000000    0
#> a_7    0   0   0   0 0.0000000 0.000000   1   1   0    0    0 0.0000000    1
#> a_8    0   0   0   0 0.0000000 0.000000   0   1   0    0    0 0.0000000    0
#> a_9    0   0   0   0 0.0000000 0.000000   0   0   1    1    0 0.0000000    0
#> a_10   0   0   0   0 0.0000000 0.000000   0   0   0    1    0 0.6107276    0
#> a_11   0   0   0   0 0.0000000 0.000000   0   0   0    0    1 1.0000000    0
#> a_12   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 1.0000000    0
#> a_13   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    1
#> a_14   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_15   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_16   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_17   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_18   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_19   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_20   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_21   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_22   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_23   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_24   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_25   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_26   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_27   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_28   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_29   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_30   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_31   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_32   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#>      a_14 a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27
#> a_1     0    0    0 0.0000000    0    0    0    0    1    0    0    0    0    0
#> a_2     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_3     0    1    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0    1 0.0000000    0    0    0    0    0    0    0    0    1    0
#> a_9     0    0    1 1.0000000    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0    0 1.0000000    0    0    0    0    0    0    0    0    0    0
#> a_14    1    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_15    0    1    1 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_16    0    0    1 0.6301399    0    1    0    0    0    0    0    0    0    0
#> a_17    0    0    0 1.0000000    0    1    1    0    0    0    0    0    0    0
#> a_18    0    0    0 0.0000000    1    0    0    0    0    0    0    0    0    0
#> a_19    0    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0
#> a_20    0    0    0 0.0000000    0    0    1    0    0    0    0    0    0    0
#> a_21    0    0    0 0.0000000    0    0    0    1    0    1    0    0    0    0
#> a_22    0    0    0 0.0000000    0    0    0    0    1    0    0    0    0    0
#> a_23    0    0    0 0.0000000    0    0    0    0    0    1    0    0    0    0
#> a_24    0    0    0 0.0000000    0    0    0    0    0    0    1    1    0    0
#> a_25    0    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0
#> a_26    0    0    0 0.0000000    0    0    0    0    0    0    0    0    1    1
#> a_27    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1
#> a_28    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_29    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0 17.963012 16.367046 15.294911 18.438750 0.7779897
#> a_2     0    0    0    0    0 17.896523 23.350692 25.108579 18.331503 0.9204810
#> a_3     0    0    0    0    0  1.535188  2.279772 21.179209 28.446852 0.3188925
#> a_4     0    0    0    0    0  2.796833  2.795881 33.215645 29.147368 0.6609736
#> a_5     0    0    0    0    0 21.853454 21.007565 25.537682 25.314682 0.5793563
#> a_6     0    0    0    0    0 20.478428 19.576316 15.082884 16.109018 0.2115191
#> a_7     0    0    0    0    0  9.392949 12.484031 24.348656 25.250819 0.5388433
#> a_8     0    0    0    0    0 15.821257 19.166391 18.729395 24.180869 0.8554426
#> a_9     0    0    0    0    0 13.126060 12.011014 15.675527 22.170899 0.6107066
#> a_10    0    0    0    0    0 12.472432 15.117179 22.624332 19.480987 0.2655275
#> a_11    0    0    0    0    0  4.096467  3.733600  2.249325  2.447361 0.6189359
#> a_12    0    0    0    0    0 20.016171 22.590791 24.137117 28.071049 0.4621166
#> a_13    0    0    0    0    0 21.044711 25.510977 24.929036 26.327399 0.6127436
#> a_14    0    0    0    0    0  9.563817  9.835013 14.132872 16.679713 0.8100986
#> a_15    0    0    0    0    0 15.770983 16.534561 24.159349 27.392868 0.6708817
#> a_16    0    0    0    0    0 11.018422 12.915831 15.679516 20.779377 0.7796220
#> a_17    0    0    0    0    0  5.679745  6.079704 12.147460 18.119581 0.7669165
#> a_18    0    0    0    0    0  6.144101  6.722276 20.797960 25.686715 0.9037182
#> a_19    0    0    0    0    0  9.028746 10.439004 31.000482 28.190993 0.2082377
#> a_20    0    0    0    0    0 20.345221 16.852888 15.670042 19.897807 0.8340823
#> a_21    0    0    0    0    0 13.780872 15.273107 14.042772 19.101690 1.0000000
#> a_22    0    0    0    0    0 21.918292 23.494092 23.535072 26.501813 0.5574692
#> a_23    0    0    0    0    0 15.982858 19.349531  3.302232  4.088450 0.7893504
#> a_24    0    0    0    0    0 11.017342 14.134810 19.005816 21.405465 0.8228377
#> a_25    0    0    0    0    0 14.941656 12.389785  4.142930  4.011166 0.2894774
#> a_26    0    0    0    0    0 18.335830 17.259144 32.056793 27.479786 0.1324645
#> a_27    0    0    0    0    0  9.359052 12.036817 21.071455 26.666063 0.3252231
#> a_28    1    1    0    0    0 16.935758 21.822831 13.030227 11.346295 0.5542215
#> a_29    0    1    0    0    0 16.281907 21.167176 19.812226 22.749990 0.8657887
#> a_30    0    0    1    0    0  5.134259  7.774489 19.250676 23.376890 0.9260183
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.7707162  4.986406  7.460420  6.606933  7.949813
#> a_2  1.0000000 21.308128 23.181124 18.662309 18.796419
#> a_3  0.3208244  8.016200  9.693084 12.714580 16.375868
#> a_4  0.7342884 19.090209 16.765026 12.376801 18.724155
#> a_5  0.7671839 20.474354 18.680104 19.319228 21.777953
#> a_6  0.2186946 19.861365 21.886926 16.486510 14.684312
#> a_7  0.6785291 14.847216 18.491200 16.175577 17.408202
#> a_8  0.9521721 16.527193 16.867908  3.391724  4.337451
#> a_9  0.7089474  9.521545 10.697819  9.378635 12.248324
#> a_10 0.3258267 14.254034 14.960525  8.200199 12.383487
#> a_11 0.5536225 13.446561 14.096289  6.532911  7.464715
#> a_12 0.4624188 11.644447 14.681521 14.532020 13.583356
#> a_13 0.5510326  6.034696  6.770783 12.883699 16.813627
#> a_14 1.0000000 19.034140 22.472536 21.159986 19.725321
#> a_15 0.7099357  8.070652  6.900784 23.476331 18.041898
#> a_16 0.5963259  9.255648  8.856880  8.906802  9.001149
#> a_17 1.0000000 12.608149 14.644083 14.514581 13.419627
#> a_18 0.8237009 20.499358 24.713960 19.280194 18.317259
#> a_19 0.2813376 15.915224 18.681348 21.692541 23.703520
#> a_20 0.7403467 13.371616 16.318936 16.941932 18.582673
#> a_21 1.0000000  8.768538 10.290281  9.974735 13.587481
#> a_22 0.6253199 12.047518 15.975204  4.752420  5.931897
#> a_23 0.8404302 18.436669 22.153346  9.562629 11.222232
#> a_24 0.9693249 12.583194 15.968286 11.381076 11.825792
#> a_25 0.3218563  8.745835 11.098897 10.066102 10.867605
#> a_26 0.1692780 19.486653 17.067673 10.132674  9.244749
#> a_27 0.3750098  4.711516  6.326396 17.979224 15.907729
#> a_28 0.6272049 10.572464 11.646934 19.894999 16.271297
#> a_29 1.0000000 18.220016 18.626828 20.741392 18.053864
#> a_30 0.8522549 10.468562 15.296005 10.972719 12.383819
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  65.80493
#> Const constraint (Cc):  507.1387
#> Score/scope constraint (Cs):  0.938891
#> Quailty constraint (Cq):  0.5037103
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  114.8203
#> Const constraint (Cc):  650.6061
#> Score/scope constraint (Cs):  0.9706556
#> Quailty constraint (Cq):  0.6096297
#> Resource constraint(s) (CR):
#>          R_1     R_2
#> TPR 221.7799 210.546

# Simulate the estimation uncertainty of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase1(PDM$PDM,a,b,"beta")
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4       a_5      a_6 a_7 a_8 a_9 a_10 a_11      a_12 a_13
#> a_1    1   1   1   0 0.0000000 0.000000   0   1   0    0    0 0.0000000    0
#> a_2    0   1   1   1 0.0000000 1.000000   0   0   0    1    0 0.0000000    0
#> a_3    0   0   1   1 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_4    0   0   0   1 0.0000000 0.000000   0   0   0    0    0 0.0000000    1
#> a_5    0   0   0   0 0.7081613 0.000000   0   0   0    0    0 0.0000000    0
#> a_6    0   0   0   0 0.0000000 0.544443   1   0   0    0    0 0.0000000    0
#> a_7    0   0   0   0 0.0000000 0.000000   1   1   0    0    0 0.0000000    1
#> a_8    0   0   0   0 0.0000000 0.000000   0   1   0    0    0 0.0000000    0
#> a_9    0   0   0   0 0.0000000 0.000000   0   0   1    1    0 0.0000000    0
#> a_10   0   0   0   0 0.0000000 0.000000   0   0   0    1    0 0.6107276    0
#> a_11   0   0   0   0 0.0000000 0.000000   0   0   0    0    1 1.0000000    0
#> a_12   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 1.0000000    0
#> a_13   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    1
#> a_14   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_15   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_16   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_17   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_18   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_19   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_20   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_21   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_22   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_23   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_24   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_25   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_26   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_27   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_28   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_29   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_30   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_31   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#> a_32   0   0   0   0 0.0000000 0.000000   0   0   0    0    0 0.0000000    0
#>      a_14 a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27
#> a_1     0    0    0 0.0000000    0    0    0    0    1    0    0    0    0    0
#> a_2     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_3     0    1    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0    1 0.0000000    0    0    0    0    0    0    0    0    1    0
#> a_9     0    0    1 1.0000000    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_13    0    0    0 1.0000000    0    0    0    0    0    0    0    0    0    0
#> a_14    1    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_15    0    1    1 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_16    0    0    1 0.6301399    0    1    0    0    0    0    0    0    0    0
#> a_17    0    0    0 1.0000000    0    1    1    0    0    0    0    0    0    0
#> a_18    0    0    0 0.0000000    1    0    0    0    0    0    0    0    0    0
#> a_19    0    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0
#> a_20    0    0    0 0.0000000    0    0    1    0    0    0    0    0    0    0
#> a_21    0    0    0 0.0000000    0    0    0    1    0    1    0    0    0    0
#> a_22    0    0    0 0.0000000    0    0    0    0    1    0    0    0    0    0
#> a_23    0    0    0 0.0000000    0    0    0    0    0    1    0    0    0    0
#> a_24    0    0    0 0.0000000    0    0    0    0    0    0    1    1    0    0
#> a_25    0    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0
#> a_26    0    0    0 0.0000000    0    0    0    0    0    0    0    0    1    1
#> a_27    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1
#> a_28    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_29    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0 18.524127 18.471824 14.965558 16.061915 0.7299031
#> a_2     0    0    0    0    0 18.347637 19.295954 22.401965 18.727246 0.9039964
#> a_3     0    0    0    0    0  1.613565  2.240596 20.343609 26.245717 0.2737948
#> a_4     0    0    0    0    0  2.849198  2.939975 33.597108 30.843721 0.7668225
#> a_5     0    0    0    0    0 20.704657 22.646572 21.851894 23.952738 0.6027756
#> a_6     0    0    0    0    0 16.041666 20.356132 15.577619 16.295004 0.2043891
#> a_7     0    0    0    0    0 10.168537  9.141458 19.049020 19.250777 0.5614823
#> a_8     0    0    0    0    0 17.638133 21.270325 20.519180 21.811966 0.9335582
#> a_9     0    0    0    0    0 12.652120 12.555926 17.633121 19.228632 0.7747038
#> a_10    0    0    0    0    0 12.730077 16.690469 18.142073 26.501551 0.2909910
#> a_11    0    0    0    0    0  4.992187  4.177027  2.420790  2.855600 0.6349788
#> a_12    0    0    0    0    0 15.231492 21.552605 26.895893 26.477932 0.4493214
#> a_13    0    0    0    0    0 19.776987 21.202869 24.928576 27.350187 0.5307776
#> a_14    0    0    0    0    0  9.263813  7.993152 15.191463 13.617741 0.7244007
#> a_15    0    0    0    0    0 15.645462 16.452122 21.768648 26.835371 0.7202918
#> a_16    0    0    0    0    0 13.265844 14.594043 23.244120 22.794373 0.6417198
#> a_17    0    0    0    0    0  5.410769  6.257827 14.477281 15.375399 0.8723796
#> a_18    0    0    0    0    0  6.519946  6.006478 19.811209 28.955883 0.8846937
#> a_19    0    0    0    0    0 10.163821 11.767750 24.554453 34.324609 0.2153345
#> a_20    0    0    0    0    0 19.001248 17.653630 18.983596 19.119799 0.7779806
#> a_21    0    0    0    0    0 15.357863 19.603443 11.580825 14.192014 1.0000000
#> a_22    0    0    0    0    0 21.021918 22.049551 30.533581 31.862986 0.5854952
#> a_23    0    0    0    0    0 18.013540 22.071603  2.446712  3.330337 0.9748771
#> a_24    0    0    0    0    0 11.534006 13.113377 18.141949 16.523563 0.8401025
#> a_25    0    0    0    0    0 11.543890 13.820141  4.509072  4.841714 0.2430732
#> a_26    0    0    0    0    0 17.547215 17.941995 33.503893 31.507616 0.1208206
#> a_27    0    0    0    0    0  8.692409 12.094400 24.430823 20.579091 0.3456277
#> a_28    1    1    0    0    0 13.676521 19.055074 12.242433 12.776103 0.5258464
#> a_29    0    1    0    0    0 13.427504 19.306387 21.319996 22.870274 1.0000000
#> a_30    0    0    1    0    0  6.148199  6.892647 21.500256 20.440905 0.8412105
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.8753813  5.829827  5.996516  7.855382  8.296136
#> a_2  1.0000000 17.306521 18.392767 20.658945 20.713082
#> a_3  0.3890060  9.036265  9.374570 14.996966 17.899478
#> a_4  0.6708186 20.862885 20.373240 13.332526 18.760386
#> a_5  0.7857402 17.708631 21.218601 19.771586 20.526769
#> a_6  0.1894406 14.926217 18.748978 20.182053 18.258339
#> a_7  0.5901424 16.228294 14.591024 19.203360 20.970374
#> a_8  0.9990183 15.337002 21.496480  3.584401  3.883513
#> a_9  0.7023996 12.967229 10.548793 10.218399 13.009221
#> a_10 0.2703057 15.859190 14.550843  8.286536 10.390667
#> a_11 0.5920147 16.982036 14.382433  6.939071  9.174009
#> a_12 0.3394011  9.726413 12.490204 14.058475 11.985381
#> a_13 0.6104587  6.836369  7.078108 12.933052 15.623317
#> a_14 0.8910700 20.212899 19.223400 20.724800 20.260963
#> a_15 0.7540616  7.009715  8.399027 24.026304 18.885124
#> a_16 0.8004765  8.143780 10.407835  7.355420  9.385121
#> a_17 1.0000000 13.533885 15.670163 10.229667 14.726690
#> a_18 1.0000000 20.213207 22.578399 19.897538 15.586859
#> a_19 0.2885060 19.722356 19.424510 20.339519 23.316911
#> a_20 1.0000000 14.778394 13.305386 20.245036 16.491849
#> a_21 1.0000000  6.795756  8.394538 11.260945 13.936179
#> a_22 0.7196203 13.253166 12.523688  5.282745  4.758981
#> a_23 1.0000000 21.836624 24.475719 11.395415  9.388611
#> a_24 0.9006678 14.799078 16.190889 11.061033 13.823384
#> a_25 0.2878541  8.745202 10.448226 10.202488  9.748981
#> a_26 0.1592555 18.160983 18.471636  9.925234 11.456174
#> a_27 0.3901374  6.193278  5.703311 14.444200 13.573603
#> a_28 0.6808150 10.232433 10.265635 18.048708 17.545189
#> a_29 0.9579886 13.461190 18.019925 17.556669 23.736523
#> a_30 0.9067978  9.968942 14.497752 11.340065 13.495815
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  65.65179
#> Const constraint (Cc):  521.7419
#> Score/scope constraint (Cs):  0.938891
#> Quailty constraint (Cq):  0.5071714
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  110.9441
#> Const constraint (Cc):  632.946
#> Score/scope constraint (Cs):  0.9706556
#> Quailty constraint (Cq):  0.6235243
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 213.7481 207.7174