Function to simulate estimation uncertainty.

Simulating the effects of the estimatio uncertainty.

phase1(x,a=-0.1,b=0.30,pdftype="uniform")

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

a

optimistic duration, the negative percent deviation from the actual duration (a negative value). The default value is a=-0.1, which means if the actual duration is 1, the minimal possible duration is 0.9.

b

pessimistic duration, the positive percent deviation from the actual duration (a positive value). The default value is b=0.3, which means if the actual duration is 1, the maximal possible duration is 1.3.

pdftype

the pdftype is either 'uniform' or 'beta' (the default value is 'uniform'), which means the generated distribution either follows uniform or a beta distribution betweem interval [a,b]

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15 a_16
#> a_1    1   1   1   1   0   0   0   1   0    0    0    0    0    0    0    0
#> a_2    0   1   0   0   0   0   1   1   0    0    0    0    0    0    0    0
#> a_3    0   0   1   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0   0    0    0    0    0    0    0    0
#> a_5    0   0   0   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   0   1   0    0    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0   0    0    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1   0    0    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0   1    0    0    0    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0   0    1    1    0    0    1    0    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    0    0    0    0    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    0    0    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    0    0    0
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1    0    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0    1    0
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    1
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#>      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28 a_29     a_30
#> a_1     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_2     0    0    0    1    0    0    0    0    0    0    0    0    0 0.000000
#> a_3     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_4     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_5     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_6     1    1    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_7     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_8     0    0    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_9     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_10    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_11    1    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_12    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_13    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_14    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_15    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_16    0    0    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_17    1    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_18    0    1    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_19    0    0    1    0    0    0    0    0    1    0    0    1    0 0.000000
#> a_20    0    0    0    1    1    0    0    0    0    0    0    0    0 0.000000
#> a_21    0    0    0    0    1    0    0    0    0    0    0    0    0 0.000000
#> a_22    0    0    0    0    0    1    0    1    0    0    0    0    0 0.000000
#> a_23    0    0    0    0    0    0    1    0    0    0    0    0    0 0.000000
#>      a_31 a_32        t_1        t_2       c_1       c_2        q_1       q_2
#> a_1     0    0 13.6801075 16.4519507 26.114620 28.741059 0.72951447 0.9005453
#> a_2     0    0 13.4819808 15.4118831 14.337580 14.514880 0.46256049 0.4852781
#> a_3     0    0 14.4788548 17.0692821  7.705581  8.988310 0.58309241 0.6782434
#> a_4     0    0 16.5076775 19.7143556  2.936490  3.092684 0.72066538 0.8029435
#> a_5     0    0  0.4897473  0.5520533 20.124256 23.106256 0.28861586 0.3550966
#> a_6     0    0 12.9326649 14.4124407 22.909274 27.775500 0.88680122 0.9009543
#> a_7     0    0 13.1246015 15.5430982 21.402488 23.502951 0.80172523 0.8218242
#> a_8     0    0  9.6888919 11.7785592  5.772779  6.232011 0.76929201 0.7998109
#> a_9     0    0 16.9687746 17.3134258 21.057519 23.412367 0.61244932 0.6222112
#> a_10    0    0 16.1867634 18.4324585 10.065812 11.165259 0.63458443 0.6579822
#> a_11    0    0 13.0346109 14.4718418 23.896146 26.077926 0.82456053 0.8394583
#> a_12    0    0 14.0701790 15.8610862 21.395339 23.660110 0.09985229 0.1194909
#> a_13    0    0 14.2608413 17.4124680 13.767487 14.944389 0.64321453 0.6435363
#> a_14    0    0 17.2691326 19.6475972 12.472020 14.583312 0.70930538 0.7992340
#> a_15    0    0  4.7332505  5.2834425 21.979791 27.206659 0.72299852 0.8816483
#> a_16    0    0 14.8585982 16.0072347 10.723625 11.259551 0.66595206 0.7826662
#> a_17    0    0  4.2734254  4.6154387 21.000727 23.583385 0.41499147 0.4586523
#> a_18    0    0 11.1421054 13.8587718 26.192097 29.124049 0.69417411 0.8565072
#> a_19    0    0  7.2768244  8.6007486 21.946710 22.459332 0.81061052 0.8421701
#> a_20    0    0  8.1972763  8.7795590 12.205498 14.753293 0.74297096 0.8426219
#> a_21    0    0 18.9815180 19.8331695 25.693727 27.870227 0.32616072 0.3964216
#> a_22    0    0  9.2169585 10.6563695 19.549865 19.988946 0.73879226 0.9229434
#> a_23    0    0  8.4688136  9.7307987 16.600077 18.075674 0.75992052 0.7907126
#>          r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  15.688495 15.878667  3.027468  3.073746
#> a_2  14.489417 16.510080 16.757659 17.122081
#> a_3   2.562573  2.910354 16.441989 19.560162
#> a_4  14.552040 14.849357  8.169377  9.879925
#> a_5  11.087432 13.853322 18.396022 19.701264
#> a_6   3.971511  4.877126 12.636748 13.467254
#> a_7  14.667600 17.407536  9.947246 12.283991
#> a_8  14.047746 16.472958  9.653321 11.806773
#> a_9  11.312362 12.611541 12.162951 14.497319
#> a_10  9.989219 10.717486 15.378682 17.481448
#> a_11 10.860625 11.148659 13.082442 13.634646
#> a_12 13.217671 15.689434  7.078815  8.322052
#> a_13  4.241430  4.838073  4.762837  5.826257
#> a_14 14.623954 18.166253 17.670376 19.277538
#> a_15  4.101152  4.783947  3.189522  3.528057
#> a_16  4.099104  4.250029 17.827518 18.796784
#> a_17 12.442642 14.262147 16.470931 19.794832
#> a_18 11.630503 13.425977 14.095745 15.035643
#> a_19 18.187273 19.000845 13.984734 14.772445
#> a_20 14.141757 17.406346 16.224193 16.372792
#> a_21  7.989905  8.629935 16.468672 17.902136
#> a_22 15.586364 18.436735 14.790957 17.850659
#> a_23 13.998687 16.712155 12.084062 12.650156
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define negative percentage deviation from actual duration.
a=-0.20

# Define negative percentage deviation from actual duration.
b=0.40

# Simulate the effects of estimation uncertainty of a project scenario using MFPP package
# with default values.
 Res<-phase1(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15 a_16
#> a_1    1   1   1   1   0   0   0   1   0    0    0    0    0    0    0    0
#> a_2    0   1   0   0   0   0   1   1   0    0    0    0    0    0    0    0
#> a_3    0   0   1   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0   0    0    0    0    0    0    0    0
#> a_5    0   0   0   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   0   1   0    0    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0   0    0    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1   0    0    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0   1    0    0    0    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0   0    1    1    0    0    1    0    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    0    0    0    0    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    0    0    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    0    0    0
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1    0    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0    1    0
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    1
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#>      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28 a_29     a_30
#> a_1     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_2     0    0    0    1    0    0    0    0    0    0    0    0    0 0.000000
#> a_3     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_4     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_5     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_6     1    1    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_7     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_8     0    0    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_9     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_10    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_11    1    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_12    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_13    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_14    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_15    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_16    0    0    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_17    1    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_18    0    1    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_19    0    0    1    0    0    0    0    0    1    0    0    1    0 0.000000
#> a_20    0    0    0    1    1    0    0    0    0    0    0    0    0 0.000000
#> a_21    0    0    0    0    1    0    0    0    0    0    0    0    0 0.000000
#> a_22    0    0    0    0    0    1    0    1    0    0    0    0    0 0.000000
#> a_23    0    0    0    0    0    0    1    0    0    0    0    0    0 0.000000
#>      a_31 a_32        t_1        t_2       c_1       c_2       q_1       q_2
#> a_1     0    0 16.2109669 20.5763609 29.293610 33.484409 0.8347079 1.0000000
#> a_2     0    0 13.7515242 15.8738093 14.341279 18.125657 0.4949246 0.5324526
#> a_3     0    0 18.0455578 18.5811472  8.103162  9.773155 0.7232325 0.8278788
#> a_4     0    0 17.3888658 23.2016645  3.575742  3.674792 0.8658981 0.8534789
#> a_5     0    0  0.5608805  0.5127032 24.586581 29.542524 0.2894826 0.4124248
#> a_6     0    0 12.8534571 16.9544781 22.955538 30.336544 1.0000000 1.0000000
#> a_7     0    0 15.3695954 14.2986259 20.024030 29.349022 0.7997067 0.8382407
#> a_8     0    0 11.7563727 15.0720839  5.735340  6.056722 0.8978404 0.9695070
#> a_9     0    0 19.3403864 17.9303372 19.913357 23.935052 0.7156523 0.7302920
#> a_10    0    0 16.2575505 22.3585800  9.757545 11.922706 0.7757074 0.7188888
#> a_11    0    0 14.6243884 16.0496064 30.343374 31.725458 0.8848486 0.8802218
#> a_12    0    0 17.8853359 19.5334012 19.268903 23.922975 0.1150369 0.1258084
#> a_13    0    0 17.3868981 17.2614939 15.252722 14.175274 0.7102607 0.7019982
#> a_14    0    0 15.7729856 22.0362740 15.455302 18.212689 0.7867477 0.8027341
#> a_15    0    0  5.1814664  4.9869628 24.699485 29.094216 0.7662195 1.0000000
#> a_16    0    0 18.9639813 17.1897595 11.070313 14.575070 0.6916053 0.9837485
#> a_17    0    0  4.7157449  4.9119732 19.822549 26.274784 0.4764882 0.4460102
#> a_18    0    0 14.1388105 13.9681704 23.828618 37.170537 0.8516314 0.8624421
#> a_19    0    0  8.8332723 11.0124111 20.385786 21.404767 0.8364806 0.8224660
#> a_20    0    0  8.4502152  8.8981340 13.781589 16.993734 0.9124138 0.8173507
#> a_21    0    0 20.9631727 21.5785156 24.037189 35.158929 0.3557470 0.3902390
#> a_22    0    0 11.7154698 12.8251377 22.926737 25.510417 0.8857638 1.0000000
#> a_23    0    0  7.6505861 12.5025998 20.959110 17.865697 0.8509457 0.7628786
#>          r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  19.581456 17.268876  2.962415  3.494374
#> a_2  13.498254 16.988352 21.319351 20.791121
#> a_3   3.054856  3.349441 17.958308 21.505005
#> a_4  18.341060 16.919525  8.956750 12.473383
#> a_5  13.574797 12.527808 22.847927 23.332984
#> a_6   5.001435  5.229895 11.416097 16.995383
#> a_7  13.936988 17.719944 10.209693 14.244126
#> a_8  13.599258 14.986775 10.345578 12.118919
#> a_9  13.559528 14.005508 11.268655 13.769706
#> a_10 11.103813 13.911272 16.760378 19.114555
#> a_11 11.945919 10.540965 14.308138 17.513325
#> a_12 17.006623 16.816659  8.843005  7.664979
#> a_13  4.033198  4.777616  5.834691  7.082113
#> a_14 18.195181 17.887213 18.966058 22.037001
#> a_15  4.291058  4.333397  2.921572  4.281043
#> a_16  4.905593  5.343710 18.476947 22.446634
#> a_17 11.251653 16.902551 16.577757 21.936139
#> a_18 12.112018 15.554614 17.397701 19.202419
#> a_19 22.047606 21.496835 16.181924 18.157688
#> a_20 15.687679 21.357441 16.639907 18.639175
#> a_21  8.632621 10.402524 19.715608 16.184805
#> a_22 15.009547 17.116992 18.593170 21.343430
#> a_23 12.598851 15.203791 14.109400 12.314808
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  52.18273
#> Const constraint (Cc):  529.6995
#> Score/scope constraint (Cs):  0.9105052
#> Quailty constraint (Cq):  0.5648136
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  63.25136
#> Const constraint (Cc):  651.8889
#> Score/scope constraint (Cs):  0.9984057
#> Quailty constraint (Cq):  0.6990293
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 219.9317 230.9678

# Simulate the estimation uncertainty of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase1(PDM$PDM,a,b,"beta")
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15 a_16
#> a_1    1   1   1   1   0   0   0   1   0    0    0    0    0    0    0    0
#> a_2    0   1   0   0   0   0   1   1   0    0    0    0    0    0    0    0
#> a_3    0   0   1   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0   0    0    0    0    0    0    0    0
#> a_5    0   0   0   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   0   1   0    0    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0   0    0    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1   0    0    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0   1    0    0    0    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0   0    1    1    0    0    1    0    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    0    0    0    0    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    0    0    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    0    0    0
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1    0    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0    1    0
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    1
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#>      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28 a_29     a_30
#> a_1     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_2     0    0    0    1    0    0    0    0    0    0    0    0    0 0.000000
#> a_3     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_4     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_5     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_6     1    1    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_7     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_8     0    0    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_9     0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_10    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_11    1    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_12    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_13    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_14    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_15    0    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_16    0    0    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_17    1    0    0    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_18    0    1    1    0    0    0    0    0    0    0    0    0    0 0.000000
#> a_19    0    0    1    0    0    0    0    0    1    0    0    1    0 0.000000
#> a_20    0    0    0    1    1    0    0    0    0    0    0    0    0 0.000000
#> a_21    0    0    0    0    1    0    0    0    0    0    0    0    0 0.000000
#> a_22    0    0    0    0    0    1    0    1    0    0    0    0    0 0.000000
#> a_23    0    0    0    0    0    0    1    0    0    0    0    0    0 0.000000
#>      a_31 a_32        t_1        t_2       c_1       c_2       q_1       q_2
#> a_1     0    0 15.9640422 14.1890161 28.341467 36.697339 0.7922726 0.9531545
#> a_2     0    0 13.2319390 13.7924692 13.845298 13.708440 0.4286165 0.5132463
#> a_3     0    0 16.8625254 18.8532836  6.567924 11.084631 0.5975876 0.7568646
#> a_4     0    0 19.3150535 21.6748900  2.932586  2.926171 0.7952834 0.8350458
#> a_5     0    0  0.5519031  0.7083468 21.880478 21.697139 0.3253755 0.3825814
#> a_6     0    0 15.3039161 16.2695165 28.071425 28.632966 1.0000000 0.8922475
#> a_7     0    0 13.7800366 18.4037772 20.918735 25.919836 0.8320624 0.8847910
#> a_8     0    0 12.1594888 12.5319495  6.525540  6.764211 0.9985982 0.7491043
#> a_9     0    0 18.0085265 22.6114121 21.398130 29.910565 0.5201415 0.7378836
#> a_10    0    0 14.6395457 18.7585318 11.713923 11.699640 0.8466423 0.7069611
#> a_11    0    0 11.8992269 13.9298388 30.704449 29.285520 0.9615236 0.9006031
#> a_12    0    0 13.0365078 19.7162086 23.506822 22.794581 0.1002145 0.1186445
#> a_13    0    0 13.7640727 19.1838215 16.191260 18.733942 0.5772904 0.6991357
#> a_14    0    0 17.2081738 18.0229248 12.236974 15.031981 0.8223078 0.9596101
#> a_15    0    0  5.5473025  5.5423017 21.528022 25.097715 0.8053098 1.0000000
#> a_16    0    0 15.5570500 21.2441086 10.185608 12.756910 0.6678324 0.6924805
#> a_17    0    0  4.5826761  4.8850376 21.630673 22.293328 0.4874203 0.5068144
#> a_18    0    0 11.9940375 16.8878515 28.581085 34.817280 0.6399717 0.8762219
#> a_19    0    0  8.1692182  9.9006202 22.887412 26.094096 0.8443163 0.9356732
#> a_20    0    0  9.0480888  8.0526246 14.629373 15.959197 0.7993380 0.9608481
#> a_21    0    0 23.3169402 22.3661926 31.939506 32.339226 0.3038136 0.3902594
#> a_22    0    0 11.6149909  9.5051453 18.601832 20.461690 0.6835750 1.0000000
#> a_23    0    0 10.0877680  8.9049070 17.000273 20.678965 0.7153602 0.7851570
#>          r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  18.688835 17.282625  3.116480  3.682767
#> a_2  14.955792 18.287355 20.209675 15.076883
#> a_3   2.966432  3.255744 15.694116 22.603505
#> a_4  17.254697 16.474219 10.405813  9.884995
#> a_5  14.300135 12.544660 22.973873 19.668899
#> a_6   3.846180  5.441191 14.036823 15.635581
#> a_7  17.105779 17.329248 10.517723 13.067047
#> a_8  16.162987 17.331895 11.242796 12.424355
#> a_9  11.727116 16.299566 15.778136 17.225002
#> a_10 11.136832 10.113022 18.483230 16.240626
#> a_11 10.250721  9.957699 13.017943 12.166608
#> a_12 16.881486 16.486759  8.020439  8.783506
#> a_13  5.223795  5.174638  5.447595  6.572237
#> a_14 12.407809 16.092537 16.619529 23.484060
#> a_15  5.045613  5.545265  2.701244  3.424073
#> a_16  4.370098  4.160368 22.720070 24.487554
#> a_17 14.521861 13.518968 19.267447 20.200708
#> a_18 11.878004 14.356279 17.759920 18.765660
#> a_19 17.888000 17.843785 15.212733 19.988089
#> a_20 15.960558 20.567170 20.775660 19.610305
#> a_21  8.613909  9.012913 17.929529 18.342249
#> a_22 17.054260 19.763645 16.210284 17.524034
#> a_23 16.901081 17.303222 12.452993 15.440383
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  52.90191
#> Const constraint (Cc):  544.6582
#> Score/scope constraint (Cs):  0.9105052
#> Quailty constraint (Cq):  0.5223186
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  64.37236
#> Const constraint (Cc):  639.0798
#> Score/scope constraint (Cs):  0.9984057
#> Quailty constraint (Cq):  0.6961556
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 222.4474 232.6403