Function to simulate shock effects.

Simulating the shock effects of the changes of customer claims. Two steps of Monte-Carlo simulation is applied for selecting (1) and changing (2) task demands and the structures.

phase2 (x,p=0.10, s=5.0)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability factor for task selection (a numeric value between 0 and 1). The default value is p=1.00

s

Scale factor: the ratio of the modification (a numeric integer). The default value is s=5.0

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15 a_16
#> a_1    1   1   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_2    0   1   0   0   1   0   0   0   0    0    0    0    0    0    0    1
#> a_3    0   0   1   1   0   0   1   0   1    0    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0   0    0    0    0    0    0    0    0
#> a_5    0   0   0   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   0   0   0    0    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0   0    0    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1   1    0    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0   1    0    1    1    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0   0    1    0    0    0    0    0    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    1    0    1    0    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    1    0    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    1    0    1
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1    0    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0    1    0
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    1
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_24   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_25   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_26   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_27   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_28   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_29   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_30   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_31   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_32   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#>      a_17 a_18 a_19 a_20 a_21 a_22      a_23 a_24 a_25 a_26      a_27 a_28 a_29
#> a_1     1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_3     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_5     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_6     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_9     0    0    0    0    0    0 0.0000000    0    1    0 0.0000000    0    0
#> a_10    0    1    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_11    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_12    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_14    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_15    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_16    1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_17    1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_18    0    1    1    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_19    0    0    1    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_20    0    0    0    1    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_21    0    0    0    0    1    0 0.0000000    1    0    1 0.0000000    0    0
#> a_22    0    0    0    0    0    1 0.0000000    0    0    0 0.0000000    0    0
#> a_23    0    0    0    0    0    0 0.6463515    0    0    0 1.0000000    0    0
#> a_24    0    0    0    0    0    0 0.0000000    1    0    0 0.0000000    0    0
#> a_25    0    0    0    0    0    0 0.0000000    0    1    0 0.0000000    0    0
#> a_26    0    0    0    0    0    0 0.0000000    0    0    1 0.0000000    0    0
#> a_27    0    0    0    0    0    0 0.0000000    0    0    0 0.8292881    0    0
#> a_28    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    1    0
#> a_29    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    1
#> a_30    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_31    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_32    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#>      a_30 a_31 a_32        t_1        t_2       c_1       c_2        q_1
#> a_1     0    0    0 11.8916130 13.3444854 20.020238 21.553242 0.53676423
#> a_2     0    0    0 14.9659260 15.1816147 22.167287 26.433042 0.63778634
#> a_3     0    0    0 10.8655287 12.6624450 10.509449 11.328609 0.35549199
#> a_4     0    0    0 13.6691601 14.7544433 17.241608 20.329701 0.31849139
#> a_5     0    0    0 16.7834686 16.8514988  4.189229  5.115331 0.11253449
#> a_6     0    0    0 14.7754776 16.6296040 18.055389 22.014823 0.65808900
#> a_7     0    0    0  5.1759319  5.9398908 15.680885 16.047019 0.70684075
#> a_8     0    0    0  4.7121688  5.0124561 21.351730 22.290717 0.55782091
#> a_9     0    0    0  9.9182979 10.7163339 27.049832 28.218362 0.71034599
#> a_10    0    0    0  8.0661975  9.3036601  7.977406  9.677134 0.74036181
#> a_11    0    0    0  9.7989421 11.3591390 13.639653 14.365024 0.47457496
#> a_12    0    0    0  8.2787449  8.6266836 25.824335 29.984597 0.83070012
#> a_13    0    0    0  3.0940336  3.8051394 13.000581 13.501725 0.85463039
#> a_14    0    0    0 13.0941929 14.3810774 13.306452 15.715983 0.79008285
#> a_15    0    0    0 15.4044329 16.6394641 12.332705 14.761988 0.61984236
#> a_16    0    0    0  6.7937667  8.0812106 13.676979 14.751001 0.91190336
#> a_17    0    0    0 12.5084295 14.1756037 18.704672 20.800134 0.64761026
#> a_18    0    0    0 11.6913437 13.7407406 15.669289 17.095781 0.62650379
#> a_19    0    0    0 11.7616985 12.4066678  5.756634  6.422804 0.03992803
#> a_20    0    0    0 15.3757309 17.7756626  3.744248  4.486643 0.47035270
#> a_21    0    0    0  6.5530621  7.9267353 18.351590 22.262507 0.63839904
#> a_22    0    0    0 13.1836807 15.7023458 12.216722 12.260121 0.63565649
#> a_23    0    0    0 15.2014328 17.3910680 27.521805 28.856404 0.53848825
#> a_24    0    0    0  5.6336056  6.7321914 20.027071 22.503683 0.94339462
#> a_25    0    0    0  0.7415563  0.8448231  9.727409 11.109698 0.88013418
#> a_26    0    0    0 14.9464393 15.3950064 20.812593 22.435027 0.55858736
#> a_27    0    0    0  5.8557259  5.9911162 19.520455 22.557716 0.76962921
#> a_28    0    0    0  4.5820341  5.6745144 20.880603 23.979216 0.50970051
#> a_29    0    0    0  8.9200439 10.6299948  9.034088  9.700537 0.91675540
#> a_30    1    0    0 12.0728560 13.6402281 12.714207 14.346181 0.81252496
#> a_31    0    0    0  0.0000000  0.0000000  0.000000  0.000000 0.00000000
#> a_32    0    0    0  0.0000000  0.0000000  0.000000  0.000000 0.00000000
#>             q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.60252536 14.888130 17.699212 15.633761 16.360819
#> a_2  0.79067074 14.351543 14.957477 18.954418 19.630162
#> a_3  0.36738331  8.057171  9.018003 11.368448 14.182842
#> a_4  0.33269949 15.016792 18.754036 17.749455 17.789004
#> a_5  0.13100629 14.808646 15.118578 10.504284 13.019049
#> a_6  0.71990439  3.855404  3.971634 18.154435 18.436216
#> a_7  0.77306560 13.497793 13.601283 16.070143 18.700060
#> a_8  0.67045594 13.013537 14.260991 14.445074 14.608508
#> a_9  0.88091984 15.049855 17.248083 14.057264 14.355346
#> a_10 0.89447865 10.929796 12.744250 10.071112 11.893246
#> a_11 0.56079175 14.848506 16.168040  2.544960  2.811392
#> a_12 0.98124631 13.388124 15.379156 15.224792 16.914068
#> a_13 0.94176260 14.919665 18.585243  7.422570  7.975259
#> a_14 0.89706826 10.103026 10.974729  4.108298  4.592476
#> a_15 0.74422444 16.276430 18.733442 15.622869 17.398372
#> a_16 0.99737883 10.817114 12.847348 15.024028 18.649785
#> a_17 0.65147501 10.915945 13.429219 15.742631 16.131658
#> a_18 0.76645859 13.432123 14.931469  8.111717  8.997034
#> a_19 0.04049661  4.316758  4.440542  5.610128  6.428969
#> a_20 0.49486653  4.702885  5.137604  7.393161  7.417742
#> a_21 0.72876878 14.622694 16.819331 11.500314 12.693353
#> a_22 0.75617118 10.450238 11.167204 13.112433 14.187162
#> a_23 0.56777702 14.675321 16.801014 10.056050 11.621755
#> a_24 0.97636418 10.595962 10.925884  6.569944  7.996124
#> a_25 0.88944214  5.673096  6.514702 12.008043 12.809806
#> a_26 0.57970078 17.236591 18.798120  3.607970  3.646844
#> a_27 0.82883314 13.617084 16.512905 12.736667 14.459075
#> a_28 0.58863677 13.780183 15.749947 10.638540 12.957629
#> a_29 0.99417185  1.158465  1.333155  8.146257  9.123871
#> a_30 0.81634531 17.005223 18.534624 13.690744 14.102891
#> a_31 0.00000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.00000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=5.0

# Simulate the shock effect of a project scenario using MFPP package
# with default values.
 Res<-phase2(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15 a_16
#> a_1    1   1   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_2    0   1   0   0   1   0   0   0   0    0    0    0    0    0    0    1
#> a_3    0   0   1   1   0   0   1   0   1    0    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0   0    0    0    0    0    0    0    0
#> a_5    0   0   0   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   0   0   0    0    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0   0    0    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1   1    0    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0   1    0    1    1    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0   0    1    0    0    0    0    0    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    1    0    1    0    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    1    0    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    1    0    1
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1    0    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0    1    0
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    1
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_24   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_25   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_26   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_27   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_28   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_29   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_30   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_31   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_32   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#>      a_17 a_18 a_19 a_20 a_21 a_22      a_23 a_24 a_25 a_26      a_27 a_28 a_29
#> a_1     1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_3     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_5     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_6     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_9     0    0    0    0    0    0 0.0000000    0    1    0 0.0000000    0    0
#> a_10    0    1    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_11    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_12    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_14    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_15    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_16    1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_17    1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_18    0    1    1    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_19    0    0    1    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_20    0    0    0    1    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_21    0    0    0    0    1    0 0.0000000    1    0    1 0.0000000    0    0
#> a_22    0    0    0    0    0    1 0.0000000    0    0    0 0.0000000    0    0
#> a_23    0    0    0    0    0    0 0.6463515    0    0    0 1.0000000    0    0
#> a_24    0    0    0    0    0    0 0.0000000    1    0    0 0.0000000    0    0
#> a_25    0    0    0    0    0    0 0.0000000    0    1    0 0.0000000    0    0
#> a_26    0    0    0    0    0    0 0.0000000    0    0    1 0.0000000    0    0
#> a_27    0    0    0    0    0    0 0.0000000    0    0    0 0.8292881    0    0
#> a_28    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    1    0
#> a_29    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    1
#> a_30    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_31    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_32    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#>      a_30 a_31 a_32        t_1        t_2        c_1        c_2       q_1
#> a_1     0    0    0 11.8916130 13.3444854  20.020238  21.553242 0.5367642
#> a_2     0    0    0 14.9659260 15.1816147  22.167287  26.433042 0.6377863
#> a_3     0    0    0 10.8655287 12.6624450  10.509449  11.328609 0.3554920
#> a_4     0    0    0 13.6691601 14.7544433  17.241608  20.329701 0.3184914
#> a_5     0    0    0 16.7834686 16.8514988  25.135373   5.115331 0.6752069
#> a_6     0    0    0 14.7754776 16.6296040  18.055389  22.014823 0.6580890
#> a_7     0    0    0  5.1759319  5.9398908  15.680885  16.047019 0.7068407
#> a_8     0    0    0 28.2730128  5.0124561  21.351730  22.290717 0.5578209
#> a_9     0    0    0  9.9182979 10.7163339 162.298990  28.218362 0.7103460
#> a_10    0    0    0  8.0661975  9.3036601   7.977406   9.677134 0.7403618
#> a_11    0    0    0  9.7989421 11.3591390  13.639653  14.365024 0.4745750
#> a_12    0    0    0  8.2787449  8.6266836  25.824335  29.984597 0.8307001
#> a_13    0    0    0  3.0940336  3.8051394  13.000581  13.501725 0.8546304
#> a_14    0    0    0 13.0941929 14.3810774  79.838714  15.715983 0.7900829
#> a_15    0    0    0 15.4044329 16.6394641  12.332705  14.761988 0.6198424
#> a_16    0    0    0  6.7937667  8.0812106  13.676979  14.751001 0.9119034
#> a_17    0    0    0 12.5084295 14.1756037  18.704672  20.800134 0.6476103
#> a_18    0    0    0 11.6913437 13.7407406  15.669289  17.095781 0.6265038
#> a_19    0    0    0 70.5701908 12.4066678  34.539802   6.422804 0.2395682
#> a_20    0    0    0 15.3757309 17.7756626   3.744248   4.486643 0.4703527
#> a_21    0    0    0  6.5530621  7.9267353  18.351590  22.262507 0.6383990
#> a_22    0    0    0 13.1836807 15.7023458  73.300331  73.560725 0.6356565
#> a_23    0    0    0 15.2014328 17.3910680  27.521805  28.856404 0.5384882
#> a_24    0    0    0  5.6336056  6.7321914  20.027071 135.022098 1.0000000
#> a_25    0    0    0  0.7415563  0.8448231   9.727409  11.109698 0.8801342
#> a_26    0    0    0 14.9464393 15.3950064  20.812593  22.435027 0.5585874
#> a_27    0    0    0  5.8557259  5.9911162  19.520455  22.557716 0.7696292
#> a_28    0    0    0  4.5820341  5.6745144  20.880603  23.979216 0.5097005
#> a_29    0    0    0  8.9200439 10.6299948   9.034088   9.700537 0.9167554
#> a_30    1    0    0 12.0728560 81.8413683  12.714207  86.077085 0.8125250
#> a_31    0    0    0  0.0000000  0.0000000   0.000000   0.000000 0.0000000
#> a_32    0    0    0  0.0000000  0.0000000   0.000000   0.000000 0.0000000
#>             q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.60252536 89.328782 17.699212 15.633761 16.360819
#> a_2  0.79067074 14.351543 14.957477 18.954418 19.630162
#> a_3  0.36738331  8.057171  9.018003 11.368448 14.182842
#> a_4  0.33269949 15.016792 18.754036 17.749455 17.789004
#> a_5  0.13100629 14.808646 15.118578 10.504284 13.019049
#> a_6  1.00000000  3.855404  3.971634 18.154435 18.436216
#> a_7  0.77306560 13.497793 13.601283 16.070143 18.700060
#> a_8  0.67045594 13.013537 14.260991 14.445074 14.608508
#> a_9  0.88091984 90.299130 17.248083 14.057264 14.355346
#> a_10 0.89447865 10.929796 76.465501 10.071112 11.893246
#> a_11 0.56079175 14.848506 16.168040  2.544960  2.811392
#> a_12 0.98124631 13.388124 15.379156 15.224792 16.914068
#> a_13 0.94176260 14.919665 18.585243  7.422570  7.975259
#> a_14 0.89706826 60.618159 10.974729  4.108298  4.592476
#> a_15 0.74422444 16.276430 18.733442 93.737211 17.398372
#> a_16 0.99737883 10.817114 12.847348 15.024028 18.649785
#> a_17 0.65147501 10.915945 13.429219 94.455788 16.131658
#> a_18 0.76645859 13.432123 89.588812  8.111717  8.997034
#> a_19 0.04049661  4.316758 26.643254  5.610128  6.428969
#> a_20 0.49486653  4.702885  5.137604  7.393161  7.417742
#> a_21 0.72876878 14.622694 16.819331 11.500314 12.693353
#> a_22 0.75617118 10.450238 11.167204 13.112433 14.187162
#> a_23 0.56777702 14.675321 16.801014 10.056050 11.621755
#> a_24 0.97636418 10.595962 10.925884  6.569944  7.996124
#> a_25 0.88944214  5.673096  6.514702 12.008043 76.858838
#> a_26 0.57970078 17.236591 18.798120  3.607970  3.646844
#> a_27 0.82883314 13.617084 16.512905 12.736667 14.459075
#> a_28 0.58863677 13.780183 15.749947 10.638540 12.957629
#> a_29 0.99417185  1.158465  1.333155  8.146257  9.123871
#> a_30 1.00000000 17.005223 18.534624 82.144466 14.102891
#> a_31 0.00000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.00000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  51.95571
#> Const constraint (Cc):  489.9168
#> Score/scope constraint (Cs):  0.9160119
#> Quailty constraint (Cq):  0.4946501
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  93.61459
#> Const constraint (Cc):  1016.795
#> Score/scope constraint (Cs):  0.9807012
#> Quailty constraint (Cq):  0.6905818
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 331.2703 495.9505

# Simulate the effect of a project scenario using MFPP package
# with specified values.
 Res<-phase2(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15 a_16
#> a_1    1   1   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_2    0   1   0   0   1   0   0   0   0    0    0    0    0    0    0    1
#> a_3    0   0   1   1   0   0   1   0   1    0    0    0    0    0    0    0
#> a_4    0   0   0   1   0   0   0   0   0    0    0    0    0    0    0    0
#> a_5    0   0   0   0   1   0   0   0   0    0    0    0    0    0    0    0
#> a_6    0   0   0   0   0   1   0   0   0    0    0    0    0    0    0    0
#> a_7    0   0   0   0   0   0   1   0   0    0    0    0    0    0    0    0
#> a_8    0   0   0   0   0   0   0   1   1    0    0    0    0    0    0    0
#> a_9    0   0   0   0   0   0   0   0   1    0    1    1    0    0    0    0
#> a_10   0   0   0   0   0   0   0   0   0    1    0    0    0    0    0    0
#> a_11   0   0   0   0   0   0   0   0   0    0    1    1    0    1    0    0
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    1    0    0
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    1    0    1
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1    0    0
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0    1    0
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    1
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_24   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_25   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_26   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_27   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_28   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_29   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_30   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_31   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#> a_32   0   0   0   0   0   0   0   0   0    0    0    0    0    0    0    0
#>      a_17 a_18 a_19 a_20 a_21 a_22      a_23 a_24 a_25 a_26      a_27 a_28 a_29
#> a_1     1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_3     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_5     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_6     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_9     0    0    0    0    0    0 0.0000000    0    1    0 0.0000000    0    0
#> a_10    0    1    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_11    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_12    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_14    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_15    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_16    1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_17    1    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_18    0    1    1    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_19    0    0    1    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_20    0    0    0    1    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_21    0    0    0    0    1    0 0.0000000    1    0    1 0.0000000    0    0
#> a_22    0    0    0    0    0    1 0.0000000    0    0    0 0.0000000    0    0
#> a_23    0    0    0    0    0    0 0.6463515    0    0    0 1.0000000    0    0
#> a_24    0    0    0    0    0    0 0.0000000    1    0    0 0.0000000    0    0
#> a_25    0    0    0    0    0    0 0.0000000    0    1    0 0.0000000    0    0
#> a_26    0    0    0    0    0    0 0.0000000    0    0    1 0.0000000    0    0
#> a_27    0    0    0    0    0    0 0.0000000    0    0    0 0.8292881    0    0
#> a_28    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    1    0
#> a_29    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    1
#> a_30    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_31    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#> a_32    0    0    0    0    0    0 0.0000000    0    0    0 0.0000000    0    0
#>      a_30 a_31 a_32        t_1        t_2        c_1        c_2        q_1
#> a_1     0    0    0  71.349678 13.3444854  20.020238  21.553242 0.53676423
#> a_2     0    0    0  14.965926 15.1816147  22.167287 158.598250 1.00000000
#> a_3     0    0    0  10.865529 12.6624450  10.509449  11.328609 0.35549199
#> a_4     0    0    0  13.669160 14.7544433  17.241608  20.329701 0.31849139
#> a_5     0    0    0 100.700812 16.8514988   4.189229   5.115331 0.67520693
#> a_6     0    0    0  14.775478 99.7776240  18.055389  22.014823 0.65808900
#> a_7     0    0    0   5.175932  5.9398908  94.085311  16.047019 0.70684075
#> a_8     0    0    0   4.712169  5.0124561  21.351730  22.290717 1.00000000
#> a_9     0    0    0   9.918298 10.7163339  27.049832  28.218362 0.71034599
#> a_10    0    0    0   8.066198 55.8219605   7.977406   9.677134 0.74036181
#> a_11    0    0    0  58.793653 11.3591390  13.639653  14.365024 0.47457496
#> a_12    0    0    0   8.278745  8.6266836  25.824335  29.984597 0.83070012
#> a_13    0    0    0   3.094034  3.8051394  13.000581  13.501725 0.85463039
#> a_14    0    0    0  13.094193 14.3810774  13.306452  15.715983 0.79008285
#> a_15    0    0    0  15.404433 16.6394641  12.332705  14.761988 0.61984236
#> a_16    0    0    0   6.793767  8.0812106  13.676979  14.751001 0.91190336
#> a_17    0    0    0  12.508429 14.1756037  18.704672  20.800134 0.64761026
#> a_18    0    0    0  70.148062 13.7407406  15.669289  17.095781 0.62650379
#> a_19    0    0    0  11.761698 12.4066678   5.756634   6.422804 0.03992803
#> a_20    0    0    0  15.375731 17.7756626   3.744248  26.919857 0.47035270
#> a_21    0    0    0   6.553062  7.9267353  18.351590  22.262507 0.63839904
#> a_22    0    0    0  13.183681 15.7023458  12.216722  12.260121 0.63565649
#> a_23    0    0    0  15.201433 17.3910680  27.521805  28.856404 0.53848825
#> a_24    0    0    0   5.633606  6.7321914  20.027071  22.503683 0.94339462
#> a_25    0    0    0   1.000000  0.8448231   9.727409  11.109698 0.88013418
#> a_26    0    0    0  14.946439 15.3950064  20.812593  22.435027 0.55858736
#> a_27    0    0    0   5.855726  5.9911162 117.122727  22.557716 0.76962921
#> a_28    0    0    0   4.582034  5.6745144  20.880603  23.979216 0.50970051
#> a_29    0    0    0   8.920044 10.6299948  54.204525   9.700537 0.91675540
#> a_30    1    0    0  12.072856 13.6402281  12.714207  14.346181 0.81252496
#> a_31    0    0    0   0.000000  0.0000000   0.000000   0.000000 0.00000000
#> a_32    0    0    0   0.000000  0.0000000   0.000000   0.000000 0.00000000
#>             q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.60252536 14.888130 17.699212 15.633761 16.360819
#> a_2  0.79067074 14.351543 14.957477 18.954418 19.630162
#> a_3  0.36738331 48.343028  9.018003 11.368448 14.182842
#> a_4  0.33269949 90.100749 18.754036 17.749455 17.789004
#> a_5  0.13100629 14.808646 15.118578 10.504284 78.114293
#> a_6  0.71990439  3.855404  3.971634 18.154435 18.436216
#> a_7  0.77306560 13.497793 13.601283 16.070143 18.700060
#> a_8  1.00000000 13.013537 14.260991 14.445074 14.608508
#> a_9  1.00000000 15.049855 17.248083 14.057264 14.355346
#> a_10 0.89447865 10.929796 12.744250 10.071112 11.893246
#> a_11 0.56079175 14.848506 16.168040  2.544960  2.811392
#> a_12 0.98124631 13.388124 15.379156 91.348749 16.914068
#> a_13 0.94176260 89.517989 18.585243  7.422570  7.975259
#> a_14 0.89706826 10.103026 10.974729  4.108298  4.592476
#> a_15 0.74422444 16.276430 18.733442 15.622869 17.398372
#> a_16 0.99737883 10.817114 12.847348 15.024028 18.649785
#> a_17 0.65147501 10.915945 13.429219 15.742631 16.131658
#> a_18 0.76645859 13.432123 14.931469  8.111717  8.997034
#> a_19 0.04049661  4.316758 26.643254  5.610128  6.428969
#> a_20 0.49486653  4.702885  5.137604  7.393161  7.417742
#> a_21 0.72876878 14.622694 16.819331 11.500314 12.693353
#> a_22 0.75617118 10.450238 11.167204 13.112433 14.187162
#> a_23 0.56777702 14.675321 16.801014 10.056050 11.621755
#> a_24 0.97636418 10.595962 65.555301  6.569944  7.996124
#> a_25 0.88944214  5.673096  6.514702 12.008043 12.809806
#> a_26 0.57970078 17.236591 18.798120  3.607970  3.646844
#> a_27 0.82883314 13.617084 16.512905 12.736667 14.459075
#> a_28 0.58863677 13.780183 15.749947 10.638540 12.957629
#> a_29 0.99417185  1.158465  1.333155 48.877544 54.743223
#> a_30 0.81634531 17.005223 18.534624 13.690744 14.102891
#> a_31 0.00000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.00000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  53.5159
#> Const constraint (Cc):  424.6955
#> Score/scope constraint (Cs):  0.9160119
#> Quailty constraint (Cq):  0.5080806
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  138.3767
#> Const constraint (Cc):  896.6105
#> Score/scope constraint (Cs):  0.9807012
#> Quailty constraint (Cq):  0.6547246
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 351.8765 319.9115