Function to simulate shock effects.

Simulating the shock effects of the changes of customer claims. Two steps of Monte-Carlo simulation is applied for selecting (1) and changing (2) task demands and the structures.

phase2 (x,p=0.10, s=5.0)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability factor for task selection (a numeric value between 0 and 1). The default value is p=1.00

s

Scale factor: the ratio of the modification (a numeric integer). The default value is s=5.0

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2       a_3      a_4       a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13
#> a_1    1   1 1.0000000 0.000000 0.0000000   0   0   1   0    0    0    0    0
#> a_2    0   1 0.8912689 0.000000 0.0000000   0   0   1   0    1    1    0    0
#> a_3    0   0 1.0000000 1.000000 0.7270103   0   0   0   0    0    0    0    0
#> a_4    0   0 0.0000000 0.652201 0.0000000   0   0   0   0    0    0    0    0
#> a_5    0   0 0.0000000 0.000000 1.0000000   0   0   0   0    0    0    0    0
#> a_6    0   0 0.0000000 0.000000 0.0000000   1   1   0   0    0    0    0    0
#> a_7    0   0 0.0000000 0.000000 0.0000000   0   1   1   0    0    0    0    0
#> a_8    0   0 0.0000000 0.000000 0.0000000   0   0   1   0    0    0    0    0
#> a_9    0   0 0.0000000 0.000000 0.0000000   0   0   0   1    1    0    0    0
#> a_10   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    1    0    0    0
#> a_11   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    1    0    1
#> a_12   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    1    0
#> a_13   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    1
#> a_14   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_15   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_16   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_17   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_18   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_19   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_20   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_21   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_22   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_23   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_24   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_25   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_26   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_27   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_28   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_29   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_30   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_31   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_32   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#>      a_14 a_15 a_16 a_17 a_18 a_19      a_20 a_21      a_22 a_23 a_24 a_25 a_26
#> a_1     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_3     0    0    0    0    0    1 0.0000000    0 0.0000000    0    0    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_5     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_6     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0 1.0000000    0    0    0    0
#> a_9     0    0    1    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_10    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_11    0    0    0    0    1    0 0.0000000    0 1.0000000    0    0    0    0
#> a_12    0    1    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_14    1    0    0    1    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_15    0    1    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_16    0    0    1    0    0    1 1.0000000    0 0.0000000    0    0    0    0
#> a_17    0    0    0    1    0    0 1.0000000    0 0.0000000    0    0    0    0
#> a_18    0    0    0    0    1    0 0.0000000    0 0.0000000    0    0    0    0
#> a_19    0    0    0    0    0    1 0.0000000    1 0.0000000    0    0    0    0
#> a_20    0    0    0    0    0    0 0.5560004    0 0.0000000    0    0    0    0
#> a_21    0    0    0    0    0    0 0.0000000    1 0.0000000    0    0    0    0
#> a_22    0    0    0    0    0    0 0.0000000    0 0.7790602    0    0    0    0
#> a_23    0    0    0    0    0    0 0.0000000    0 0.0000000    1    0    0    0
#> a_24    0    0    0    0    0    0 0.0000000    0 0.0000000    0    1    0    0
#> a_25    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    1    1
#> a_26    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    1
#> a_27    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_28    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_29    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_30    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_31    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_32    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#>      a_27 a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2
#> a_1     0    0    0    0    0    0 14.141973 15.864184  6.556538  7.809308
#> a_2     0    0    0    0    0    0 14.539539 17.719771 14.441926 14.466368
#> a_3     0    0    0    0    0    0 15.112950 15.499594 22.560535 25.086064
#> a_4     0    0    0    0    0    0 10.565089 12.231109 12.163830 14.281124
#> a_5     0    0    0    0    0    0 17.267819 17.574151 17.064326 18.823600
#> a_6     0    0    0    0    0    0 10.886153 11.671429 13.054599 16.114114
#> a_7     0    0    0    0    0    0 13.927073 15.672236 14.275622 16.372875
#> a_8     0    0    0    0    0    0  8.480818 10.071631 21.984575 23.178130
#> a_9     0    0    0    0    0    0 15.460040 16.225211  5.727427  6.212365
#> a_10    0    0    0    0    0    0 18.995293 19.593060 24.547870 25.807115
#> a_11    0    0    0    0    0    0  9.551921 11.483743 13.533505 14.919836
#> a_12    0    0    0    0    0    0 15.902919 17.229141 25.672002 28.766794
#> a_13    0    0    0    0    0    0 12.034688 13.762720 23.505602 25.979076
#> a_14    0    0    0    0    0    0 12.806432 14.098797 28.984527 28.985274
#> a_15    0    0    0    0    0    0 11.571560 12.479330  4.860567  5.430846
#> a_16    0    0    0    0    0    0 11.731891 13.536144 10.408207 10.803226
#> a_17    0    1    0    0    0    0 11.216304 12.156917 14.191440 16.717510
#> a_18    0    0    0    0    0    0 15.036931 17.164897 27.605309 29.217308
#> a_19    0    0    0    0    0    0  4.398083  5.202067 13.461574 15.663273
#> a_20    0    0    0    0    0    0 15.403487 15.485915 18.516178 20.399904
#> a_21    0    0    0    0    0    0 19.111545 19.439112 18.980393 23.388480
#> a_22    0    0    0    0    0    0 17.266546 18.994056 10.130730 11.623305
#> a_23    0    0    0    0    0    0  7.351911  7.735180 21.262619 25.780150
#> a_24    0    0    0    0    0    0  4.090452  4.697967 22.172188 22.511402
#> a_25    0    0    0    0    0    0  3.222618  3.659738 19.934531 19.989345
#> a_26    0    0    0    0    0    0 17.635777 18.103814 25.055494 28.537777
#> a_27    1    0    0    0    0    0  6.121262  6.928335 20.962243 25.774095
#> a_28    0    1    0    0    0    0  2.926171  3.007856 12.595089 14.249092
#> a_29    0    0    1    0    0    0 11.300497 11.525740 23.395497 28.062672
#> a_30    0    0    0    1    0    0  8.013793  8.225475 16.197630 19.924594
#> a_31    0    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000
#> a_32    0    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000
#>            q_1       q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.4748316 0.5874260  9.014427 10.540018 11.835147 13.654070
#> a_2  0.3745058 0.3817889  8.162290  8.564753 13.840732 16.409338
#> a_3  0.3563028 0.4096981  7.025636  8.560349 15.587440 18.311130
#> a_4  0.2597835 0.3243661 12.123623 14.410088 16.084003 16.103416
#> a_5  0.8178978 0.9311107  6.445753  7.152970 18.384865 18.650128
#> a_6  0.6033084 0.6185609 17.002332 19.032843 16.086038 18.016533
#> a_7  0.3049933 0.3307353 15.008904 17.384890  4.937528  5.940808
#> a_8  0.1198387 0.1248406  4.339723  4.658304 14.502759 16.716788
#> a_9  0.2966110 0.3033478 14.105501 15.410357  8.066015  8.998139
#> a_10 0.8866809 0.9310738 11.376568 12.899927  4.947412  5.813560
#> a_11 0.2028855 0.2392913  4.227334  5.058233 14.613183 14.984981
#> a_12 0.6492997 0.7086933  8.480081  9.989669 13.616338 13.621442
#> a_13 0.8598113 0.9748014 16.324805 19.478692  2.239468  2.407388
#> a_14 0.7710879 0.9499077 15.346569 17.834254 16.217022 16.455971
#> a_15 0.8702971 0.9606455  7.254702  8.726825 11.291184 13.975367
#> a_16 0.5268925 0.5761784 12.699713 15.840776  1.466751  1.775187
#> a_17 0.6776788 0.8440271 15.833434 17.977267  7.814955  8.641824
#> a_18 0.8015126 0.9701227  7.391692  8.249332 14.092508 15.788154
#> a_19 0.5679507 0.6290227 13.511103 13.975925  4.675126  5.824442
#> a_20 0.8196071 0.8902852 15.702859 19.352274 17.057635 17.854341
#> a_21 0.6882839 0.8421307  7.230257  7.819900 12.085235 14.041898
#> a_22 0.8381491 0.9427598  9.586990 10.640572 17.728693 19.059325
#> a_23 0.8902557 0.9337935  5.716440  6.819869 14.087202 14.797669
#> a_24 0.7864419 0.8265575 16.309577 16.633090 11.886484 11.920572
#> a_25 0.5036039 0.5550415  6.195775  7.645585 18.132948 19.395568
#> a_26 0.5076584 0.5765441 11.487337 13.221729 14.954972 16.700592
#> a_27 0.6431294 0.7247906 15.735192 19.610216 15.829174 16.523626
#> a_28 0.7826274 0.9418112 11.192893 11.304415 13.124469 14.512932
#> a_29 0.8988236 0.9373761 15.962895 17.208447 12.782893 13.503333
#> a_30 0.6465016 0.6596656 13.775296 14.015743 13.365503 14.330496
#> a_31 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=5.0

# Simulate the shock effect of a project scenario using MFPP package
# with default values.
 Res<-phase2(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2       a_3      a_4       a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13
#> a_1    1   1 1.0000000 0.000000 0.0000000   0   0   1   0    0    0    0    0
#> a_2    0   1 0.8912689 0.000000 0.0000000   0   0   1   0    1    1    0    0
#> a_3    0   0 1.0000000 1.000000 0.7270103   0   0   0   0    0    0    0    0
#> a_4    0   0 0.0000000 0.652201 0.0000000   0   0   0   0    0    0    0    0
#> a_5    0   0 0.0000000 0.000000 1.0000000   0   0   0   0    0    0    0    0
#> a_6    0   0 0.0000000 0.000000 0.0000000   1   1   0   0    0    0    0    0
#> a_7    0   0 0.0000000 0.000000 0.0000000   0   1   1   0    0    0    0    0
#> a_8    0   0 0.0000000 0.000000 0.0000000   0   0   1   0    0    0    0    0
#> a_9    0   0 0.0000000 0.000000 0.0000000   0   0   0   1    1    0    0    0
#> a_10   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    1    0    0    0
#> a_11   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    1    0    1
#> a_12   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    1    0
#> a_13   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    1
#> a_14   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_15   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_16   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_17   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_18   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_19   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_20   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_21   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_22   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_23   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_24   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_25   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_26   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_27   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_28   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_29   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_30   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_31   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_32   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#>      a_14 a_15 a_16 a_17 a_18 a_19      a_20 a_21      a_22 a_23 a_24 a_25 a_26
#> a_1     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_3     0    0    0    0    0    1 0.0000000    0 0.0000000    0    0    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_5     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_6     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0 1.0000000    0    0    0    0
#> a_9     0    0    1    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_10    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_11    0    0    0    0    1    0 0.0000000    0 1.0000000    0    0    0    0
#> a_12    0    1    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_14    1    0    0    1    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_15    0    1    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_16    0    0    1    0    0    1 1.0000000    0 0.0000000    0    0    0    0
#> a_17    0    0    0    1    0    0 1.0000000    0 0.0000000    0    0    0    0
#> a_18    0    0    0    0    1    0 0.0000000    0 0.0000000    0    0    0    0
#> a_19    0    0    0    0    0    1 0.0000000    1 0.0000000    0    0    0    0
#> a_20    0    0    0    0    0    0 0.5560004    0 0.0000000    0    0    0    0
#> a_21    0    0    0    0    0    0 0.0000000    1 0.0000000    0    0    0    0
#> a_22    0    0    0    0    0    0 0.0000000    0 0.7790602    0    0    0    0
#> a_23    0    0    0    0    0    0 0.0000000    0 0.0000000    1    0    0    0
#> a_24    0    0    0    0    0    0 0.0000000    0 0.0000000    0    1    0    0
#> a_25    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    1    1
#> a_26    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    1
#> a_27    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_28    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_29    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_30    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_31    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_32    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#>      a_27 a_28 a_29 a_30 a_31 a_32        t_1        t_2        c_1        c_2
#> a_1     0    0    0    0    0    0  14.141973  15.864184   6.556538   7.809308
#> a_2     0    0    0    0    0    0  14.539539  17.719771  14.441926  86.798209
#> a_3     0    0    0    0    0    0  15.112950  15.499594  22.560535  25.086064
#> a_4     0    0    0    0    0    0  10.565089  12.231109  12.163830  14.281124
#> a_5     0    0    0    0    0    0  17.267819  17.574151 102.385954  18.823600
#> a_6     0    0    0    0    0    0  10.886153  11.671429  13.054599  16.114114
#> a_7     0    0    0    0    0    0  13.927073  15.672236  14.275622  16.372875
#> a_8     0    0    0    0    0    0   8.480818  10.071631  21.984575  23.178130
#> a_9     0    0    0    0    0    0  92.760239  16.225211   5.727427   6.212365
#> a_10    0    0    0    0    0    0 113.971755 117.558360  24.547870  25.807115
#> a_11    0    0    0    0    0    0   9.551921  68.902460  13.533505  14.919836
#> a_12    0    0    0    0    0    0  15.902919  17.229141  25.672002 172.600763
#> a_13    0    0    0    0    0    0  12.034688  13.762720  23.505602 155.874457
#> a_14    0    0    0    0    0    0  76.838592  14.098797  28.984527  28.985274
#> a_15    0    0    0    0    0    0  11.571560  12.479330   4.860567   5.430846
#> a_16    0    0    0    0    0    0  11.731891  13.536144  62.449244  10.803226
#> a_17    0    1    0    0    0    0  11.216304  12.156917  14.191440  16.717510
#> a_18    0    0    0    0    0    0  15.036931  17.164897  27.605309 175.303849
#> a_19    0    0    0    0    0    0   4.398083   5.202067  80.769441  93.979638
#> a_20    0    0    0    0    0    0  15.403487  92.915489  18.516178  20.399904
#> a_21    0    0    0    0    0    0  19.111545  19.439112  18.980393  23.388480
#> a_22    0    0    0    0    0    0  17.266546  18.994056  10.130730  11.623305
#> a_23    0    0    0    0    0    0   7.351911   7.735180 127.575715  25.780150
#> a_24    0    0    0    0    0    0   4.090452   4.697967  22.172188 135.068414
#> a_25    0    0    0    0    0    0  19.335711   3.659738  19.934531  19.989345
#> a_26    0    0    0    0    0    0  17.635777 108.622885  25.055494  28.537777
#> a_27    1    0    0    0    0    0   6.121262   6.928335  20.962243  25.774095
#> a_28    0    1    0    0    0    0   2.926171   3.007856  12.595089  14.249092
#> a_29    0    0    1    0    0    0  11.300497  11.525740  23.395497  28.062672
#> a_30    0    0    0    1    0    0   8.013793   8.225475  16.197630  19.924594
#> a_31    0    0    0    0    0    0   0.000000   0.000000   0.000000   0.000000
#> a_32    0    0    0    0    0    0   0.000000   0.000000   0.000000   0.000000
#>            q_1       q_2     r_1.1      r_2.1      r_1.2     r_2.2
#> a_1  0.4748316 0.5874260  9.014427  10.540018  11.835147 81.924420
#> a_2  0.3745058 0.3817889  8.162290   8.564753  13.840732 16.409338
#> a_3  0.3563028 0.4096981  7.025636   8.560349  15.587440 18.311130
#> a_4  0.2597835 1.0000000 12.123623  14.410088  16.084003 16.103416
#> a_5  0.8178978 0.9311107  6.445753   7.152970  18.384865 18.650128
#> a_6  0.6033084 0.6185609 17.002332 114.197059  16.086038 18.016533
#> a_7  1.0000000 0.3307353 90.053421  17.384890   4.937528 35.644845
#> a_8  0.1198387 0.1248406  4.339723  27.949826  87.016554 16.716788
#> a_9  0.2966110 0.3033478 14.105501  15.410357   8.066015  8.998139
#> a_10 0.8866809 0.9310738 11.376568  12.899927   4.947412  5.813560
#> a_11 0.2028855 0.2392913  4.227334   5.058233  14.613183 14.984981
#> a_12 0.6492997 0.7086933  8.480081   9.989669  13.616338 13.621442
#> a_13 0.8598113 0.9748014 16.324805  19.478692   2.239468  2.407388
#> a_14 0.7710879 0.9499077 15.346569 107.005527  16.217022 16.455971
#> a_15 0.8702971 0.9606455  7.254702   8.726825  11.291184 13.975367
#> a_16 0.5268925 0.5761784 12.699713  15.840776   1.466751  1.775187
#> a_17 1.0000000 1.0000000 15.833434  17.977267   7.814955  8.641824
#> a_18 0.8015126 0.9701227  7.391692   8.249332  14.092508 15.788154
#> a_19 0.5679507 1.0000000 13.511103  13.975925   4.675126  5.824442
#> a_20 0.8196071 0.8902852 15.702859  19.352274 102.345809 17.854341
#> a_21 0.6882839 0.8421307  7.230257   7.819900  12.085235 84.251388
#> a_22 1.0000000 0.9427598  9.586990  10.640572 106.372160 19.059325
#> a_23 0.8902557 0.9337935  5.716440   6.819869  14.087202 14.797669
#> a_24 1.0000000 0.8265575 16.309577  16.633090  11.886484 11.920572
#> a_25 0.5036039 0.5550415 37.174649   7.645585 108.797685 19.395568
#> a_26 0.5076584 0.5765441 11.487337  79.330377  89.729831 16.700592
#> a_27 0.6431294 0.7247906 15.735192  19.610216  15.829174 16.523626
#> a_28 0.7826274 0.9418112 11.192893  11.304415  13.124469 14.512932
#> a_29 0.8988236 0.9373761 15.962895  17.208447  76.697358 13.503333
#> a_30 0.6465016 0.6596656 82.651779  14.015743  13.365503 14.330496
#> a_31 0.0000000 0.0000000  0.000000   0.000000   0.000000  0.000000
#> a_32 0.0000000 0.0000000  0.000000   0.000000   0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  130.197
#> Const constraint (Cc):  556.9715
#> Score/scope constraint (Cs):  0.8998203
#> Quailty constraint (Cq):  0.4912283
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  210.3186
#> Const constraint (Cc):  1504.9
#> Score/scope constraint (Cs):  0.9612683
#> Quailty constraint (Cq):  0.6562681
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 493.1345 604.2199

# Simulate the effect of a project scenario using MFPP package
# with specified values.
 Res<-phase2(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2       a_3      a_4       a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13
#> a_1    1   1 1.0000000 0.000000 0.0000000   0   0   1   0    0    0    0    0
#> a_2    0   1 0.8912689 0.000000 0.0000000   0   0   1   0    1    1    0    0
#> a_3    0   0 1.0000000 1.000000 0.7270103   0   0   0   0    0    0    0    0
#> a_4    0   0 0.0000000 0.652201 0.0000000   0   0   0   0    0    0    0    0
#> a_5    0   0 0.0000000 0.000000 1.0000000   0   0   0   0    0    0    0    0
#> a_6    0   0 0.0000000 0.000000 0.0000000   1   1   0   0    0    0    0    0
#> a_7    0   0 0.0000000 0.000000 0.0000000   0   1   1   0    0    0    0    0
#> a_8    0   0 0.0000000 0.000000 0.0000000   0   0   1   0    0    0    0    0
#> a_9    0   0 0.0000000 0.000000 0.0000000   0   0   0   1    1    0    0    0
#> a_10   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    1    0    0    0
#> a_11   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    1    0    1
#> a_12   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    1    0
#> a_13   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    1
#> a_14   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_15   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_16   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_17   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_18   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_19   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_20   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_21   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_22   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_23   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_24   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_25   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_26   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_27   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_28   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_29   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_30   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_31   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#> a_32   0   0 0.0000000 0.000000 0.0000000   0   0   0   0    0    0    0    0
#>      a_14 a_15 a_16 a_17 a_18 a_19      a_20 a_21      a_22 a_23 a_24 a_25 a_26
#> a_1     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_3     0    0    0    0    0    1 0.0000000    0 0.0000000    0    0    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_5     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_6     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0 1.0000000    0    0    0    0
#> a_9     0    0    1    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_10    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_11    0    0    0    0    1    0 0.0000000    0 1.0000000    0    0    0    0
#> a_12    0    1    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_14    1    0    0    1    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_15    0    1    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_16    0    0    1    0    0    1 1.0000000    0 0.0000000    0    0    0    0
#> a_17    0    0    0    1    0    0 1.0000000    0 0.0000000    0    0    0    0
#> a_18    0    0    0    0    1    0 0.0000000    0 0.0000000    0    0    0    0
#> a_19    0    0    0    0    0    1 0.0000000    1 0.0000000    0    0    0    0
#> a_20    0    0    0    0    0    0 0.5560004    0 0.0000000    0    0    0    0
#> a_21    0    0    0    0    0    0 0.0000000    1 0.0000000    0    0    0    0
#> a_22    0    0    0    0    0    0 0.0000000    0 0.7790602    0    0    0    0
#> a_23    0    0    0    0    0    0 0.0000000    0 0.0000000    1    0    0    0
#> a_24    0    0    0    0    0    0 0.0000000    0 0.0000000    0    1    0    0
#> a_25    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    1    1
#> a_26    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    1
#> a_27    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_28    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_29    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_30    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_31    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#> a_32    0    0    0    0    0    0 0.0000000    0 0.0000000    0    0    0    0
#>      a_27 a_28 a_29 a_30 a_31 a_32        t_1        t_2        c_1        c_2
#> a_1     0    0    0    0    0    0  14.141973  15.864184   6.556538   7.809308
#> a_2     0    0    0    0    0    0  14.539539  17.719771  14.441926  14.466368
#> a_3     0    0    0    0    0    0  90.677701  92.997564  22.560535  25.086064
#> a_4     0    0    0    0    0    0  10.565089  12.231109  12.163830  14.281124
#> a_5     0    0    0    0    0    0  17.267819 105.444908  17.064326  18.823600
#> a_6     0    0    0    0    0    0  65.316916  11.671429  13.054599  16.114114
#> a_7     0    0    0    0    0    0  13.927073  15.672236  14.275622  16.372875
#> a_8     0    0    0    0    0    0   8.480818  10.071631  21.984575  23.178130
#> a_9     0    0    0    0    0    0  15.460040  97.351264   5.727427   6.212365
#> a_10    0    0    0    0    0    0  18.995293  19.593060  24.547870 154.842692
#> a_11    0    0    0    0    0    0   9.551921  11.483743  13.533505  89.519017
#> a_12    0    0    0    0    0    0  15.902919  17.229141  25.672002  28.766794
#> a_13    0    0    0    0    0    0  12.034688  13.762720  23.505602  25.979076
#> a_14    0    0    0    0    0    0  12.806432  14.098797  28.984527  28.985274
#> a_15    0    0    0    0    0    0  11.571560  12.479330   4.860567   5.430846
#> a_16    0    0    0    0    0    0  11.731891  13.536144  10.408207  10.803226
#> a_17    0    1    0    0    0    0  11.216304  12.156917  14.191440  16.717510
#> a_18    0    0    0    0    0    0  15.036931  17.164897  27.605309  29.217308
#> a_19    0    0    0    0    0    0   4.398083   5.202067  13.461574  15.663273
#> a_20    0    0    0    0    0    0  15.403487  15.485915  18.516178 122.399421
#> a_21    0    0    0    0    0    0  19.111545 116.634671 113.882360  23.388480
#> a_22    0    0    0    0    0    0  17.266546  18.994056  10.130730  11.623305
#> a_23    0    0    0    0    0    0   7.351911   7.735180  21.262619  25.780150
#> a_24    0    0    0    0    0    0   4.090452   4.697967  22.172188  22.511402
#> a_25    0    0    0    0    0    0   3.222618   3.659738  19.934531  19.989345
#> a_26    0    0    0    0    0    0 105.814662  18.103814  25.055494  28.537777
#> a_27    1    0    0    0    0    0   6.121262   6.928335  20.962243  25.774095
#> a_28    0    1    0    0    0    0   2.926171   3.007856  12.595089  14.249092
#> a_29    0    0    1    0    0    0  11.300497  11.525740  23.395497  28.062672
#> a_30    0    0    0    1    0    0   8.013793   8.225475  16.197630  19.924594
#> a_31    0    0    0    0    0    0   0.000000   0.000000   0.000000   0.000000
#> a_32    0    0    0    0    0    0   0.000000   0.000000   0.000000   0.000000
#>            q_1       q_2     r_1.1      r_2.1     r_1.2      r_2.2
#> a_1  0.4748316 0.5874260  9.014427  63.240106 11.835147  13.654070
#> a_2  0.3745058 0.3817889  8.162290   8.564753 13.840732  16.409338
#> a_3  0.3563028 0.4096981  7.025636   8.560349 15.587440  18.311130
#> a_4  0.2597835 0.3243661 12.123623  14.410088 16.084003  16.103416
#> a_5  0.8178978 0.9311107 38.674517   7.152970 18.384865 111.900770
#> a_6  0.6033084 0.6185609 17.002332  19.032843 16.086038  18.016533
#> a_7  0.3049933 1.0000000 15.008904  17.384890  4.937528   5.940808
#> a_8  0.7190319 0.7490434  4.339723  27.949826 14.502759  16.716788
#> a_9  0.2966110 0.3033478 14.105501  15.410357  8.066015   8.998139
#> a_10 0.8866809 0.9310738 68.259410  12.899927 29.684472   5.813560
#> a_11 0.2028855 0.2392913  4.227334   5.058233 14.613183  14.984981
#> a_12 0.6492997 0.7086933  8.480081   9.989669 13.616338  13.621442
#> a_13 0.8598113 0.9748014 16.324805  19.478692  2.239468  14.444326
#> a_14 1.0000000 0.9499077 15.346569  17.834254 97.302133  16.455971
#> a_15 0.8702971 0.9606455  7.254702   8.726825 11.291184  13.975367
#> a_16 0.5268925 0.5761784 76.198276  15.840776  1.466751   1.775187
#> a_17 0.6776788 0.8440271 95.000605  17.977267  7.814955   8.641824
#> a_18 0.8015126 0.9701227  7.391692   8.249332 14.092508  15.788154
#> a_19 0.5679507 0.6290227 81.066619  13.975925  4.675126   5.824442
#> a_20 0.8196071 0.8902852 15.702859  19.352274 17.057635  17.854341
#> a_21 0.6882839 0.8421307  7.230257  46.919399 12.085235  14.041898
#> a_22 0.8381491 0.9427598  9.586990  10.640572 17.728693  19.059325
#> a_23 1.0000000 0.9337935 34.298637   6.819869 14.087202  14.797669
#> a_24 1.0000000 0.8265575 16.309577  16.633090 11.886484  11.920572
#> a_25 1.0000000 0.5550415  6.195775   7.645585 18.132948  19.395568
#> a_26 0.5076584 0.5765441 11.487337  13.221729 14.954972 100.203550
#> a_27 0.6431294 0.7247906 15.735192  19.610216 15.829174  16.523626
#> a_28 0.7826274 0.9418112 11.192893  11.304415 13.124469  14.512932
#> a_29 1.0000000 0.9373761 15.962895 103.250680 12.782893  13.503333
#> a_30 0.6465016 0.6596656 13.775296  14.015743 13.365503  14.330496
#> a_31 0.0000000 0.0000000  0.000000   0.000000  0.000000   0.000000
#> a_32 0.0000000 0.0000000  0.000000   0.000000  0.000000   0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  114.1873
#> Const constraint (Cc):  487.3999
#> Score/scope constraint (Cs):  0.8998203
#> Quailty constraint (Cq):  0.521712
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  232.7241
#> Const constraint (Cc):  981.0032
#> Score/scope constraint (Cs):  0.9612683
#> Quailty constraint (Cq):  0.6828662
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 447.0456 455.8387