Function to simulate shock effects.

Simulating the shock effects of the changes of customer claims. Two steps of Monte-Carlo simulation is applied for selecting (1) and changing (2) task demands and the structures.

phase2 (x,p=0.10, s=5.0)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability factor for task selection (a numeric value between 0 and 1). The default value is p=1.00

s

Scale factor: the ratio of the modification (a numeric integer). The default value is s=5.0

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>            a_1       a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14
#> a_1  0.7091605 0.5590394   0   0   1   0   0   0   0    0    0    0    0    0
#> a_2  0.0000000 1.0000000   1   0   0   0   0   0   0    0    0    1    0    1
#> a_3  0.0000000 0.0000000   1   0   0   0   0   0   0    0    0    0    0    0
#> a_4  0.0000000 0.0000000   0   1   1   0   0   0   0    0    0    0    0    0
#> a_5  0.0000000 0.0000000   0   0   1   1   0   0   0    0    0    0    0    0
#> a_6  0.0000000 0.0000000   0   0   0   1   1   0   0    0    0    0    0    0
#> a_7  0.0000000 0.0000000   0   0   0   0   1   0   0    0    0    0    0    0
#> a_8  0.0000000 0.0000000   0   0   0   0   0   1   0    0    1    0    0    0
#> a_9  0.0000000 0.0000000   0   0   0   0   0   0   1    1    0    0    0    0
#> a_10 0.0000000 0.0000000   0   0   0   0   0   0   0    1    0    0    0    0
#> a_11 0.0000000 0.0000000   0   0   0   0   0   0   0    0    1    1    0    0
#> a_12 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    1    0    1
#> a_13 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    1    0
#> a_14 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    1
#> a_15 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_16 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_17 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_18 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_19 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_20 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_21 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_22 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_23 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#>      a_15 a_16 a_17 a_18 a_19 a_20      a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0    0    0    0    1    0    0
#> a_5     0    0    1    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_6     0    0    1    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0 1.0000000    0    0    0    1    0    0    0
#> a_11    0    1    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_12    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_15    1    1    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_16    0    1    1    0    0    0 1.0000000    0    0    0    0    0    0    0
#> a_17    0    0    1    0    0    0 1.0000000    0    0    0    0    0    0    0
#> a_18    0    0    0    1    0    1 0.0000000    0    0    0    0    0    0    1
#> a_19    0    0    0    0    1    0 0.0000000    0    0    0    1    0    0    0
#> a_20    0    0    0    0    0    1 1.0000000    0    1    1    0    0    0    0
#> a_21    0    0    0    0    0    0 0.7256702    0    0    0    0    0    0    0
#> a_22    0    0    0    0    0    0 0.0000000    1    0    0    0    0    0    0
#> a_23    0    0    0    0    0    0 0.0000000    0    1    0    0    0    0    0
#>      a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0 17.531259 19.084087 27.311911 29.926325 0.8536874
#> a_2     0    0    0    0  4.820224  5.495589 10.566209 10.780705 0.7534860
#> a_3     0    0    0    0  2.107527  2.221010 25.974909 26.958987 0.6401849
#> a_4     0    0    0    0  7.434504  7.827378 11.823290 12.561168 0.6978778
#> a_5     0    0    0    0  6.182150  7.012429 22.827424 26.009324 0.7814043
#> a_6     0    0    0    0  5.794647  6.391731 16.597288 17.034923 0.5503067
#> a_7     0    0    0    0 18.459025 19.706978 22.620154 25.713495 0.8238043
#> a_8     0    0    0    0 11.851330 14.444878 21.899749 24.371011 0.3650242
#> a_9     0    0    0    0 14.745404 17.465784 24.091172 25.497549 0.5534493
#> a_10    1    0    0    0 17.266286 18.681364 26.296066 29.860031 0.6054410
#> a_11    0    0    0    0 15.189539 16.895615 28.813190 29.881916 0.2824529
#> a_12    0    0    0    0 12.647876 12.749813 29.380480 29.883561 0.5688418
#> a_13    0    0    0    0 16.973873 17.976284 23.673675 28.576314 0.8732880
#> a_14    0    0    0    0 16.420097 19.028072 10.367477 12.720721 0.3548655
#> a_15    0    0    0    0 10.412552 10.992751  1.092802  1.299897 0.2072740
#> a_16    0    0    0    0 14.431334 15.676016 24.402873 27.552072 0.5206630
#> a_17    0    0    0    0 11.598776 13.328027  9.546253 11.543589 0.8175856
#> a_18    0    0    0    0  5.412674  6.717889 24.618485 29.541237 0.7893475
#> a_19    0    0    0    0 17.938751 19.551489 14.966086 18.680369 0.7698803
#> a_20    0    0    0    0  9.695153 11.980482 17.222686 19.395370 0.4310363
#> a_21    0    0    0    0  6.084737  7.466236 28.388642 28.736639 0.4809787
#> a_22    0    0    0    0  9.610395 11.676783 21.858951 26.166284 0.7756851
#> a_23    0    0    0    0  6.137923  6.805665 11.554483 13.125116 0.5772860
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.8565043  1.611395  1.925636 11.122589 11.746663
#> a_2  0.8168378 15.342167 16.746353  9.106504  9.535491
#> a_3  0.7247593 12.098467 14.111539 13.037738 14.098185
#> a_4  0.7223266 15.713614 16.428359  8.715772 10.348599
#> a_5  0.9655829 13.369735 14.526219  8.222085  9.166269
#> a_6  0.6306857 16.731277 18.164199 16.803829 16.909413
#> a_7  0.8496001 16.504574 19.993735 17.905696 19.325907
#> a_8  0.4394869  9.385456 10.856166 17.429635 18.725057
#> a_9  0.5968096 18.922304 19.028511 12.388925 12.683863
#> a_10 0.6790242 13.127022 14.352992 15.328886 18.413303
#> a_11 0.3330379 17.045448 17.805777  2.968248  3.655802
#> a_12 0.5817666 16.324459 18.233667 17.944877 19.445782
#> a_13 0.9033101  3.394323  3.627144 17.469085 17.660076
#> a_14 0.4121944 11.372953 12.244415  8.745459  8.796890
#> a_15 0.2396054 17.455557 18.247644  4.932539  5.084040
#> a_16 0.6241376  5.721384  7.120396 17.435565 18.191698
#> a_17 0.8178535 14.509811 16.123752 10.424341 10.828033
#> a_18 0.9171962 15.110415 17.008045 11.504178 11.638061
#> a_19 0.8365813 18.367175 19.201476 15.751592 18.795417
#> a_20 0.4417708  3.637973  3.677787 13.941825 17.260527
#> a_21 0.5364657  2.786912  3.047366  6.491206  7.667756
#> a_22 0.8000588  8.001641  8.068024 15.309268 16.635095
#> a_23 0.6193800  4.324907  4.973385  7.698138  9.077293
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=5.0

# Simulate the shock effect of a project scenario using MFPP package
# with default values.
 Res<-phase2(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>            a_1       a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14
#> a_1  0.7091605 0.5590394   0   0   1   0   0   0   0    0    0    0    0    0
#> a_2  0.0000000 1.0000000   1   0   0   0   0   0   0    0    0    1    0    1
#> a_3  0.0000000 0.0000000   1   0   0   0   0   0   0    0    0    0    0    0
#> a_4  0.0000000 0.0000000   0   1   1   0   0   0   0    0    0    0    0    0
#> a_5  0.0000000 0.0000000   0   0   1   1   0   0   0    0    0    0    0    0
#> a_6  0.0000000 0.0000000   0   0   0   1   1   0   0    0    0    0    0    0
#> a_7  0.0000000 0.0000000   0   0   0   0   1   0   0    0    0    0    0    0
#> a_8  0.0000000 0.0000000   0   0   0   0   0   1   0    0    1    0    0    0
#> a_9  0.0000000 0.0000000   0   0   0   0   0   0   1    1    0    0    0    0
#> a_10 0.0000000 0.0000000   0   0   0   0   0   0   0    1    0    0    0    0
#> a_11 0.0000000 0.0000000   0   0   0   0   0   0   0    0    1    1    0    0
#> a_12 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    1    0    1
#> a_13 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    1    0
#> a_14 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    1
#> a_15 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_16 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_17 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_18 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_19 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_20 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_21 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_22 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_23 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#>      a_15 a_16 a_17 a_18 a_19 a_20      a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0    0    0    0    1    0    0
#> a_5     0    0    1    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_6     0    0    1    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0 1.0000000    0    0    0    1    0    0    0
#> a_11    0    1    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_12    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_15    1    1    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_16    0    1    1    0    0    0 1.0000000    0    0    0    0    0    0    0
#> a_17    0    0    1    0    0    0 1.0000000    0    0    0    0    0    0    0
#> a_18    0    0    0    1    0    1 0.0000000    0    0    0    0    0    0    1
#> a_19    0    0    0    0    1    0 0.0000000    0    0    0    1    0    0    0
#> a_20    0    0    0    0    0    1 1.0000000    0    1    1    0    0    0    0
#> a_21    0    0    0    0    0    0 0.7256702    0    0    0    0    0    0    0
#> a_22    0    0    0    0    0    0 0.0000000    1    0    0    0    0    0    0
#> a_23    0    0    0    0    0    0 0.0000000    0    1    0    0    0    0    0
#>      a_29 a_30 a_31 a_32        t_1       t_2        c_1        c_2       q_1
#> a_1     0    0    0    0  17.531259 19.084087  27.311911 179.557951 0.8536874
#> a_2     0    0    0    0  28.921345  5.495589  10.566209  10.780705 0.7534860
#> a_3     0    0    0    0   2.107527  2.221010  25.974909  26.958987 1.0000000
#> a_4     0    0    0    0   7.434504  7.827378  11.823290  12.561168 0.6978778
#> a_5     0    0    0    0   6.182150  7.012429 136.964544  26.009324 0.7814043
#> a_6     0    0    0    0  34.767883  6.391731  16.597288  17.034923 0.5503067
#> a_7     0    0    0    0 110.754147 19.706978  22.620154  25.713495 0.8238043
#> a_8     0    0    0    0  11.851330 14.444878  21.899749  24.371011 0.3650242
#> a_9     0    0    0    0  88.472426 17.465784  24.091172  25.497549 0.5534493
#> a_10    1    0    0    0  17.266286 18.681364  26.296066  29.860031 0.6054410
#> a_11    0    0    0    0  15.189539 16.895615  28.813190 179.291495 0.2824529
#> a_12    0    0    0    0  12.647876 12.749813  29.380480  29.883561 0.5688418
#> a_13    0    0    0    0  16.973873 17.976284  23.673675  28.576314 0.8732880
#> a_14    0    0    0    0  16.420097 19.028072  10.367477  12.720721 0.3548655
#> a_15    0    0    0    0  10.412552 10.992751   1.092802   1.299897 0.2072740
#> a_16    0    0    0    0  14.431334 15.676016  24.402873 165.312434 1.0000000
#> a_17    0    0    0    0  11.598776 13.328027   9.546253  11.543589 0.8175856
#> a_18    0    0    0    0   5.412674  6.717889  24.618485  29.541237 1.0000000
#> a_19    0    0    0    0 107.632503 19.551489  14.966086  18.680369 0.7698803
#> a_20    0    0    0    0   9.695153 11.980482  17.222686  19.395370 0.4310363
#> a_21    0    0    0    0   6.084737  7.466236  28.388642  28.736639 0.4809787
#> a_22    0    0    0    0   9.610395 11.676783  21.858951 156.997706 0.7756851
#> a_23    0    0    0    0   6.137923 40.833988  11.554483  13.125116 0.5772860
#>            q_2      r_1.1      r_2.1     r_1.2      r_2.2
#> a_1  0.8565043   1.611395  11.553815 11.122589  70.479979
#> a_2  0.8168378  15.342167  16.746353  9.106504   9.535491
#> a_3  0.7247593  12.098467  14.111539 13.037738  14.098185
#> a_4  0.7223266  15.713614  98.570152  8.715772  10.348599
#> a_5  0.9655829  13.369735  14.526219  8.222085   9.166269
#> a_6  0.6306857  16.731277 108.985194 16.803829  16.909413
#> a_7  0.8496001  16.504574  19.993735 17.905696  19.325907
#> a_8  0.4394869   9.385456  10.856166 17.429635  18.725057
#> a_9  0.5968096  18.922304  19.028511 12.388925  12.683863
#> a_10 0.6790242  13.127022  14.352992 15.328886  18.413303
#> a_11 0.3330379  17.045448  17.805777  2.968248   3.655802
#> a_12 0.5817666  16.324459 109.402002 17.944877  19.445782
#> a_13 0.9033101   3.394323   3.627144 17.469085  17.660076
#> a_14 0.4121944  11.372953  12.244415  8.745459   8.796890
#> a_15 1.0000000  17.455557  18.247644  4.932539   5.084040
#> a_16 0.6241376   5.721384   7.120396 17.435565 109.150185
#> a_17 0.8178535  14.509811  16.123752 10.424341  10.828033
#> a_18 0.9171962  15.110415  17.008045 11.504178  11.638061
#> a_19 0.8365813  18.367175  19.201476 15.751592  18.795417
#> a_20 0.4417708   3.637973  22.066720 13.941825  17.260527
#> a_21 0.5364657   2.786912   3.047366  6.491206   7.667756
#> a_22 0.8000588   8.001641   8.068024 15.309268  16.635095
#> a_23 1.0000000   4.324907   4.973385  7.698138   9.077293
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  58.92405
#> Const constraint (Cc):  516.6595
#> Score/scope constraint (Cs):  0.9240281
#> Quailty constraint (Cq):  0.5166689
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  160.3618
#> Const constraint (Cc):  1303.516
#> Score/scope constraint (Cs):  0.9794536
#> Quailty constraint (Cq):  0.6803562
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 377.3839 218.8304

# Simulate the effect of a project scenario using MFPP package
# with specified values.
 Res<-phase2(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>            a_1       a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14
#> a_1  0.7091605 0.5590394   0   0   1   0   0   0   0    0    0    0    0    0
#> a_2  0.0000000 1.0000000   1   0   0   0   0   0   0    0    0    1    0    1
#> a_3  0.0000000 0.0000000   1   0   0   0   0   0   0    0    0    0    0    0
#> a_4  0.0000000 0.0000000   0   1   1   0   0   0   0    0    0    0    0    0
#> a_5  0.0000000 0.0000000   0   0   1   1   0   0   0    0    0    0    0    0
#> a_6  0.0000000 0.0000000   0   0   0   1   1   0   0    0    0    0    0    0
#> a_7  0.0000000 0.0000000   0   0   0   0   1   0   0    0    0    0    0    0
#> a_8  0.0000000 0.0000000   0   0   0   0   0   1   0    0    1    0    0    0
#> a_9  0.0000000 0.0000000   0   0   0   0   0   0   1    1    0    0    0    0
#> a_10 0.0000000 0.0000000   0   0   0   0   0   0   0    1    0    0    0    0
#> a_11 0.0000000 0.0000000   0   0   0   0   0   0   0    0    1    1    0    0
#> a_12 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    1    0    1
#> a_13 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    1    0
#> a_14 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    1
#> a_15 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_16 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_17 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_18 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_19 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_20 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_21 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_22 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#> a_23 0.0000000 0.0000000   0   0   0   0   0   0   0    0    0    0    0    0
#>      a_15 a_16 a_17 a_18 a_19 a_20      a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0 0.0000000    0    0    0    0    1    0    0
#> a_5     0    0    1    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_6     0    0    1    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0 1.0000000    0    0    0    1    0    0    0
#> a_11    0    1    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_12    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_13    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_15    1    1    0    0    0    0 0.0000000    0    0    0    0    0    0    0
#> a_16    0    1    1    0    0    0 1.0000000    0    0    0    0    0    0    0
#> a_17    0    0    1    0    0    0 1.0000000    0    0    0    0    0    0    0
#> a_18    0    0    0    1    0    1 0.0000000    0    0    0    0    0    0    1
#> a_19    0    0    0    0    1    0 0.0000000    0    0    0    1    0    0    0
#> a_20    0    0    0    0    0    1 1.0000000    0    1    1    0    0    0    0
#> a_21    0    0    0    0    0    0 0.7256702    0    0    0    0    0    0    0
#> a_22    0    0    0    0    0    0 0.0000000    1    0    0    0    0    0    0
#> a_23    0    0    0    0    0    0 0.0000000    0    1    0    0    0    0    0
#>      a_29 a_30 a_31 a_32        t_1        t_2        c_1        c_2       q_1
#> a_1     0    0    0    0  17.531259  19.084087  27.311911  29.926325 0.8536874
#> a_2     0    0    0    0   4.820224   5.495589  10.566209  10.780705 0.7534860
#> a_3     0    0    0    0   2.107527   2.221010  25.974909  26.958987 0.6401849
#> a_4     0    0    0    0   7.434504   7.827378  11.823290  75.367011 0.6978778
#> a_5     0    0    0    0  37.092899  42.074573  22.827424  26.009324 0.7814043
#> a_6     0    0    0    0   5.794647   6.391731  16.597288  17.034923 0.5503067
#> a_7     0    0    0    0  18.459025  19.706978  22.620154  25.713495 0.8238043
#> a_8     0    0    0    0  11.851330  14.444878  21.899749  24.371011 0.3650242
#> a_9     0    0    0    0  14.745404  17.465784  24.091172  25.497549 0.5534493
#> a_10    1    0    0    0  17.266286 112.088187  26.296066  29.860031 0.6054410
#> a_11    0    0    0    0  15.189539  16.895615  28.813190  29.881916 0.2824529
#> a_12    0    0    0    0  12.647876  12.749813  29.380480  29.883561 0.5688418
#> a_13    0    0    0    0  16.973873  17.976284  23.673675  28.576314 0.8732880
#> a_14    0    0    0    0  16.420097  19.028072  10.367477  12.720721 0.3548655
#> a_15    0    0    0    0  62.475310  10.992751   1.092802   1.299897 0.2072740
#> a_16    0    0    0    0  14.431334  15.676016  24.402873  27.552072 1.0000000
#> a_17    0    0    0    0  11.598776  13.328027   9.546253  69.261535 0.8175856
#> a_18    0    0    0    0   5.412674   6.717889  24.618485  29.541237 0.7893475
#> a_19    0    0    0    0 107.632503  19.551489  14.966086 112.082216 0.7698803
#> a_20    0    0    0    0   9.695153  11.980482  17.222686  19.395370 0.4310363
#> a_21    0    0    0    0   6.084737   7.466236  28.388642  28.736639 0.4809787
#> a_22    0    0    0    0   9.610395  11.676783  21.858951  26.166284 0.7756851
#> a_23    0    0    0    0   6.137923   6.805665  69.326899  13.125116 0.5772860
#>            q_2      r_1.1      r_2.1     r_1.2      r_2.2
#> a_1  0.8565043   1.611395   1.925636 11.122589  11.746663
#> a_2  0.8168378  15.342167  16.746353  9.106504   9.535491
#> a_3  0.7247593  12.098467  14.111539 13.037738  14.098185
#> a_4  0.7223266  94.281681  16.428359  8.715772  10.348599
#> a_5  0.9655829  13.369735  14.526219  8.222085  54.997615
#> a_6  0.6306857  16.731277  18.164199 16.803829  16.909413
#> a_7  0.8496001  16.504574  19.993735 17.905696  19.325907
#> a_8  0.4394869  56.312735  10.856166 17.429635  18.725057
#> a_9  0.5968096  18.922304 114.171064 12.388925  12.683863
#> a_10 0.6790242  78.762134  14.352992 15.328886  18.413303
#> a_11 0.3330379  17.045448 106.834665  2.968248  21.934809
#> a_12 1.0000000  16.324459 109.402002 17.944877 116.674693
#> a_13 0.9033101   3.394323   3.627144 17.469085  17.660076
#> a_14 0.4121944  11.372953  12.244415  8.745459   8.796890
#> a_15 0.2396054  17.455557  18.247644  4.932539   5.084040
#> a_16 0.6241376   5.721384  42.722377 17.435565  18.191698
#> a_17 0.8178535  14.509811  16.123752 10.424341  64.968197
#> a_18 0.9171962  15.110415  17.008045 11.504178  11.638061
#> a_19 0.8365813  18.367175  19.201476 15.751592  18.795417
#> a_20 0.4417708   3.637973   3.677787 13.941825  17.260527
#> a_21 0.5364657   2.786912   3.047366  6.491206   7.667756
#> a_22 0.8000588   8.001641   8.068024 15.309268  16.635095
#> a_23 0.6193800   4.324907   4.973385  7.698138   9.077293
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  68.78107
#> Const constraint (Cc):  515.0954
#> Score/scope constraint (Cs):  0.9240281
#> Quailty constraint (Cq):  0.5116339
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  200.3302
#> Const constraint (Cc):  1000.53
#> Score/scope constraint (Cs):  0.9794536
#> Quailty constraint (Cq):  0.6312785
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 433.9869 287.5713