Function to simulate shock effects.

Phase 2 simulates the shock effects. Where p percent of task demands can be increased up to s times.

phase2 (x,p=0.10, s=5.0)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability factor for task selection (a numeric value between 0 and 1). The default value is p=1.00

s

Scale factor: the ratio of the modification (a numeric integer). The default value is s=5.0

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5       a_6 a_7 a_8       a_9 a_10 a_11 a_12 a_13 a_14
#> a_1    1   1   0   1   1 0.0000000   0   1 0.0000000    0    0    0    0    0
#> a_2    0   1   0   0   0 1.0000000   0   0 0.0000000    0    0    0    0    0
#> a_3    0   0   1   0   1 0.0000000   0   0 0.8099181    0    0    0    0    0
#> a_4    0   0   0   1   0 0.8889752   0   0 0.0000000    0    0    0    0    0
#> a_5    0   0   0   0   1 0.0000000   1   0 0.0000000    0    0    0    0    0
#> a_6    0   0   0   0   0 1.0000000   0   0 0.0000000    1    0    0    0    0
#> a_7    0   0   0   0   0 0.0000000   1   0 0.0000000    0    1    0    0    0
#> a_8    0   0   0   0   0 0.0000000   0   1 0.0000000    0    0    0    0    0
#> a_9    0   0   0   0   0 0.0000000   0   0 1.0000000    0    0    0    1    1
#> a_10   0   0   0   0   0 0.0000000   0   0 0.0000000    1    0    0    0    0
#> a_11   0   0   0   0   0 0.0000000   0   0 0.0000000    0    1    0    0    0
#> a_12   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    1    1    0
#> a_13   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    1    0
#> a_14   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    1
#> a_15   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_16   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_17   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_18   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_19   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_20   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_21   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_22   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_23   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_24   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_25   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_26   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_27   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_28   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_29   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_30   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_31   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_32   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#>      a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_2     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0 0.0000000    0    0    0    0    1    0    1    0    0    0    0
#> a_4     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0 0.0000000    0    1    0    0    0    0    0    0    0    1    0
#> a_13    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_14    1    0 0.0000000    1    0    0    0    0    0    0    0    0    0    0
#> a_15    1    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_16    0    1 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_17    0    0 0.5413422    1    0    0    0    0    0    0    0    0    0    0
#> a_18    0    0 0.0000000    1    1    0    0    0    0    0    0    0    0    0
#> a_19    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_20    0    0 0.0000000    0    0    1    0    0    0    1    0    0    0    0
#> a_21    0    0 0.0000000    0    0    0    1    0    0    0    0    0    0    0
#> a_22    0    0 0.0000000    0    0    0    0    1    0    0    0    0    1    0
#> a_23    0    0 0.0000000    0    0    0    0    0    1    1    0    0    0    0
#> a_24    0    0 0.0000000    0    0    0    0    0    0    1    0    0    0    0
#> a_25    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0    0
#> a_26    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_27    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1    0
#> a_28    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    1
#> a_29    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#>      a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0 14.948606 15.509474 25.839698 27.661218 0.3189828
#> a_2     0    0    0    0  8.015039  8.610136 22.694062 27.868539 0.5636123
#> a_3     1    0    0    0 14.378779 16.035126 24.167441 25.581428 0.7538660
#> a_4     0    0    0    0 15.984962 17.418123 21.369905 24.323209 0.8202957
#> a_5     0    0    0    0 12.460300 14.062938 21.917021 25.045319 0.5195803
#> a_6     0    0    0    0 11.487382 13.718946 15.912714 18.964777 0.6713669
#> a_7     0    0    0    0 17.043953 17.505885 19.535440 19.588522 0.9099388
#> a_8     0    0    0    0 14.034081 15.192699 16.028586 18.825257 0.6388246
#> a_9     0    0    0    0  9.784912 10.516337 22.894152 23.730792 0.6890539
#> a_10    0    0    0    0 13.813661 16.523457 25.380819 27.488169 0.2118509
#> a_11    0    0    0    0  9.689921 10.315362 24.739549 27.012964 0.7763219
#> a_12    0    0    0    0 11.647210 14.555639  3.172138  3.738315 0.6606608
#> a_13    0    0    0    0 12.680739 13.149514 21.394432 21.829955 0.6707058
#> a_14    0    0    0    0 16.298166 18.605439 20.171863 23.586129 0.8223687
#> a_15    0    0    0    0  8.006016  9.332448 15.623857 18.814216 0.8645843
#> a_16    0    0    0    0 12.615394 13.509555 27.107708 27.784410 0.7268910
#> a_17    0    0    0    0  2.586533  2.735768 13.031572 13.349336 0.6420963
#> a_18    0    0    0    0  5.789435  6.014969 16.306777 20.062337 0.8612720
#> a_19    0    0    0    0 13.302378 16.023874 11.029903 12.667627 0.6894588
#> a_20    0    0    0    0 11.790252 11.944369 26.398043 29.614265 0.6453486
#> a_21    0    0    0    0 19.302409 19.572119 24.285041 27.432357 0.6364141
#> a_22    0    0    0    0  4.627545  5.147432 18.476467 19.789844 0.7330079
#> a_23    0    0    0    0  7.208269  8.690340 15.351357 17.514282 0.6205058
#> a_24    0    0    0    0 14.684770 18.225222  9.205831 11.348799 0.6085489
#> a_25    0    0    0    0 14.356393 15.672058 21.058190 22.166735 0.8347402
#> a_26    0    0    0    0 14.268645 14.390181 22.611480 26.486363 0.8540616
#> a_27    0    0    0    0 18.857848 19.619736 20.406933 23.657674 0.8381981
#> a_28    0    0    0    0 18.759000 19.967306 23.477050 27.910253 0.3339515
#> a_29    1    0    0    0 13.041483 13.547579 24.341543 26.296472 0.9476242
#> a_30    0    1    0    0 12.384963 13.646933 22.409980 27.522101 0.5685394
#> a_31    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.3488904 10.896618 13.004269  7.044239  7.050787
#> a_2  0.6839927  6.834836  7.642222 11.635026 11.867491
#> a_3  0.8614303  5.738716  6.482854  6.276912  6.861716
#> a_4  0.9619901 10.298963 11.665335  5.950441  6.870794
#> a_5  0.5943242 10.401427 11.316182 10.139439 11.764015
#> a_6  0.8332478  5.604013  6.982132  2.012332  2.073576
#> a_7  0.9654047 14.750424 14.778479 15.884191 17.122667
#> a_8  0.7463287 17.955709 19.569087 14.545078 15.231286
#> a_9  0.7997577  9.701433 10.688470  3.025964  3.598624
#> a_10 0.2637172 14.809999 17.427866 13.867885 15.872598
#> a_11 0.8737655  9.515580 11.621850 16.951672 18.546340
#> a_12 0.7083543 14.553840 15.323590 14.066109 17.342391
#> a_13 0.7880661  4.464986  5.227570  3.499813  3.786263
#> a_14 0.8273084 10.299797 10.589226  5.424895  6.387007
#> a_15 0.8767224 14.945804 16.869269 14.246273 16.342309
#> a_16 0.7677229  6.375455  6.943772 18.331007 19.985932
#> a_17 0.7791434 16.600627 16.788744  7.074296  7.225289
#> a_18 0.9987428 18.282815 19.261126 14.835957 17.631363
#> a_19 0.7255009 10.599220 11.741673  9.155824 11.319332
#> a_20 0.7538648 11.932785 11.992854  8.847819  9.999989
#> a_21 0.6368662 13.406808 14.126326 16.047519 19.572732
#> a_22 0.8814433  7.190460  8.525937  6.841606  7.680390
#> a_23 0.6503214 15.552743 18.640250 17.752896 18.694721
#> a_24 0.6990211 13.160844 15.619482 17.617548 18.376356
#> a_25 0.8488998 15.600341 15.971226  8.590777 10.154614
#> a_26 0.8631246 10.338348 11.088681 16.976292 18.544546
#> a_27 0.9865557 15.117499 17.115331 16.895264 19.562199
#> a_28 0.3832898 14.739937 16.607151  9.248820  9.525623
#> a_29 0.9722495 17.363076 17.501495  6.378482  6.463294
#> a_30 0.5705287 13.754077 16.583640 15.604195 19.404988
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=5.0

# Simulate the shock effect of a project scenario using MFPP package
# with default values.
 Res<-phase2(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5       a_6 a_7 a_8       a_9 a_10 a_11 a_12 a_13 a_14
#> a_1    1   1   0   1   1 0.0000000   0   1 0.0000000    0    0    0    0    0
#> a_2    0   1   0   0   0 1.0000000   0   0 0.0000000    0    0    0    0    0
#> a_3    0   0   1   0   1 0.0000000   0   0 0.8099181    0    0    0    0    0
#> a_4    0   0   0   1   0 0.8889752   0   0 0.0000000    0    0    0    0    0
#> a_5    0   0   0   0   1 0.0000000   1   0 0.0000000    0    0    0    0    0
#> a_6    0   0   0   0   0 1.0000000   0   0 0.0000000    1    0    0    0    0
#> a_7    0   0   0   0   0 0.0000000   1   0 0.0000000    0    1    0    0    0
#> a_8    0   0   0   0   0 0.0000000   0   1 0.0000000    0    0    0    0    0
#> a_9    0   0   0   0   0 0.0000000   0   0 1.0000000    0    0    0    1    1
#> a_10   0   0   0   0   0 0.0000000   0   0 0.0000000    1    0    0    0    0
#> a_11   0   0   0   0   0 0.0000000   0   0 0.0000000    0    1    0    0    0
#> a_12   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    1    1    0
#> a_13   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    1    0
#> a_14   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    1
#> a_15   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_16   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_17   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_18   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_19   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_20   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_21   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_22   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_23   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_24   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_25   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_26   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_27   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_28   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_29   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_30   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_31   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_32   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#>      a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_2     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0 0.0000000    0    0    0    0    1    0    1    0    0    0    0
#> a_4     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0 0.0000000    0    1    0    0    0    0    0    0    0    1    0
#> a_13    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_14    1    0 0.0000000    1    0    0    0    0    0    0    0    0    0    0
#> a_15    1    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_16    0    1 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_17    0    0 0.5413422    1    0    0    0    0    0    0    0    0    0    0
#> a_18    0    0 0.0000000    1    1    0    0    0    0    0    0    0    0    0
#> a_19    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_20    0    0 0.0000000    0    0    1    0    0    0    1    0    0    0    0
#> a_21    0    0 0.0000000    0    0    0    1    0    0    0    0    0    0    0
#> a_22    0    0 0.0000000    0    0    0    0    1    0    0    0    0    1    0
#> a_23    0    0 0.0000000    0    0    0    0    0    1    1    0    0    0    0
#> a_24    0    0 0.0000000    0    0    0    0    0    0    1    0    0    0    0
#> a_25    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0    0
#> a_26    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_27    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1    0
#> a_28    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    1
#> a_29    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#>      a_29 a_30 a_31 a_32       t_1        t_2        c_1        c_2       q_1
#> a_1     0    0    0    0 14.948606  15.509474  25.839698  27.661218 0.3189828
#> a_2     0    0    0    0  8.015039   8.610136 136.164373  27.868539 0.5636123
#> a_3     1    0    0    0 14.378779  16.035126  24.167441  25.581428 0.7538660
#> a_4     0    0    0    0 15.984962  17.418123  21.369905  24.323209 0.8202957
#> a_5     0    0    0    0 12.460300  14.062938  21.917021 150.271916 0.5195803
#> a_6     0    0    0    0 11.487382  13.718946  15.912714  18.964777 0.6713669
#> a_7     0    0    0    0 17.043953  17.505885  19.535440  19.588522 0.9099388
#> a_8     0    0    0    0 14.034081  15.192699  16.028586  18.825257 0.6388246
#> a_9     0    0    0    0  9.784912  10.516337  22.894152  23.730792 0.6890539
#> a_10    0    0    0    0 13.813661  16.523457  25.380819  27.488169 0.2118509
#> a_11    0    0    0    0  9.689921  10.315362  24.739549  27.012964 0.7763219
#> a_12    0    0    0    0 11.647210  14.555639   3.172138   3.738315 0.6606608
#> a_13    0    0    0    0 12.680739  13.149514  21.394432  21.829955 0.6707058
#> a_14    0    0    0    0 16.298166  18.605439  20.171863  23.586129 1.0000000
#> a_15    0    0    0    0  8.006016   9.332448  15.623857  18.814216 1.0000000
#> a_16    0    0    0    0 12.615394  13.509555  27.107708  27.784410 0.7268910
#> a_17    0    0    0    0  2.586533   2.735768  13.031572  13.349336 0.6420963
#> a_18    0    0    0    0  5.789435   6.014969  16.306777  20.062337 0.8612720
#> a_19    0    0    0    0 13.302378  16.023874  11.029903  12.667627 0.6894588
#> a_20    0    0    0    0 11.790252  11.944369  26.398043  29.614265 0.6453486
#> a_21    0    0    0    0 19.302409  19.572119  24.285041  27.432357 0.6364141
#> a_22    0    0    0    0  4.627545   5.147432  18.476467  19.789844 0.7330079
#> a_23    0    0    0    0  7.208269   8.690340  15.351357  17.514282 0.6205058
#> a_24    0    0    0    0 14.684770  18.225222   9.205831  11.348799 0.6085489
#> a_25    0    0    0    0 14.356393  15.672058  21.058190  22.166735 0.8347402
#> a_26    0    0    0    0 85.611873  14.390181  22.611480  26.486363 0.8540616
#> a_27    0    0    0    0 18.857848  19.619736  20.406933  23.657674 0.8381981
#> a_28    0    0    0    0 18.759000 119.803837  23.477050  27.910253 1.0000000
#> a_29    1    0    0    0 13.041483  13.547579  24.341543  26.296472 0.9476242
#> a_30    0    1    0    0 12.384963  13.646933  22.409980  27.522101 0.5685394
#> a_31    0    0    0    0  0.000000   0.000000   0.000000   0.000000 0.0000000
#> a_32    0    0    0    0  0.000000   0.000000   0.000000   0.000000 0.0000000
#>            q_2     r_1.1      r_2.1     r_1.2     r_2.2
#> a_1  0.3488904 10.896618  13.004269  7.044239 42.304721
#> a_2  0.6839927  6.834836   7.642222 11.635026 11.867491
#> a_3  0.8614303  5.738716   6.482854  6.276912  6.861716
#> a_4  0.9619901 10.298963  11.665335  5.950441  6.870794
#> a_5  0.5943242 10.401427  11.316182 10.139439 11.764015
#> a_6  1.0000000 33.624077   6.982132  2.012332 12.441458
#> a_7  0.9654047 14.750424  14.778479 15.884191 17.122667
#> a_8  0.7463287 17.955709 117.414524 14.545078 15.231286
#> a_9  0.7997577  9.701433  10.688470  3.025964  3.598624
#> a_10 0.2637172 14.809999 104.567198 13.867885 15.872598
#> a_11 0.8737655 57.093483  11.621850 16.951672 18.546340
#> a_12 0.7083543 14.553840  15.323590 14.066109 17.342391
#> a_13 0.7880661  4.464986   5.227570  3.499813  3.786263
#> a_14 0.8273084 10.299797  10.589226  5.424895  6.387007
#> a_15 0.8767224 14.945804  16.869269 14.246273 16.342309
#> a_16 0.7677229  6.375455   6.943772 18.331007 19.985932
#> a_17 0.7791434 16.600627  16.788744  7.074296  7.225289
#> a_18 0.9987428 18.282815 115.566755 14.835957 17.631363
#> a_19 0.7255009 10.599220  11.741673  9.155824 11.319332
#> a_20 0.7538648 11.932785  11.992854 53.086916  9.999989
#> a_21 1.0000000 80.440848  14.126326 96.285114 19.572732
#> a_22 0.8814433  7.190460   8.525937  6.841606  7.680390
#> a_23 0.6503214 15.552743 111.841502 17.752896 18.694721
#> a_24 0.6990211 13.160844  15.619482 17.617548 18.376356
#> a_25 0.8488998 15.600341  15.971226  8.590777 60.927684
#> a_26 0.8631246 10.338348  11.088681 16.976292 18.544546
#> a_27 0.9865557 15.117499  17.115331 16.895264 19.562199
#> a_28 0.3832898 14.739937  16.607151  9.248820 57.153736
#> a_29 1.0000000 17.363076  17.501495  6.378482  6.463294
#> a_30 0.5705287 13.754077  16.583640 15.604195 19.404988
#> a_31 0.0000000  0.000000   0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000   0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  53.57295
#> Const constraint (Cc):  588.4825
#> Score/scope constraint (Cs):  0.9759364
#> Quailty constraint (Cq):  0.6352587
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  125.3129
#> Const constraint (Cc):  901.1841
#> Score/scope constraint (Cs):  0.9810045
#> Quailty constraint (Cq):  0.7598244
#> Resource constraint(s) (CR):
#>         R_1      R_2
#> TPR 520.835 400.8865

# Simulate the effect of a project scenario using MFPP package
# with specified values.
 Res<-phase2(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5       a_6 a_7 a_8       a_9 a_10 a_11 a_12 a_13 a_14
#> a_1    1   1   0   1   1 0.0000000   0   1 0.0000000    0    0    0    0    0
#> a_2    0   1   0   0   0 1.0000000   0   0 0.0000000    0    0    0    0    0
#> a_3    0   0   1   0   1 0.0000000   0   0 0.8099181    0    0    0    0    0
#> a_4    0   0   0   1   0 0.8889752   0   0 0.0000000    0    0    0    0    0
#> a_5    0   0   0   0   1 0.0000000   1   0 0.0000000    0    0    0    0    0
#> a_6    0   0   0   0   0 1.0000000   0   0 0.0000000    1    0    0    0    0
#> a_7    0   0   0   0   0 0.0000000   1   0 0.0000000    0    1    0    0    0
#> a_8    0   0   0   0   0 0.0000000   0   1 0.0000000    0    0    0    0    0
#> a_9    0   0   0   0   0 0.0000000   0   0 1.0000000    0    0    0    1    1
#> a_10   0   0   0   0   0 0.0000000   0   0 0.0000000    1    0    0    0    0
#> a_11   0   0   0   0   0 0.0000000   0   0 0.0000000    0    1    0    0    0
#> a_12   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    1    1    0
#> a_13   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    1    0
#> a_14   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    1
#> a_15   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_16   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_17   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_18   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_19   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_20   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_21   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_22   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_23   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_24   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_25   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_26   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_27   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_28   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_29   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_30   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_31   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#> a_32   0   0   0   0   0 0.0000000   0   0 0.0000000    0    0    0    0    0
#>      a_15 a_16      a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28
#> a_1     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_2     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0 0.0000000    0    0    0    0    1    0    1    0    0    0    0
#> a_4     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_10    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_12    0    0 0.0000000    0    1    0    0    0    0    0    0    0    1    0
#> a_13    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_14    1    0 0.0000000    1    0    0    0    0    0    0    0    0    0    0
#> a_15    1    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_16    0    1 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_17    0    0 0.5413422    1    0    0    0    0    0    0    0    0    0    0
#> a_18    0    0 0.0000000    1    1    0    0    0    0    0    0    0    0    0
#> a_19    0    0 0.0000000    0    1    0    0    0    0    0    0    0    0    0
#> a_20    0    0 0.0000000    0    0    1    0    0    0    1    0    0    0    0
#> a_21    0    0 0.0000000    0    0    0    1    0    0    0    0    0    0    0
#> a_22    0    0 0.0000000    0    0    0    0    1    0    0    0    0    1    0
#> a_23    0    0 0.0000000    0    0    0    0    0    1    1    0    0    0    0
#> a_24    0    0 0.0000000    0    0    0    0    0    0    1    0    0    0    0
#> a_25    0    0 0.0000000    0    0    0    0    0    0    0    1    0    0    0
#> a_26    0    0 0.0000000    0    0    0    0    0    0    0    0    1    0    0
#> a_27    0    0 0.0000000    0    0    0    0    0    0    0    0    0    1    0
#> a_28    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    1
#> a_29    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_30    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_31    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#> a_32    0    0 0.0000000    0    0    0    0    0    0    0    0    0    0    0
#>      a_29 a_30 a_31 a_32       t_1        t_2        c_1        c_2       q_1
#> a_1     0    0    0    0 14.948606  15.509474  25.839698  27.661218 1.0000000
#> a_2     0    0    0    0  8.015039   8.610136  22.694062  27.868539 0.5636123
#> a_3     1    0    0    0 14.378779  16.035126  24.167441  25.581428 0.7538660
#> a_4     0    0    0    0 15.984962  17.418123  21.369905 145.939251 1.0000000
#> a_5     0    0    0    0 74.761803  14.062938  21.917021  25.045319 0.5195803
#> a_6     0    0    0    0 11.487382  13.718946  95.476286  18.964777 0.6713669
#> a_7     0    0    0    0 17.043953  17.505885  19.535440  19.588522 0.9099388
#> a_8     0    0    0    0 14.034081  91.156193  16.028586 112.951545 0.6388246
#> a_9     0    0    0    0  9.784912  10.516337  22.894152  23.730792 0.6890539
#> a_10    0    0    0    0 13.813661  16.523457  25.380819  27.488169 0.2118509
#> a_11    0    0    0    0  9.689921  10.315362  24.739549  27.012964 0.7763219
#> a_12    0    0    0    0 11.647210  14.555639   3.172138   3.738315 1.0000000
#> a_13    0    0    0    0 12.680739  13.149514  21.394432  21.829955 0.6707058
#> a_14    0    0    0    0 16.298166  18.605439  20.171863  23.586129 1.0000000
#> a_15    0    0    0    0  8.006016   9.332448  15.623857  18.814216 0.8645843
#> a_16    0    0    0    0 12.615394  13.509555  27.107708  27.784410 0.7268910
#> a_17    0    0    0    0 15.519197   2.735768  13.031572  13.349336 0.6420963
#> a_18    0    0    0    0  5.789435   6.014969  97.840665  20.062337 0.8612720
#> a_19    0    0    0    0 79.814267  16.023874  11.029903  12.667627 0.6894588
#> a_20    0    0    0    0 11.790252  71.666217  26.398043  29.614265 0.6453486
#> a_21    0    0    0    0 19.302409  19.572119  24.285041  27.432357 0.6364141
#> a_22    0    0    0    0  4.627545   5.147432  18.476467  19.789844 0.7330079
#> a_23    0    0    0    0  7.208269   8.690340  15.351357  17.514282 0.6205058
#> a_24    0    0    0    0 14.684770 109.351334   9.205831  11.348799 0.6085489
#> a_25    0    0    0    0 14.356393  15.672058  21.058190  22.166735 0.8347402
#> a_26    0    0    0    0 14.268645  14.390181 135.668882  26.486363 0.8540616
#> a_27    0    0    0    0 18.857848  19.619736  20.406933  23.657674 0.8381981
#> a_28    0    0    0    0 18.759000  19.967306  23.477050  27.910253 0.3339515
#> a_29    1    0    0    0 13.041483  13.547579  24.341543  26.296472 0.9476242
#> a_30    0    1    0    0 12.384963  81.881597  22.409980  27.522101 0.5685394
#> a_31    0    0    0    0  0.000000   0.000000   0.000000   0.000000 0.0000000
#> a_32    0    0    0    0  0.000000   0.000000   0.000000   0.000000 0.0000000
#>            q_2      r_1.1     r_2.1      r_1.2     r_2.2
#> a_1  0.3488904  10.896618 13.004269   7.044239  7.050787
#> a_2  0.6839927   6.834836  7.642222  11.635026 11.867491
#> a_3  0.8614303   5.738716 38.897124  37.661470  6.861716
#> a_4  0.9619901  10.298963 11.665335   5.950441  6.870794
#> a_5  0.5943242  10.401427 11.316182  10.139439 11.764015
#> a_6  0.8332478   5.604013  6.982132  12.073992 12.441458
#> a_7  0.9654047  14.750424 14.778479  15.884191 17.122667
#> a_8  0.7463287 107.734256 19.569087  14.545078 15.231286
#> a_9  0.7997577   9.701433 10.688470   3.025964  3.598624
#> a_10 0.2637172  14.809999 17.427866  13.867885 95.235586
#> a_11 0.8737655   9.515580 11.621850  16.951672 18.546340
#> a_12 0.7083543  14.553840 15.323590  14.066109 17.342391
#> a_13 0.7880661  26.789916  5.227570   3.499813  3.786263
#> a_14 0.8273084  10.299797 10.589226   5.424895  6.387007
#> a_15 0.8767224  14.945804 16.869269  14.246273 16.342309
#> a_16 0.7677229   6.375455  6.943772  18.331007 19.985932
#> a_17 0.7791434  16.600627 16.788744   7.074296  7.225289
#> a_18 0.9987428  18.282815 19.261126  89.015744 17.631363
#> a_19 0.7255009  10.599220 11.741673  54.934946 11.319332
#> a_20 0.7538648  11.932785 11.992854   8.847819  9.999989
#> a_21 0.6368662  13.406808 14.126326  16.047519 19.572732
#> a_22 0.8814433   7.190460  8.525937  41.049636  7.680390
#> a_23 1.0000000  15.552743 18.640250 106.517375 18.694721
#> a_24 0.6990211  13.160844 15.619482  17.617548 18.376356
#> a_25 0.8488998  15.600341 15.971226  51.544664 10.154614
#> a_26 1.0000000  10.338348 11.088681  16.976292 18.544546
#> a_27 0.9865557  15.117499 17.115331  16.895264 19.562199
#> a_28 0.3832898  14.739937 99.642903   9.248820  9.525623
#> a_29 0.9722495  17.363076 17.501495   6.378482  6.463294
#> a_30 1.0000000  13.754077 16.583640  15.604195 19.404988
#> a_31 0.0000000   0.000000  0.000000   0.000000  0.000000
#> a_32 0.0000000   0.000000  0.000000   0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  55.17559
#> Const constraint (Cc):  593.9905
#> Score/scope constraint (Cs):  0.9759364
#> Quailty constraint (Cq):  0.6389258
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  181.0176
#> Const constraint (Cc):  1146.876
#> Score/scope constraint (Cs):  0.9810045
#> Quailty constraint (Cq):  0.7812887
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 392.6407 390.1416