Function to simulate the effects of the change of customer claims.

Phase 3 selects P percent of nodes (i.e. tasks) or arcs (i.e. dependencies), where the score value can be changed up to the maximal change effect S.

phase3 (x,p=0.10, s=0.50)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability of structural change (default value is p=0.10).

s

Increase/decrease ratio of priorities (default value is s=0.50).

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2       a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14 a_15
#> a_1    1   1 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_2    0   1 0.7541521   0   0   0   0   0   0    0    0    0    0    0    0
#> a_3    0   0 1.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_4    0   0 0.0000000   1   1   0   0   0   0    0    0    0    0    0    0
#> a_5    0   0 0.0000000   0   1   1   0   0   1    0    0    0    0    0    0
#> a_6    0   0 0.0000000   0   0   1   1   0   0    0    0    0    0    0    0
#> a_7    0   0 0.0000000   0   0   0   1   1   1    0    0    0    0    0    0
#> a_8    0   0 0.0000000   0   0   0   0   1   1    0    0    0    0    0    0
#> a_9    0   0 0.0000000   0   0   0   0   0   1    1    0    0    0    0    0
#> a_10   0   0 0.0000000   0   0   0   0   0   0    1    1    0    0    0    0
#> a_11   0   0 0.0000000   0   0   0   0   0   0    0    1    0    0    0    0
#> a_12   0   0 0.0000000   0   0   0   0   0   0    0    0    1    0    0    0
#> a_13   0   0 0.0000000   0   0   0   0   0   0    0    0    0    1    0    0
#> a_14   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    1    1
#> a_15   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    1
#> a_16   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_17   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_18   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_19   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_20   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_21   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_22   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_23   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_24   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_25   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_26   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_27   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_28   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_29   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_30   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_31   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#> a_32   0   0 0.0000000   0   0   0   0   0   0    0    0    0    0    0    0
#>           a_16 a_17 a_18 a_19 a_20      a_21 a_22 a_23 a_24 a_25 a_26      a_27
#> a_1  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_2  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_3  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_4  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_5  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_6  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_7  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_8  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_9  0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_10 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_11 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_12 0.6277054    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_13 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_14 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_15 0.0000000    0    0    0    0 0.0000000    0    0    0    0    1 0.0000000
#> a_16 1.0000000    0    0    0    1 0.0000000    0    0    0    0    0 0.0000000
#> a_17 0.0000000    1    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_18 0.0000000    0    1    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_19 0.0000000    0    0    1    0 0.0000000    0    0    0    0    0 0.0000000
#> a_20 0.0000000    0    0    0    1 0.0000000    0    0    0    0    0 0.0000000
#> a_21 0.0000000    0    0    0    0 0.7845888    0    0    0    0    0 1.0000000
#> a_22 0.0000000    0    0    0    0 0.0000000    1    0    0    0    0 0.0000000
#> a_23 0.0000000    0    0    0    0 0.0000000    0    1    0    0    0 0.0000000
#> a_24 0.0000000    0    0    0    0 0.0000000    0    0    1    0    0 0.0000000
#> a_25 0.0000000    0    0    0    0 0.0000000    0    0    0    1    0 0.0000000
#> a_26 0.0000000    0    0    0    0 0.0000000    0    0    0    0    1 0.0000000
#> a_27 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.5365539
#> a_28 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_29 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_30 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_31 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#> a_32 0.0000000    0    0    0    0 0.0000000    0    0    0    0    0 0.0000000
#>      a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1     0    0    0    0    0 17.988832 19.127005 20.568303 22.279613 0.8882406
#> a_2     0    0    0    0    0 15.081810 18.302111 17.534816 20.761163 0.8769479
#> a_3     0    0    0    0    0 14.240011 15.143593 11.577610 13.218178 0.7723235
#> a_4     0    0    0    0    0 15.674017 17.582841  6.110607  7.359889 0.8102371
#> a_5     0    0    0    0    0 12.833843 15.384379 26.324067 28.643350 0.4448718
#> a_6     0    0    1    0    0  5.566151  6.535407 14.289272 15.545962 0.7882429
#> a_7     0    0    0    0    0  2.578501  2.894432  7.132353  7.620537 0.8809800
#> a_8     0    0    0    0    0 11.557863 13.183659 14.729273 16.154947 0.6386941
#> a_9     0    1    0    0    0  4.798449  4.856830  7.680459  8.152240 0.4625235
#> a_10    1    0    0    0    0 10.323926 11.088535 21.187469 25.992627 0.5017891
#> a_11    0    0    0    0    0 13.536387 14.094251 21.863119 22.897879 0.4522190
#> a_12    0    0    0    0    0 19.321407 19.857673 11.145619 13.128061 0.4175142
#> a_13    0    0    0    0    0 18.086163 19.745135 14.783013 15.728130 0.6451717
#> a_14    0    0    0    0    0 11.876159 14.037685 14.098883 17.169118 0.4603812
#> a_15    0    0    0    0    0 16.558382 18.490874 29.162846 29.217674 0.5940545
#> a_16    0    0    0    0    0  9.974855 12.384611 18.944719 21.449154 0.6500486
#> a_17    0    0    0    0    0 17.208546 18.376092 10.605456 11.462907 0.2852379
#> a_18    0    0    0    0    0 12.688786 13.025876 27.010155 28.293452 0.8583351
#> a_19    0    0    0    0    0 12.468721 12.620358 22.133452 25.796955 0.7901498
#> a_20    0    0    0    0    0 11.966749 14.511960 27.676817 28.914975 0.7820463
#> a_21    0    0    0    0    0 14.319445 14.768940  7.652318  9.218587 0.7491675
#> a_22    0    0    0    0    0  8.369767  8.733741 20.266409 21.717795 0.5125975
#> a_23    0    0    0    0    0 19.597091 19.694865 22.920383 27.749743 0.2874026
#> a_24    0    0    0    0    0 15.672876 16.049774 12.439768 15.248019 0.4017195
#> a_25    0    0    0    0    0 14.447613 14.675790 23.832971 28.077046 0.2115085
#> a_26    0    0    0    0    0 14.029555 15.231050  8.190733  9.411844 0.4061392
#> a_27    0    0    0    0    0 10.639661 12.755563 10.559789 11.030585 0.6424797
#> a_28    1    0    0    0    0 16.077929 18.823855 14.766706 16.815879 0.4069954
#> a_29    0    1    0    0    0 16.423084 17.618492 11.803370 12.130163 0.8255076
#> a_30    0    0    1    0    0 15.218708 16.118045 14.077923 16.588334 0.7268704
#> a_31    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.9314763 16.120854 17.943786 14.780207 16.964018
#> a_2  0.9849212 17.859135 17.934834  9.162576  9.419533
#> a_3  0.8631712 14.320133 14.786456 12.617844 13.235226
#> a_4  0.8426560 10.401469 10.929051  9.793280 12.038743
#> a_5  0.4842879  4.108790  4.949332  8.393751  9.810822
#> a_6  0.9530216 15.485598 17.739522 15.048862 17.310016
#> a_7  0.9793607 12.650253 13.571309 13.512534 13.534097
#> a_8  0.6953015  8.687989  9.976771 12.908844 15.377938
#> a_9  0.4784249 11.732152 13.208918 16.830624 19.182113
#> a_10 0.5955873 17.828818 19.261987 16.811946 17.934174
#> a_11 0.5623470 16.215054 19.854431  7.756503  8.188351
#> a_12 0.4864934  6.507619  6.658726 18.796041 19.406844
#> a_13 0.7645040 14.235694 17.113786 15.293594 15.395986
#> a_14 0.5595988 12.008204 12.633985 13.316641 14.360756
#> a_15 0.5965020 14.351037 16.674273 10.327789 11.324895
#> a_16 0.7765281 15.761161 19.282968 16.254917 19.478570
#> a_17 0.2960431 14.151702 16.034958  8.689863 10.685093
#> a_18 0.9041203 19.156483 19.527958 12.493678 12.953977
#> a_19 0.9538794 15.000989 15.313814 10.932904 11.204853
#> a_20 0.8167196  8.899455 10.681522 15.530620 18.730379
#> a_21 0.8965643 13.190576 13.239351 17.414158 19.705267
#> a_22 0.6357680  2.290598  2.311401 17.463129 18.235447
#> a_23 0.3346774 14.938295 15.614826 15.294982 19.111532
#> a_24 0.4869194  8.537860  8.818208  9.019290 10.798356
#> a_25 0.2304049 14.811971 15.254109  6.504949  7.703983
#> a_26 0.4445152  9.374992  9.520159 13.295970 16.363044
#> a_27 0.7675355 17.133481 17.610679  9.542570 10.804002
#> a_28 0.4145064  6.925574  8.463678  9.015884  9.425910
#> a_29 0.9963289 11.303581 11.436490  7.151009  8.702367
#> a_30 0.7894047  7.461584  9.168640  6.569233  6.846343
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection and dependency selecton of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=0.10

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with default values.

 Res<-phase3(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2       a_3 a_4 a_5 a_6 a_7       a_8       a_9      a_10      a_11
#> a_1    1   1 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.3474337
#> a_2    0   1 0.7541521   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_3    0   0 1.0000000   0   0   0   0 0.0000000 0.0000000 0.1290182 0.0000000
#> a_4    0   0 0.0000000   1   1   0   0 0.3911996 0.0000000 0.0000000 0.0000000
#> a_5    0   0 0.0000000   0   1   1   0 0.0000000 1.0000000 0.0000000 0.0000000
#> a_6    0   0 0.0000000   0   0   1   1 0.0000000 0.4588544 0.0000000 0.0000000
#> a_7    0   0 0.0000000   0   0   0   1 1.0000000 1.0000000 0.0000000 0.0000000
#> a_8    0   0 0.0000000   0   0   0   0 1.0000000 1.0000000 0.0000000 0.0000000
#> a_9    0   0 0.0000000   0   0   0   0 0.0000000 1.0000000 1.0000000 0.2603595
#> a_10   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 1.0000000 1.0000000
#> a_11   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 1.0000000
#> a_12   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_13   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_14   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_15   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_16   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_17   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_18   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_19   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_20   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_21   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_22   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_23   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_24   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_25   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_26   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_27   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_28   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_29   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_30   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_31   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#> a_32   0   0 0.0000000   0   0   0   0 0.0000000 0.0000000 0.0000000 0.0000000
#>           a_12       a_13      a_14      a_15      a_16      a_17       a_18
#> a_1  0.1064774 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_2  0.0000000 0.00000000 0.0000000 0.2972175 0.0000000 0.0000000 0.00000000
#> a_3  0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.04923612
#> a_4  0.0000000 0.00000000 0.1847581 0.0000000 0.0000000 0.0000000 0.00000000
#> a_5  0.0000000 0.01390212 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_6  0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_7  0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_8  0.0000000 0.00000000 0.0000000 0.0000000 0.2003860 0.0000000 0.00000000
#> a_9  0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.4691198 0.00000000
#> a_10 0.3512347 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.43836636
#> a_11 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_12 1.0000000 0.00000000 0.0000000 0.0000000 0.6277054 0.0000000 0.00000000
#> a_13 0.0000000 1.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_14 0.0000000 0.00000000 1.0000000 1.0000000 0.0000000 0.0000000 0.00000000
#> a_15 0.0000000 0.00000000 0.0000000 1.0000000 0.0000000 0.0000000 0.00000000
#> a_16 0.0000000 0.00000000 0.0000000 0.0000000 1.0000000 0.0000000 0.00000000
#> a_17 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 1.0000000 0.00000000
#> a_18 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> a_19 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_20 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_21 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_22 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_23 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_24 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_25 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_26 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_27 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_28 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_29 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_30 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_31 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> a_32 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#>            a_19      a_20      a_21       a_22      a_23       a_24 a_25
#> a_1  0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_2  0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_3  0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_4  0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.41763210    0
#> a_5  0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.39902114    0
#> a_6  0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_7  0.00000000 0.2723871 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_8  0.00000000 0.3997749 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_9  0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.09887907    0
#> a_10 0.02328474 0.1266091 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_11 0.00000000 0.0000000 0.0000000 0.05660445 0.0000000 0.00000000    0
#> a_12 0.32289512 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_13 0.00000000 0.4729142 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_14 0.00000000 0.0000000 0.0000000 0.30199984 0.0000000 0.47411629    0
#> a_15 0.00000000 0.4070416 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_16 0.00000000 1.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_17 0.00000000 0.4663831 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_18 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_19 1.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_20 0.00000000 1.0000000 0.1271964 0.00000000 0.0000000 0.00000000    0
#> a_21 0.00000000 0.0000000 0.7845888 0.00000000 0.3294665 0.00000000    0
#> a_22 0.00000000 0.0000000 0.0000000 1.00000000 0.0000000 0.00000000    0
#> a_23 0.00000000 0.0000000 0.0000000 0.00000000 1.0000000 0.00000000    0
#> a_24 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 1.00000000    0
#> a_25 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    1
#> a_26 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_27 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_28 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_29 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_30 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_31 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#> a_32 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000    0
#>            a_26      a_27      a_28         a_29       a_30 a_31 a_32       t_1
#> a_1  0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 17.988832
#> a_2  0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 15.081810
#> a_3  0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 14.240011
#> a_4  0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 15.674017
#> a_5  0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 12.833843
#> a_6  0.00000000 0.3736902 0.0000000 0.0000000000 1.00000000    0    0  5.566151
#> a_7  0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0  2.578501
#> a_8  0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 11.557863
#> a_9  0.00000000 0.0000000 0.0000000 1.0000000000 0.00000000    0    0  4.798449
#> a_10 0.00000000 0.0000000 1.0000000 0.0000000000 0.00000000    0    0 10.323926
#> a_11 0.00000000 0.0000000 0.0000000 0.0824997809 0.06248601    0    0 13.536387
#> a_12 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 19.321407
#> a_13 0.00000000 0.3832502 0.3468358 0.0000000000 0.00000000    0    0 18.086163
#> a_14 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 11.876159
#> a_15 1.00000000 0.0000000 0.0000000 0.4545065156 0.00000000    0    0 16.558382
#> a_16 0.00000000 0.0000000 0.0000000 0.0008759787 0.18268494    0    0  9.974855
#> a_17 0.00000000 0.2678426 0.0000000 0.0000000000 0.00000000    0    0 17.208546
#> a_18 0.00000000 0.0000000 0.1204766 0.0000000000 0.00000000    0    0 12.688786
#> a_19 0.00000000 0.0000000 0.0000000 0.4608583699 0.00000000    0    0 12.468721
#> a_20 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 11.966749
#> a_21 0.00000000 1.0000000 0.0000000 0.0000000000 0.00000000    0    0 14.319445
#> a_22 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0  8.369767
#> a_23 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 19.597091
#> a_24 0.07491386 0.3712684 0.0000000 0.0000000000 0.00000000    0    0 15.672876
#> a_25 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 14.447613
#> a_26 1.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0 14.029555
#> a_27 0.00000000 0.5365539 0.0000000 0.0000000000 0.00000000    0    0 10.639661
#> a_28 0.00000000 0.0000000 1.0000000 0.0000000000 0.36914930    0    0 16.077929
#> a_29 0.00000000 0.0000000 0.0000000 1.0000000000 0.00000000    0    0 16.423084
#> a_30 0.00000000 0.0000000 0.0000000 0.0000000000 1.00000000    0    0 15.218708
#> a_31 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0  0.000000
#> a_32 0.00000000 0.0000000 0.0000000 0.0000000000 0.00000000    0    0  0.000000
#>            t_2       c_1       c_2       q_1       q_2     r_1.1     r_2.1
#> a_1  19.127005 20.568303 22.279613 0.8882406 0.9314763 16.120854 17.943786
#> a_2  18.302111 17.534816 20.761163 0.8769479 0.9849212 17.859135 17.934834
#> a_3  15.143593 11.577610 13.218178 0.7723235 0.8631712 14.320133 14.786456
#> a_4  17.582841  6.110607  7.359889 0.8102371 0.8426560 10.401469 10.929051
#> a_5  15.384379 26.324067 28.643350 0.4448718 0.4842879  4.108790  4.949332
#> a_6   6.535407 14.289272 15.545962 0.7882429 0.9530216 15.485598 17.739522
#> a_7   2.894432  7.132353  7.620537 0.8809800 0.9793607 12.650253 13.571309
#> a_8  13.183659 14.729273 16.154947 0.6386941 0.6953015  8.687989  9.976771
#> a_9   4.856830  7.680459  8.152240 0.4625235 0.4784249 11.732152 13.208918
#> a_10 11.088535 21.187469 25.992627 0.5017891 0.5955873 17.828818 19.261987
#> a_11 14.094251 21.863119 22.897879 0.4522190 0.5623470 16.215054 19.854431
#> a_12 19.857673 11.145619 13.128061 0.4175142 0.4864934  6.507619  6.658726
#> a_13 19.745135 14.783013 15.728130 0.6451717 0.7645040 14.235694 17.113786
#> a_14 14.037685 14.098883 17.169118 0.4603812 0.5595988 12.008204 12.633985
#> a_15 18.490874 29.162846 29.217674 0.5940545 0.5965020 14.351037 16.674273
#> a_16 12.384611 18.944719 21.449154 0.6500486 0.7765281 15.761161 19.282968
#> a_17 18.376092 10.605456 11.462907 0.2852379 0.2960431 14.151702 16.034958
#> a_18 13.025876 27.010155 28.293452 0.8583351 0.9041203 19.156483 19.527958
#> a_19 12.620358 22.133452 25.796955 0.7901498 0.9538794 15.000989 15.313814
#> a_20 14.511960 27.676817 28.914975 0.7820463 0.8167196  8.899455 10.681522
#> a_21 14.768940  7.652318  9.218587 0.7491675 0.8965643 13.190576 13.239351
#> a_22  8.733741 20.266409 21.717795 0.5125975 0.6357680  2.290598  2.311401
#> a_23 19.694865 22.920383 27.749743 0.2874026 0.3346774 14.938295 15.614826
#> a_24 16.049774 12.439768 15.248019 0.4017195 0.4869194  8.537860  8.818208
#> a_25 14.675790 23.832971 28.077046 0.2115085 0.2304049 14.811971 15.254109
#> a_26 15.231050  8.190733  9.411844 0.4061392 0.4445152  9.374992  9.520159
#> a_27 12.755563 10.559789 11.030585 0.6424797 0.7675355 17.133481 17.610679
#> a_28 18.823855 14.766706 16.815879 0.4069954 0.4145064  6.925574  8.463678
#> a_29 17.618492 11.803370 12.130163 0.8255076 0.9963289 11.303581 11.436490
#> a_30 16.118045 14.077923 16.588334 0.7268704 0.7894047  7.461584  9.168640
#> a_31  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000  0.000000
#> a_32  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000  0.000000
#>          r_1.2     r_2.2
#> a_1  14.780207 16.964018
#> a_2   9.162576  9.419533
#> a_3  12.617844 13.235226
#> a_4   9.793280 12.038743
#> a_5   8.393751  9.810822
#> a_6  15.048862 17.310016
#> a_7  13.512534 13.534097
#> a_8  12.908844 15.377938
#> a_9  16.830624 19.182113
#> a_10 16.811946 17.934174
#> a_11  7.756503  8.188351
#> a_12 18.796041 19.406844
#> a_13 15.293594 15.395986
#> a_14 13.316641 14.360756
#> a_15 10.327789 11.324895
#> a_16 16.254917 19.478570
#> a_17  8.689863 10.685093
#> a_18 12.493678 12.953977
#> a_19 10.932904 11.204853
#> a_20 15.530620 18.730379
#> a_21 17.414158 19.705267
#> a_22 17.463129 18.235447
#> a_23 15.294982 19.111532
#> a_24  9.019290 10.798356
#> a_25  6.504949  7.703983
#> a_26 13.295970 16.363044
#> a_27  9.542570 10.804002
#> a_28  9.015884  9.425910
#> a_29  7.151009  8.702367
#> a_30  6.569233  6.846343
#> a_31  0.000000  0.000000
#> a_32  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  79.41068
#> Const constraint (Cc):  472.8566
#> Score/scope constraint (Cs):  0.930523
#> Quailty constraint (Cq):  0.5129009
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  152.7441
#> Const constraint (Cc):  547.7748
#> Score/scope constraint (Cs):  0.9733252
#> Quailty constraint (Cq):  0.6206245
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 238.0234 227.2112

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase3(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2       a_3 a_4 a_5 a_6         a_7 a_8        a_9 a_10       a_11
#> a_1    1   1 0.0000000   0   0   0 0.008224458   0 0.00000000    0 0.00000000
#> a_2    0   1 0.7541521   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_3    0   0 1.0000000   0   0   0 0.000000000   0 0.07781858    0 0.00000000
#> a_4    0   0 0.0000000   1   1   0 0.068864404   0 0.00000000    0 0.00000000
#> a_5    0   0 0.0000000   0   1   1 0.000000000   0 1.00000000    0 0.07892963
#> a_6    0   0 0.0000000   0   0   1 1.000000000   0 0.00000000    0 0.00000000
#> a_7    0   0 0.0000000   0   0   0 1.000000000   1 1.00000000    0 0.04593324
#> a_8    0   0 0.0000000   0   0   0 0.000000000   1 1.00000000    0 0.00000000
#> a_9    0   0 0.0000000   0   0   0 0.000000000   0 1.00000000    1 0.00000000
#> a_10   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    1 1.00000000
#> a_11   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 1.00000000
#> a_12   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_13   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_14   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_15   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_16   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_17   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_18   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_19   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_20   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_21   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_22   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_23   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_24   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_25   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_26   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_27   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_28   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_29   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_30   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_31   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#> a_32   0   0 0.0000000   0   0   0 0.000000000   0 0.00000000    0 0.00000000
#>      a_12       a_13      a_14       a_15      a_16       a_17 a_18 a_19
#> a_1     0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_2     0 0.00000000 0.0000000 0.00000000 0.0000000 0.05064688    0    0
#> a_3     0 0.02382738 0.0000000 0.08749346 0.0000000 0.00000000    0    0
#> a_4     0 0.00000000 0.0259076 0.08294762 0.0000000 0.00000000    0    0
#> a_5     0 0.00000000 0.0000000 0.00000000 0.0000000 0.02417457    0    0
#> a_6     0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_7     0 0.00000000 0.0000000 0.00000000 0.0000000 0.04192390    0    0
#> a_8     0 0.00000000 0.0000000 0.00000000 0.0000000 0.05403563    0    0
#> a_9     0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_10    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_11    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_12    1 0.00000000 0.0000000 0.00000000 0.6277054 0.00000000    0    0
#> a_13    0 1.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_14    0 0.00000000 1.0000000 1.00000000 0.0000000 0.00000000    0    0
#> a_15    0 0.00000000 0.0000000 1.00000000 0.0000000 0.00000000    0    0
#> a_16    0 0.00000000 0.0000000 0.00000000 1.0000000 0.00000000    0    0
#> a_17    0 0.00000000 0.0000000 0.00000000 0.0000000 1.00000000    0    0
#> a_18    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    1    0
#> a_19    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    1
#> a_20    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_21    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_22    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_23    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_24    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_25    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_26    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_27    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_28    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_29    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_30    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_31    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#> a_32    0 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000    0    0
#>            a_20       a_21        a_22        a_23        a_24       a_25
#> a_1  0.00000000 0.03130673 0.000000000 0.015124499 0.000000000 0.00000000
#> a_2  0.00000000 0.00000000 0.000000000 0.050741643 0.000000000 0.00000000
#> a_3  0.04340674 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_4  0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_5  0.00000000 0.00000000 0.001821982 0.000000000 0.000000000 0.00000000
#> a_6  0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.04397720
#> a_7  0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_8  0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_9  0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.01265850
#> a_10 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.06636382
#> a_11 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_12 0.00000000 0.00000000 0.000000000 0.000000000 0.082508436 0.00000000
#> a_13 0.00000000 0.00000000 0.000000000 0.091436756 0.000000000 0.00000000
#> a_14 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_15 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_16 1.00000000 0.04779039 0.000000000 0.000000000 0.000000000 0.00000000
#> a_17 0.00000000 0.00000000 0.000000000 0.015628677 0.000000000 0.00000000
#> a_18 0.00000000 0.09695283 0.004872158 0.000000000 0.000000000 0.00000000
#> a_19 0.00000000 0.00000000 0.000000000 0.000000000 0.007924553 0.00000000
#> a_20 1.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_21 0.00000000 0.78458880 0.000000000 0.000000000 0.000000000 0.07807322
#> a_22 0.00000000 0.00000000 1.000000000 0.009068413 0.027637723 0.00000000
#> a_23 0.00000000 0.00000000 0.000000000 1.000000000 0.000000000 0.00000000
#> a_24 0.00000000 0.00000000 0.000000000 0.000000000 1.000000000 0.00000000
#> a_25 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 1.00000000
#> a_26 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_27 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_28 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_29 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_30 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_31 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#> a_32 0.00000000 0.00000000 0.000000000 0.000000000 0.000000000 0.00000000
#>            a_26      a_27       a_28      a_29        a_30 a_31 a_32       t_1
#> a_1  0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 17.988832
#> a_2  0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 15.081810
#> a_3  0.00000000 0.0000000 0.09726718 0.0000000 0.000000000    0    0 14.240011
#> a_4  0.09602802 0.0000000 0.00000000 0.0000000 0.000000000    0    0 15.674017
#> a_5  0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 12.833843
#> a_6  0.00000000 0.0000000 0.00000000 0.0000000 1.000000000    0    0  5.566151
#> a_7  0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0  2.578501
#> a_8  0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 11.557863
#> a_9  0.00000000 0.0000000 0.00000000 1.0000000 0.000000000    0    0  4.798449
#> a_10 0.00000000 0.0000000 1.00000000 0.0000000 0.000000000    0    0 10.323926
#> a_11 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 13.536387
#> a_12 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 19.321407
#> a_13 0.00000000 0.0000000 0.00000000 0.0000000 0.056384567    0    0 18.086163
#> a_14 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 11.876159
#> a_15 1.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 16.558382
#> a_16 0.00000000 0.0000000 0.00000000 0.0000000 0.004301501    0    0  9.974855
#> a_17 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 17.208546
#> a_18 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 12.688786
#> a_19 0.00000000 0.0000000 0.00000000 0.0000000 0.036458413    0    0 12.468721
#> a_20 0.00000000 0.0000000 0.00000000 0.0141135 0.000000000    0    0 11.966749
#> a_21 0.00000000 1.0000000 0.00000000 0.0000000 0.000000000    0    0 14.319445
#> a_22 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0  8.369767
#> a_23 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0 19.597091
#> a_24 0.00000000 0.0000000 0.08934116 0.0000000 0.000000000    0    0 15.672876
#> a_25 0.00000000 0.0000000 0.05379624 0.0000000 0.000000000    0    0 14.447613
#> a_26 1.00000000 0.0000000 0.00000000 0.0000000 0.061130420    0    0 14.029555
#> a_27 0.00000000 0.5365539 0.00000000 0.0000000 0.000000000    0    0 10.639661
#> a_28 0.00000000 0.0000000 1.00000000 0.0000000 0.000000000    0    0 16.077929
#> a_29 0.00000000 0.0000000 0.00000000 1.0000000 0.000000000    0    0 16.423084
#> a_30 0.00000000 0.0000000 0.00000000 0.0000000 1.000000000    0    0 15.218708
#> a_31 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0  0.000000
#> a_32 0.00000000 0.0000000 0.00000000 0.0000000 0.000000000    0    0  0.000000
#>            t_2       c_1       c_2       q_1       q_2     r_1.1     r_2.1
#> a_1  19.127005 20.568303 22.279613 0.8882406 0.9314763 16.120854 17.943786
#> a_2  18.302111 17.534816 20.761163 0.8769479 0.9849212 17.859135 17.934834
#> a_3  15.143593 11.577610 13.218178 0.7723235 0.8631712 14.320133 14.786456
#> a_4  17.582841  6.110607  7.359889 0.8102371 0.8426560 10.401469 10.929051
#> a_5  15.384379 26.324067 28.643350 0.4448718 0.4842879  4.108790  4.949332
#> a_6   6.535407 14.289272 15.545962 0.7882429 0.9530216 15.485598 17.739522
#> a_7   2.894432  7.132353  7.620537 0.8809800 0.9793607 12.650253 13.571309
#> a_8  13.183659 14.729273 16.154947 0.6386941 0.6953015  8.687989  9.976771
#> a_9   4.856830  7.680459  8.152240 0.4625235 0.4784249 11.732152 13.208918
#> a_10 11.088535 21.187469 25.992627 0.5017891 0.5955873 17.828818 19.261987
#> a_11 14.094251 21.863119 22.897879 0.4522190 0.5623470 16.215054 19.854431
#> a_12 19.857673 11.145619 13.128061 0.4175142 0.4864934  6.507619  6.658726
#> a_13 19.745135 14.783013 15.728130 0.6451717 0.7645040 14.235694 17.113786
#> a_14 14.037685 14.098883 17.169118 0.4603812 0.5595988 12.008204 12.633985
#> a_15 18.490874 29.162846 29.217674 0.5940545 0.5965020 14.351037 16.674273
#> a_16 12.384611 18.944719 21.449154 0.6500486 0.7765281 15.761161 19.282968
#> a_17 18.376092 10.605456 11.462907 0.2852379 0.2960431 14.151702 16.034958
#> a_18 13.025876 27.010155 28.293452 0.8583351 0.9041203 19.156483 19.527958
#> a_19 12.620358 22.133452 25.796955 0.7901498 0.9538794 15.000989 15.313814
#> a_20 14.511960 27.676817 28.914975 0.7820463 0.8167196  8.899455 10.681522
#> a_21 14.768940  7.652318  9.218587 0.7491675 0.8965643 13.190576 13.239351
#> a_22  8.733741 20.266409 21.717795 0.5125975 0.6357680  2.290598  2.311401
#> a_23 19.694865 22.920383 27.749743 0.2874026 0.3346774 14.938295 15.614826
#> a_24 16.049774 12.439768 15.248019 0.4017195 0.4869194  8.537860  8.818208
#> a_25 14.675790 23.832971 28.077046 0.2115085 0.2304049 14.811971 15.254109
#> a_26 15.231050  8.190733  9.411844 0.4061392 0.4445152  9.374992  9.520159
#> a_27 12.755563 10.559789 11.030585 0.6424797 0.7675355 17.133481 17.610679
#> a_28 18.823855 14.766706 16.815879 0.4069954 0.4145064  6.925574  8.463678
#> a_29 17.618492 11.803370 12.130163 0.8255076 0.9963289 11.303581 11.436490
#> a_30 16.118045 14.077923 16.588334 0.7268704 0.7894047  7.461584  9.168640
#> a_31  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000  0.000000
#> a_32  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000  0.000000
#>          r_1.2     r_2.2
#> a_1  14.780207 16.964018
#> a_2   9.162576  9.419533
#> a_3  12.617844 13.235226
#> a_4   9.793280 12.038743
#> a_5   8.393751  9.810822
#> a_6  15.048862 17.310016
#> a_7  13.512534 13.534097
#> a_8  12.908844 15.377938
#> a_9  16.830624 19.182113
#> a_10 16.811946 17.934174
#> a_11  7.756503  8.188351
#> a_12 18.796041 19.406844
#> a_13 15.293594 15.395986
#> a_14 13.316641 14.360756
#> a_15 10.327789 11.324895
#> a_16 16.254917 19.478570
#> a_17  8.689863 10.685093
#> a_18 12.493678 12.953977
#> a_19 10.932904 11.204853
#> a_20 15.530620 18.730379
#> a_21 17.414158 19.705267
#> a_22 17.463129 18.235447
#> a_23 15.294982 19.111532
#> a_24  9.019290 10.798356
#> a_25  6.504949  7.703983
#> a_26 13.295970 16.363044
#> a_27  9.542570 10.804002
#> a_28  9.015884  9.425910
#> a_29  7.151009  8.702367
#> a_30  6.569233  6.846343
#> a_31  0.000000  0.000000
#> a_32  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  79.41068
#> Const constraint (Cc):  472.8566
#> Score/scope constraint (Cs):  0.930523
#> Quailty constraint (Cq):  0.5129009
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  105.0257
#> Const constraint (Cc):  547.7748
#> Score/scope constraint (Cs):  0.9733252
#> Quailty constraint (Cq):  0.6206245
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 238.0234 227.2112