Function to simulate the effects of the change of customer claims.

Phase 3 selects P percent of nodes (i.e. tasks) or arcs (i.e. dependencies), where the score value can be changed up to the maximal change effect S.

phase3 (x,p=0.10, s=0.50)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability of structural change (default value is p=0.10).

s

Increase/decrease ratio of priorities (default value is s=0.50).

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>            a_1 a_2       a_3 a_4 a_5 a_6 a_7 a_8 a_9     a_10 a_11 a_12 a_13
#> a_1  0.9066439   1 0.9903896   0   1   0   0   0   0 0.000000    0    0    0
#> a_2  0.0000000   1 0.0000000   0   0   1   0   0   0 0.000000    1    1    0
#> a_3  0.0000000   0 1.0000000   1   0   0   0   0   0 0.000000    0    0    0
#> a_4  0.0000000   0 0.0000000   1   0   1   1   0   0 0.000000    0    0    0
#> a_5  0.0000000   0 0.0000000   0   1   0   1   0   0 0.000000    0    0    0
#> a_6  0.0000000   0 0.0000000   0   0   1   0   0   0 0.000000    0    0    0
#> a_7  0.0000000   0 0.0000000   0   0   0   1   1   0 0.000000    0    0    0
#> a_8  0.0000000   0 0.0000000   0   0   0   0   1   0 0.000000    0    1    0
#> a_9  0.0000000   0 0.0000000   0   0   0   0   0   1 0.000000    0    0    0
#> a_10 0.0000000   0 0.0000000   0   0   0   0   0   0 0.759131    0    1    0
#> a_11 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    1    0    0
#> a_12 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    1    0
#> a_13 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    1
#> a_14 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_15 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_16 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_17 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_18 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_19 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_20 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_21 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_22 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_23 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_24 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_25 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_26 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_27 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_28 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_29 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_30 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_31 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#> a_32 0.0000000   0 0.0000000   0   0   0   0   0   0 0.000000    0    0    0
#>      a_14 a_15      a_16      a_17 a_18 a_19 a_20 a_21 a_22      a_23 a_24 a_25
#> a_1     0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_2     0    0 0.0000000 0.0000000    0    0    0    0    0 1.0000000    0    0
#> a_3     0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_4     0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_5     0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_6     0    0 0.0000000 1.0000000    0    0    0    0    0 0.0000000    0    0
#> a_7     0    0 0.0000000 0.0000000    0    0    0    0    0 1.0000000    0    0
#> a_8     0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_9     0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_10    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_11    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_12    0    0 0.0000000 0.0000000    0    0    1    0    0 0.0000000    0    1
#> a_13    1    0 1.0000000 0.0000000    1    0    0    0    0 0.0000000    0    0
#> a_14    1    0 0.0000000 0.0000000    1    0    0    0    0 0.0000000    0    0
#> a_15    0    1 0.0000000 0.8995704    0    0    0    0    0 0.0000000    0    0
#> a_16    0    0 0.8830332 1.0000000    0    0    0    0    0 0.0000000    0    0
#> a_17    0    0 0.0000000 1.0000000    1    0    0    0    0 0.0000000    0    0
#> a_18    0    0 0.0000000 0.0000000    1    1    0    0    0 0.0000000    0    0
#> a_19    0    0 0.0000000 0.0000000    0    1    0    0    0 0.0000000    0    0
#> a_20    0    0 0.0000000 0.0000000    0    0    1    0    0 0.0000000    0    0
#> a_21    0    0 0.0000000 0.0000000    0    0    0    1    1 0.0000000    0    0
#> a_22    0    0 0.0000000 0.0000000    0    0    0    0    1 0.0000000    1    0
#> a_23    0    0 0.0000000 0.0000000    0    0    0    0    0 0.7084684    0    0
#> a_24    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    1    1
#> a_25    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    1
#> a_26    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_27    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_28    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_29    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_30    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_31    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#> a_32    0    0 0.0000000 0.0000000    0    0    0    0    0 0.0000000    0    0
#>      a_26 a_27 a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2
#> a_1     0    0    0    0    0    0    0 11.614037 12.784342 16.587077 16.761759
#> a_2     0    0    0    0    0    0    0  4.121886  5.054966 11.147324 12.559384
#> a_3     0    0    0    0    0    0    0 15.358559 16.115031 16.834592 17.741677
#> a_4     0    0    0    0    0    0    0 17.894203 18.102969  4.060621  4.214420
#> a_5     0    0    1    0    0    0    0 15.085095 17.695008 20.958541 26.072822
#> a_6     0    0    0    0    0    0    0 14.319120 14.409424 26.893106 28.910187
#> a_7     0    0    0    0    0    0    0  8.864099  9.682644 13.345231 15.607719
#> a_8     0    0    0    0    0    0    0 10.755433 11.331403 17.588337 18.527621
#> a_9     0    0    0    0    0    0    0 10.516914 11.384112 14.292003 16.989090
#> a_10    0    0    0    0    0    0    0 14.220198 15.169210 23.920712 29.307066
#> a_11    0    0    0    0    0    0    0 11.038457 11.242332 13.785746 16.921660
#> a_12    0    0    0    0    0    0    0  9.192852  9.822248  5.672920  6.779559
#> a_13    0    0    0    0    0    0    0 14.422218 15.299377 12.280684 13.099977
#> a_14    0    0    0    0    0    0    0  8.908309 10.362681 23.341656 26.874973
#> a_15    0    0    0    0    0    0    0 14.819874 14.977339 24.588875 29.630167
#> a_16    0    0    0    0    0    0    0 12.997007 13.202506  3.660839  3.737921
#> a_17    0    0    0    0    0    0    0  4.733456  4.900503 19.860798 23.051479
#> a_18    0    0    1    0    0    0    0 10.830267 10.916130 11.578831 13.693574
#> a_19    0    0    0    0    0    0    0  9.629171 11.147297 18.430871 19.367208
#> a_20    0    0    0    0    0    0    0  8.902413 10.329525 16.351691 20.051984
#> a_21    0    0    0    0    0    0    0 16.507873 19.122966 18.563874 21.901635
#> a_22    1    0    0    0    0    0    0 16.338715 19.228619  8.525770  9.653850
#> a_23    0    0    0    0    0    0    0  8.222119  9.392752 25.898586 26.583637
#> a_24    0    0    0    0    0    0    0 14.098285 14.504880 19.762919 21.096080
#> a_25    0    0    0    0    0    0    0 18.947202 19.887677 24.072288 27.863023
#> a_26    1    0    0    0    0    0    0 13.363519 14.822240  8.694058 10.321637
#> a_27    0    1    1    0    0    0    0  4.319870  5.003505 19.510276 23.272497
#> a_28    0    0    1    0    0    0    0  8.176533  9.439393  4.394097  4.912596
#> a_29    0    0    0    1    0    0    0 10.782633 12.806951  7.275584  8.930004
#> a_30    0    0    0    0    1    0    0 14.702158 18.318025  4.961777  5.126808
#> a_31    0    0    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000
#> a_32    0    0    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000
#>            q_1       q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.7289126 0.8477518 12.868761 16.048514  8.706383 10.123616
#> a_2  0.5854930 0.5894973 14.622542 15.815271 13.840134 14.662799
#> a_3  0.3343637 0.3360570 13.972806 14.820612 14.477571 15.282634
#> a_4  0.8227868 0.8576877 14.550214 16.675306 13.856342 14.120806
#> a_5  0.5547276 0.5850978  5.395939  6.086078 12.794855 14.195348
#> a_6  0.7835201 0.9181038 15.813909 18.881590  5.254581  6.278343
#> a_7  0.8832772 0.9666211  9.413988 11.443397 16.069254 19.384906
#> a_8  0.5038670 0.6285794 18.639018 19.858078 17.044689 17.118024
#> a_9  0.6604714 0.7334535 18.062541 19.932492  7.163287  8.348913
#> a_10 0.6298894 0.6767447  7.079100  8.134003  4.516859  5.281954
#> a_11 0.2381067 0.2623417 12.234648 13.709939  4.386387  4.422093
#> a_12 0.5527829 0.6238006  2.630720  2.692537 13.348104 13.831426
#> a_13 0.4467724 0.5313077  5.266463  5.717537 14.145286 16.620365
#> a_14 0.5135463 0.6000324 10.617989 11.036023 16.708685 17.062276
#> a_15 0.5603901 0.5769788  8.134992  8.551720 11.713389 13.226298
#> a_16 0.9392598 0.9959738  9.897001 11.231088 15.628864 16.478439
#> a_17 0.4056641 0.4096660 15.195832 16.561368 16.759475 18.066766
#> a_18 0.3743712 0.4311201 15.222897 17.213583 13.404487 14.789017
#> a_19 0.8250651 0.8574533  7.484585  8.959685 14.124891 14.549392
#> a_20 0.5699819 0.6989571  3.087767  3.830567 11.718285 12.471498
#> a_21 0.3693066 0.4576709 10.611177 11.612678 11.168838 13.746469
#> a_22 0.2336509 0.2396473  7.998636  8.903443  5.626902  6.320560
#> a_23 0.6325585 0.6513369  6.206415  7.650701 10.195350 10.592203
#> a_24 0.7562783 0.9343720 14.483746 15.556542  5.067421  6.279849
#> a_25 0.8347796 0.8526454 16.229821 18.461789  9.312767  9.398486
#> a_26 0.8875873 0.9911888  3.193926  3.286785 11.834751 14.670908
#> a_27 0.7884143 0.9429358 11.784616 13.664660 10.405138 10.733779
#> a_28 0.7469609 0.8005259  7.857809  9.591398 12.565678 14.153038
#> a_29 0.7013851 0.7645998  7.994833  8.845410  8.345609  9.161047
#> a_30 0.4228657 0.4476316 11.807770 13.552836 11.414386 12.386624
#> a_31 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection and dependency selecton of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=0.10

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with default values.

 Res<-phase3(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>            a_1 a_2       a_3 a_4 a_5 a_6 a_7 a_8      a_9       a_10      a_11
#> a_1  0.9066439   1 0.9903896   0   1   0   0   0 0.000000 0.00000000 0.0000000
#> a_2  0.0000000   1 0.0000000   0   0   1   0   0 0.000000 0.00000000 1.0000000
#> a_3  0.0000000   0 1.0000000   1   0   0   0   0 0.000000 0.00000000 0.2332981
#> a_4  0.0000000   0 0.0000000   1   0   1   1   0 0.000000 0.00000000 0.0000000
#> a_5  0.0000000   0 0.0000000   0   1   0   1   0 0.000000 0.05790103 0.0000000
#> a_6  0.0000000   0 0.0000000   0   0   1   0   0 0.000000 0.00000000 0.0000000
#> a_7  0.0000000   0 0.0000000   0   0   0   1   1 0.312219 0.00000000 0.0000000
#> a_8  0.0000000   0 0.0000000   0   0   0   0   1 0.000000 0.00000000 0.0000000
#> a_9  0.0000000   0 0.0000000   0   0   0   0   0 1.000000 0.00000000 0.0000000
#> a_10 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.75913096 0.0000000
#> a_11 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 1.0000000
#> a_12 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_13 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_14 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_15 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_16 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_17 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_18 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_19 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_20 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_21 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_22 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_23 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_24 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_25 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_26 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_27 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_28 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_29 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_30 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_31 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#> a_32 0.0000000   0 0.0000000   0   0   0   0   0 0.000000 0.00000000 0.0000000
#>      a_12 a_13      a_14      a_15       a_16      a_17      a_18       a_19
#> a_1     0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.06842952
#> a_2     1    0 0.0000000 0.0000000 0.05962415 0.0000000 0.0000000 0.00000000
#> a_3     0    0 0.3373984 0.0000000 0.03308181 0.0000000 0.0000000 0.00000000
#> a_4     0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_5     0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_6     0    0 0.0000000 0.0000000 0.00000000 1.0000000 0.0000000 0.38861712
#> a_7     0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_8     1    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_9     0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_10    1    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_11    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_12    1    0 0.0000000 0.2272724 0.00000000 0.0000000 0.2905604 0.00000000
#> a_13    0    1 1.0000000 0.0000000 1.00000000 0.0000000 1.0000000 0.00000000
#> a_14    0    0 1.0000000 0.0000000 0.00000000 0.0000000 1.0000000 0.00000000
#> a_15    0    0 0.0000000 1.0000000 0.00000000 0.8995704 0.0000000 0.00000000
#> a_16    0    0 0.0000000 0.0000000 0.88303318 1.0000000 0.0000000 0.00000000
#> a_17    0    0 0.0000000 0.0000000 0.00000000 1.0000000 1.0000000 0.18829586
#> a_18    0    0 0.0000000 0.0000000 0.00000000 0.0000000 1.0000000 1.00000000
#> a_19    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 1.00000000
#> a_20    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_21    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_22    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_23    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_24    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_25    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_26    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_27    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_28    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_29    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_30    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_31    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#> a_32    0    0 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000
#>      a_20 a_21      a_22       a_23      a_24      a_25       a_26       a_27
#> a_1     0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_2     0    0 0.0000000 1.00000000 0.0000000 0.3358313 0.00000000 0.36128526
#> a_3     0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_4     0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_5     0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.35866626 0.00000000
#> a_6     0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_7     0    0 0.0000000 1.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_8     0    0 0.2842479 0.00000000 0.1310253 0.0000000 0.00000000 0.10204670
#> a_9     0    0 0.3253921 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_10    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_11    0    0 0.0000000 0.48252162 0.0000000 0.0000000 0.00000000 0.00000000
#> a_12    1    0 0.0000000 0.00000000 0.0000000 1.0000000 0.00000000 0.00000000
#> a_13    0    0 0.0000000 0.04812533 0.4832196 0.0000000 0.00000000 0.00000000
#> a_14    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.01301117 0.00000000
#> a_15    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.45929862
#> a_16    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.25810187 0.00000000
#> a_17    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_18    0    0 0.0000000 0.00000000 0.0000000 0.4322431 0.00000000 0.00000000
#> a_19    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_20    1    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_21    0    1 1.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_22    0    0 1.0000000 0.00000000 1.0000000 0.0000000 1.00000000 0.00000000
#> a_23    0    0 0.0000000 0.70846842 0.0000000 0.0000000 0.00000000 0.00000000
#> a_24    0    0 0.0000000 0.00000000 1.0000000 1.0000000 0.00000000 0.00000000
#> a_25    0    0 0.0000000 0.00000000 0.0000000 1.0000000 0.00000000 0.00000000
#> a_26    0    0 0.0000000 0.00000000 0.0000000 0.0000000 1.00000000 0.06849255
#> a_27    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 1.00000000
#> a_28    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_29    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_30    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_31    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> a_32    0    0 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#>           a_28      a_29      a_30 a_31 a_32       t_1       t_2       c_1
#> a_1  0.0000000 0.0000000 0.0000000    0    0 11.614037 12.784342 16.587077
#> a_2  0.0000000 0.0000000 0.0000000    0    0  4.121886  5.054966 11.147324
#> a_3  0.0000000 0.0000000 0.0000000    0    0 15.358559 16.115031 16.834592
#> a_4  0.0000000 0.0000000 0.0000000    0    0 17.894203 18.102969  4.060621
#> a_5  1.0000000 0.0000000 0.0000000    0    0 15.085095 17.695008 20.958541
#> a_6  0.0000000 0.0000000 0.0000000    0    0 14.319120 14.409424 26.893106
#> a_7  0.0000000 0.0000000 0.0000000    0    0  8.864099  9.682644 13.345231
#> a_8  0.0000000 0.4700416 0.0000000    0    0 10.755433 11.331403 17.588337
#> a_9  0.0000000 0.0000000 0.0000000    0    0 10.516914 11.384112 14.292003
#> a_10 0.0000000 0.0000000 0.0000000    0    0 14.220198 15.169210 23.920712
#> a_11 0.0000000 0.0000000 0.0000000    0    0 11.038457 11.242332 13.785746
#> a_12 0.0000000 0.0000000 0.2031586    0    0  9.192852  9.822248  5.672920
#> a_13 0.0000000 0.0000000 0.0000000    0    0 14.422218 15.299377 12.280684
#> a_14 0.0000000 0.0000000 0.0000000    0    0  8.908309 10.362681 23.341656
#> a_15 0.0000000 0.0000000 0.0000000    0    0 14.819874 14.977339 24.588875
#> a_16 0.0000000 0.0000000 0.3250618    0    0 12.997007 13.202506  3.660839
#> a_17 0.4686582 0.0000000 0.0000000    0    0  4.733456  4.900503 19.860798
#> a_18 1.0000000 0.0000000 0.0000000    0    0 10.830267 10.916130 11.578831
#> a_19 0.0000000 0.4986326 0.0635813    0    0  9.629171 11.147297 18.430871
#> a_20 0.0000000 0.0000000 0.0000000    0    0  8.902413 10.329525 16.351691
#> a_21 0.1921974 0.0000000 0.0000000    0    0 16.507873 19.122966 18.563874
#> a_22 0.0000000 0.0000000 0.0000000    0    0 16.338715 19.228619  8.525770
#> a_23 0.0000000 0.0000000 0.0000000    0    0  8.222119  9.392752 25.898586
#> a_24 0.0000000 0.0000000 0.0000000    0    0 14.098285 14.504880 19.762919
#> a_25 0.2764421 0.0000000 0.0000000    0    0 18.947202 19.887677 24.072288
#> a_26 0.0000000 0.0000000 0.4682150    0    0 13.363519 14.822240  8.694058
#> a_27 1.0000000 0.0000000 0.0000000    0    0  4.319870  5.003505 19.510276
#> a_28 1.0000000 0.0000000 0.0000000    0    0  8.176533  9.439393  4.394097
#> a_29 0.0000000 1.0000000 0.0000000    0    0 10.782633 12.806951  7.275584
#> a_30 0.0000000 0.0000000 1.0000000    0    0 14.702158 18.318025  4.961777
#> a_31 0.0000000 0.0000000 0.0000000    0    0  0.000000  0.000000  0.000000
#> a_32 0.0000000 0.0000000 0.0000000    0    0  0.000000  0.000000  0.000000
#>            c_2       q_1       q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  16.761759 0.7289126 0.8477518 12.868761 16.048514  8.706383 10.123616
#> a_2  12.559384 0.5854930 0.5894973 14.622542 15.815271 13.840134 14.662799
#> a_3  17.741677 0.3343637 0.3360570 13.972806 14.820612 14.477571 15.282634
#> a_4   4.214420 0.8227868 0.8576877 14.550214 16.675306 13.856342 14.120806
#> a_5  26.072822 0.5547276 0.5850978  5.395939  6.086078 12.794855 14.195348
#> a_6  28.910187 0.7835201 0.9181038 15.813909 18.881590  5.254581  6.278343
#> a_7  15.607719 0.8832772 0.9666211  9.413988 11.443397 16.069254 19.384906
#> a_8  18.527621 0.5038670 0.6285794 18.639018 19.858078 17.044689 17.118024
#> a_9  16.989090 0.6604714 0.7334535 18.062541 19.932492  7.163287  8.348913
#> a_10 29.307066 0.6298894 0.6767447  7.079100  8.134003  4.516859  5.281954
#> a_11 16.921660 0.2381067 0.2623417 12.234648 13.709939  4.386387  4.422093
#> a_12  6.779559 0.5527829 0.6238006  2.630720  2.692537 13.348104 13.831426
#> a_13 13.099977 0.4467724 0.5313077  5.266463  5.717537 14.145286 16.620365
#> a_14 26.874973 0.5135463 0.6000324 10.617989 11.036023 16.708685 17.062276
#> a_15 29.630167 0.5603901 0.5769788  8.134992  8.551720 11.713389 13.226298
#> a_16  3.737921 0.9392598 0.9959738  9.897001 11.231088 15.628864 16.478439
#> a_17 23.051479 0.4056641 0.4096660 15.195832 16.561368 16.759475 18.066766
#> a_18 13.693574 0.3743712 0.4311201 15.222897 17.213583 13.404487 14.789017
#> a_19 19.367208 0.8250651 0.8574533  7.484585  8.959685 14.124891 14.549392
#> a_20 20.051984 0.5699819 0.6989571  3.087767  3.830567 11.718285 12.471498
#> a_21 21.901635 0.3693066 0.4576709 10.611177 11.612678 11.168838 13.746469
#> a_22  9.653850 0.2336509 0.2396473  7.998636  8.903443  5.626902  6.320560
#> a_23 26.583637 0.6325585 0.6513369  6.206415  7.650701 10.195350 10.592203
#> a_24 21.096080 0.7562783 0.9343720 14.483746 15.556542  5.067421  6.279849
#> a_25 27.863023 0.8347796 0.8526454 16.229821 18.461789  9.312767  9.398486
#> a_26 10.321637 0.8875873 0.9911888  3.193926  3.286785 11.834751 14.670908
#> a_27 23.272497 0.7884143 0.9429358 11.784616 13.664660 10.405138 10.733779
#> a_28  4.912596 0.7469609 0.8005259  7.857809  9.591398 12.565678 14.153038
#> a_29  8.930004 0.7013851 0.7645998  7.994833  8.845410  8.345609  9.161047
#> a_30  5.126808 0.4228657 0.4476316 11.807770 13.552836 11.414386 12.386624
#> a_31  0.000000 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32  0.000000 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  81.01235
#> Const constraint (Cc):  386.7725
#> Score/scope constraint (Cs):  0.7991826
#> Quailty constraint (Cq):  0.425422
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  125.3136
#> Const constraint (Cc):  519.562
#> Score/scope constraint (Cs):  0.9740115
#> Quailty constraint (Cq):  0.6139134
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 126.7333 133.6462

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase3(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>            a_1 a_2       a_3 a_4        a_5 a_6          a_7        a_8
#> a_1  0.9066439   1 0.9903896   0 1.00000000   0 0.0000000000 0.00000000
#> a_2  0.0000000   1 0.0000000   0 0.00000000   1 0.0006527663 0.00000000
#> a_3  0.0000000   0 1.0000000   1 0.00000000   0 0.0000000000 0.00000000
#> a_4  0.0000000   0 0.0000000   1 0.05861502   1 1.0000000000 0.00000000
#> a_5  0.0000000   0 0.0000000   0 1.00000000   0 1.0000000000 0.00000000
#> a_6  0.0000000   0 0.0000000   0 0.00000000   1 0.0000000000 0.09110412
#> a_7  0.0000000   0 0.0000000   0 0.00000000   0 1.0000000000 1.00000000
#> a_8  0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 1.00000000
#> a_9  0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_10 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_11 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_12 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_13 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_14 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_15 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_16 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_17 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_18 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_19 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_20 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_21 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_22 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_23 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_24 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_25 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_26 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_27 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_28 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_29 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_30 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_31 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#> a_32 0.0000000   0 0.0000000   0 0.00000000   0 0.0000000000 0.00000000
#>             a_9       a_10 a_11       a_12 a_13       a_14       a_15
#> a_1  0.07567425 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_2  0.00000000 0.00000000    1 1.00000000    0 0.00000000 0.00000000
#> a_3  0.06402703 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_4  0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_5  0.00000000 0.05856717    0 0.00000000    0 0.00000000 0.00000000
#> a_6  0.00000000 0.07605866    0 0.00000000    0 0.00000000 0.08561595
#> a_7  0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_8  0.00000000 0.00000000    0 1.00000000    0 0.00000000 0.00000000
#> a_9  1.00000000 0.00000000    0 0.03470862    0 0.00000000 0.00000000
#> a_10 0.00000000 0.75913096    0 1.00000000    0 0.02964603 0.00000000
#> a_11 0.00000000 0.00000000    1 0.00000000    0 0.00000000 0.00000000
#> a_12 0.00000000 0.00000000    0 1.00000000    0 0.00000000 0.00000000
#> a_13 0.00000000 0.00000000    0 0.00000000    1 1.00000000 0.00000000
#> a_14 0.00000000 0.00000000    0 0.00000000    0 1.00000000 0.00000000
#> a_15 0.00000000 0.00000000    0 0.00000000    0 0.00000000 1.00000000
#> a_16 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_17 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_18 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_19 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_20 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_21 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_22 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_23 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_24 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_25 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_26 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_27 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_28 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_29 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_30 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_31 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#> a_32 0.00000000 0.00000000    0 0.00000000    0 0.00000000 0.00000000
#>            a_16      a_17       a_18      a_19 a_20        a_21       a_22
#> a_1  0.09268611 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_2  0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_3  0.00000000 0.0000000 0.03952720 0.0787032    0 0.000000000 0.05165228
#> a_4  0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_5  0.00000000 0.0219778 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_6  0.00000000 1.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_7  0.01522010 0.0000000 0.00000000 0.0000000    0 0.000000000 0.04407499
#> a_8  0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_9  0.00000000 0.0000000 0.05789741 0.0000000    0 0.033628041 0.00000000
#> a_10 0.00000000 0.0000000 0.00000000 0.0000000    0 0.067977307 0.00000000
#> a_11 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_12 0.00000000 0.0000000 0.00000000 0.0000000    1 0.000000000 0.00000000
#> a_13 1.00000000 0.0000000 1.00000000 0.0000000    0 0.000000000 0.00000000
#> a_14 0.00000000 0.0000000 1.00000000 0.0000000    0 0.000000000 0.06335328
#> a_15 0.00000000 0.8995704 0.00000000 0.0000000    0 0.088246803 0.00000000
#> a_16 0.88303318 1.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_17 0.00000000 1.0000000 1.00000000 0.0000000    0 0.000000000 0.00000000
#> a_18 0.00000000 0.0000000 1.00000000 1.0000000    0 0.000000000 0.00000000
#> a_19 0.00000000 0.0000000 0.00000000 1.0000000    0 0.000000000 0.00000000
#> a_20 0.00000000 0.0000000 0.00000000 0.0000000    1 0.005833573 0.07825650
#> a_21 0.00000000 0.0000000 0.00000000 0.0000000    0 1.000000000 1.00000000
#> a_22 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 1.00000000
#> a_23 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_24 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_25 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_26 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_27 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_28 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_29 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_30 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_31 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#> a_32 0.00000000 0.0000000 0.00000000 0.0000000    0 0.000000000 0.00000000
#>            a_23 a_24       a_25       a_26        a_27       a_28       a_29
#> a_1  0.00000000    0 0.07477879 0.00000000 0.000000000 0.00000000 0.00000000
#> a_2  1.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_3  0.00000000    0 0.00000000 0.08864175 0.000000000 0.00000000 0.06602636
#> a_4  0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_5  0.00000000    0 0.00000000 0.00000000 0.000000000 1.00000000 0.00000000
#> a_6  0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_7  1.00000000    0 0.02217449 0.00000000 0.000000000 0.00000000 0.00000000
#> a_8  0.03025056    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_9  0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_10 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_11 0.08910179    0 0.00000000 0.00000000 0.007031336 0.00000000 0.00000000
#> a_12 0.00000000    0 1.00000000 0.00000000 0.000000000 0.08927729 0.00000000
#> a_13 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_14 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_15 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_16 0.00000000    0 0.06444531 0.00000000 0.000000000 0.06560094 0.00000000
#> a_17 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_18 0.00000000    0 0.00000000 0.00000000 0.000000000 1.00000000 0.00000000
#> a_19 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_20 0.00000000    0 0.00000000 0.05252232 0.000000000 0.00000000 0.00000000
#> a_21 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_22 0.00000000    1 0.00000000 1.00000000 0.000000000 0.00000000 0.00000000
#> a_23 0.70846842    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_24 0.00000000    1 1.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_25 0.00000000    0 1.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_26 0.00000000    0 0.00000000 1.00000000 0.000000000 0.00000000 0.00000000
#> a_27 0.00000000    0 0.00000000 0.00000000 1.000000000 1.00000000 0.00000000
#> a_28 0.00000000    0 0.00000000 0.00000000 0.000000000 1.00000000 0.00000000
#> a_29 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 1.00000000
#> a_30 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_31 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_32 0.00000000    0 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#>            a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1  0.00000000    0    0 11.614037 12.784342 16.587077 16.761759 0.7289126
#> a_2  0.00000000    0    0  4.121886  5.054966 11.147324 12.559384 0.5854930
#> a_3  0.00000000    0    0 15.358559 16.115031 16.834592 17.741677 0.3343637
#> a_4  0.00000000    0    0 17.894203 18.102969  4.060621  4.214420 0.8227868
#> a_5  0.00000000    0    0 15.085095 17.695008 20.958541 26.072822 0.5547276
#> a_6  0.00000000    0    0 14.319120 14.409424 26.893106 28.910187 0.7835201
#> a_7  0.00000000    0    0  8.864099  9.682644 13.345231 15.607719 0.8832772
#> a_8  0.00000000    0    0 10.755433 11.331403 17.588337 18.527621 0.5038670
#> a_9  0.00000000    0    0 10.516914 11.384112 14.292003 16.989090 0.6604714
#> a_10 0.00000000    0    0 14.220198 15.169210 23.920712 29.307066 0.6298894
#> a_11 0.00000000    0    0 11.038457 11.242332 13.785746 16.921660 0.2381067
#> a_12 0.00000000    0    0  9.192852  9.822248  5.672920  6.779559 0.5527829
#> a_13 0.00000000    0    0 14.422218 15.299377 12.280684 13.099977 0.4467724
#> a_14 0.00000000    0    0  8.908309 10.362681 23.341656 26.874973 0.5135463
#> a_15 0.04641284    0    0 14.819874 14.977339 24.588875 29.630167 0.5603901
#> a_16 0.00000000    0    0 12.997007 13.202506  3.660839  3.737921 0.9392598
#> a_17 0.00000000    0    0  4.733456  4.900503 19.860798 23.051479 0.4056641
#> a_18 0.09991238    0    0 10.830267 10.916130 11.578831 13.693574 0.3743712
#> a_19 0.00000000    0    0  9.629171 11.147297 18.430871 19.367208 0.8250651
#> a_20 0.08690410    0    0  8.902413 10.329525 16.351691 20.051984 0.5699819
#> a_21 0.00000000    0    0 16.507873 19.122966 18.563874 21.901635 0.3693066
#> a_22 0.00000000    0    0 16.338715 19.228619  8.525770  9.653850 0.2336509
#> a_23 0.00000000    0    0  8.222119  9.392752 25.898586 26.583637 0.6325585
#> a_24 0.00000000    0    0 14.098285 14.504880 19.762919 21.096080 0.7562783
#> a_25 0.00000000    0    0 18.947202 19.887677 24.072288 27.863023 0.8347796
#> a_26 0.08237705    0    0 13.363519 14.822240  8.694058 10.321637 0.8875873
#> a_27 0.00000000    0    0  4.319870  5.003505 19.510276 23.272497 0.7884143
#> a_28 0.00000000    0    0  8.176533  9.439393  4.394097  4.912596 0.7469609
#> a_29 0.00000000    0    0 10.782633 12.806951  7.275584  8.930004 0.7013851
#> a_30 1.00000000    0    0 14.702158 18.318025  4.961777  5.126808 0.4228657
#> a_31 0.00000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32 0.00000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.8477518 12.868761 16.048514  8.706383 10.123616
#> a_2  0.5894973 14.622542 15.815271 13.840134 14.662799
#> a_3  0.3360570 13.972806 14.820612 14.477571 15.282634
#> a_4  0.8576877 14.550214 16.675306 13.856342 14.120806
#> a_5  0.5850978  5.395939  6.086078 12.794855 14.195348
#> a_6  0.9181038 15.813909 18.881590  5.254581  6.278343
#> a_7  0.9666211  9.413988 11.443397 16.069254 19.384906
#> a_8  0.6285794 18.639018 19.858078 17.044689 17.118024
#> a_9  0.7334535 18.062541 19.932492  7.163287  8.348913
#> a_10 0.6767447  7.079100  8.134003  4.516859  5.281954
#> a_11 0.2623417 12.234648 13.709939  4.386387  4.422093
#> a_12 0.6238006  2.630720  2.692537 13.348104 13.831426
#> a_13 0.5313077  5.266463  5.717537 14.145286 16.620365
#> a_14 0.6000324 10.617989 11.036023 16.708685 17.062276
#> a_15 0.5769788  8.134992  8.551720 11.713389 13.226298
#> a_16 0.9959738  9.897001 11.231088 15.628864 16.478439
#> a_17 0.4096660 15.195832 16.561368 16.759475 18.066766
#> a_18 0.4311201 15.222897 17.213583 13.404487 14.789017
#> a_19 0.8574533  7.484585  8.959685 14.124891 14.549392
#> a_20 0.6989571  3.087767  3.830567 11.718285 12.471498
#> a_21 0.4576709 10.611177 11.612678 11.168838 13.746469
#> a_22 0.2396473  7.998636  8.903443  5.626902  6.320560
#> a_23 0.6513369  6.206415  7.650701 10.195350 10.592203
#> a_24 0.9343720 14.483746 15.556542  5.067421  6.279849
#> a_25 0.8526454 16.229821 18.461789  9.312767  9.398486
#> a_26 0.9911888  3.193926  3.286785 11.834751 14.670908
#> a_27 0.9429358 11.784616 13.664660 10.405138 10.733779
#> a_28 0.8005259  7.857809  9.591398 12.565678 14.153038
#> a_29 0.7645998  7.994833  8.845410  8.345609  9.161047
#> a_30 0.4476316 11.807770 13.552836 11.414386 12.386624
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  81.01235
#> Const constraint (Cc):  386.7725
#> Score/scope constraint (Cs):  0.7991826
#> Quailty constraint (Cq):  0.425422
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  165.823
#> Const constraint (Cc):  519.562
#> Score/scope constraint (Cs):  0.9740115
#> Quailty constraint (Cq):  0.6139134
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 126.7333 133.6462