Function to simulate the effects of the change of customer claims.

Simulating the effects of the changes of customer claims.

phase3 (x,p=0.10, s=0.50)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability of structural change (default value is p=0.10).

s

Increase/decrease ratio of priorities (default value is s=0.50).

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9 a_10 a_11 a_12 a_13 a_14      a_15
#> a_1    1   1   0   0   0   0   0   0   0    1    0    0    0    0 0.0000000
#> a_2    0   1   0   0   0   1   0   0   0    0    0    0    0    0 0.0000000
#> a_3    0   0   1   0   0   1   0   0   0    0    0    0    0    0 0.0000000
#> a_4    0   0   0   1   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_5    0   0   0   0   1   1   0   0   0    0    0    0    0    0 0.0000000
#> a_6    0   0   0   0   0   1   0   0   0    0    0    0    0    0 0.0000000
#> a_7    0   0   0   0   0   0   1   0   0    0    1    1    0    0 0.0000000
#> a_8    0   0   0   0   0   0   0   1   0    0    0    0    1    1 0.0000000
#> a_9    0   0   0   0   0   0   0   0   1    0    0    0    0    0 0.0000000
#> a_10   0   0   0   0   0   0   0   0   0    1    0    0    0    0 0.0000000
#> a_11   0   0   0   0   0   0   0   0   0    0    1    0    0    0 0.0000000
#> a_12   0   0   0   0   0   0   0   0   0    0    0    1    0    0 0.0000000
#> a_13   0   0   0   0   0   0   0   0   0    0    0    0    1    1 0.0000000
#> a_14   0   0   0   0   0   0   0   0   0    0    0    0    0    1 0.0000000
#> a_15   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.8530965
#> a_16   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_17   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_18   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_19   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_20   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_21   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_22   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#> a_23   0   0   0   0   0   0   0   0   0    0    0    0    0    0 0.0000000
#>      a_16 a_17 a_18 a_19 a_20 a_21 a_22 a_23 a_24 a_25 a_26 a_27 a_28 a_29 a_30
#> a_1     0    0    0    0    0    0    1    0    0    0    0    0    0    0    0
#> a_2     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_3     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_4     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_5     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_6     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_7     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_8     0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_9     0    0    0    0    0    0    0    1    0    0    0    0    0    0    0
#> a_10    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_11    0    0    0    0    0    1    0    0    0    0    0    0    0    0    0
#> a_12    0    0    0    0    1    0    0    0    0    0    0    0    0    0    0
#> a_13    1    0    1    0    0    0    0    0    0    0    0    0    0    0    0
#> a_14    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_15    1    1    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_16    1    0    1    1    1    0    0    0    0    0    0    0    0    0    0
#> a_17    0    1    0    0    0    0    0    0    0    0    0    0    0    0    0
#> a_18    0    0    1    0    0    0    0    0    0    0    0    0    0    0    0
#> a_19    0    0    0    1    1    0    0    0    0    0    0    0    0    0    0
#> a_20    0    0    0    0    1    0    0    0    0    0    0    0    0    0    0
#> a_21    0    0    0    0    0    1    1    0    0    0    0    0    0    0    0
#> a_22    0    0    0    0    0    0    1    0    1    0    0    0    0    0    0
#> a_23    0    0    0    0    0    0    0    1    0    0    0    0    0    0    0
#>      a_31 a_32       t_1       t_2       c_1       c_2       q_1       q_2
#> a_1     0    0 17.950089 18.403540 29.393410 29.915396 0.3518513 0.4375988
#> a_2     0    0 14.224597 14.338678 10.496919 12.283090 0.5892250 0.5960014
#> a_3     0    0 14.915847 15.423241 23.590195 25.621071 0.7408263 0.7841152
#> a_4     0    0  7.976957  9.244181 11.534110 14.400851 0.6954735 0.7639299
#> a_5     0    0  2.999100  3.138785 24.584350 27.146246 0.4472131 0.4886791
#> a_6     0    0 11.137984 12.785473 17.223476 21.260846 0.8445624 0.9864297
#> a_7     0    0  9.438422 10.170768 24.022290 28.887484 0.6164655 0.6273984
#> a_8     0    0 15.186307 18.172617 18.994147 21.560808 0.1999571 0.2037743
#> a_9     0    0 18.290131 19.651121 17.761613 20.076858 0.3946370 0.4507824
#> a_10    0    0 15.703013 16.588448 29.666567 29.953144 0.6357273 0.7914785
#> a_11    0    0 18.318786 19.685810 11.871342 12.103631 0.5126610 0.5908640
#> a_12    0    0  2.849316  3.061432 19.513801 21.825049 0.8753781 0.9177854
#> a_13    0    0 15.213259 16.214729 21.509859 24.074931 0.1383885 0.1695088
#> a_14    0    0 14.831473 16.228743 15.289471 16.395286 0.6077309 0.6128823
#> a_15    0    0 16.046108 18.488964 11.047075 12.364153 0.8200307 0.9722890
#> a_16    0    0  2.725662  3.343454 27.403287 28.940116 0.8180288 0.8755376
#> a_17    0    0  8.342210  9.139146 20.813924 24.962954 0.4740241 0.5881275
#> a_18    0    0 10.613578 12.340723 23.316655 28.809769 0.9603816 0.9887022
#> a_19    0    0 10.093572 11.460653 21.138849 21.852112 0.7573691 0.8134369
#> a_20    0    0 10.834191 13.413347 23.654581 28.776225 0.9747351 0.9784398
#> a_21    0    0 15.253282 18.025119  7.729552  8.576484 0.8003580 0.8401704
#> a_22    0    0 12.123242 13.320236 29.306171 29.582284 0.1472393 0.1718037
#> a_23    0    0 15.337821 16.463816  6.556768  8.058097 0.7461687 0.8474826
#>          r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  14.256467 15.870656 10.676853 11.414132
#> a_2  14.831556 17.399726 19.403114 19.742534
#> a_3   9.433229 11.491184 13.869534 15.971039
#> a_4  16.301763 18.956335  8.725636  9.277427
#> a_5  12.119493 13.466753 13.243236 16.165971
#> a_6  17.367283 19.710280 11.478353 13.629785
#> a_7   4.703478  5.809796  6.452266  6.550161
#> a_8   9.190084 10.959761 17.547510 17.697489
#> a_9  16.621850 18.333866 14.359293 17.512703
#> a_10  7.786789  8.237959 14.278388 16.801156
#> a_11 16.011753 19.905956 16.794595 18.169388
#> a_12 10.166530 10.208056  5.706215  5.782083
#> a_13 12.495945 13.620985  9.770938 10.201374
#> a_14 14.001403 14.644320 11.139883 11.169362
#> a_15 13.805073 14.125530 15.370891 18.878566
#> a_16  5.498801  6.320464 16.511005 17.461193
#> a_17 16.598158 17.227799 17.832324 18.094659
#> a_18  7.418644  8.238836 11.463007 13.632362
#> a_19 15.784555 18.879254 15.484162 17.485637
#> a_20 13.623084 13.983234 10.040093 10.573649
#> a_21 16.136340 17.108077 12.722006 13.473890
#> a_22  3.223824  4.014121 12.873498 15.461772
#> a_23  7.998266  8.095491  9.388756 10.398365
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection and dependency selecton of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=0.10

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with default values.

 Res<-phase3(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7        a_8     a_9 a_10      a_11      a_12
#> a_1    1   1   0   0   0   0   0 0.00000000 0.00000    1 0.0000000 0.0000000
#> a_2    0   1   0   0   0   1   0 0.09378039 0.00000    0 0.3647205 0.0000000
#> a_3    0   0   1   0   0   1   0 0.26690216 0.00000    0 0.0000000 0.0000000
#> a_4    0   0   0   1   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_5    0   0   0   0   1   1   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_6    0   0   0   0   0   1   0 0.00000000 0.18204    0 0.0000000 0.3891076
#> a_7    0   0   0   0   0   0   1 0.00000000 0.00000    0 1.0000000 1.0000000
#> a_8    0   0   0   0   0   0   0 1.00000000 0.00000    0 0.0000000 0.0000000
#> a_9    0   0   0   0   0   0   0 0.00000000 1.00000    0 0.0000000 0.0000000
#> a_10   0   0   0   0   0   0   0 0.00000000 0.00000    1 0.0000000 0.0000000
#> a_11   0   0   0   0   0   0   0 0.00000000 0.00000    0 1.0000000 0.0000000
#> a_12   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 1.0000000
#> a_13   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_14   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_15   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_16   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_17   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_18   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_19   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_20   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_21   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_22   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#> a_23   0   0   0   0   0   0   0 0.00000000 0.00000    0 0.0000000 0.0000000
#>           a_13      a_14      a_15       a_16    a_17      a_18      a_19
#> a_1  0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_2  0.0000000 0.0000000 0.0000000 0.09123494 0.00000 0.0000000 0.2581268
#> a_3  0.2330093 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_4  0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_5  0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.4897376 0.0000000
#> a_6  0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_7  0.0000000 0.1592614 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_8  1.0000000 1.0000000 0.0000000 0.00000000 0.00000 0.2493888 0.0000000
#> a_9  0.0000000 0.3345220 0.0000000 0.00000000 0.46309 0.0000000 0.0000000
#> a_10 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_11 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_12 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_13 1.0000000 1.0000000 0.0000000 1.00000000 0.00000 1.0000000 0.0000000
#> a_14 0.0000000 1.0000000 0.0000000 0.00000000 0.00000 0.2041598 0.0000000
#> a_15 0.0000000 0.0000000 0.8530965 1.00000000 1.00000 0.0000000 0.0000000
#> a_16 0.0000000 0.0000000 0.0000000 1.00000000 0.00000 1.0000000 1.0000000
#> a_17 0.0000000 0.0000000 0.0000000 0.00000000 1.00000 0.0000000 0.0000000
#> a_18 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 1.0000000 0.0000000
#> a_19 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 1.0000000
#> a_20 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_21 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_22 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#> a_23 0.0000000 0.0000000 0.0000000 0.00000000 0.00000 0.0000000 0.0000000
#>            a_20     a_21      a_22       a_23      a_24       a_25      a_26
#> a_1  0.00000000 0.000000 1.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_2  0.00000000 0.000000 0.4807274 0.00000000 0.0000000 0.00000000 0.0000000
#> a_3  0.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_4  0.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_5  0.07661179 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_6  0.00000000 0.000000 0.0000000 0.40709118 0.0000000 0.00000000 0.0000000
#> a_7  0.00000000 0.406256 0.0000000 0.13214633 0.0000000 0.00000000 0.0000000
#> a_8  0.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_9  0.45740300 0.000000 0.0000000 1.00000000 0.0000000 0.00000000 0.0000000
#> a_10 0.00000000 0.000000 0.0000000 0.08208039 0.0000000 0.00000000 0.3285186
#> a_11 0.00000000 1.000000 0.4739995 0.06085997 0.4742517 0.00000000 0.0000000
#> a_12 1.00000000 0.000000 0.3284388 0.00000000 0.0000000 0.02853908 0.0000000
#> a_13 0.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_14 0.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.2944392
#> a_15 0.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_16 1.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_17 0.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_18 0.00000000 0.000000 0.3058534 0.00000000 0.0000000 0.00000000 0.0000000
#> a_19 1.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_20 1.00000000 0.000000 0.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_21 0.00000000 1.000000 1.0000000 0.00000000 0.0000000 0.00000000 0.0000000
#> a_22 0.00000000 0.000000 1.0000000 0.00000000 1.0000000 0.00000000 0.0000000
#> a_23 0.00000000 0.000000 0.0000000 1.00000000 0.0000000 0.00000000 0.0000000
#>            a_27       a_28      a_29      a_30 a_31 a_32       t_1       t_2
#> a_1  0.00000000 0.00000000 0.0000000 0.0000000    0    0 17.950089 18.403540
#> a_2  0.00000000 0.00000000 0.0000000 0.0000000    0    0 14.224597 14.338678
#> a_3  0.00000000 0.00000000 0.0000000 0.0000000    0    0 14.915847 15.423241
#> a_4  0.00000000 0.00000000 0.0000000 0.0000000    0    0  7.976957  9.244181
#> a_5  0.00000000 0.00000000 0.0000000 0.0000000    0    0  2.999100  3.138785
#> a_6  0.00000000 0.00000000 0.0000000 0.0000000    0    0 11.137984 12.785473
#> a_7  0.00000000 0.06750604 0.0000000 0.0000000    0    0  9.438422 10.170768
#> a_8  0.00000000 0.00000000 0.0000000 0.0000000    0    0 15.186307 18.172617
#> a_9  0.00000000 0.00000000 0.0000000 0.0000000    0    0 18.290131 19.651121
#> a_10 0.00000000 0.00000000 0.1715791 0.0000000    0    0 15.703013 16.588448
#> a_11 0.00000000 0.00000000 0.0000000 0.0000000    0    0 18.318786 19.685810
#> a_12 0.23171119 0.00000000 0.0000000 0.0000000    0    0  2.849316  3.061432
#> a_13 0.00000000 0.00000000 0.0000000 0.0000000    0    0 15.213259 16.214729
#> a_14 0.00000000 0.00000000 0.0000000 0.0000000    0    0 14.831473 16.228743
#> a_15 0.00000000 0.00000000 0.0000000 0.0000000    0    0 16.046108 18.488964
#> a_16 0.00000000 0.00000000 0.0000000 0.4920913    0    0  2.725662  3.343454
#> a_17 0.00000000 0.00000000 0.0000000 0.1055166    0    0  8.342210  9.139146
#> a_18 0.00000000 0.00000000 0.0000000 0.0000000    0    0 10.613578 12.340723
#> a_19 0.00000000 0.00000000 0.0000000 0.0000000    0    0 10.093572 11.460653
#> a_20 0.00000000 0.00000000 0.0000000 0.0000000    0    0 10.834191 13.413347
#> a_21 0.01572215 0.33992560 0.0000000 0.0000000    0    0 15.253282 18.025119
#> a_22 0.00000000 0.00000000 0.0000000 0.0000000    0    0 12.123242 13.320236
#> a_23 0.00000000 0.00000000 0.0000000 0.0000000    0    0 15.337821 16.463816
#>            c_1       c_2       q_1       q_2     r_1.1     r_2.1     r_1.2
#> a_1  29.393410 29.915396 0.3518513 0.4375988 14.256467 15.870656 10.676853
#> a_2  10.496919 12.283090 0.5892250 0.5960014 14.831556 17.399726 19.403114
#> a_3  23.590195 25.621071 0.7408263 0.7841152  9.433229 11.491184 13.869534
#> a_4  11.534110 14.400851 0.6954735 0.7639299 16.301763 18.956335  8.725636
#> a_5  24.584350 27.146246 0.4472131 0.4886791 12.119493 13.466753 13.243236
#> a_6  17.223476 21.260846 0.8445624 0.9864297 17.367283 19.710280 11.478353
#> a_7  24.022290 28.887484 0.6164655 0.6273984  4.703478  5.809796  6.452266
#> a_8  18.994147 21.560808 0.1999571 0.2037743  9.190084 10.959761 17.547510
#> a_9  17.761613 20.076858 0.3946370 0.4507824 16.621850 18.333866 14.359293
#> a_10 29.666567 29.953144 0.6357273 0.7914785  7.786789  8.237959 14.278388
#> a_11 11.871342 12.103631 0.5126610 0.5908640 16.011753 19.905956 16.794595
#> a_12 19.513801 21.825049 0.8753781 0.9177854 10.166530 10.208056  5.706215
#> a_13 21.509859 24.074931 0.1383885 0.1695088 12.495945 13.620985  9.770938
#> a_14 15.289471 16.395286 0.6077309 0.6128823 14.001403 14.644320 11.139883
#> a_15 11.047075 12.364153 0.8200307 0.9722890 13.805073 14.125530 15.370891
#> a_16 27.403287 28.940116 0.8180288 0.8755376  5.498801  6.320464 16.511005
#> a_17 20.813924 24.962954 0.4740241 0.5881275 16.598158 17.227799 17.832324
#> a_18 23.316655 28.809769 0.9603816 0.9887022  7.418644  8.238836 11.463007
#> a_19 21.138849 21.852112 0.7573691 0.8134369 15.784555 18.879254 15.484162
#> a_20 23.654581 28.776225 0.9747351 0.9784398 13.623084 13.983234 10.040093
#> a_21  7.729552  8.576484 0.8003580 0.8401704 16.136340 17.108077 12.722006
#> a_22 29.306171 29.582284 0.1472393 0.1718037  3.223824  4.014121 12.873498
#> a_23  6.556768  8.058097 0.7461687 0.8474826  7.998266  8.095491  9.388756
#>          r_2.2
#> a_1  11.414132
#> a_2  19.742534
#> a_3  15.971039
#> a_4   9.277427
#> a_5  16.165971
#> a_6  13.629785
#> a_7   6.550161
#> a_8  17.697489
#> a_9  17.512703
#> a_10 16.801156
#> a_11 18.169388
#> a_12  5.782083
#> a_13 10.201374
#> a_14 11.169362
#> a_15 18.878566
#> a_16 17.461193
#> a_17 18.094659
#> a_18 13.632362
#> a_19 17.485637
#> a_20 10.573649
#> a_21 13.473890
#> a_22 15.461772
#> a_23 10.398365
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  69.04086
#> Const constraint (Cc):  564.6838
#> Score/scope constraint (Cs):  0.941824
#> Quailty constraint (Cq):  0.5261166
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  122.3245
#> Const constraint (Cc):  642.5535
#> Score/scope constraint (Cs):  0.9950472
#> Quailty constraint (Cq):  0.6263071
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 212.2644 235.0653

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase3(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3        a_4 a_5        a_6 a_7        a_8        a_9 a_10
#> a_1    1   1   0 0.00000000   0 0.03743107   0 0.00000000 0.00000000    1
#> a_2    0   1   0 0.00000000   0 1.00000000   0 0.00000000 0.00926436    0
#> a_3    0   0   1 0.04140854   0 1.00000000   0 0.00000000 0.00000000    0
#> a_4    0   0   0 1.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_5    0   0   0 0.00000000   1 1.00000000   0 0.02011732 0.00000000    0
#> a_6    0   0   0 0.00000000   0 1.00000000   0 0.00000000 0.00000000    0
#> a_7    0   0   0 0.00000000   0 0.00000000   1 0.09231142 0.03587906    0
#> a_8    0   0   0 0.00000000   0 0.00000000   0 1.00000000 0.00000000    0
#> a_9    0   0   0 0.00000000   0 0.00000000   0 0.00000000 1.00000000    0
#> a_10   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    1
#> a_11   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_12   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_13   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_14   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_15   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_16   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_17   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_18   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_19   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_20   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_21   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_22   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#> a_23   0   0   0 0.00000000   0 0.00000000   0 0.00000000 0.00000000    0
#>            a_11       a_12 a_13 a_14      a_15       a_16 a_17       a_18
#> a_1  0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_2  0.00000000 0.00000000    0    0 0.0000000 0.02864478    0 0.00000000
#> a_3  0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.07703813
#> a_4  0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_5  0.00000000 0.00000000    0    0 0.0000000 0.08726387    0 0.00000000
#> a_6  0.04823289 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_7  1.00000000 1.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_8  0.00000000 0.00000000    1    1 0.0000000 0.00000000    0 0.00000000
#> a_9  0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_10 0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_11 1.00000000 0.06528968    0    0 0.0000000 0.00000000    0 0.00000000
#> a_12 0.00000000 1.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_13 0.00000000 0.00000000    1    1 0.0000000 1.00000000    0 1.00000000
#> a_14 0.00000000 0.00000000    0    1 0.0000000 0.00000000    0 0.00000000
#> a_15 0.00000000 0.00000000    0    0 0.8530965 1.00000000    1 0.00000000
#> a_16 0.00000000 0.00000000    0    0 0.0000000 1.00000000    0 1.00000000
#> a_17 0.00000000 0.00000000    0    0 0.0000000 0.00000000    1 0.01473725
#> a_18 0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 1.00000000
#> a_19 0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_20 0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_21 0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_22 0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#> a_23 0.00000000 0.00000000    0    0 0.0000000 0.00000000    0 0.00000000
#>            a_19 a_20       a_21       a_22       a_23       a_24        a_25
#> a_1  0.00000000    0 0.00000000 1.00000000 0.00000000 0.00000000 0.000000000
#> a_2  0.00000000    0 0.00000000 0.00000000 0.00000000 0.01924324 0.000000000
#> a_3  0.00000000    0 0.02135156 0.00000000 0.00000000 0.03945265 0.000000000
#> a_4  0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_5  0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_6  0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_7  0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.005638083
#> a_8  0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_9  0.00000000    0 0.00000000 0.00000000 1.00000000 0.00000000 0.000000000
#> a_10 0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_11 0.00000000    0 1.00000000 0.00000000 0.08560364 0.00000000 0.000000000
#> a_12 0.00000000    1 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_13 0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_14 0.00000000    0 0.00000000 0.09369808 0.00000000 0.00000000 0.000000000
#> a_15 0.02626278    0 0.00000000 0.00000000 0.00000000 0.01452579 0.004875396
#> a_16 1.00000000    1 0.00000000 0.00000000 0.00000000 0.00000000 0.048833053
#> a_17 0.00000000    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_18 0.06623964    0 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_19 1.00000000    1 0.05145328 0.00000000 0.00000000 0.00000000 0.000000000
#> a_20 0.00000000    1 0.00000000 0.00000000 0.00000000 0.00000000 0.000000000
#> a_21 0.00000000    0 1.00000000 1.00000000 0.00000000 0.00000000 0.000000000
#> a_22 0.00000000    0 0.00000000 1.00000000 0.00000000 1.00000000 0.000000000
#> a_23 0.00000000    0 0.00000000 0.00000000 1.00000000 0.00000000 0.000000000
#>            a_26       a_27         a_28       a_29       a_30 a_31 a_32
#> a_1  0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_2  0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_3  0.00000000 0.03409966 0.0511908876 0.00000000 0.00000000    0    0
#> a_4  0.00000000 0.00000000 0.0004663314 0.00000000 0.01812122    0    0
#> a_5  0.03834907 0.00000000 0.0000000000 0.00000000 0.06159500    0    0
#> a_6  0.00000000 0.00000000 0.0351249481 0.00000000 0.00000000    0    0
#> a_7  0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_8  0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_9  0.03311288 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_10 0.07046379 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_11 0.00000000 0.00000000 0.0000000000 0.04070656 0.00000000    0    0
#> a_12 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_13 0.01688706 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_14 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_15 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_16 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_17 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_18 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_19 0.00000000 0.00000000 0.0000000000 0.00000000 0.09107246    0    0
#> a_20 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_21 0.00000000 0.00000000 0.0000000000 0.00000000 0.00000000    0    0
#> a_22 0.00000000 0.06670188 0.0000000000 0.00000000 0.00000000    0    0
#> a_23 0.00000000 0.00000000 0.0000000000 0.06068504 0.00000000    0    0
#>            t_1       t_2       c_1       c_2       q_1       q_2     r_1.1
#> a_1  17.950089 18.403540 29.393410 29.915396 0.3518513 0.4375988 14.256467
#> a_2  14.224597 14.338678 10.496919 12.283090 0.5892250 0.5960014 14.831556
#> a_3  14.915847 15.423241 23.590195 25.621071 0.7408263 0.7841152  9.433229
#> a_4   7.976957  9.244181 11.534110 14.400851 0.6954735 0.7639299 16.301763
#> a_5   2.999100  3.138785 24.584350 27.146246 0.4472131 0.4886791 12.119493
#> a_6  11.137984 12.785473 17.223476 21.260846 0.8445624 0.9864297 17.367283
#> a_7   9.438422 10.170768 24.022290 28.887484 0.6164655 0.6273984  4.703478
#> a_8  15.186307 18.172617 18.994147 21.560808 0.1999571 0.2037743  9.190084
#> a_9  18.290131 19.651121 17.761613 20.076858 0.3946370 0.4507824 16.621850
#> a_10 15.703013 16.588448 29.666567 29.953144 0.6357273 0.7914785  7.786789
#> a_11 18.318786 19.685810 11.871342 12.103631 0.5126610 0.5908640 16.011753
#> a_12  2.849316  3.061432 19.513801 21.825049 0.8753781 0.9177854 10.166530
#> a_13 15.213259 16.214729 21.509859 24.074931 0.1383885 0.1695088 12.495945
#> a_14 14.831473 16.228743 15.289471 16.395286 0.6077309 0.6128823 14.001403
#> a_15 16.046108 18.488964 11.047075 12.364153 0.8200307 0.9722890 13.805073
#> a_16  2.725662  3.343454 27.403287 28.940116 0.8180288 0.8755376  5.498801
#> a_17  8.342210  9.139146 20.813924 24.962954 0.4740241 0.5881275 16.598158
#> a_18 10.613578 12.340723 23.316655 28.809769 0.9603816 0.9887022  7.418644
#> a_19 10.093572 11.460653 21.138849 21.852112 0.7573691 0.8134369 15.784555
#> a_20 10.834191 13.413347 23.654581 28.776225 0.9747351 0.9784398 13.623084
#> a_21 15.253282 18.025119  7.729552  8.576484 0.8003580 0.8401704 16.136340
#> a_22 12.123242 13.320236 29.306171 29.582284 0.1472393 0.1718037  3.223824
#> a_23 15.337821 16.463816  6.556768  8.058097 0.7461687 0.8474826  7.998266
#>          r_2.1     r_1.2     r_2.2
#> a_1  15.870656 10.676853 11.414132
#> a_2  17.399726 19.403114 19.742534
#> a_3  11.491184 13.869534 15.971039
#> a_4  18.956335  8.725636  9.277427
#> a_5  13.466753 13.243236 16.165971
#> a_6  19.710280 11.478353 13.629785
#> a_7   5.809796  6.452266  6.550161
#> a_8  10.959761 17.547510 17.697489
#> a_9  18.333866 14.359293 17.512703
#> a_10  8.237959 14.278388 16.801156
#> a_11 19.905956 16.794595 18.169388
#> a_12 10.208056  5.706215  5.782083
#> a_13 13.620985  9.770938 10.201374
#> a_14 14.644320 11.139883 11.169362
#> a_15 14.125530 15.370891 18.878566
#> a_16  6.320464 16.511005 17.461193
#> a_17 17.227799 17.832324 18.094659
#> a_18  8.238836 11.463007 13.632362
#> a_19 18.879254 15.484162 17.485637
#> a_20 13.983234 10.040093 10.573649
#> a_21 17.108077 12.722006 13.473890
#> a_22  4.014121 12.873498 15.461772
#> a_23  8.095491  9.388756 10.398365
#>  [ reached getOption("max.print") -- omitted 9 rows ]
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  69.04086
#> Const constraint (Cc):  564.6838
#> Score/scope constraint (Cs):  0.941824
#> Quailty constraint (Cq):  0.5261166
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  164.0273
#> Const constraint (Cc):  642.5535
#> Score/scope constraint (Cs):  0.9950472
#> Quailty constraint (Cq):  0.6263071
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 212.2644 235.0653