Function to simulate the effects of the change of customer claims.

Simulating the effects of the changes of customer claims.

phase3 (x,p=0.10, s=0.50)

Arguments

x

is either a Project Domain Matrix (a numeric matrix), where PDM=[LD,TD, CD, QD,RD]. LD is an N by N upper triangular matrix of logic domain (a numeric matrix); TD is an N by w matrix of task durations (a numeric matrix); CD is an N by w matrix of cost demands (a numeric matrix); QD is an optinal N by w matrix of quality parameters (a numeric optional matrix); RD is an N by w*nR matrix of resource demands (a numeric optional matrix), or PDM list, which contain a PDM matrix, and the number of completion modes (w) and the number of resources.

p

Probability of structural change (default value is p=0.10).

s

Increase/decrease ratio of priorities (default value is s=0.50).

Value

PDMout

PDM matrix with same structure as the input PDM matrix (a matrix).

References

Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.

Author

Zsolt T. Kosztyan*, Aamir Saghir

e-mail: kzst@gtk.uni-pannon.hu

See also

Examples

# Simulation of project domain matrix based on customer claims
# of a project scenario using MFPP package.

# Generate a project domain matrix based on
# logic domain, task duration, cost demans,
# quality parameters (optional),  resource demand (optional) of a project.

PDM=generatepdm(30,0.05,0,20,30,20,2,2,2,QD=TRUE,lst=TRUE)
PDM
#> $PDM
#>      a_1 a_2 a_3 a_4 a_5 a_6 a_7 a_8 a_9      a_10 a_11 a_12      a_13 a_14
#> a_1    1   1   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_2    0   1   0   0   0   0   1   0   0 0.0000000    0    0 0.0000000    0
#> a_3    0   0   1   1   1   0   1   0   0 0.0000000    0    0 0.0000000    0
#> a_4    0   0   0   1   1   0   0   0   0 0.0000000    1    0 0.0000000    0
#> a_5    0   0   0   0   1   1   0   0   0 0.0000000    0    0 0.0000000    0
#> a_6    0   0   0   0   0   1   0   0   0 0.0000000    0    0 0.0000000    0
#> a_7    0   0   0   0   0   0   1   0   0 0.0000000    0    0 0.0000000    0
#> a_8    0   0   0   0   0   0   0   1   0 0.0000000    0    0 0.0000000    0
#> a_9    0   0   0   0   0   0   0   0   1 0.0000000    0    0 0.0000000    1
#> a_10   0   0   0   0   0   0   0   0   0 0.9024786    0    0 0.0000000    0
#> a_11   0   0   0   0   0   0   0   0   0 0.0000000    1    0 0.0000000    0
#> a_12   0   0   0   0   0   0   0   0   0 0.0000000    0    1 0.0000000    0
#> a_13   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.5481191    0
#> a_14   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    1
#> a_15   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_16   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_17   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_18   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_19   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_20   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_21   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_22   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_23   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_24   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_25   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_26   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_27   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_28   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_29   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_30   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_31   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#> a_32   0   0   0   0   0   0   0   0   0 0.0000000    0    0 0.0000000    0
#>      a_15 a_16      a_17      a_18 a_19 a_20 a_21 a_22 a_23 a_24      a_25 a_26
#> a_1     0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_2     0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_3     0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_4     0    1 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_5     0    0 0.6528603 0.0000000    0    0    0    0    0    0 0.6999984    0
#> a_6     0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_7     0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_8     0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_9     0    0 0.0000000 0.0000000    0    0    0    0    1    0 0.0000000    1
#> a_10    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_11    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_12    0    1 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_13    1    0 0.0000000 0.0000000    0    0    0    1    0    0 0.0000000    0
#> a_14    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_15    1    1 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_16    0    1 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_17    0    0 1.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_18    0    0 0.0000000 0.8960271    0    0    1    0    0    0 0.0000000    0
#> a_19    0    0 0.0000000 0.0000000    1    0    0    0    0    0 0.0000000    0
#> a_20    0    0 0.0000000 0.0000000    0    1    0    0    1    0 0.0000000    0
#> a_21    0    0 0.0000000 0.0000000    0    0    1    1    0    0 0.0000000    0
#> a_22    0    0 0.0000000 0.0000000    0    0    0    1    0    0 0.0000000    0
#> a_23    0    0 0.0000000 0.0000000    0    0    0    0    1    1 0.0000000    0
#> a_24    0    0 0.0000000 0.0000000    0    0    0    0    0    1 0.0000000    0
#> a_25    0    0 0.0000000 0.0000000    0    0    0    0    0    0 1.0000000    0
#> a_26    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    1
#> a_27    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_28    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_29    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_30    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_31    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#> a_32    0    0 0.0000000 0.0000000    0    0    0    0    0    0 0.0000000    0
#>      a_27 a_28 a_29 a_30 a_31 a_32       t_1       t_2       c_1       c_2
#> a_1     0    0    0    0    0    0 10.426394 11.213613 25.272786 29.278435
#> a_2     0    0    0    0    0    0 13.105591 13.438846 25.846428 27.452301
#> a_3     0    0    0    0    0    0 11.092262 11.774449 17.660762 17.979756
#> a_4     0    0    0    0    0    0 15.377575 18.256570 23.697080 26.927072
#> a_5     0    0    0    0    0    0  5.880284  7.059765 23.660577 28.038196
#> a_6     0    0    0    0    0    0 12.955886 15.605068 24.260865 28.563617
#> a_7     0    0    0    0    0    0 17.312229 17.889400 23.328771 28.099881
#> a_8     0    0    0    0    0    0 10.095609 11.420714 17.307105 19.324963
#> a_9     0    0    0    0    0    0 14.323377 17.726446 26.080725 26.979852
#> a_10    0    0    0    0    0    0 14.389777 14.450874 13.397790 15.177083
#> a_11    0    0    0    0    0    0  2.349280  2.698947 18.036972 22.139972
#> a_12    0    0    0    0    0    0 18.283023 19.342973 24.651206 27.763070
#> a_13    0    0    0    1    0    0 17.355668 19.484392 24.125596 27.073167
#> a_14    0    0    0    0    0    0 10.695193 12.782222 25.986798 29.885576
#> a_15    0    0    0    0    0    0 17.623265 17.743676 22.009532 27.304979
#> a_16    0    0    0    0    0    0 11.760587 12.580820 12.636675 13.417895
#> a_17    0    0    0    0    0    0 12.373478 14.701087 16.666252 18.508923
#> a_18    0    0    0    0    0    0 14.101937 16.275002 25.743913 26.430538
#> a_19    0    0    0    0    0    0 16.364481 18.176138 24.411479 25.779186
#> a_20    0    0    0    0    0    0  5.427402  5.878149 14.884306 16.370090
#> a_21    0    0    0    0    0    0 12.088231 14.650307 20.456462 20.838941
#> a_22    0    0    0    0    0    0 11.350533 11.981891 15.921049 18.628447
#> a_23    0    0    0    0    0    0 15.223248 18.586725 10.940493 12.092901
#> a_24    0    0    0    0    0    0  4.721662  5.408775 11.383334 13.818671
#> a_25    0    0    0    0    0    0  6.334040  7.358711  5.224045  5.564825
#> a_26    0    0    0    0    0    0 12.466044 13.711882 24.645728 25.487309
#> a_27    1    1    0    0    0    0  6.379926  6.403626 10.735392 12.339701
#> a_28    0    1    0    1    0    0 13.367389 16.124879 23.202514 25.408944
#> a_29    0    0    1    0    0    0  8.637099  8.820786 12.617950 14.024660
#> a_30    0    0    0    1    0    0 15.788665 18.839957 27.840631 29.770100
#> a_31    0    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000
#> a_32    0    0    0    0    0    0  0.000000  0.000000  0.000000  0.000000
#>            q_1       q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.7009487 0.7051269 19.142279 19.575510 18.281120 19.042007
#> a_2  0.8337533 0.8542729  4.933475  6.138483 12.595107 14.770672
#> a_3  0.6466953 0.7473228  6.817175  7.795564 14.735410 16.598127
#> a_4  0.9513618 0.9889359 13.839999 15.343899  5.972538  7.055062
#> a_5  0.5250799 0.5330658  6.267628  7.411396  6.450301  6.538009
#> a_6  0.6940692 0.7668453  3.950216  4.107459  9.289265 10.502049
#> a_7  0.2618209 0.2888903  4.077908  4.846183  4.817711  5.608249
#> a_8  0.6095883 0.7544222  4.728853  5.899045 13.174409 16.157249
#> a_9  0.4366356 0.4771481 10.559574 11.932932 16.995914 19.598302
#> a_10 0.6548767 0.6610499 18.809822 19.099694 16.335881 18.721584
#> a_11 0.6781702 0.7804477  8.630257  8.981465  4.738021  5.131397
#> a_12 0.3863120 0.4457430 11.256330 12.318239  8.124279  8.351317
#> a_13 0.7853239 0.8901634 17.201101 17.482861 12.443029 15.103766
#> a_14 0.7955224 0.8181046  4.241842  4.890299 11.593107 12.264379
#> a_15 0.7971173 0.8776033 11.900768 14.407444 14.275121 15.259057
#> a_16 0.8801418 0.9780843  6.013225  7.249531 17.433895 19.119162
#> a_17 0.7189923 0.8195444  3.541876  3.979613 18.450799 18.963920
#> a_18 0.8309514 0.9709035 15.996472 19.471148 11.156134 13.645428
#> a_19 0.5461420 0.6730773 13.621231 14.150737 17.828150 18.213156
#> a_20 0.7280329 0.8049074 13.956646 16.717431  8.400169  9.157287
#> a_21 0.4898880 0.5898161  4.368548  4.980145 12.558942 15.663518
#> a_22 0.6089206 0.6848562 18.950535 19.561355 11.995298 12.514193
#> a_23 0.8146085 0.8658141 13.806063 13.979610  6.918098  7.049938
#> a_24 0.6248423 0.6943628  4.201647  4.538177 18.030566 19.739217
#> a_25 0.8094851 0.9142680  2.909079  3.224480 10.236169 11.541350
#> a_26 0.7267285 0.8897467 17.646002 18.493574  9.463414 11.457806
#> a_27 0.5243992 0.6453788  4.493721  4.869261  9.683876 10.565663
#> a_28 0.7979987 0.9685787 16.593085 19.644321 15.756230 16.485419
#> a_29 0.7071938 0.8549488 19.797639 19.965000 17.470078 17.821747
#> a_30 0.2856292 0.3137830 13.714524 14.211467  9.185085  9.472837
#> a_31 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> $w
#> [1] 2
#> 
#> $Rs
#> [1] 2
#> 
#> attr(,"class")
#> [1] "PDM_list"

# Define probability factor for task selection and dependency selecton of a project.
p=0.10

# Choose a scale factor for the ration of modification.
s=0.10

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with default values.

 Res<-phase3(PDM)
 summary(Res)
#> 
#> summary PDM list:
#> 
#> Number of completion modes (w):  2
#> Number of resources (Rs):  2
#> summary PDM matrix:
#>      a_1 a_2       a_3 a_4       a_5 a_6        a_7 a_8 a_9       a_10
#> a_1    1   1 0.2701228   0 0.1266482   0 0.00000000   0   0 0.00000000
#> a_2    0   1 0.0000000   0 0.0000000   0 1.00000000   0   0 0.06057166
#> a_3    0   0 1.0000000   1 1.0000000   0 1.00000000   0   0 0.00000000
#> a_4    0   0 0.0000000   1 1.0000000   0 0.09876852   0   0 0.00000000
#> a_5    0   0 0.0000000   0 1.0000000   1 0.00000000   0   0 0.00000000
#> a_6    0   0 0.0000000   0 0.0000000   1 0.00000000   0   0 0.00000000
#> a_7    0   0 0.0000000   0 0.0000000   0 1.00000000   0   0 0.00000000
#> a_8    0   0 0.0000000   0 0.0000000   0 0.00000000   1   0 0.45841724
#> a_9    0   0 0.0000000   0 0.0000000   0 0.00000000   0   1 0.00000000
#> a_10   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.90247858
#> a_11   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_12   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_13   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_14   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_15   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_16   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_17   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_18   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_19   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_20   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_21   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_22   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_23   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_24   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_25   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_26   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_27   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_28   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_29   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_30   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_31   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#> a_32   0   0 0.0000000   0 0.0000000   0 0.00000000   0   0 0.00000000
#>           a_11      a_12      a_13      a_14      a_15 a_16      a_17
#> a_1  0.0000000 0.0000000 0.0000000 0.0000000 0.3937069    0 0.1961161
#> a_2  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_3  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_4  1.0000000 0.4228603 0.0000000 0.0000000 0.0000000    1 0.4102113
#> a_5  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.6528603
#> a_6  0.2073590 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_7  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_8  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_9  0.4437162 0.0000000 0.0000000 1.0000000 0.0000000    0 0.0000000
#> a_10 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_11 1.0000000 0.0000000 0.0000000 0.4400885 0.2054426    0 0.0000000
#> a_12 0.0000000 1.0000000 0.0000000 0.4749555 0.0000000    1 0.0000000
#> a_13 0.0000000 0.0000000 0.5481191 0.3050066 1.0000000    0 0.0000000
#> a_14 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000    0 0.0000000
#> a_15 0.0000000 0.0000000 0.0000000 0.0000000 1.0000000    1 0.0000000
#> a_16 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    1 0.0000000
#> a_17 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 1.0000000
#> a_18 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_19 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_20 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_21 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_22 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_23 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_24 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_25 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_26 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_27 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_28 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_29 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_30 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_31 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#> a_32 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0 0.0000000
#>            a_18      a_19      a_20       a_21 a_22      a_23       a_24
#> a_1  0.00000000 0.0000000 0.1763197 0.00000000    0 0.0000000 0.09261452
#> a_2  0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_3  0.09868951 0.0000000 0.1773874 0.00000000    0 0.4426674 0.00000000
#> a_4  0.00000000 0.1465937 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_5  0.00000000 0.0000000 0.2085450 0.00000000    0 0.0000000 0.00000000
#> a_6  0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_7  0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_8  0.00000000 0.3735821 0.4841506 0.00000000    0 0.0000000 0.00000000
#> a_9  0.00000000 0.0000000 0.0000000 0.00000000    0 1.0000000 0.00000000
#> a_10 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_11 0.20306103 0.0000000 0.0000000 0.04712291    0 0.0000000 0.00000000
#> a_12 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_13 0.00000000 0.0000000 0.0000000 0.00000000    1 0.0000000 0.00000000
#> a_14 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_15 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_16 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.18829840
#> a_17 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_18 1.00000000 0.0000000 0.0000000 1.00000000    0 0.0000000 0.00000000
#> a_19 0.00000000 1.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_20 0.00000000 0.0000000 1.0000000 0.00000000    0 1.0000000 0.32589791
#> a_21 0.00000000 0.0000000 0.0000000 1.00000000    1 0.0000000 0.00000000
#> a_22 0.00000000 0.0000000 0.0000000 0.00000000    1 0.0000000 0.00000000
#> a_23 0.00000000 0.0000000 0.0000000 0.00000000    0 1.0000000 1.00000000
#> a_24 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 1.00000000
#> a_25 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_26 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_27 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_28 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_29 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_30 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_31 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#> a_32 0.00000000 0.0000000 0.0000000 0.00000000    0 0.0000000 0.00000000
#>            a_25      a_26      a_27      a_28      a_29       a_30 a_31 a_32
#> a_1  0.49358278 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_2  0.15756834 0.4181672 0.1289029 0.0000000 0.0000000 0.00000000    0    0
#> a_3  0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_4  0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_5  0.69999840 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_6  0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_7  0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.28705802    0    0
#> a_8  0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_9  0.00000000 1.0000000 0.0000000 0.0000000 0.3298465 0.00000000    0    0
#> a_10 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_11 0.00000000 0.0000000 0.0000000 0.0000000 0.4950046 0.00000000    0    0
#> a_12 0.05310398 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_13 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000    0    0
#> a_14 0.19927216 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_15 0.00000000 0.0000000 0.0000000 0.0000000 0.1068230 0.00000000    0    0
#> a_16 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_17 0.28051967 0.0000000 0.1981716 0.0000000 0.0000000 0.00000000    0    0
#> a_18 0.00000000 0.0000000 0.1779113 0.0000000 0.0000000 0.00000000    0    0
#> a_19 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_20 0.00000000 0.0000000 0.0000000 0.0000000 0.1863339 0.00000000    0    0
#> a_21 0.00000000 0.2602709 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_22 0.00000000 0.0000000 0.4818296 0.0000000 0.0000000 0.37167080    0    0
#> a_23 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.03237803    0    0
#> a_24 0.00000000 0.0000000 0.0000000 0.3255324 0.0000000 0.00000000    0    0
#> a_25 1.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_26 0.00000000 1.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_27 0.00000000 0.0000000 1.0000000 1.0000000 0.0000000 0.00000000    0    0
#> a_28 0.00000000 0.0000000 0.0000000 1.0000000 0.0000000 1.00000000    0    0
#> a_29 0.00000000 0.0000000 0.0000000 0.0000000 1.0000000 0.00000000    0    0
#> a_30 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000    0    0
#> a_31 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#> a_32 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000    0    0
#>            t_1       t_2       c_1       c_2       q_1       q_2     r_1.1
#> a_1  10.426394 11.213613 25.272786 29.278435 0.7009487 0.7051269 19.142279
#> a_2  13.105591 13.438846 25.846428 27.452301 0.8337533 0.8542729  4.933475
#> a_3  11.092262 11.774449 17.660762 17.979756 0.6466953 0.7473228  6.817175
#> a_4  15.377575 18.256570 23.697080 26.927072 0.9513618 0.9889359 13.839999
#> a_5   5.880284  7.059765 23.660577 28.038196 0.5250799 0.5330658  6.267628
#> a_6  12.955886 15.605068 24.260865 28.563617 0.6940692 0.7668453  3.950216
#> a_7  17.312229 17.889400 23.328771 28.099881 0.2618209 0.2888903  4.077908
#> a_8  10.095609 11.420714 17.307105 19.324963 0.6095883 0.7544222  4.728853
#> a_9  14.323377 17.726446 26.080725 26.979852 0.4366356 0.4771481 10.559574
#> a_10 14.389777 14.450874 13.397790 15.177083 0.6548767 0.6610499 18.809822
#> a_11  2.349280  2.698947 18.036972 22.139972 0.6781702 0.7804477  8.630257
#> a_12 18.283023 19.342973 24.651206 27.763070 0.3863120 0.4457430 11.256330
#> a_13 17.355668 19.484392 24.125596 27.073167 0.7853239 0.8901634 17.201101
#> a_14 10.695193 12.782222 25.986798 29.885576 0.7955224 0.8181046  4.241842
#> a_15 17.623265 17.743676 22.009532 27.304979 0.7971173 0.8776033 11.900768
#> a_16 11.760587 12.580820 12.636675 13.417895 0.8801418 0.9780843  6.013225
#> a_17 12.373478 14.701087 16.666252 18.508923 0.7189923 0.8195444  3.541876
#> a_18 14.101937 16.275002 25.743913 26.430538 0.8309514 0.9709035 15.996472
#> a_19 16.364481 18.176138 24.411479 25.779186 0.5461420 0.6730773 13.621231
#> a_20  5.427402  5.878149 14.884306 16.370090 0.7280329 0.8049074 13.956646
#> a_21 12.088231 14.650307 20.456462 20.838941 0.4898880 0.5898161  4.368548
#> a_22 11.350533 11.981891 15.921049 18.628447 0.6089206 0.6848562 18.950535
#> a_23 15.223248 18.586725 10.940493 12.092901 0.8146085 0.8658141 13.806063
#> a_24  4.721662  5.408775 11.383334 13.818671 0.6248423 0.6943628  4.201647
#> a_25  6.334040  7.358711  5.224045  5.564825 0.8094851 0.9142680  2.909079
#> a_26 12.466044 13.711882 24.645728 25.487309 0.7267285 0.8897467 17.646002
#> a_27  6.379926  6.403626 10.735392 12.339701 0.5243992 0.6453788  4.493721
#> a_28 13.367389 16.124879 23.202514 25.408944 0.7979987 0.9685787 16.593085
#> a_29  8.637099  8.820786 12.617950 14.024660 0.7071938 0.8549488 19.797639
#> a_30 15.788665 18.839957 27.840631 29.770100 0.2856292 0.3137830 13.714524
#> a_31  0.000000  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000
#> a_32  0.000000  0.000000  0.000000  0.000000 0.0000000 0.0000000  0.000000
#>          r_2.1     r_1.2     r_2.2
#> a_1  19.575510 18.281120 19.042007
#> a_2   6.138483 12.595107 14.770672
#> a_3   7.795564 14.735410 16.598127
#> a_4  15.343899  5.972538  7.055062
#> a_5   7.411396  6.450301  6.538009
#> a_6   4.107459  9.289265 10.502049
#> a_7   4.846183  4.817711  5.608249
#> a_8   5.899045 13.174409 16.157249
#> a_9  11.932932 16.995914 19.598302
#> a_10 19.099694 16.335881 18.721584
#> a_11  8.981465  4.738021  5.131397
#> a_12 12.318239  8.124279  8.351317
#> a_13 17.482861 12.443029 15.103766
#> a_14  4.890299 11.593107 12.264379
#> a_15 14.407444 14.275121 15.259057
#> a_16  7.249531 17.433895 19.119162
#> a_17  3.979613 18.450799 18.963920
#> a_18 19.471148 11.156134 13.645428
#> a_19 14.150737 17.828150 18.213156
#> a_20 16.717431  8.400169  9.157287
#> a_21  4.980145 12.558942 15.663518
#> a_22 19.561355 11.995298 12.514193
#> a_23 13.979610  6.918098  7.049938
#> a_24  4.538177 18.030566 19.739217
#> a_25  3.224480 10.236169 11.541350
#> a_26 18.493574  9.463414 11.457806
#> a_27  4.869261  9.683876 10.565663
#> a_28 19.644321 15.756230 16.485419
#> a_29 19.965000 17.470078 17.821747
#> a_30 14.211467  9.185085  9.472837
#> a_31  0.000000  0.000000  0.000000
#> a_32  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  45.30601
#> Const constraint (Cc):  555.1098
#> Score/scope constraint (Cs):  0.9070451
#> Quailty constraint (Cq):  0.5609304
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  139.6705
#> Const constraint (Cc):  660.4691
#> Score/scope constraint (Cs):  0.9782441
#> Quailty constraint (Cq):  0.6951121
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 163.0445 209.2096

# Simulate the effects of change of customer claims of a project scenario using MFPP package
# with specified values.
 Res<-phase3(PDM$PDM,p,s)
 summary(Res,w=2,Rs=2)
#> 
#> summary PDM matrix:
#>      a_1 a_2 a_3 a_4 a_5 a_6        a_7       a_8        a_9       a_10
#> a_1    1   1   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_2    0   1   0   0   0   0 1.00000000 0.0000000 0.00000000 0.00000000
#> a_3    0   0   1   1   1   0 1.00000000 0.0000000 0.00000000 0.00000000
#> a_4    0   0   0   1   1   0 0.04963507 0.0000000 0.00000000 0.00000000
#> a_5    0   0   0   0   1   1 0.00000000 0.0387342 0.05645316 0.00000000
#> a_6    0   0   0   0   0   1 0.00000000 0.0000000 0.00000000 0.00000000
#> a_7    0   0   0   0   0   0 1.00000000 0.0000000 0.09005704 0.00000000
#> a_8    0   0   0   0   0   0 0.00000000 1.0000000 0.00000000 0.00000000
#> a_9    0   0   0   0   0   0 0.00000000 0.0000000 1.00000000 0.02916104
#> a_10   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.90247858
#> a_11   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_12   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_13   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_14   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_15   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_16   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_17   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_18   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_19   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_20   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_21   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_22   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_23   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_24   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_25   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_26   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_27   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_28   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_29   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_30   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_31   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#> a_32   0   0   0   0   0   0 0.00000000 0.0000000 0.00000000 0.00000000
#>           a_11        a_12       a_13       a_14       a_15 a_16       a_17
#> a_1  0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_2  0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_3  0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.04631642
#> a_4  1.0000000 0.000000000 0.00000000 0.00000000 0.00000000    1 0.00000000
#> a_5  0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.65286027
#> a_6  0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_7  0.0137806 0.041033654 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_8  0.0000000 0.045013676 0.00000000 0.00000000 0.00000000    0 0.05319121
#> a_9  0.0000000 0.089978062 0.00000000 1.00000000 0.00000000    0 0.00000000
#> a_10 0.0000000 0.000000000 0.04738656 0.00000000 0.05501991    0 0.00000000
#> a_11 1.0000000 0.008228273 0.00000000 0.05470975 0.00000000    0 0.00000000
#> a_12 0.0000000 1.000000000 0.00000000 0.00000000 0.00000000    1 0.00000000
#> a_13 0.0000000 0.000000000 0.54811912 0.00000000 1.00000000    0 0.00000000
#> a_14 0.0000000 0.000000000 0.00000000 1.00000000 0.00000000    0 0.00000000
#> a_15 0.0000000 0.000000000 0.00000000 0.00000000 1.00000000    1 0.00000000
#> a_16 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    1 0.00000000
#> a_17 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 1.00000000
#> a_18 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_19 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_20 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_21 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_22 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_23 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_24 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_25 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_26 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_27 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_28 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_29 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_30 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_31 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#> a_32 0.0000000 0.000000000 0.00000000 0.00000000 0.00000000    0 0.00000000
#>           a_18      a_19      a_20       a_21       a_22         a_23
#> a_1  0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_2  0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_3  0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_4  0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_5  0.0435177 0.0382995 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_6  0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_7  0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_8  0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_9  0.0000000 0.0000000 0.0000000 0.02071694 0.00000000 1.0000000000
#> a_10 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0004555495
#> a_11 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_12 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_13 0.0000000 0.0000000 0.0000000 0.00000000 1.00000000 0.0000000000
#> a_14 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_15 0.0000000 0.0000000 0.0000000 0.04305140 0.00000000 0.0000000000
#> a_16 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_17 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_18 0.8960271 0.0000000 0.0894797 1.00000000 0.05054363 0.0000000000
#> a_19 0.0000000 1.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_20 0.0000000 0.0000000 1.0000000 0.00000000 0.00000000 1.0000000000
#> a_21 0.0000000 0.0000000 0.0000000 1.00000000 1.00000000 0.0380569515
#> a_22 0.0000000 0.0000000 0.0000000 0.00000000 1.00000000 0.0000000000
#> a_23 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 1.0000000000
#> a_24 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_25 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_26 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_27 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_28 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_29 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_30 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_31 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#> a_32 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000000
#>            a_24       a_25       a_26        a_27       a_28       a_29
#> a_1  0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_2  0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_3  0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_4  0.00000000 0.00000000 0.00000000 0.026902934 0.00000000 0.00000000
#> a_5  0.00000000 0.69999840 0.00000000 0.000000000 0.00000000 0.00000000
#> a_6  0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_7  0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_8  0.00000000 0.00000000 0.00000000 0.004622067 0.00000000 0.00000000
#> a_9  0.00000000 0.00000000 1.00000000 0.000000000 0.06719387 0.00000000
#> a_10 0.00000000 0.00000000 0.00000000 0.098773714 0.00000000 0.00000000
#> a_11 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_12 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_13 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.09717937
#> a_14 0.00000000 0.00000000 0.07803186 0.000000000 0.00000000 0.01627064
#> a_15 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_16 0.00000000 0.00000000 0.00000000 0.000000000 0.02886792 0.00000000
#> a_17 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_18 0.08588694 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_19 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_20 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_21 0.00000000 0.07216616 0.00000000 0.000000000 0.00000000 0.00000000
#> a_22 0.00000000 0.00000000 0.07368523 0.000000000 0.00000000 0.00000000
#> a_23 1.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_24 1.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_25 0.00000000 1.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_26 0.00000000 0.00000000 1.00000000 0.000000000 0.00000000 0.00000000
#> a_27 0.00000000 0.00000000 0.00000000 1.000000000 1.00000000 0.00000000
#> a_28 0.00000000 0.00000000 0.00000000 0.000000000 1.00000000 0.00000000
#> a_29 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 1.00000000
#> a_30 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_31 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#> a_32 0.00000000 0.00000000 0.00000000 0.000000000 0.00000000 0.00000000
#>            a_30 a_31 a_32       t_1       t_2       c_1       c_2       q_1
#> a_1  0.00000000    0    0 10.426394 11.213613 25.272786 29.278435 0.7009487
#> a_2  0.02602340    0    0 13.105591 13.438846 25.846428 27.452301 0.8337533
#> a_3  0.00000000    0    0 11.092262 11.774449 17.660762 17.979756 0.6466953
#> a_4  0.09735773    0    0 15.377575 18.256570 23.697080 26.927072 0.9513618
#> a_5  0.00000000    0    0  5.880284  7.059765 23.660577 28.038196 0.5250799
#> a_6  0.00000000    0    0 12.955886 15.605068 24.260865 28.563617 0.6940692
#> a_7  0.02334401    0    0 17.312229 17.889400 23.328771 28.099881 0.2618209
#> a_8  0.00000000    0    0 10.095609 11.420714 17.307105 19.324963 0.6095883
#> a_9  0.00000000    0    0 14.323377 17.726446 26.080725 26.979852 0.4366356
#> a_10 0.00000000    0    0 14.389777 14.450874 13.397790 15.177083 0.6548767
#> a_11 0.00000000    0    0  2.349280  2.698947 18.036972 22.139972 0.6781702
#> a_12 0.09118293    0    0 18.283023 19.342973 24.651206 27.763070 0.3863120
#> a_13 1.00000000    0    0 17.355668 19.484392 24.125596 27.073167 0.7853239
#> a_14 0.00000000    0    0 10.695193 12.782222 25.986798 29.885576 0.7955224
#> a_15 0.00000000    0    0 17.623265 17.743676 22.009532 27.304979 0.7971173
#> a_16 0.00000000    0    0 11.760587 12.580820 12.636675 13.417895 0.8801418
#> a_17 0.00000000    0    0 12.373478 14.701087 16.666252 18.508923 0.7189923
#> a_18 0.00000000    0    0 14.101937 16.275002 25.743913 26.430538 0.8309514
#> a_19 0.00000000    0    0 16.364481 18.176138 24.411479 25.779186 0.5461420
#> a_20 0.00000000    0    0  5.427402  5.878149 14.884306 16.370090 0.7280329
#> a_21 0.00000000    0    0 12.088231 14.650307 20.456462 20.838941 0.4898880
#> a_22 0.00000000    0    0 11.350533 11.981891 15.921049 18.628447 0.6089206
#> a_23 0.00000000    0    0 15.223248 18.586725 10.940493 12.092901 0.8146085
#> a_24 0.02350503    0    0  4.721662  5.408775 11.383334 13.818671 0.6248423
#> a_25 0.00000000    0    0  6.334040  7.358711  5.224045  5.564825 0.8094851
#> a_26 0.00000000    0    0 12.466044 13.711882 24.645728 25.487309 0.7267285
#> a_27 0.00000000    0    0  6.379926  6.403626 10.735392 12.339701 0.5243992
#> a_28 1.00000000    0    0 13.367389 16.124879 23.202514 25.408944 0.7979987
#> a_29 0.00000000    0    0  8.637099  8.820786 12.617950 14.024660 0.7071938
#> a_30 1.00000000    0    0 15.788665 18.839957 27.840631 29.770100 0.2856292
#> a_31 0.00000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#> a_32 0.00000000    0    0  0.000000  0.000000  0.000000  0.000000 0.0000000
#>            q_2     r_1.1     r_2.1     r_1.2     r_2.2
#> a_1  0.7051269 19.142279 19.575510 18.281120 19.042007
#> a_2  0.8542729  4.933475  6.138483 12.595107 14.770672
#> a_3  0.7473228  6.817175  7.795564 14.735410 16.598127
#> a_4  0.9889359 13.839999 15.343899  5.972538  7.055062
#> a_5  0.5330658  6.267628  7.411396  6.450301  6.538009
#> a_6  0.7668453  3.950216  4.107459  9.289265 10.502049
#> a_7  0.2888903  4.077908  4.846183  4.817711  5.608249
#> a_8  0.7544222  4.728853  5.899045 13.174409 16.157249
#> a_9  0.4771481 10.559574 11.932932 16.995914 19.598302
#> a_10 0.6610499 18.809822 19.099694 16.335881 18.721584
#> a_11 0.7804477  8.630257  8.981465  4.738021  5.131397
#> a_12 0.4457430 11.256330 12.318239  8.124279  8.351317
#> a_13 0.8901634 17.201101 17.482861 12.443029 15.103766
#> a_14 0.8181046  4.241842  4.890299 11.593107 12.264379
#> a_15 0.8776033 11.900768 14.407444 14.275121 15.259057
#> a_16 0.9780843  6.013225  7.249531 17.433895 19.119162
#> a_17 0.8195444  3.541876  3.979613 18.450799 18.963920
#> a_18 0.9709035 15.996472 19.471148 11.156134 13.645428
#> a_19 0.6730773 13.621231 14.150737 17.828150 18.213156
#> a_20 0.8049074 13.956646 16.717431  8.400169  9.157287
#> a_21 0.5898161  4.368548  4.980145 12.558942 15.663518
#> a_22 0.6848562 18.950535 19.561355 11.995298 12.514193
#> a_23 0.8658141 13.806063 13.979610  6.918098  7.049938
#> a_24 0.6943628  4.201647  4.538177 18.030566 19.739217
#> a_25 0.9142680  2.909079  3.224480 10.236169 11.541350
#> a_26 0.8897467 17.646002 18.493574  9.463414 11.457806
#> a_27 0.6453788  4.493721  4.869261  9.683876 10.565663
#> a_28 0.9685787 16.593085 19.644321 15.756230 16.485419
#> a_29 0.8549488 19.797639 19.965000 17.470078 17.821747
#> a_30 0.3137830 13.714524 14.211467  9.185085  9.472837
#> a_31 0.0000000  0.000000  0.000000  0.000000  0.000000
#> a_32 0.0000000  0.000000  0.000000  0.000000  0.000000
#> attr(,"class")
#> [1] "PDM_matrix"
#> 
#> Minimal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  45.30601
#> Const constraint (Cc):  529.3659
#> Score/scope constraint (Cs):  0.8450991
#> Quailty constraint (Cq):  0.513768
#> 
#> Maximal constraints:
#> 
#> Summary of the PDM constraints structure:
#> 
#> Time constraint (Ct):  174.8116
#> Const constraint (Cc):  660.4691
#> Score/scope constraint (Cs):  0.9748937
#> Quailty constraint (Cq):  0.6919657
#> Resource constraint(s) (CR):
#>          R_1      R_2
#> TPR 163.0445 209.2096